Constraints

A guide and manual for defining and manipulating constraints in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

Constraints are an integral part of infinite dimensional problems and serve as a fundamental utility of InfiniteOpt. In particular, InfiniteOpt supports finite constraints that entail finite variables and/or measures that fully any infinite parameter dependencies (e.g., first stage constraints), infinite constraints that are enforced over the entire domain of its infinite parameter dependencies (e.g., path constraints), and bounded infinite constraints which are enforced over some specified sub-domain of its infinite parameter dependencies (e.g., initial conditions). This page will highlight how to implement these types of constraints in InfiniteOpt.

Basic Usage

Principally, the @constraint and @BDconstraint macros should be employed to specify constraints. Here @constraint is used to specify typical finite and infinite constraints and @BDconstraint is used to specify bounded constraints (i.e., infinite constraints with a constrained sub-domain of its full infinite sub-domain). First, let's setup an infinite model with variables that we can add constraints to:

julia> model = InfiniteModel();

julia> @infinite_parameter(model, t in [0, 10]);

julia> @infinite_parameter(model, x[1:2] in Normal());

julia> @infinite_variable(model, T(t, x));

julia> @infinite_variable(model, g(t));

julia> @hold_variable(model, z[1:2]);

julia> @hold_variable(model, w, parameter_bounds = (t in [0, 5]));

Unbounded Constraints

Unbounded constraints don't entail user-defined curtailing of dependent infinite parameter domains and are defined using @constraint. This is done following typical JuMP syntax. Anonymous constraints (constraints without associated register names) are defined via 2 arguments where the first is the infinite model and the second is the constraint expression. For example, let's define the constraint $2T(t, x) + ||z||^2 \leq 0$:

julia> cref = @constraint(model, 2T + sum(z[i]^2 for i = 1:2) <= 0)
z[1]² + z[2]² + 2 T(t, x) ≤ 0.0, ∀ t ∈ [0, 10], x[1] ~ Normal, x[2] ~ Normal

Thus, we added an infinite constraint (which infinite with respect to t and x) to model and stored the corresponding constraint reference to cref. The allowed constraint operators are ==, <=, ≤, >=, and ≥.

Note

Currently, InfiniteOpt only supports scalar constraint expressions. Other types such as vector constraints are not currently extended.

Similarly, we can define an array of constraints with varied indexes by including an additional argument before the constraint expression. For example, let's define $3z[i] - 14 == 0, \forall i \in \{1,2\}$:

julia> crefs = @constraint(model, [i = 1:2], 3z[i] - 14 == 0)
2-element Array{InfOptConstraintRef{ScalarShape},1}:
 3 z[1] = 14.0
 3 z[2] = 14.0

Thus, we added two constraints to model and stored a vector of the corresponding constraint references to the Julia variable crefs.

Named constraints are defined by including a name as part of the second argument. For example, let's add the constraint $\int_0^10 g(t) dt == 4$:

julia> @constraint(model, measure_constr, integral(g, t) == 4)
measure_constr : integral{t ∈ [0, 10]}[g(t)] = 4.0

Thus, we added another constraint named measure_constr and created a Julia variable measure_constr that contains a reference to that constraint.

Note

Linear algebra constraints can also be used when defining constraints when . is added in front of the constraint operators (e.g., .<=). This behavior is further explained in JuMP's documentation here. However, note that that vector array sets such as MOI.Nonnegatives are not currently supported.

Bounded Constraints

Bounded constraints denote constraints with a particular sub-domain of an infinite domain. Such constraints might pertain to point constraints evaluated at a particular infinite parameter values, constraints limited to a certain range of infinite parameter values, and/or entail bounded hold variables whose sub-domains are also taken into consideration. These types of constraints are defined using @BDconstraint. For example, let's add the initial condition $g(0) == 0$:

julia> @BDconstraint(model, initial(t == 0), g == 0)
initial : g(t) = 0.0, ∀ t = 0

Thus, we have added a constraint to model defined over the sub-domain $t = 0$ in accordance with the initial condition. This is referred to as a bounded constraint named initial where a Julia variable initial has been defined to store a reference to that constraint.

More complex sub-domains can be specified by simply adding more conditions in the second argument. To illustrate this, let's define an anonymous constraint for $2T^2(t, x) + w \geq 3, \ \forall t = 0, \ x \in [-1, 1]^2$:

julia> cref = @BDconstraint(model, (t == 0, x in [-1, 1]), 2T^2 + w >= 3)
2 T(t, x)² + w ≥ 3.0, ∀ t = 0, x[1] ~ Normal ∩ [-1, 1], x[2] ~ Normal ∩ [-1, 1]

where cref contains the corresponding constraint reference.

Now we have added constraints to our model and it is ready to be transcribed!

Data Structure

Here we detail the data structures used to store constraints in InfiniteOpt. In general, constraints in JuMP are of the form: [function] in [set] where function corresponds to a JuMP expression and set corresponds to a MOI set. Since InfiniteOpt only supports scalar constraints currently, expressions must be inherited from JuMP.AbstractJuMPScalar and the supported MOI sets are MOI.EqualTo, MOI.LessThan, MOI.GreaterThan, MOI.Integer, and MOI.ZeroOne. Where the last 2 are intended for single variable constraints.

Furthermore, constraints in the form mentioned above are stored in appropriate constraint object inherited from JuMP.AbstractConstraint. Typical scalar constraints use JuMP.ScalarConstraint which has the fields:

func::JuMP.AbstractJuMPScalar The constraint expression

set::MOI.AbstractScalarSet The MOI set.

Similarly, bounded constraints are stored in BoundedScalarConstraint which has bounds and orig_bounds fields in addition to func and set. These additional fields store the current constraint bounds and the bounds originally given, respectively. This distinction is needed to facilitate deletion methods such as deleting a bounded hold variable.

These constraint objects are what store constraints in InfiniteModels. And these are referred to by an InfOptConstraintRef.

Definition

In this section, we describe the ins and outs of defining constraints. Note that this process is analogous to the manner in which variables are defined and added to the model.

Manual Definition

The @constraint and @BDconstraint both follow a similar methodology behind the scenes and these commonalities are discussed in this section for conciseness. Defining/initializing a constraint principally involves the following steps:

Define the constraint information (i.e., function, set, and parameter bounds)

Construct a concrete subtype of JuMP.AbstractConstraint to store the constraint information

Add the AbstractConstraint object to an InfiniteModel and assign a name

Create an InfOptConstraintRef that points to the constraint object stored in the model.

The constraint objects are specified via JuMP.build_constraint which requires that the user provides a JuMP.AbstractJuMPScalar, a MOI.AbstractScalarSet, and any keyword arguments such as ParameterBound. For example, let's build a scalar constraint for $3T(t, x) - g^2(t) \leq 0$:

julia> constr = build_constraint(error, 3T - g^2, MOI.LessThan(0.0))
ScalarConstraint{GenericQuadExpr{Float64,GeneralVariableRef},MathOptInterface.LessThan{Float64}}(-g(t)² + 3 T(t, x), MathOptInterface.LessThan{Float64}(0.0))

Now the built constraint object can be added to the infinite model via add_constraint. Let's do so with our example and assign it the name of c1 (note that adding a name is optional):

julia> cref = add_constraint(model, constr, "c1")
c1 : -g(t)² + 3 T(t, x) ≤ 0.0, ∀ t ∈ [0, 10], x[1] ~ Normal, x[2] ~ Normal

Thus, we have made our constraint and added it model and now have a constraint reference cref that we can use to access it.

The @constraint and @BDconstraint automate the above steps where @BDconstraint is specifically used to implement the parameter_bounds keyword argument in build_constraint.

Macro Definition

As mentioned above in the Basic Usage section, the @constraint macro should be used for defining all constraints that do not entail specified parameter bounds. In general, the syntax follows the form @constraint([InfiniteModel], [name][indexing expr], [scalar constr expr]). The second argument is optional and is used to assign a name and/or define indexing variables to be used in the constraint expr. When a name is provided it is registered and cannot be used again for another constraint or variable name. The indexing expression can be used to produce an array of constraints as shown below (notice this is equivalent to looping over individual @constraint calls):

julia> crefs = @constraint(model, [i = 1:2], 2z[i] - g == 0)
2-element Array{InfOptConstraintRef{ScalarShape},1}:
 2 z[1] - g(t) = 0.0, ∀ t ∈ [0, 10]
 2 z[2] - g(t) = 0.0, ∀ t ∈ [0, 10]

julia> crefs = Vector{InfOptConstraintRef{ScalarShape}}(undef, 2);

julia> for i = 1:2
           crefs[i] = @constraint(model, 2z[i] - g == 0)
       end

julia> crefs
2-element Array{InfOptConstraintRef{ScalarShape},1}:
 2 z[1] - g(t) = 0.0, ∀ t ∈ [0, 10]
 2 z[2] - g(t) = 0.0, ∀ t ∈ [0, 10]

Again, please note that only scalar constraints are currently supported and thus the [scalar constr expr] must be scalar.

An interesting corollary is that @constraint in certain cases will produce bound constraints even if the user doesn't specify any parameter bounds. This, occurs when a bounded hold variable is included. For example, if we want to make the constraint $g(t) + 2.5w \leg 2$ it would only be valid over the sub-domain $t \in [0, 5] \subsetneq [0, 10]$ due to the restrictions on w. Thus, @constraint automatically takes this into account and produces a bounded constraint:

julia> @constraint(model, bounded_example, g + 2.5w <= 2)
bounded_example : g(t) + 2.5 w ≤ 2.0, ∀ t ∈ [0, 5]

If this behavior is not desired, the parameter bounds associated with the variable should be deleted and then should be managed individually for each constraint.

The @BDconstraint is very similar except that it adds the capability of symbolically specifying parameter bounds. Thus the syntax is of the form @BDconstraint([InfiniteModel], [name][indexing expr](bound expr), [scalar constr expr]). Note that here the name and [indexing expr] are optional, but the (param bounds) tuple is required. Thus, three arguments must always be given. The (bound expr) can be of the form:

(param in [lb, ub], ...) enforcing param to be in a sub-domain from lb to ub (note ∈ can be used in place of in)

(params in [lb, ub], ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(lb <= param <= ub, ...) enforcing param to be in a sub-domain from lb to ub

(lb <= params <= ub, ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(param == value, ...) enforcing param to be equal to value

(params == value, ...) enforcing that all parameter references in params each to be equal to value

Any combination of the above forms. Must be inside parentheses and comma separated.

Please refer to the Basic Usage section and the manual for examples.

Finally, the @constraint and @BDconstraint macros allow the user to specify the container keyword argument when defining an array of constraints. For example, we can force a group of bounded constraint references to be stored in a JuMP.SparseAxisArray:

julia> @BDconstraint(model, [i = 1:2](x[i] in [0, 1]), T^2 + z[i] <= 1,
                     container = SparseAxisArray)
JuMP.Containers.SparseAxisArray{InfOptConstraintRef{ScalarShape},1,Tuple{Int64}} with 2 entries:
  [2]  =  T(t, x)² + z[2] ≤ 1.0, ∀ t ∈ [0, 10], x[1] ~ Normal, x[2] ~ Normal ∩ [0, 1] 
  [1]  =  T(t, x)² + z[1] ≤ 1.0, ∀ t ∈ [0, 10], x[1] ~ Normal ∩ [0, 1], x[2] ~ Normal

For more information on JuMP containers please visit their page here.

Queries

In this section, we describe a variety of methods to extract constraint information.

Basic

A number of constraint properties can be extracted via constraint references. Principally, the validity, name, model, index, and constraint object can be queried via is_valid, name, owner_model, index, and constraint_object, respectively. These methods all constitute extensions of JuMP methods and follow exactly the same behavior. Let's try them out with the following example:

julia> is_valid(model, measure_constr) # check if contained in model
true

julia> name(measure_constr) # get the name
"measure_constr"

julia> m = owner_model(measure_constr); # get the model it is added to

julia> index(measure_constr) # get the constraint's index
ConstraintIndex(4)

julia> constraint_object(measure_constr) # get the raw constraint datatype
ScalarConstraint{GenericAffExpr{Float64,GeneralVariableRef},MathOptInterface.EqualTo{Float64}}(integral{t ∈ [0, 10]}[g(t)], MathOptInterface.EqualTo{Float64}(4.0))

Also, constraint_by_name can be used to retrieve a constraint reference if only the name is known and its name is unique. For example, let's extract the reference for "c1":

julia> cref = constraint_by_name(model, "c1")
c1 : -g(t)² + 3 T(t, x) ≤ 0.0, ∀ t ∈ [0, 10], x[1] ~ Normal, x[2] ~ Normal

Parameter Bounds

As explained above, bounded constraints serve as an integral capability of InfiniteOpt. Information about parameter bounds can be obtained via has_parameter_bounds and parameter_bounds which indicate if a constraint is bounded and what its ParameterBounds are, respectively. These are exemplified below:

julia> has_parameter_bounds(measure_constr) # check if constraint is bounded
false

julia> has_parameter_bounds(initial)
true

julia> parameter_bounds(initial)
Subdomain bounds (1): t = 0

Note that parameter_bounds will error if the constraint is not bounded.

General

Constraints can be defined in a number of ways symbolically that differ from how it is actually stored in the model. This principally occurs since like terms and constants are combined together where possible with the variable terms on the left hand side and the constant on the right hand side. For instance, the constraint $2g(t) + 3g(t) - 2 \leq 1 + z_1$ would be normalized $5g(t) - z_1 \leq 3$. In accordance with this behavior normalized_rhs and normalized_coefficient can be used to query the normalized right hand side and the coefficient of a particular variable reference, respectively. Let's employ the above example to illustrate this:

julia> @constraint(model, constr, 2g + 3g - 2 <= 1 + z[1])
constr : 5 g(t) - z[1] ≤ 3.0, ∀ t ∈ [0, 10]

julia> normalized_rhs(constr)
3.0

julia> normalized_coefficient(constr, g)
5.0

There also exist a number of methods for querying an infinite model about what constraints it contains. list_of_constraint_types can be used query what types of constraints have been added to a model. This is provided as a list of tuples where the first element is the expression type and the second element is the set type (recall that constraints are stored in the form func-in-set). Thus, for our current model we obtain:

julia> list_of_constraint_types(model)
4-element Array{Tuple{DataType,DataType},1}:
 (GenericQuadExpr{Float64,GeneralVariableRef}, MathOptInterface.LessThan{Float64})
 (GenericQuadExpr{Float64,GeneralVariableRef}, MathOptInterface.GreaterThan{Float64})
 (GenericAffExpr{Float64,GeneralVariableRef}, MathOptInterface.LessThan{Float64})
 (GenericAffExpr{Float64,GeneralVariableRef}, MathOptInterface.EqualTo{Float64})

This information is useful when in combination with the num_constraints and all_constraints methods which can take the expression type and/or the set type as inputs. Here num_constraints provides the number of constraints that match a certain type and all_constraints returns a list of constraint references matching the criteria provided. These have been extended beyond JuMP functionality such additional methods have been provided for the cases in which one wants to query solely off of set or off expression type. Let's illustrate this with num_constraints:

julia> num_constraints(model) # total number of constraints
15

julia> num_constraints(model, GenericQuadExpr{Float64,GeneralVariableRef})
5

julia> num_constraints(model, MathOptInterface.LessThan{Float64})
6

julia> num_constraints(model, GenericQuadExpr{Float64,GeneralVariableRef},
                       MathOptInterface.LessThan{Float64})
4                   

Modification

In this section, we highlight a number of methods that can be used to modify existing constraints.

Deletion

All constraints in InfiniteOpt can be removed in like manner to typical JuMP constraints with the appropriate extension of delete. This will remove the corresponding constraint object from the model. However, please note any registered names will remain registered in the infinite model. This means that a constraint with a registered name cannot be repeatedly added and removed using the same name. To exemplify this, let's delete the constraint c1:

julia> cref = constraint_by_name(model, "c1")
c1 : -g(t)² + 3 T(t, x) ≤ 0.0, ∀ t ∈ [0, 10], x[1] ~ Normal, x[2] ~ Normal

julia> delete(model, cref)

General

There also are a number of ways to modify information and characteristics of constraints. First, set_name can be used to specify a new name for a particular constraint. For instance, let's update the name of initial to "init_cond":

julia> set_name(initial, "init_cond")

julia> initial
init_cond : g(t) = 0.0, ∀ t = 0

We can also update the normalized right hand side constant value or normalized left hand side variable coefficient value using set_normalized_rhs and set_normalized_coefficient , respectively. Let's again consider the constraint $5g(t) - z_1 \leq 3$ as an example. Let's change the constant term to -1 and the g(t) coefficient to 2.5:

julia> set_normalized_rhs(constr, -1)

julia> set_normalized_coefficient(constr, g, 2.5)

julia> constr
constr : 2.5 g(t) - z[1] ≤ -1.0, ∀ t ∈ [0, 10]

Note

In some cases, it may be more convenient to dynamically modify coefficients and other values via the use of finite parameters. This provides an avenue to update parameters without having to be concerned about the normalized form. For more information, see the Finite Parameters page.

Parameter Bounds

Parameter bounds can be added to, modified, or removed from any constraint in InfiniteOpt. Principally, this is accomplished via @add_parameter_bounds, @set_parameter_bounds, and delete_parameter_bounds in like manner to hold variables.

First, parameter bounds can be added to a constraint in an intuitive symbolic syntax via @add_parameter_bounds which follows form @add_parameter_bounds(ref, bound_expr) where (bound_expr) can be of the form:

(param in [lb, ub], ...) enforcing param to be in a sub-domain from lb to ub (note ∈ can be used in place of in)

(params in [lb, ub], ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(lb <= param <= ub, ...) enforcing param to be in a sub-domain from lb to ub

(lb <= params <= ub, ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(param == value, ...) enforcing param to be equal to value

(params == value, ...) enforcing that all parameter references in params each to be equal to value

Any combination of the above forms. Must be inside parentheses and comma separated.

For example, let's add the bound $t \in [0, 1]$ to constr:

julia> @add_parameter_bounds(constr, (t in [0, 1]))

julia> constr
constr : 2.5 g(t) - z[1] ≤ -1.0, ∀ t ∈ [0, 1]

In similar manner, @set_parameter_bounds can be employed to specify what bounds a constraint has (overwriting any existing ones if forced). It follows the same syntax, so let's use it to change the bounds on t to $t = 0$:

julia> @set_parameter_bounds(constr, (t == 0), force = true)

julia> constr
constr : 2.5 g(t) - z[1] ≤ -1.0, ∀ t = 0

Note

Constraint parameters bounds are intersected with those of any hold variables that are part of the constraint. If no such intersection exists then an error is thrown.

Finally, constraint bounds can be deleted via delete_parameter_bounds which deletes all parameter bounds. Again, note the parameter bounds associated with hold variables will be unaffected and can only be removed by deleting them from the variables directly. Now let's delete the parameter bounds associated with our example:

julia> delete_parameter_bounds(constr)

julia> constr
constr : 2.5 g(t) - z[1] ≤ -1.0, ∀ t ∈ [0, 10]

Datatypes

InfiniteOpt.BoundedScalarConstraint

InfiniteOpt.ConstraintData

InfiniteOpt.ConstraintIndex

InfiniteOpt.InfOptConstraintRef

InfiniteOpt.BoundedScalarConstraint — Type

BoundedScalarConstraint{F <: JuMP.AbstractJuMPScalar,
                        S <: MOI.AbstractScalarSet,
                        P <: GeneralVariableRef
                        } <: JuMP.AbstractConstraint

A DataType that stores scalar constraints that are defined over a sub-domain of infinite parameters.

Fields

func::F The JuMP object.

set::S The MOI set.

bounds::ParameterBounds{P} Set of valid parameter sub-domains that further boundconstraint.

orig_bounds::ParameterBounds{P} Set of the constraint's original parameter sub-domains (not considering hold variables)

InfiniteOpt.ConstraintData — Type

ConstraintData <: AbstractDataObject

A mutable DataType for storing constraints and their data.

Fields

constraint::JuMP.AbstractConstraint: The scalar constraint.

object_nums::Vector{Int}: The object numbers of the parameter objects that the constraint depends on.

name::String: The name used for printing.

measure_indices::Vector{MeasureIndex}: Indices of dependent measures.

is_info_constraint::Bool: Is this is constraint based on variable info (e.g., lower bound)

InfiniteOpt.ConstraintIndex — Type

ConstraintIndex <: ObjectIndex

A DataType for storing the index of constraint objects.

Fields

value::Int: The index value.

InfiniteOpt.InfOptConstraintRef — Type

InfOptConstraintRef{S <: JuMP.AbstractShape}

A DataType for constraints that are in InfiniteModels

Fields

model::InfiniteModel: Infinite model.

index::ConstraintIndex: Index of the constraint in model.

shape::JuMP.AbstractShape: Shape of the constraint

Methods/Macros

InfiniteOpt.@BDconstraint

InfiniteOpt.add_parameter_bounds

InfiniteOpt.delete_parameter_bounds

InfiniteOpt.has_parameter_bounds

InfiniteOpt.parameter_bounds

InfiniteOpt.parameter_refs

InfiniteOpt.set_parameter_bounds

JuMP.add_constraint

JuMP.add_to_function_constant

JuMP.all_constraints

JuMP.all_constraints

JuMP.all_constraints

JuMP.all_constraints

JuMP.build_constraint

JuMP.constraint_by_name

JuMP.constraint_object

JuMP.delete

JuMP.index

JuMP.is_valid

JuMP.list_of_constraint_types

JuMP.name

JuMP.normalized_coefficient

JuMP.normalized_rhs

JuMP.num_constraints

JuMP.num_constraints

JuMP.num_constraints

JuMP.num_constraints

JuMP.owner_model

JuMP.set_name

JuMP.set_normalized_coefficient

JuMP.set_normalized_rhs

InfiniteOpt.@BDconstraint — Macro

@BDconstraint(model::InfiniteModel, [i = ..., ...](bound_expr), constr_expr;
              [kw_args...])::InfOptConstraintRef

Add an anonymous bounded constraint to model and return an appropriate container of constraint reference(s).

    @BDconstraint(model::InfiniteModel, name[i = ..., ...](bound_expr),
                  constr_expr; [kw_args...])::InfOptConstraintRef

Add a named bounded constraint to model and return an appropriate container of constraint reference(s). This defines the constraint as expressed in constr_expr over some sub-domain of parameters as indicated by bound_expr. The format of constr_expr must follow the same syntax as that specified in JuMP.@constraint. For example, if we want to express the constraint 2T(t, x) + 3y = 42, the constraint expression constr_expr would be 2 * T + 3 * y == 42.

By default, JuMP.@constraint would express the above example constraint over the whole domain of t and x. However, we can use @BDconstraint to express this constraint over some sub-domain(s) via bound_expr. Here (bound_expr) can be of the form:

(param in [lb, ub], ...) enforcing param to be in a sub-domain from lb to ub (note ∈ can be used in place of in)

(params in [lb, ub], ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(lb <= param <= ub, ...) enforcing param to be in a sub-domain from lb to ub

(lb <= params <= ub, ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(param == value, ...) enforcing param to be equal to value

(params == value, ...) enforcing that all parameter references in params each to be equal to value

Any combination of the above forms. Must be inside parentheses and comma separated.

Please note that when specifying value conditions on dependent infinite parameters, a value must be provided for each parameter in the dependent container otherwise an error will be thrown.

Like typical constraints, the container keyword argument can be used to specify the JuMP container type used to store the constraint references. Note this macro errors if bound_expr is ommited or if some unrecognized syntax is used.

Examples

julia> @BDconstraint(model, c1(t in [0, 1]), T^2 + z <= 1)
c1 : T(x, t)² + z ≤ 1.0, ∀ t ∈ [0, 1], x[1] ∈ [-1, 1], x[2] ∈ [-1, 1], x[3] ∈ [-1, 1],

julia> @BDconstraint(model, c2[i = 1:3](x[i] in [0, 1]), T^2 + z + x[i] <= 1)
3-element Array{InfiniteConstraintRef,1}:
 c2[1] : T(x, t)² + z + x[1] ≤ 1.0, ∀ t ∈ [0, 1], x[1] ∈ [0, 1], x[2] ∈ [-1, 1], x[3] ∈ [-1, 1]
 c2[2] : T(x, t)² + z + x[2] ≤ 1.0, ∀ t ∈ [0, 1], x[1] ∈ [-1, 1], x[2] ∈ [0, 1], x[3] ∈ [-1, 1]
 c2[3] : T(x, t)² + z + x[3] ≤ 1.0, ∀ t ∈ [0, 1], x[1] ∈ [-1, 1], x[2] ∈ [-1, 1], x[3] ∈ [0, 1]

julia> @BDconstraint(model, (x == 0, t == 0), T^2 + z <= 1)
T(x, t)² + z ≤ 1.0, ∀ t = 0, x[1] = 0, x[2] = 0, x[3] = 0

julia> @BDconstraint(model, [i = 1:3](x[i] == 0), T^2 + z <= 1,
                     container = SparseAxisArray)
JuMP.Containers.SparseAxisArray{InfiniteConstraintRef,1,Tuple{Any}} with 3 entries:
  [3]  =  T(x, t)² + z ≤ 1.0, ∀ t ∈ [0, 1], x[1] = 0, x[2] ∈ [-1, 1], x[3] ∈ [-1, 1]
  [2]  =  T(x, t)² + z ≤ 1.0, ∀ t ∈ [0, 1], x[1] ∈ [-1, 1], x[2] = 0, x[3] ∈ [-1, 1]
  [1]  =  T(x, t)² + z ≤ 1.0, ∀ t ∈ [0, 1], x[1] ∈ [-1, 1], x[2] ∈ [-1, 1], x[3] = 0

JuMP.build_constraint — Method

JuMP.build_constraint(_error::Function, expr::JuMP.AbstractJuMPScalar,
                      set::MOI.AbstractScalarSet;
                      [parameter_bounds::ParameterBounds = ParameterBounds()]
                      )::JuMP.AbstractConstraint

Extend JuMP.build_constraint to accept the parameter_bounds argument and return a BoundedScalarConstraint if the parameter_bounds keyword argument is specifed or return a JuMP.ScalarConstraint otherwise. This is primarily intended to work as an internal function for constraint macros.

Example

julia> @infinite_parameter(model, t in [0, 10]);

julia> @infinite_variable(model, g(t));

julia> @hold_variable(model, x);

julia> constr = build_constraint(error, g + x, MOI.EqualTo(42.0),
              parameter_bounds = ParameterBounds((t => IntervalSet(0, 1),)));

julia> isa(constr, BoundedScalarConstraint)
true

JuMP.add_constraint — Method

JuMP.add_constraint(model::InfiniteModel, c::JuMP.AbstractConstraint,
                    [name::String = ""])::InfOptConstraintRef

Extend JuMP.add_constraint to add a constraint c to an infinite model model with name name. Returns an appropriate constraint reference whose type depends on what variables are used to define the constraint. Errors if a vector constraint is used, the constraint only constains parameters, or if any variables do not belong to model. This is primarily used as an internal method for the cosntraint macros.

Example

julia> @infinite_parameter(model, t in [0, 10]);

julia> @infinite_variable(model, g(t));

julia> @hold_variable(model, x);

julia> constr = build_constraint(error, g + x, MOI.EqualTo(42));

julia> cref = add_constraint(model, constr, "name")
name : g(t) + x = 42.0, ∀ t ∈ [0, 10]

JuMP.owner_model — Method

JuMP.owner_model(cref::InfOptConstraintRef)::InfiniteModel

Extend JuMP.owner_model to return the infinite model associated with cref.

Example

julia> model = owner_model(cref)
An InfiniteOpt Model
Minimization problem with:
Finite Parameters: 0
Infinite Parameters: 3
Variables: 3
Measures: 0
Objective function type: GeneralVariableRef
`GenericAffExpr{Float64,GeneralVariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
Names registered in the model: g, t, h, x
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

JuMP.index — Method

JuMP.index(cref::InfOptConstraintRef)::ConstraintIndex

Extend JuMP.index to return the index of an InfiniteOpt constraint cref.

Example

julia> index(cref)
ConstrainIndex(2)

JuMP.constraint_object — Method

JuMP.constraint_object(cref::InfOptConstraintRef)::JuMP.AbstractConstraint

Extend JuMP.constraint_object to return the constraint object associated with cref.

Example

julia> @infinite_parameter(model, t in [0, 10]);

julia> @hold_variable(model, x <= 1);

julia> cref = UpperBoundRef(x);

julia> obj = constraint_object(cref)
ScalarConstraint{GeneralVariableRef,MathOptInterface.LessThan{Float64}}(x,
MathOptInterface.LessThan{Float64}(1.0))

JuMP.name — Method

JuMP.name(cref::InfOptConstraintRef)::String

Extend [JuMP.name](@ref JuMP.name(::JuMP.ConstraintRef{JuMP.Model,<:JuMP._MOICON}) to return the name of an InfiniteOpt constraint.

Example

julia> name(cref)
"constr_name"

JuMP.set_name — Method

JuMP.set_name(cref::InfOptConstraintRef, name::String)::Nothing

Extend JuMP.set_name to specify the name of a constraint cref.

Example

julia> set_name(cref, "new_name")

julia> name(cref)
"new_name"

JuMP.is_valid — Method

JuMP.is_valid(model::InfiniteModel, cref::InfOptConstraintRef)::Bool

Extend JuMP.is_valid to return Bool whether an InfiniteOpt constraint reference is valid.

Example

julia> is_valid(model, cref)
true

JuMP.delete — Method

JuMP.delete(model::InfiniteModel, cref::InfOptConstraintRef)::Nothing

Extend JuMP.delete to delete an InfiniteOpt constraint and all associated information. Errors if cref is invalid.

Example

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 z ≥ 0.0
 g(t) + z ≥ 42.0, ∀ t ∈ [0, 6]

julia> delete(model, cref)

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 z ≥ 0.0

InfiniteOpt.parameter_refs — Method

parameter_refs(cref::InfOptConstraintRef)::Tuple

Return the tuple of infinite parameter references that determine the infinite dependencies of cref.

Example

julia> parameter_refs(cref)
(t,)

InfiniteOpt.has_parameter_bounds — Method

has_parameter_bounds(cref::InfOptConstraintRef)::Bool

Return a Bool indicating if cref is limited to a sub-domain as defined by a ParameterBounds object.

Example

julia> has_parameter_bounds(cref)
true

InfiniteOpt.parameter_bounds — Method

parameter_bounds(cref::InfOptConstraintRef)::ParameterBounds{GeneralVariableRef}

Return the ParameterBounds object associated with the constraint cref. Errors if cref does not have parameter bounds.

Example

julia> parameter_bounds(cref)
Subdomain bounds (1): t ∈ [0, 2]

InfiniteOpt.set_parameter_bounds — Method

set_parameter_bounds(cref::InfOptConstraintRef,
                     bounds:ParameterBounds{GeneralVariableRef};
                     [force::Bool = false])::Nothing

Specify a new ParameterBounds object bounds for the constraint cref. This is meant to be primarily used by @set_parameter_bounds which provides a more intuitive syntax.

Example

julia> set_parameter_bounds(cref, ParameterBounds((t => IntervalSet(0, 2),)))

julia> parameter_bounds(cref)
Subdomain bounds (1): t ∈ [0, 2]

InfiniteOpt.add_parameter_bounds — Method

add_parameter_bounds(cref::InfOptConstraintRef,
                     new_bounds::ParameterBounds{GeneralVariableRef}
                     )::Nothing

Add an additional parameter bound to cref such that it is defined over the sub-domain based on pref from lower to upper. This is primarily meant to be used by @add_parameter_bounds.

julia> add_parameter_bounds(cref, ParameterBounds((t => IntervalSet(0, 2),))

julia> parameter_bounds(cref)
Subdomain bounds (1): t ∈ [0, 2]

InfiniteOpt.delete_parameter_bounds — Method

delete_parameter_bounds(cref::InfOptConstraintRef)::Nothing

Delete all the parameter bounds of the constraint cref. Note any bounds that are needed for hold variables inside in cref will be unaffected.

Example

julia> @BDconstraint(model, c1(x == 0), y <= 42)
c1 : y(x) ≤ 42, ∀ x[1] = 0, x[2] = 0

julia> delete_parameter_bounds(c1)

julia> c1
c1 : y(x) ≤ 42, ∀ x[1] ∈ [-1, 1], x[2] ∈ [-1, 1]

JuMP.set_normalized_rhs — Method

JuMP.set_normalized_rhs(cref::InfOptConstraintRef, value::Real)::Nothing

Set the right-hand side term of constraint to value. Note that prior to this step, JuMP will aggregate all constant terms onto the right-hand side of the constraint. For example, given a constraint 2x + 1 <= 2, set_normalized_rhs(con, 4) will create the constraint 2x <= 4, not 2x + 1 <= 4.

julia> @constraint(model, con, 2x + 1 <= 2)
con : 2 x ≤ 1.0

julia> set_normalized_rhs(con, 4)

julia> con
con : 2 x ≤ 4.0

JuMP.normalized_rhs — Method

JuMP.normalized_rhs(cref::InfOptConstraintRef)::Float64

Return the right-hand side term of cref after JuMP has converted the constraint into its normalized form.

JuMP.add_to_function_constant — Method

JuMP.add_to_function_constant(cref::InfOptConstraintRef, value::Real)::Nothing

Add value to the function constant term. Note that for scalar constraints, JuMP will aggregate all constant terms onto the right-hand side of the constraint so instead of modifying the function, the set will be translated by -value. For example, given a constraint 2x <= 3, add_to_function_constant(c, 4) will modify it to 2x <= -1. ```

JuMP.set_normalized_coefficient — Method

JuMP.set_normalized_coefficient(cref::InfOptConstraintRef,
                                variable::GeneralVariableRef,
                                value::Real)::Nothing

Set the coefficient of variable in the constraint constraint to value. Note that prior to this step, JuMP will aggregate multiple terms containing the same variable. For example, given a constraint 2x + 3x <= 2, set_normalized_coefficient(con, x, 4) will create the constraint 4x <= 2.

julia> con
con : 5 x ≤ 2.0

julia> set_normalized_coefficient(con, x, 4)

julia> con
con : 4 x ≤ 2.0

JuMP.normalized_coefficient — Method

JuMP.normalized_coefficient(cref::InfOptConstraintRef,
                            variable::GeneralVariableRef)::Float64

Return the coefficient associated with variable in constraint after JuMP has normalized the constraint into its standard form.

JuMP.constraint_by_name — Method

JuMP.constraint_by_name(model::InfiniteModel,
                        name::String)::Union{InfOptConstraintRef, Nothing}

Extend JuMP.constraint_by_name to return the constraint reference associated with name if one exists or returns nothing. Errors if more than one constraint uses the same name.

Example

julia> constraint_by_name(model, "constr_name")
constr_name : x + pt = 3.0

JuMP.list_of_constraint_types — Method

JuMP.list_of_constraint_types(model::InfiniteModel)::Vector{Tuple{DataType, DataType}}

Extend JuMP.list_of_constraint_types to return a list of tuples that contain all the used combinations of function types and set types in the model.

julia> all_constraints(model)
3-element Array{Tuple{DataType,DataType},1}:
 (GeneralVariableRef, MathOptInterface.LessThan{Float64})
 (GeneralVariableRef, MathOptInterface.GreaterThan{Float64})
 (GeneralVariableRef, MathOptInterface.Integer)

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel,
                     function_type::Type{<:JuMP.AbstractJuMPScalar},
                     set_type::Type{<:MOI.AbstractSet})::Int

Extend JuMP.num_constraints to return the number of constraints with a partiuclar function type and set type.

Example

julia> num_constraints(model, HoldVariableRef, MOI.LessThan)
1

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel,
                     function_type::Type{<:JuMP.AbstractJuMPScalar})::Int

Extend JuMP.num_constraints to search by function types for all MOI sets and return the total number of constraints with a particular function type.

julia> num_constraints(model, HoldVariableRef)
3

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel,
                     function_type::Type{<:MOI.AbstractSet})::Int

Extend JuMP.num_constraints to search by MOI set type for all function types and return the total number of constraints that use a particular MOI set type.

julia> num_constraints(model, MOI.LessThan)
2

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel)::Int

Extend JuMP.num_constraints to return the total number of constraints in an infinite model model.

julia> num_constraints(model)
4

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel,
                     function_type::Type{<:JuMP.AbstractJuMPScalar},
                     set_type::Type{<:MOI.AbstractSet}
                     )::Vector{InfOptConstraintRef}

Extend JuMP.all_constraints to return a list of all the constraints with a particular function type and set type.

julia> all_constraints(model, GeneralVariableRef, MOI.LessThan)
1-element Array{InfOptConstraintRef,1}:
 x ≤ 1.0

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel,
                     function_type::Type{<:JuMP.AbstractJuMPScalar}
                     )::Vector{InfOptConstraintRef}

Extend JuMP.all_constraints to search by function types for all MOI sets and return a list of all constraints use a particular function type.

julia> all_constraints(model, GeneralVariableRef)
3-element Array{InfOptConstraintRef,1}:
 x ≥ 0.0
 x ≤ 3.0
 x integer

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel,
                     set_type::Type{<:MOI.AbstractSet}
                     )::Vector{InfOptConstraintRef}

Extend JuMP.all_constraints to search by MOI set type for all function types and return a list of all constraints that use a particular set type.

julia> all_constraints(model, MOI.GreaterThan)
3-element Array{InfOptConstraintRef,1}:
 x ≥ 0.0
 g(t) ≥ 0.0, ∀ t ∈ [0, 6]
 g(0.5) ≥ 0.0

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel)::Vector{InfOptConstraintRef}

Extend JuMP.all_constraints to return all a list of all the constraints in an infinite model model.

julia> all_constraints(model)
5-element Array{InfOptConstraintRef,1}:
 x ≥ 0.0
 x ≤ 3.0
 x integer
 g(t) ≥ 0.0, ∀ t ∈ [0, 6]
 g(0.5) ≥ 0.0