Finite Parameters

A guide and manual to the definition and use of finite parameters in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

Often a mathematical model needs to be optimized several times in accordance with a set of fixed parameter values. In such cases, it is typically preferable to modify these values in place without having to redefine the entire model. This ability is provided in InfiniteOpt via @finite_parameter which permits users to define finite parameters whose values can later be modified as needed. Furthermore, at the optimization step these parameters are replaced with their numeric values. Thus, not adding unnecessary decision variables as is typically done in JuMP models using JuMP.fix on placeholder variables.

Basic Usage

Once an InfiniteModel model has been defined we can add a finite parameter via @finite_parameter. For example, let's define a maximum cost parameter called max_cost with an initial value of 42:

julia> @finite_parameter(model, max_cost, 42)
max_cost

Notice that a Julia variable called max_cost is returned that contains a GeneralVariableRef that points to the finite parameter we have just created. An array of parameters can also be defined following standard JuMP macro syntax:

julia> values = [2, 3.2, 1];

julia> @finite_parameter(model, params[i = 1:3], values[i])
3-element Array{GeneralVariableRef,1}:
 params[1]
 params[2]
 params[3]

The @finite_parameter macro emulates all typical JuMP functionality and can define anonymous parameters, use JuMP containers and more. We refer to its documentation below to learn more. Once a finite parameter is defined the corresponding GeneralVariableRef can be used in expressions, objectives, measures, and constraints just like infinite parameters.

The value of a finite parameter can be checked using parameter_value and can modified using JuMP.set_value. For example, let's update the value of max_cost to be now be 10.2:

julia> parameter_value(max_cost)
42.0

julia> set_value(max_cost, 10.2)

julia> parameter_value(max_cost)
10.2

Advanced Details

The ability to implement finite parameters stems from its ability to support mixed variable types using by using the GeneralVariableRef buffer. As such, finite parameters will be treated as variables until the model is transcribed. For example, this means that the expression max_cost * x will be treated as a quadratic expression when it is expressed in its InfiniteOpt form, however it is converted into the appropriate affine expression when transcripted.

Note

In previous versions finite parameters were just special cases of infinite parameters. However, they now have their own distinct underlying data structure.

Warning

InfiniteOpt's implementation of finite parameters should not be a reason to use InfiniteOpt to model non-infinite-dimensional problems, since the added overhead will make it slower than just iteratively building JuMP models.

Datatypes

InfiniteOpt.FiniteParameter

InfiniteOpt.FiniteParameterIndex

InfiniteOpt.FiniteParameterRef

InfiniteOpt.FiniteParameter — Type

FiniteParameter <: ScalarParameter

A DataType for storing finite parameters meant to be nested in expressions and replaced with their values at runtime.

Fields

value::Float64: The parameter value.

InfiniteOpt.FiniteParameterIndex — Type

FiniteParameterIndex <: ObjectIndex

A DataType for storing the index of a FiniteParameter.

Fields

value::Int64: The index value.

InfiniteOpt.FiniteParameterRef — Type

FiniteParameterRef <: FiniteVariableRef

A DataType for finite parameters references who are replaced with their values at the transcription step.

Fields

model::InfiniteModel: Infinite model.

index::FiniteParameterIndex: Index of the parameter in model.

Methods/Macros

InfiniteOpt.@finite_parameter

InfiniteOpt.parameter_value

JuMP.set_value

InfiniteOpt.@finite_parameter — Macro

@finite_parameter(model::InfiniteModel, value)::GeneralVariableRef

Define and add an anonymous finite parameter to model and return its parameter reference. Its value is equal to value.

    @finite_parameter(model::InfiniteModel, param_expr,
                      value_expr)::GeneralVariableRef

Define and add a finite parameter(s) to model and return appropriate parameter reference(s). The parameter(s) has/have value(s) indicated by the value_expr. The expression param_expr can be of the form:

paramname creating a scalar parameter of name paramname

paramname[...] or [...] creating a container of parameters

The expression value_expr simply expresses the value of the parameter(s). This is typically a number but could be an array indexed using an index defined in param_expr.

The recognized keyword arguments in kw_args are the following:

base_name: Sets the name prefix used to generate parameter names. It corresponds to the parameter name for scalar parameter, otherwise, the parameter names are set to base_name[...] for each index ... of the axes axes.

container: Specify the container type.

Examples

julia> par = @finite_parameter(model, 2)
noname

julia> vals = [3, 2];

julia> pars = @finite_parameter(model, [i = 1:2], vals[i], base_name = "par")
2-element Array{ParameterRef,1}:
 par[1]
 par[2]

julia> @finite_parameter(model, par2, 42)
par2

InfiniteOpt.parameter_value — Method

parameter_value(pref::FiniteParameterRef)::Float64

Return the value of a finite parameter reference pref. Errors if it is an infinite parameter.

Example

julia> value(cost)
42.0

JuMP.set_value — Method

JuMP.set_value(pref::FiniteParameterRef, value::Real)::Nothing

Set the value of pref so long as it is a finite parameter. Errors if it is an infinite parameter.

Example

julia> set_value(cost, 27)

julia> value(cost)
27.0