Measures

A guide and manual for defining and using measures in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

Basic Usage

Theoretical Abstraction

Preset Evaluations

Custom Evaluations

Expansion

Reduced Infinite Variables

Datatypes

InfiniteOpt.AbstractMeasureData

InfiniteOpt.AbstractReducedInfo

InfiniteOpt.DiscreteMeasureData

InfiniteOpt.DiscreteMeasureData

InfiniteOpt.DiscreteMeasureData

InfiniteOpt.Measure

InfiniteOpt.MeasureRef

InfiniteOpt.MultiDiscreteMeasureData

InfiniteOpt.ReducedInfiniteInfo

InfiniteOpt.ReducedInfiniteVariableRef

InfiniteOpt.AbstractMeasureData — Type

AbstractMeasureData

An abstract type to define data for measures to define the behavior of Measure.

InfiniteOpt.DiscreteMeasureData — Type

DiscreteMeasureData <: AbstractMeasureData

A DataType for one dimensional measure abstraction data where the measure abstraction is of the form: $measure = \int_{\tau \in T} f(\tau) w(\tau) d\tau \approx \sum_{i = 1}^N \alpha_i f(\tau_i) w(\tau_i)$.

Fields

parameter_ref::ParameterRef The infinite parameter over which the integration occurs.

coefficients::Vector{<:Number} Coefficients $\alpha_i$ for the above measure abstraction.

supports::Vector{<:Number} Support points $\tau_i$ for the above measure abstraction.

name::String Name of the measure that will be implemented.

weight_function::Function Weighting function $w$ must map support value input value of type Number to a scalar value.

InfiniteOpt.MultiDiscreteMeasureData — Type

MultiDiscreteMeasureData <: AbstractMeasureData

A DataType for multi-dimensional measure abstraction data where the measure abstraction is of the form: $measure = \int_{\tau \in T} f(\tau) w(\tau) d\tau \approx \sum_{i = 1}^N \alpha_i f(\tau_i) w(\tau_i)$.

Fields

parameter_ref::JuMP.Containers.SparseAxisArray{<:ParameterRef} The infinite parameters over which the integration occurs.

coefficients::Vector{<:Number} Coefficients $\alpha_i$ for the above measure abstraction.

supports::Vector{<:JuMP.Containers.SparseAxisArray{<:Number}} Support points $\tau_i$ for the above measure abstraction.

name::String Name of the measure that will be implemented.

weight_function::Function Weighting function $w$ must map a numerical support of type JuMP.Containers.SparseAxisArray to a scalar value.

InfiniteOpt.Measure — Type

Measure{T <: JuMP.AbstractJuMPScalar, V <: AbstractMeasureData}

A DataType for measure abstractions.

Fields

func::T Infinite variable expression.

data::V Data of the abstraction as described in a AbstractMeasureData subtype.

InfiniteOpt.MeasureRef — Type

MeasureRef <: FiniteVariableRef

A DataType for referring to measure abstractions.

Fields

model::InfiniteModel Infinite model.

index::Int Index of variable in model.

InfiniteOpt.AbstractReducedInfo — Type

AbstractReducedInfo

An abstract type reduced variable information.

InfiniteOpt.ReducedInfiniteInfo — Type

ReducedInfiniteInfo <: AbstractReducedInfo

A DataType for storing reduced infinite variable information.

Fields

infinite_variable_ref::InfiniteVariableRef The original infinite variable.

eval_supports::Dict{Int, Union{Number, JuMP.Containers.SparseAxisArray{<:Number}}} The original parameter tuple indices to the evaluation supports.

InfiniteOpt.ReducedInfiniteVariableRef — Type

ReducedInfiniteVariableRef <: GeneralVariableRef

A DataType for partially transcripted infinite dimensional variable references. This is used to expand measures that contain infinite variables that are not fully transcripted by the measure.

Fields

model::InfiniteModel Infinite model.

index::Int Index of variable in model.

Methods

InfiniteOpt.add_measure

InfiniteOpt.eval_supports

InfiniteOpt.expand

InfiniteOpt.expand_all_measures!

InfiniteOpt.expect

InfiniteOpt.infinite_variable_ref

InfiniteOpt.is_used

InfiniteOpt.measure

InfiniteOpt.measure

InfiniteOpt.measure_data

InfiniteOpt.measure_function

InfiniteOpt.parameter_refs

InfiniteOpt.support_sum

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_objective

JuMP.BinaryRef

JuMP.FixRef

JuMP.IntegerRef

JuMP.LowerBoundRef

JuMP.UpperBoundRef

JuMP.delete

JuMP.delete

JuMP.fix_value

JuMP.has_lower_bound

JuMP.has_upper_bound

JuMP.is_binary

JuMP.is_fixed

JuMP.is_integer

JuMP.is_valid

JuMP.is_valid

JuMP.lower_bound

JuMP.name

JuMP.name

JuMP.set_name

JuMP.start_value

JuMP.upper_bound

InfiniteOpt.expect — Function

expect(expr::JuMP.AbstractJuMPScalar,
       [params::Union{ParameterRef, AbstractArray{<:ParameterRef},
                      Nothing} = nothing];
       [num_supports::Int = 50,
        use_existing_supports::Bool = false])::MeasureRef

Creates a measure that represents the expected value of an expression in a random parameter involved in the expression. Return the MeasureRef of the created measure.

Example

julia> @infinite_parameter(model, x in Normal(0., 1.))
x

julia> @infinite_variable(model, f(x))
f(x)

julia> meas = expect(f, num_supports = 2)
expect(f(x))

julia> expand(meas)
0.5 f(0.6791074260357777) + 0.5 f(0.8284134829000359)

InfiniteOpt.support_sum — Function

support_sum(expr::JuMP.AbstractJuMPScalar,
            [params::Union{ParameterRef, AbstractArray{<:ParameterRef},
                           Nothing} = nothing])::MeasureRef

Creates a measure that represents the sum of the expression over a parameter using its existing supports. Return the MeasureRef of the created measure.

Example

julia> @infinite_parameter(model, x in [0, 1], supports = [0.3, 0.7])
x

julia> @infinite_variable(model, f(x))
f(x)

julia> meas = support_sum(f)
sum(f(x))

julia> expand(meas)
f(0.3) + f(0.7)

InfiniteOpt.measure — Method

measure(expr::JuMP.AbstractJuMPScalar,
        [params::Union{ParameterRef, AbstractArray{<:ParameterRef},
                      Nothing} = nothing,
        lb::Union{Number, AbstractArray{<:Number}, Nothing},
        ub::Union{Number, AbstractArray{<:Number}, Nothing}];
        [eval_method::Function = mc_sampling, num_supports::Int = 50,
        weight_func::Function = _w, name = "measure",
        use_existing_supports::Bool = false,
        call_from_expect::Bool = false])::MeasureRef

Returns a measure reference that evaluates expr without using an object of AbstractMeasureData type. Similar to the main measure method, this function aims to implement measures of the form: $\int_{p \in P} expr(p) w(p) dp$ where $p$ is an infinite parameter (scalar or vector) and $w$ is the weight function. This function serves as a flexible interface where users only have to provide necessary data about the integration. Instead of taking an AbstractMeasureData object as input, this function constructs the AbstractMeasureData object using some default numerical integration schemes. By default, the function will generate points by Monte Carlo sampling from the interval if the parameter is in an IntervalSet, or from the underlying distribution if the parameter is in a DistributionSet. If the expression involves multiple groups of parameters, the user needs to specify the parameter. The user can also specify lower bounds and upper bounds for the parameters, number of supports to generate, and the function to generate the supports with. The last option makes the method extendible for more schemes to generate supports.

Example

julia> @infinite_parameter(model, x in [0., 1.])
x

julia> @infinite_variable(model, f(x))
f(x)

julia> meas = measure(f, num_supports = 5)
(f(x))

julia> expand(meas)
0.2 f(0.8236475079774124) + 0.2 f(0.9103565379264364) + 0.2 f(0.16456579813368521) + 0.2 f(0.17732884646626457) + 0.2 f(0.278880109331201)

InfiniteOpt.DiscreteMeasureData — Method

DiscreteMeasureData(parameter_ref::ParameterRef,
                    coefficients::Vector{<:Number},
                    supports::Vector{<:Number}; name::String = "measure",
                    weight_function::Function = w(t) = 1)::DiscreteMeasureData

Returns a DiscreteMeasureData object that can be utilized to define measures using measure. This accepts input for a scalar (single) parameter. Note that name is used for printing purposes and a description of the other arguments is provided in the documentation for DiscreteMeasureData. Errors if supports are out bounds or an unequal number of supports and coefficients are given.

Example

julia> data = DiscreteMeasureData(pref, [0.5, 0.5], [1, 2], name = "example")
DiscreteMeasureData(pref, [0.5, 0.5], [1, 2], "example", InfiniteOpt._w)

InfiniteOpt.DiscreteMeasureData — Method

DiscreteMeasureData(parameter_ref::AbstractArray{<:ParameterRef},
                    coefficients::Vector{<:Number},
                    supports::Vector{<:AbstractArray{<:Number}};
                    name::String = "measure",
                    weight_function::Function = w(t) = 1
                    )::MultiDiscreteMeasureData

Returns a MultiDiscreteMeasureData object that can be utilized to define measures using measure. This accepts input for an array (multi) parameter. The inner arrays in the supports vector need to match the formatting of the array used for parameter_ref. Note that name is used for printing purposes and a description of the other arguments is provided in the documentation for MultiDiscreteMeasureData. Errors if supports are out bounds, an unequal number of supports and coefficients are given, the array formats do not match, or the parameters have different group IDs.

Example

julia> data = DiscreteMeasureData(prefs, [0.5, 0.5], [[1, 1], [2, 2]], name = "example");

julia> typeof(data)
MultiDiscreteMeasureData

InfiniteOpt.measure — Method

measure(expr::JuMP.AbstractJuMPScalar, data::AbstractMeasureData)::MeasureRef

Return a measure reference that evaluates expr using according to data. This is the preferred method for implementing measures which follow the form: $\int_{p \in P} expr(p) w(p) dp$ where $p$ is an infinite parameter (scalar or vector) and $w$ is the weight function. The measure data data determines how the measure is to be evaluated. Typically, the DiscreteMeasureData constructor can be used to for data. The variable expression expr can contain InfiniteOpt variables, infinite parameters, other measure references (meaning measures can be nested), and constants. Errors if expr does not contain infinite variables, infinite parameters, or measure references. Also errors if the measure parameter specified in data is not in expr and is not in any the nested measure references. Typically, this is called inside of JuMP.@expression, JuMP.@objective, and JuMP.@constraint in a manner similar to sum. Note measures are not explicitly evaluated until build_optimizer_model! is called or unless they are expanded via expand or expand_all_measures!.

Example

julia> tdata = DiscreteMeasureData(t, [0.5, 0.5], [1, 2], name = "name1");

julia> xdata = DiscreteMeasureData(xs, [0.5, 0.5], [[-1, -1], [1, 1]],
                                   name = "name2");

julia> constr_RHS = @expression(model, measure(g - s + 2, tdata) + s^2)
name1(g(t) - s + 2) + s²

julia> @objective(model, Min, measure(g - 1  + measure(T, xdata), tdata))
name1(g(t) - 1 + name2(T(t, x)))

InfiniteOpt.add_measure — Function

add_measure(model::InfiniteModel, meas::Measure)::MeasureRef

Add a measure to model and return the corresponding measure reference. This operates in a manner similar to JuMP.add_variable.

InfiniteOpt.measure_function — Function

measure_function(mref::MeasureRef)::JuMP.AbstractJuMPScalar

Return the function associated with mref.

InfiniteOpt.measure_data — Function

measure_data(mref::MeasureRef)::AbstractMeasureData

Return the measure data associated with mref.

InfiniteOpt.used_by_constraint — Method

used_by_constraint(mref::MeasureRef)::Bool

Return a Bool indicating if mref is used by a constraint.

Example

julia> used_by_constraint(mref)
false

InfiniteOpt.used_by_measure — Method

used_by_measure(mref::MeasureRef)::Bool

Return a Bool indicating if mref is used by a measure.

Example

julia> used_by_measure(mref)
true

InfiniteOpt.used_by_objective — Method

used_by_objective(vmref::MeasureRef)::Bool

Return a Bool indicating if mref is used by the objective.

Example

julia> used_by_objective(mref)
true

InfiniteOpt.is_used — Method

is_used(mref::MeasureRef)::Bool

Return a Bool indicating if mref is used in the model.

Example

julia> is_used(mref)
true

JuMP.is_valid — Method

JuMP.is_valid(model::InfiniteModel, mref::MeasureRef)::Bool

Extend JuMP.is_valid to return Bool whether mref is valid.

JuMP.delete — Method

JuMP.delete(model::InfiniteModel, mref::MeasureRef)

Extend JuMP.delete to delete measures. Errors if measure is invalid, meaning it does not belong to the model or it has already been deleted.

Example

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 z >= 0.0
 measure(g(t)) == 0
 g(t) + z >= 42.0
 g(0.5) == 0
 t in [0, 6]

julia> delete(model, meas)

julia> print(model)
Min z
Subject to
 z >= 0.0
 0 == 0
 g(t) + z >= 42.0
 g(0.5) == 0
 t in [0, 6]

JuMP.name — Method

JuMP.name(mref::MeasureRef)::String

Extend JuMP.name to return the name associated with a measure reference.

JuMP.set_name — Method

JuMP.set_name(mref::MeasureRef, name::String)

Extend JuMP.set_name to specify the name of a measure reference.

InfiniteOpt.expand — Function

expand(mref::MeasureRef)::JuMP.AbstractJuMPScalar

Return a JuMP scalar function containing the explicit expansion of the measure mref. This expansion is done according to the measure data. Note that variables are added to the model as necessary to accomodate the expansion (i.e., point variables and reduced infinite variables are made as needed). Errors if expansion is undefined for the measure data and/or the measure expression. If desired this can be used in combination with measure to expand measures on the fly.

This is useful for extensions that employ a custom optimizer model since it can be used evaluate measures before expressions are translated to the new model. This method can also be extended to handle custom measure data types by extending InfiniteOpt._expand_measure which should be of the form:

InfiniteOpt._expand_measure(expr::AbstractJuMPScalar, data::AbstractMeasureData,
                            model::InfiniteModel, point_mapper::Function)

See the source code in InfiniteOpt/src/measures.jl for examples of how to do this.

Example

julia> tdata = DiscreteMeasureData(t, [0.5, 0.5], [0, 1])

julia> expr = expand(measure(g + z + T - h - 2, tdata))
0.5 g(0) + 0.5 g(1) + z + 0.5 T(0, x) + 0.5 T(1, x) - h(x) - 2

InfiniteOpt.expand_all_measures! — Function

expand_all_measures!(model::InfiniteModel)

Expand all of the measures used in the objective and/or constraints of model. The objective and constraints are updated accordingly. Note that variables are added to the model as necessary to accomodate the expansion (i.e., point variables and reduced infinite variables are made as needed). Errors if expansion is undefined for the measure data and/or the measure expression. Also errors if the expanded objective function is not finite.

This is useful for extensions that employ a custom optimizer model since it can be used evaluate measures before model is translated into the new model. This method can also be extended to handle custom measure data types by extending InfiniteOpt._expand_measure which should be of the form:

InfiniteOpt._expand_measure(expr::AbstractJuMPScalar, data::AbstractMeasureData,
                            model::InfiniteModel, point_mapper::Function)

See the source code in InfiniteOpt/src/measures.jl for examples of how to do this.

Example

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 T(t, xi) >= 0.0
 z >= 0.0
 g(t) + z >= 42.0
 measure(T(t, xi)) >= 0.0, for all xi in [-1, 1]
 t in [0, 6]
 xi in Normal(μ=0.0, σ=1.0)

julia> expand_all_measures!(model)

julia> print(model)
Min 3 g(6) + z
Subject to
 T(t, xi) >= 0.0
 z >= 0.0
 g(t) + z >= 42.0
 0.5 T(0, xi) + 0.5 T(6, xi) >= 0.0, for all xi in [-1, 1]
 t in [0, 6]
 xi in Normal(μ=0.0, σ=1.0)

InfiniteOpt.infinite_variable_ref — Method

infinite_variable_ref(vref::ReducedInfiniteVariableRef)::InfiniteVariableRef

Return the InfiniteVariableRef associated with the reduced infinite variable vref.

Example

julia> infinite_variable_ref(vref)
g(t, x)

InfiniteOpt.eval_supports — Method

eval_supports(vref::ReducedInfiniteVariableRef)::Dict

Return the evaluation supports associated with the reduced infinite variable vref.

Example

julia> eval_supports(vref)
Dict{Int64,Float64} with 1 entry:
  1 => 0.5

InfiniteOpt.parameter_refs — Method

parameter_refs(vref::ReducedInfiniteVariableRef)::Tuple

Return the ParameterRef(s) associated with the reduced infinite variable vref. This is formatted as a Tuple of containing the parameter references as they were inputted to define the untracripted infinite variable except, the evaluated parameters are excluded.

Example

julia> parameter_refs(vref)
(t,   [2]  =  x[2]
  [1]  =  x[1])

JuMP.name — Method

JuMP.name(vref::ReducedInfiniteVariableRef)::String

Extend JuMP.name to return name of reduced infinite variable references. This is used when displaying measure expansions that contain such variables.

Exanple

julia> name(rvref)
g(1.25, x)

JuMP.has_lower_bound — Method

JuMP.has_lower_bound(vref::ReducedInfiniteVariableRef)::Bool

Extend JuMP.has_lower_bound to return a Bool whether the original infinite variable of vref has a lower bound.

Example

julia> has_lower_bound(vref)
true

JuMP.lower_bound — Method

JuMP.lower_bound(vref::ReducedInfiniteVariableRef)::Float64

Extend JuMP.lower_bound to return the lower bound of the original infinite variable of vref. Errors if vref doesn't have a lower bound.

Example

julia> lower_bound(vref)
0.0

JuMP.LowerBoundRef — Method

JuMP.LowerBoundRef(vref::ReducedInfiniteVariableRef)::GeneralConstraintRef

Extend JuMP.LowerBoundRef to extract a constraint reference for the lower bound of the original infinite variable of vref.

Example

julia> cref = LowerBoundRef(vref)
var >= 0.0

JuMP.has_upper_bound — Method

JuMP.has_upper_bound(vref::ReducedInfiniteVariableRef)::Bool

Extend JuMP.has_upper_bound to return a Bool whether the original infinite variable of vref has an upper bound.

Example

julia> has_upper_bound(vref)
true

JuMP.upper_bound — Method

JuMP.upper_bound(vref::ReducedInfiniteVariableRef)::Float64

Extend JuMP.upper_bound to return the upper bound of the original infinite variable of vref. Errors if vref doesn't have a upper bound.

Example

julia> upper_bound(vref)
0.0

JuMP.UpperBoundRef — Method

JuMP.UpperBoundRef(vref::ReducedInfiniteVariableRef)::GeneralConstraintRef

Extend JuMP.UpperBoundRef to extract a constraint reference for the upper bound of the original infinite variable of vref.

Example

julia> cref = UpperBoundRef(vref)
var <= 1.0

JuMP.is_fixed — Method

JuMP.is_fixed(vref::ReducedInfiniteVariableRef)::Bool

Extend JuMP.is_fixed to return Bool whether the original infinite variable of vref is fixed.

Example

julia> is_fixed(vref)
true

JuMP.fix_value — Method

JuMP.fix_value(vref::ReducedInfiniteVariableRef)::Float64

Extend JuMP.fix_value to return the fix value of the original infinite variable of vref. Errors if variable is not fixed.

Example

julia> fix_value(vref)
0.0

JuMP.FixRef — Method

JuMP.FixRef(vref::ReducedInfiniteVariableRef)::GeneralConstraintRef

Extend JuMP.FixRef to return the constraint reference of the fix constraint associated with the original infinite variable of vref. Errors vref is not fixed.

Examples

julia> cref = FixRef(vref)
var == 1.0

JuMP.start_value — Method

JuMP.start_value(vref::ReducedInfiniteVariableRef)::Union{Nothing, Float64}

Extend JuMP.start_value to return starting value of the original infinite variable of vref if it has one. Returns nothing otherwise.

Example

julia> start_value(vref)
0.0

JuMP.is_binary — Method

JuMP.is_binary(vref::ReducedInfiniteVariableRef)::Bool

Extend JuMP.is_binary to return Bool whether the original infinite variable of vref is binary.

Example

julia> is_binary(vref)
true

JuMP.BinaryRef — Method

JuMP.BinaryRef(vref::ReducedInfiniteVariableRef)::GeneralConstraintRef

Extend JuMP.BinaryRef to return a constraint reference to the constraint constrainting the original infinite variable of vref to be binary. Errors if one does not exist.

Example

julia> cref = BinaryRef(vref)
var binary

JuMP.is_integer — Method

JuMP.is_integer(vref::ReducedInfiniteVariableRef)::Bool

Extend JuMP.is_integer to return Bool whether the original infinite variable of vref is integer.

Example

julia> is_integer(vref)
true

JuMP.IntegerRef — Method

JuMP.IntegerRef(vref::ReducedInfiniteVariableRef)::GeneralConstraintRef

Extend JuMP.IntegerRef to return a constraint reference to the constraint constrainting the original infinite variable of vref to be integer. Errors if one does not exist.

Example

julia> cref = IntegerRef(vref)
var integer

InfiniteOpt.used_by_constraint — Method

used_by_constraint(vref::ReducedInfiniteVariableRef)::Bool

Return a Bool indicating if vref is used by a constraint.

Example

julia> used_by_constraint(vref)
false

InfiniteOpt.used_by_measure — Method

used_by_measure(vref::ReducedInfiniteVariableRef)::Bool

Return a Bool indicating if vref is used by a measure.

Example

julia> used_by_measure(vref)
true

JuMP.is_valid — Method

JuMP.is_valid(model::InfiniteModel, vref::ReducedInfiniteVariableRef)::Bool

Extend JuMP.is_valid to accomodate reduced infinite variables.

Example

julia> is_valid(model, vref)
true

JuMP.delete — Method

JuMP.delete(model::InfiniteModel, vref::ReducedInfiniteVariableRef)

Extend JuMP.delete to delete reduced infinite variables and its dependencies. Errors if vref is invalid, meaning it has already been deleted or it belongs to another model.

Example

julia> print(model)
Min measure(g(0, t)*t + g(1, t)*t) + z
Subject to
 z >= 0.0
 g(0, t) + g(1, t) == 0
 g(x, t) + z >= 42.0
 g(0.5, 0.5) == 0
 t in [0, 6]
 x in [0, 1]

julia> delete(model, rvref1)

julia> print(model)
Min measure(t + g(1, t)*t) + z
Subject to
 z >= 0.0
 g(1, t) == 0
 g(x, t) + z >= 42.0
 g(0.5, 0.5) == 0
 t in [0, 6]
 x in [0, 1]

MeasureEvalMethods Methods

InfiniteOpt.MeasureEvalMethods.gauss_hermite

InfiniteOpt.MeasureEvalMethods.gauss_laguerre

InfiniteOpt.MeasureEvalMethods.gauss_legendre

InfiniteOpt.MeasureEvalMethods.generate_measure_data

InfiniteOpt.MeasureEvalMethods.infinite_transform

InfiniteOpt.MeasureEvalMethods.mc_sampling

InfiniteOpt.MeasureEvalMethods.mc_sampling

InfiniteOpt.MeasureEvalMethods.generate_measure_data — Function

generate_measure_data(params::Union{InfiniteOpt.ParameterRef,
                      AbstractArray{<:InfiniteOpt.ParameterRef}},
                      num_supports::Int,
                      lb::Union{Number, JuMPC.SparseAxisArray, Nothing} = nothing,
                      ub::Union{Number, JuMPC.SparseAxisArray, Nothing} = nothing;
                      method::Function = mc_sampling, name::String = "",
                      weight_func::Function = InfiniteOpt._w, kwargs...
                      )::InfiniteOpt.AbstractMeasureData

Generate an AbstractMeasureData object that automatically generate supports based on a set of given information. The information is required to include parameters and number of supports. Other optional information to input includes lower and upper bounds, measure name, support generation method, and weight functions. The users could supply extra keyword arguments if necessary for their custom support generation methods.

Example

julia> @infinite_parameter(m, x in [0., 1.])
x

julia> measure_data = generate_measure_data(x, 3, 0.3, 0.7, method = gauss_legendre)
DiscreteMeasureData(x, [0.1111111111111111, 0.17777777777777776, 0.1111111111111111], [0.3450806661517033, 0.5, 0.6549193338482967], "", InfiniteOpt._w)

InfiniteOpt.MeasureEvalMethods.mc_sampling — Method

mc_sampling(lb::Union{JuMPC.SparseAxisArray, Number},
            ub::Union{JuMPC.SparseAxisArray, Number},
            num_supports::Int; kwargs...)::Tuple

Return a tuple that contains supports and coefficients generated by Monte Carlo sampling from a uniform distribution between the lower and upper bounds provided.

Example

julia> (supps, coeffs) = mc_sampling(0., 1., 5)
([0.8236475079774124, 0.9103565379264364, 0.16456579813368521, 0.17732884646626457, 0.278880109331201], [0.2, 0.2, 0.2, 0.2, 0.2])

julia> supps
5-element Array{Float64,1}:
 0.8236475079774124
 0.9103565379264364
 0.16456579813368521
 0.17732884646626457
 0.278880109331201

InfiniteOpt.MeasureEvalMethods.mc_sampling — Method

mc_sampling(dist::Distributions.NonMatrixDistribution,
            param::InfiniteOpt.ParameterRef,
            num_supports::Int; kwargs...)::Tuple

Return a tuple that contains supports and coefficients generated by Monte Carlo sampling from a given distribution.

Example

julia> dist = Normal(0., 1.)
Normal{Float64}(μ=0.0, σ=1.0)

julia> @infinite_parameter(m, x in dist)
x

julia> mc_sampling(dist, x, 10)
([0.6791074260357777, 0.8284134829000359, -0.3530074003005963, -0.13485387193052173, 0.5866170746331097, 0.29733585084941616, 0.06494754854834232, -0.10901738508171745, -0.514210390833322, 1.5743302021369892], [0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1])

InfiniteOpt.MeasureEvalMethods.gauss_legendre — Function

gauss_legendre(lb::Number, ub::Number, num_supports::Int; kwargs...)::Tuple

Return a tuple that contains supports and coefficients generated using Gauss-Legendre quadrature method. This is useful for univariate parameter in a finite interval.

Example

julia> (supps, coeffs) = gauss_legendre(0., 1., 5)
([0.04691007703066802, 0.23076534494715845, 0.5, 0.7692346550528415, 0.9530899229693319], [0.11846344252809454, 0.23931433524968324, 0.28444444444444444, 0.23931433524968324, 0.11846344252809454])

julia> supps
5-element Array{Float64,1}:
 0.04691007703066802
 0.23076534494715845
 0.5
 0.7692346550528415
 0.9530899229693319

InfiniteOpt.MeasureEvalMethods.gauss_hermite — Function

gauss_hermite(lb::Number, ub::Number, num_supports::Int; kwargs...)::Tuple

Return a tuple that contains supports and coefficients generated using Gauss-Hermite quadrature method. This is useful for univariate parameter in an infinite interval.

Example

julia> (supps, coeffs) = gauss_hermite(-Inf, Inf, 5)
([-2.0201828704560856, -0.9585724646138196, -8.881784197001252e-16, 0.9585724646138196, 2.0201828704560856], [1.1814886255359844, 0.986580996751429, 0.9453087204829428, 0.986580996751429, 1.1814886255359844])

julia> supps
5-element Array{Float64,1}:
 -2.0201828704560856
 -0.9585724646138196
 -8.881784197001252e-16
  0.9585724646138196
  2.0201828704560856

InfiniteOpt.MeasureEvalMethods.gauss_laguerre — Function

gauss_laguerre(lb::Number, ub::Number, num_supports::Int; kwargs...)::Tuple

Return a tuple that contains supports and coefficients generated using Gauss-Laguerre quadrature method. This is useful for univariate parameter in a semi-infinite interval.

Example

julia> (supps, coeffs) = gauss_laguerre(-Inf, 0., 5)
([-0.2635603197181408, -1.413403059106515, -3.596425771040715, -7.08581000585883, -12.640800844275773], [0.6790940422077494, 1.638487873602747, 2.7694432423708255, 4.3156569009208585, 7.219186354354335])

julia> supps
5-element Array{Float64,1}:
  -0.2635603197181408
  -1.413403059106515
  -3.596425771040715
  -7.08581000585883
 -12.640800844275773

InfiniteOpt.MeasureEvalMethods.infinite_transform — Function

infinite_transform(lb::Number, ub::Number, num_supports::Int;
                   [sub_method::Function = mc_sampling,
                   transform_x::Function = _default_x,
                   transform_dx::Function = _default_dx,
                   t_lb::Number = -convert(Number, lb == -Inf && ub == Inf),
                   t_ub::Number = 1., kwargs...])::Tuple

Returns a tuple that contains supports and coefficients generated for a parameter in an infinite or semi-infinite interval. It works by transforming the original unbounded interval to a finite interval, on which a support generation method for finite intervals is applied. Then, the generated supports are transformed back to the original interval. The user is allowed to specify the support generation method for finite intevals to use, as well as the transform function. The default transform function is $t \in [-\infty, \infty] \rightarrow x \in [-1, 1]: t(x) = \frac{t}{1-t^2}$ $t \in [a, \infty] \rightarrow x \in [0, 1]: t(x) = a + \frac{t}{1-t}$ $t \in [-\infty, a] \rightarrow x \in [0, 1]: t(x) = a - \frac{1-t}{t}$

Example

julia> (supps, coeffs) = infinite_transform(-Inf, Inf, 5, sub_method = gauss_legendre)
([-5.06704059565454, -0.7583532171678754, 0.0, 0.7583532171678754, 5.06704059565454], [13.490960583398396, 1.2245949721571516, 0.5688888888888889, 1.2245949721571516, 13.490960583398396])