Objectives

A guide and manual for specifying and modifying objective functions in InfiniteOpt. The Methods section at the end comprises the manual, and the above sections comprise the guide.

Overview

Naturally, objective functions serve as a key aspect of optimization problems in general and this is certainly the case with infinite dimensional ones. In InfiniteOpt objectives are defined in much the same way they are in JuMP. One key idea to keep in mind is that the objective must evaluate to a finite expression. Note this means that objectives can only explicitly contain hold variables and point variables. Infinite expressions must be evaluated in a measure to be included (e.g., evaluate the expectation of a random variable).

Basic Usage

Principally, the objective function is specified via @objective as is done in JuMP. For example, let's define the stochastic objective to minimize $0.5 x_1 + 0.5 x_2 + \mathbb{E}_\xi [y^2 - y]$:

julia> @infinite_parameter(model, ξ in Normal())
ξ

julia> @infinite_variable(model, y(ξ))
y(ξ)

julia> @hold_variable(model, x[1:2])
2-element Array{HoldVariableRef,1}:
 x[1]
 x[2]

julia> @objective(model, Min, 0.5x[1] + 0.5x[2] + expect(y^2 - y, ξ))
0.5 x[1] + 0.5 x[2] + expect(y(ξ)² - y(ξ))

Thus, we have defined an objective using InfiniteOpt's straightforward syntax. Note that the second argument indicates the objective sense which can be expressed Min for minimization problems and Max for maximization problems. The objective function (expression) must be finite containing only hold variables, point variables, and/or measures. Also, any included measures must fully integrate over all the infinite parameters contained in its input function. For example, if we define had an infinite variable z(ξ, t) then the measure expect(z, ξ) could not be included since the resulting expression would still be infinite with respect to t. However, adding a measure for t would result in a valid object to add to an objective: integral(expect(z, ξ), t).

Now we can add objectives to our infinite models. For more detailed information, please review the information below.

Queries

This section will highlight the available methods for extracting objective information. These are all derived from extensions to JuMP functions and thus follow syntax.

Principally, these methods correspond to objective_sense, objective_funcion, and objective_function_type which return the objective sense (a subtype of MOI.OptimizationSense), the objective function (expression), and the objective function type, respectively. These methods are demonstrated in accordance with the example presented above in the Basic Usage section:

julia> objective_sense(model)
MIN_SENSE::OptimizationSense = 0

julia> objective_function(model)
0.5 x[1] + 0.5 x[2] + expect(y(ξ)² - y(ξ))

julia> objective_function_type(model)
GenericAffExpr{Float64,MeasureFiniteVariableRef}

The objective sense can be one of three possibilities: MIN_SENSE, MAX_SENSE, or FEASIBILITY_SENSE. The later sense applies to models that contain no objective function.

Modification

This section will review the methods that can be used to modify the objective. First, we'll consider the useful set_objective_coefficient method and then we'll explain the methods that enable @objective.

The coefficient of a particular variable in an objective can be readily updated via set_objective_coefficient. This is useful repeatedly optimizing an infinite model with varied objective coefficients (e.g., varied tradeoff parameters). For example, let's consider updating the coefficient of x[1] in the previous example from 0.5 to 0.25:

julia> set_objective_coefficient(model, x[1], 0.25)

julia> objective_function(model)
0.25 x[1] + 0.5 x[2] + expect(y(ξ)² - y(ξ))

Now let's consider the modification methods that enable the @objective macro. The objective function is specified via set_objective_function which simply updates the expression stored in the objective. For example, let's update out objective to simply be $0.5x_1 + 0.5x_2$:

julia> set_objective_function(model, 0.5x[1] + 0.5x[2])

julia> objective_function(model)
0.5 x[1] + 0.5 x[2]

The objective sense is updated via set_objective_sense which can specify the sense as one of the MOI.OptimizationSense subtypes. For example, let's change the current objective to be maximization problem: ``jldoctest obj julia> using MathOptInterface

julia> const MOI = MathOptInterface;

julia> setobjectivesense(model, MOI.MAX_SENSE)

julia> objectivesense(model) MAXSENSE::OptimizationSense = 1

The above 2 methods are both called via [`set_objective`](@ref JuMP.set_objective(::InfiniteModel, ::MOI.OptimizationSense, ::Union{JuMP.AbstractJuMPScalar, Real})). This is the
function that enables [`@objective`](@ref) behind the scenes. Thus, the previous
2 examples could have been implemented equivalently in the following ways:
``jldoctest obj
julia> set_objective(model, MOI.MAX_SENSE, 0.5x[1] + 0.5x[2])

julia> @objective(model, Max, 0.5x[1] + 0.5x[2])
0.5 x[1] + 0.5 x[2]

Notice that @objective offers a more intuitive syntax and is also more efficient at parsing expressions.

Note

When possible, the @objective since it is more stable and efficient than the set_objective_[aspect] methods due to its enhanced methodology for parsing expressions.

Methods

JuMP.objective_function

JuMP.objective_function_type

JuMP.objective_sense

JuMP.set_objective

JuMP.set_objective_coefficient

JuMP.set_objective_function

JuMP.set_objective_function

JuMP.set_objective_sense

JuMP.objective_sense — Method

JuMP.objective_sense(model::InfiniteModel)::MOI.OptimizationSense

Extend JuMP.objective_sense to return the objective sense of the infinite model model.

Example

julia> objective_sense(model)
MIN_SENSE::OptimizationSense = 0

JuMP.objective_function_type — Method

JuMP.objective_function_type(model::InfiniteModel)::Type{<:JuMP.AbstractJuMPScalar}

Extend JuMP.objective_function_type to return the objective function type of infinite model model.

Example

julia> objective_function_type(model)
GenericAffExpr{Float64,HoldVariableRef}

JuMP.objective_function — Method

JuMP.objective_function(model::InfiniteModel)::JuMP.AbstractJuMPScalar

Extend JuMP.objective_function to return the objective of infinite model model.

Example

julia> objective_function(model)
1

JuMP.set_objective_function — Method

JuMP.set_objective_function(model::InfiniteModel,
                            func::JuMP.AbstractJuMPScalar)

Extend JuMP.set_objective_function to set the objective expression of infinite model model. Errors if func contains infinite variables and/or parameters. Also errors if func contains invalid variables.

Example

julia> set_objective_function(model, 2x + 1)

julia> objective_function(model)
2 x + 1

JuMP.set_objective_function — Method

JuMP.set_objective_function(model::InfiniteModel, func::Real)

Extend JuMP.set_objective_function to set the objective expression of model with a number.

Example

julia> set_objective_function(model, 3)

julia> objective_function(model)
3

JuMP.set_objective_sense — Method

JuMP.set_objective_sense(model::InfiniteModel, sense::MOI.OptimizationSense)

Extend JuMP.set_objective_sense to set the objective sense of infinite model model.

Example

julia> set_objective_sense(model, MOI.MIN_SENSE)

julia> objective_sense(model)
MIN_SENSE::OptimizationSense = 0

JuMP.set_objective — Method

JuMP.set_objective(model::InfiniteModel, sense::MOI.OptimizationSense,
                   func::Union{JuMP.AbstractJuMPScalar, Real})

Extend JuMP.set_objective to set the objective of infinite model model. Errors if func contains infinite variables and/or parameters, or if it does not belong to the model.

Example

julia> set_objective(model, MOI.MIN_SENSE, 2x + 1)

julia> objective_function(model)
2 x + 1

JuMP.set_objective_coefficient — Method

JuMP.set_objective_coefficient(model::InfiniteModel,
                               variable::GeneralVariableRef,
                               coefficient::Real)

Extend JuMP.set_objective_coefficient Set the linear objective coefficient associated with variable to coefficient. Errors if the function type is unsupported.

Example

julia> @hold_variable(model, x)
x

julia> @hold_variable(model, y)
y

julia> @objective(model, x + y)
x + y

julia> set_objective_coefficient(model, y, 2)

julia> objective_function(model)
x + 2 y