Optimization

A guide and manual for optimizing (solving) InfiniteOpt models. The Methods section at the end comprise the manual, and the above sections comprise the guide.

Overview

Fundamentally, we seek to optimize a given infinite optimization model that we have defined and this is the very reason why InfiniteOpt was created. Thus, InfiniteOpt offers a general and intuitive platform to do just this. This is made up of transforming the InfiniteModel into a standard optimization problem stored as a JuMP.Model (referred to as the optimizer_model) that is then optimized via a compatible optimizer. By default this is done via a TranscriptionModel as described on the previous page. However, user-defined reformulation strategies can readily be implemented as described in the Optimizer Models section on the extensions page.

Basic Usage

For most users, optimize! is the only method required to optimize an InfiniteModel. This is exactly analogous to that of any JuMP.Model and is designed to provide a similar user experience. Let's first define an InfiniteModel with an appropriate optimizer:

julia> using InfiniteOpt, JuMP, Ipopt;

julia> model = InfiniteModel(Ipopt.Optimizer);

julia> set_optimizer_attribute(model, "print_level", 0);

julia> @infinite_parameter(model, t in [0, 10], num_supports = 10);

julia> @infinite_variable(model, x(t) >= 0);

julia> @hold_variable(model, z >= 0);

julia> @objective(model, Min, 2z);

julia> @constraint(model, c1, z >= x);

julia> @BDconstraint(model, c2(t == 0), x == 42);

julia> print(model)
Min 2 z
Subject to
 x(t) ≥ 0.0, ∀ t ∈ [0, 10]
 z ≥ 0.0
 c1 : z - x(t) ≥ 0.0, ∀ t ∈ [0, 10]
 c2 : x(t) = 42.0, ∀ t = 0

Now we optimize the model using optimize!:

julia> optimize!(model);

julia> termination_status(model)
LOCALLY_SOLVED::TerminationStatusCode = 4

Now our model has been solved and we can query the solution. How to query the solution is explained on the Results page.

If no optimizer has been specified for the InfiniteModel, one can be provided via set_optimizer:

julia> set_optimizer(model, Ipopt.Optimizer)

A number of methods also exist to adjust the optimizer settings such as suppressing output. This is explained below in the Optimizer Settings section.

Optimizer Models

As discussed previously, InfiniteModels contain an optimizer_model field which stores a transformed finite version of the model in a JuMP.Model that contains a data struct (that stores a mapping between the transformed model and the infinite model) in the Model.ext dictionary with an associated key. By default a JuMP.Model using TranscriptionData stored under the key :TransData is used and is referred to as a TranscriptionModel. The optimizer model is then what is used to optimize the infinite model and it provides the information exacted by solution queries mapped back to the infinite model using the mapping data structure.

The process for optimizing an InfiniteModel is summarized in the following steps:

fully define the InfiniteModel

build the optimizer model via build_optimizer_model!

optimize the optimizer_model via optimize!.

Here build_optimizer_model! creates a reformulated finite version of the InfiniteModel, stores it in InfiniteModel.optimizer_model via set_optimizer_model, and indicates that the optimizer model is ready via set_optimizer_model_ready. These steps are all automated when optimize! is invoked on the InfiniteModel.

The optimizer_model can be queried/extracted at any time from an InfiniteModel via optimizer_model. For example, let's extract the optimizer model from the example above in the basic usage section:

julia> trans_model = optimizer_model(model)
A JuMP Model
Minimization problem with:
Variables: 11
Objective function type: GenericAffExpr{Float64,VariableRef}
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.GreaterThan{Float64}`: 10 constraints
`VariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 11 constraints
Model mode: AUTOMATIC
CachingOptimizer state: ATTACHED_OPTIMIZER
Solver name: Ipopt

The JuMP variable(s) stored in the optimizer model that correspond to a particular InfiniteOpt variable can be queried via optimizer_model_variable. Using a TranscriptionModel this equivalent to calling transcription_variable. Thus, using the going example we get:

julia> optimizer_model_variable(x) # infinite variable
10-element Array{VariableRef,1}:
 x(support: 1)
 x(support: 2)
 x(support: 3)
 x(support: 4)
 x(support: 5)
 x(support: 6)
 x(support: 7)
 x(support: 8)
 x(support: 9)
 x(support: 10)

julia> optimizer_model_variable(z) # hold variable
z

In like manner, we get the JuMP constraints corresponding to a particular InfiniteOpt constraint via optimizer_model_constraint. Using a TranscriptionModel this equivalent to calling transcription_constraint. Thus, using going example we get:

julia> optimizer_model_constraint(c1) # infinite constraint
10-element Array{ConstraintRef,1}:
 c1(support: 1) : z - x(support: 1) ≥ 0.0
 c1(support: 2) : z - x(support: 2) ≥ 0.0
 c1(support: 3) : z - x(support: 3) ≥ 0.0
 c1(support: 4) : z - x(support: 4) ≥ 0.0
 c1(support: 5) : z - x(support: 5) ≥ 0.0
 c1(support: 6) : z - x(support: 6) ≥ 0.0
 c1(support: 7) : z - x(support: 7) ≥ 0.0
 c1(support: 8) : z - x(support: 8) ≥ 0.0
 c1(support: 9) : z - x(support: 9) ≥ 0.0
 c1(support: 10) : z - x(support: 10) ≥ 0.0

We can also query the expressions via optimizer_model_expression:

julia> optimizer_model_expression(z - x^2 + 3) # infinite expression
10-element Array{AbstractJuMPScalar,1}:
 -x(support: 1)² + z + 3
 -x(support: 2)² + z + 3
 -x(support: 3)² + z + 3
 -x(support: 4)² + z + 3
 -x(support: 5)² + z + 3
 -x(support: 6)² + z + 3
 -x(support: 7)² + z + 3
 -x(support: 8)² + z + 3
 -x(support: 9)² + z + 3
 -x(support: 10)² + z + 3

Note

Like supports the optimizer_model_[obj] methods also employ the label::Type{AbstractSupportLabel} = PublicLabel keyword argument that by default will return variables/expressions/constraints associated with public supports. The full set (e.g., ones corresponding to internal collocation nodes) is obtained via label = All.

These methods also employ the ndarray::Bool keyword argument that will cause the output to be formatted as a n-dimensional array where the dimensions correspond to the infinite parameter dependencies. For example, if we have an infinite variable y(t, ξ) and we invoke a query method with ndarray = true then we'll get a matrix whose dimensions correspond to the supports of t and ξ, respectively. Also, if ndarray = true then label correspond to the intersection of supports labels in contrast to its default of invoking the union of the labels.

The purpose of this optimizer_model abstraction is to readily enable user-defined reformulation extensions (e.g., using polynomial chaos expansion theory). However, this is all handled behind the scenes such that most users can interact with InfiniteModels like any JuMP.Model.

Optimizer Settings

A few optimizer settings can be set in a consistent way agnostic of particular solver keywords. One such setting is that of suppressing and unsuppressing optimizer verbose output. This is accomplished via set_silent and unset_silent. The syntax is exemplified below:

julia> set_silent(model)
true

julia> unset_silent(model)
false

We can also adjust the time limit in a solver independent fashion via set_time_limit_sec, unset_time_limit_sec, and time_limit_sec. These methods are illustrated below:

julia> set_time_limit_sec(model, 100)
100

julia> time_limit_sec(model)
100.0

julia> unset_time_limit_sec(model)

Other optimizer specific settings can be set via set_optimizer_attribute. For example, let's set the maximum CPU time for Ipopt:

julia> set_optimizer_attribute(model, "max_cpu_time", 60.)
60.0

Multiple settings can be specified via set_optimizer_attributes. For example, let's specify the tolerance and the maximum number of iterations:

julia> set_optimizer_attributes(model, "tol" => 1e-4, "max_iter" => 100)

Finally, we can query optimizer settings via get_optimizer_attribute. For example, let's query the maximum number of iterations:

julia> get_optimizer_attribute(model, "max_iter")
100

Note this only works if the attribute has been previously specified.

Methods

InfiniteOpt.add_infinite_model_optimizer

InfiniteOpt.build_optimizer_model!

InfiniteOpt.build_optimizer_model!

InfiniteOpt.clear_optimizer_model_build!

InfiniteOpt.clear_optimizer_model_build!

InfiniteOpt.constraint_supports

InfiniteOpt.expression_supports

InfiniteOpt.optimizer_model

InfiniteOpt.optimizer_model_constraint

InfiniteOpt.optimizer_model_constraint

InfiniteOpt.optimizer_model_expression

InfiniteOpt.optimizer_model_expression

InfiniteOpt.optimizer_model_key

InfiniteOpt.optimizer_model_key

InfiniteOpt.optimizer_model_ready

InfiniteOpt.optimizer_model_variable

InfiniteOpt.optimizer_model_variable

InfiniteOpt.set_optimizer_model

InfiniteOpt.set_optimizer_model_ready

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.variable_supports

JuMP.add_bridge

JuMP.backend

JuMP.bridge_constraints

JuMP.get_optimizer_attribute

JuMP.get_optimizer_attribute

JuMP.mode

JuMP.optimize!

JuMP.set_optimizer

JuMP.set_optimizer_attribute

JuMP.set_optimizer_attribute

JuMP.set_optimizer_attributes

JuMP.set_silent

JuMP.set_time_limit_sec

JuMP.solver_name

JuMP.time_limit_sec

JuMP.unset_silent

JuMP.unset_time_limit_sec

JuMP.optimize! — Method

JuMP.optimize!(model::InfiniteModel;
               bridge_constraints::Bool=true, kwargs...])

Extend JuMP.optimize! to optimize infinite models using the internal optimizer model. Will call build_optimizer_model! if the optimizer model isn't up to date. The kwargs correspond to keyword arguments passed to build_optimizer_model! if any are defined.

Example

julia> optimize!(model)

julia> has_values(model)
true

JuMP.set_optimizer — Method

JuMP.set_optimizer(model::InfiniteModel,
                   [optimizer_constructor;
                   bridge_constraints::Bool = true])

Extend JuMP.set_optimizer to set optimizer of infinite models. Specifically, the optimizer of the optimizer model is modified.

Example

julia> set_optimizer(model, Clp.Optimizer)

julia> optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: SolverName() attribute not implemented by the optimizer.

JuMP.set_silent — Method

JuMP.set_silent(model::InfiniteModel)

Extend JuMP.set_silent for infinite models to take precedence over any other attribute controlling verbosity and requires the solver to produce no output.

Example

julia> set_silent(model)
true

JuMP.unset_silent — Method

JuMP.unset_silent(model::InfiniteModel)

Extend JuMP.unset_silent for infinite models to neutralize the effect of the set_silent function and let the solver attributes control the verbosity.

Example

julia> unset_silent(model)
false

JuMP.set_time_limit_sec — Method

JuMP.set_time_limit_sec(model::InfiniteModel, limit)

Extend set_time_limit_sec to set the time limit (in seconds) of the solver. Can be unset using unset_time_limit_sec or with limit set to nothing.

Example

julia> set_time_limit_sec(model, 100)
100

JuMP.unset_time_limit_sec — Method

JuMP.unset_time_limit_sec(model::InfiniteModel)

Extend unset_time_limit_sec to unset the time limit of the solver. Can be set using set_time_limit_sec.

Example

julia> unset_time_limit_sec(model)

JuMP.time_limit_sec — Method

JuMP.time_limit_sec(model::InfiniteModel)

Extend [time_limit_sec](@ref JuMP.timelimitsec(::JuMP.Model) to get the time limit (in seconds) of the solve used by the optimizer model (nothing if unset). Can be set using set_time_limit_sec.

Example

julia> time_limit_sec(model)
100

JuMP.set_optimizer_attribute — Method

JuMP.set_optimizer_attribute(model::InfiniteModel, name::String, value)

Extend set_optimizer_attribute to specify a solver-specific attribute identified by name to value.

Example

julia> set_optimizer_attribute(model, "SolverSpecificAttributeName", true)
true

JuMP.set_optimizer_attribute — Method

JuMP.set_optimizer_attribute(model::InfiniteModel,
                             attr::MOI.AbstractOptimizerAttribute,
                             value)

Extend set_optimizer_attribute to set the solver-specific attribute attr in model to value.

Example

julia> set_optimizer_attribute(model, MOI.Silent(), true)
true

JuMP.set_optimizer_attributes — Method

JuMP.set_optimizer_attributes(model::InfiniteModel, pairs::Pair...)

Extend set_optimizer_attributes to set multiple solver attributes given a list of attribute => value pairs. Calls set_optimizer_attribute(model, attribute, value) for each pair.

Example

julia> model = Model(Ipopt.Optimizer);

julia> set_optimizer_attributes(model, "tol" => 1e-4, "max_iter" => 100)

is equivalent to:

julia> set_optimizer_attribute(model, "tol", 1e-4);

julia> set_optimizer_attribute(model, "max_iter", 100);

JuMP.get_optimizer_attribute — Method

JuMP.get_optimizer_attribute(model::InfiniteModel, name::String)

Extend get_optimizer_attribute to return the value associated with the solver-specific attribute named name.

Example julia-repl julia> get_optimizer_attribute(model, "tol") 0.0001`

JuMP.get_optimizer_attribute — Method

JuMP.get_optimizer_attribute(model::InfiniteModel,
                             attr::MOI.AbstractOptimizerAttribute)

Extend get_optimizer_attribute to return the value of the solver-specific attribute attr in model.

Example julia-repl julia> get_optimizer_attribute(model, MOI.Silent()) true`

JuMP.solver_name — Method

JuMP.solver_name(model::InfiniteModel)

Extend solver_name to return the name of the solver being used if there is an optimizer selected and it has a name attribute. Otherwise, an error is thrown.

Example

julia> solver_name(model)
"Gurobi"

JuMP.backend — Method

JuMP.backend(model::InfiniteModel)

Extend backend to return the MathOptInterface backend associated with the optimizer model. Note this will be empty if the optimizer model has not been build yet.

Example

julia> moi_model = backend(model);

JuMP.mode — Method

JuMP.mode(model::InfiniteModel)

Extend mode to return the MathOptInterface mode the optimizer model is in.

Example

julia> mode(model)
AUTOMATIC::ModelMode = 0

InfiniteOpt.optimizer_model — Function

optimizer_model(model::InfiniteModel)::JuMP.Model

Return the JuMP model stored in model that is used to solve it.

Example

julia> opt_model = optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

InfiniteOpt.set_optimizer_model — Function

set_optimizer_model(inf_model::InfiniteModel, opt_model::JuMP.Model;
                    inherit_optimizer::Bool = true)

Specify the JuMP model that is used to solve inf_model. This is intended for internal use and extensions. Note that opt_model should contain extension data to allow it to map to inf_model in a manner similar to TranscriptionModel. inherit_optimizer indicates whether add_infinite_model_optimizer should be invoked on the new optimizer mode to inherit the optimizer constuctor and attributes currently stored in inf_model.

Example

julia> set_optimizer_model(model, TranscriptionModel())

julia> optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

InfiniteOpt.optimizer_model_key — Method

optimizer_model_key(model::InfiniteModel)::Any

Return the extension key used in the optimizer model of model. Errors if optimizer_model.ext contains more than one key. This is intended for internal use and extensions. For extensions this is used to dispatch to the appropriate optmizer model functions such as extensions to build_optimizer_model!.

Example

julia> optimizer_model_key(model)
:TransData

InfiniteOpt.optimizer_model_key — Method

optimizer_model_key(model::JuMP.Model)::Any

Return the extension key used in the optimizer model model. Errors if model.ext contains more than one key. This is intended for internal use and extensions. For extensions this is used to dispatch to the appropriate optmizer model functions such as extensions to build_optimizer_model!. This is intended as an internal method. See optimizer_model_key for the public method

InfiniteOpt.build_optimizer_model! — Method

build_optimizer_model!(model::InfiniteModel; [kwargs...])

Build the optimizer model stored in model such that it can be treated as a normal JuMP model. Specifically, translate the variables and constraints stored in model into ones that are stored in the optimizer model and can be solved. This is provided generally to accomodate extensions that use custom optimizer model types in accordance with optimizer_model_key. However, it may be useful in certain applications when the user desires to force a build without calling optimize!. Extensions will need to implement their own version of the function build_optimizer_model!(model::InfiniteModel, key::Val{ext_key_name}; kwargs...).

Example

julia> build_optimizer_model!(model)

julia> optimizer_model_ready(model)
true

InfiniteOpt.build_optimizer_model! — Function

build_optimizer_model!(model::InfiniteModel, key::Val{ext_key_name};
                       [kwargs...])

Build the optimizer model stored in model such that it can be treated as a normal JuMP model, where the Model.ext field contains a key that points to a datastructure that appropriately maps the data between the two models. The key argument should be be typed to Val{ext_key_name}. This should also use clear_optimizer_model_build! to empty the out the current optimizer model. Ultimately, set_optimizer_model should be called to insert the build optimizer model into model and set_optimizer_model_ready should be used to update the optimizer model's status.

InfiniteOpt.clear_optimizer_model_build! — Method

clear_optimizer_model_build!(model::InfiniteModel)::JuMP.Model

Empty the optimizer model using appropriate calls of Base.empty!. This effectively resets model.optimizer_model except the optimizer, its attributes, and an an emptied optimizer model data struct are maintained. This is intended as an internal method for use by build_optimizer_model!.

InfiniteOpt.clear_optimizer_model_build! — Method

clear_optimizer_model_build!(model::JuMP.Model)::JuMP.Model

Empty the optimizer model using appropriate calls of Base.empty!. This effectively resets model except the optimizer, its attributes, and an an emptied optimizer model data struct are maintained. This is intended as an internal method for use by build_optimizer_model!.

InfiniteOpt.add_infinite_model_optimizer — Function

add_infinite_model_optimizer(opt_model::JuMP.Model, inf_model::InfiniteModel)

Parse the current optimizer and its attributes associated with inf_model and load them into opt_model. This is intended to be used as an internal method for set_optimizer_model.

InfiniteOpt.optimizer_model_variable — Method

optimizer_model_variable(vref::GeneralVariableRef; 
                         [label::Type{<:AbstractSupportLabel} = PublicLabel, 
                         ndarray::Bool = false,
                         kwargs...])

Return the reformulation variable(s) stored in the optimizer model that correspond to vref. Also errors if no such variable can be found in the optimizer model.

The keyword arugments label and ndarray are what TranscriptionOpt employ and kwargs denote extra ones that user extensions may employ in accordance with their implementation of optimizer_model_variable. Errors if such an extension has not been written.

By default only the variables associated with public supports are returned, the full set can be accessed via label = All. Moreover, infinite variables are returned as a list corresponding to their supports. However, a n-dimensional array can be obtained via ndarray = true which is handy when the variable has multiple infinite parameter dependencies. The corresponding supports are obtained via supports using the same keyword arguments.

Example

julia> optimizer_model_variable(x) # infinite variable
2-element Array{VariableRef,1}:
 x(support: 1)
 x(support: 2)

julia> optimizer_model_variable(z) # hold variable
z

InfiniteOpt.optimizer_model_variable — Function

optimizer_model_variable(vref::GeneralVariableRef, key::Val{ext_key_name};
                         [kwargs...])

Return the reformulation variable(s) stored in the optimizer model that correspond to vref. This needs to be defined for extensions that implement a custom optimizer model type. Principally, this is accomplished by typed the key argument to Val{ext_key_name}. Keyword arguments can be added as needed.

InfiniteOpt.supports — Method

supports(vref::DecisionVariableRef; 
         [label::Type{<:AbstractSupportLabel} = PublicLabel, 
         ndarray::Bool = false,
         kwargs...])

Return the supports associated with vref in the optimizer model. Errors if InfiniteOpt.variable_supports has not been extended for the optimizer model type or if vref is not be reformulated in the optimizer model.

The keyword arugments label and ndarray are what TranscriptionOpt employ and kwargs denote extra ones that user extensions may employ in accordance with their implementation of variable_supports. Errors if such an extension has not been written.

By default only the public supports are returned, the full set can be accessed via label = All. Moreover, the supports of infinite variables are returned as a list. However, a n-dimensional array can be obtained via ndarray = true which is handy when the variable has multiple infinite parameter dependencies.

Example

julia> supports(vref)
2-element Array{Tuple{Float64},1}:
 (0.0,)
 (1.0,)

InfiniteOpt.variable_supports — Function

variable_supports(optimizer_model::JuMP.Model, vref,
                  key::Val{ext_key_name}; 
                  [kwargs...])::Vector

Return the supports associated with the mappings of vref in optimizer_model. This dispatches off of key which permits optimizer model extensions. This should throw an error if vref is not associated with the variable mappings stored in optimizer_model. Keyword arguments can be added as needed. Note that no extension is necessary for point or hold variables.

InfiniteOpt.optimizer_model_expression — Method

optimizer_model_expression(expr::JuMP.AbstractJuMPScalar; 
                           [label::Type{<:AbstractSupportLabel} = PublicLabel,
                           ndarray::Bool = false, 
                           kwargs...])

Return the reformulation expression(s) stored in the optimizer model that correspond to expr. Also errors if no such expression can be found in the optimizer model (meaning one or more of the underlying variables have not been transcribed).

The keyword arugments label and ndarray are what TranscriptionOpt employ and kwargs denote extra ones that user extensions may employ in accordance with their implementation of optimizer_model_expression. Errors if such an extension has not been written.

By default only the expressions associated with public supports are returned, the full set can be accessed via label = All. Moreover, infinite expressions are returned as a list corresponding to their supports. However, a n-dimensional array can be obtained via ndarray = true which is handy when the expression has multiple infinite parameter dependencies. The corresponding supports are obtained via supports using the same keyword arguments.

Example

julia> optimizer_model_expression(my_expr) # finite expression
x(support: 1) - y

InfiniteOpt.optimizer_model_expression — Function

optimizer_model_expression(expr, key::Val{ext_key_name}; [kwargs...])

Return the reformulation expression(s) stored in the optimizer model that correspond to expr. This needs to be defined for extensions that implement a custom optimizer model type. Principally, this is accomplished by typed the key argument to Val{ext_key_name}. Keyword arguments can be added as needed. Note that if expr is a GeneralVariableRef this just dispatches to optimizer_model_variable.

InfiniteOpt.supports — Method

supports(expr::JuMP.AbstractJuMPScalar; 
         [label::Type{<:AbstractSupportLabel} = PublicLabel,
         ndarray::Bool = false,
         kwargs...])

Return the support associated with expr. Errors if expr is not associated with the constraint mappings stored in optimizer_model.

The keyword arugments label and ndarray are what TranscriptionOpt employ and kwargs denote extra ones that user extensions may employ in accordance with their implementation of expression_supports. Errors if such an extension has not been written.

By default only the public supports are returned, the full set can be accessed via label = All. Moreover, the supports of infinite expressions are returned as a list. However, a n-dimensional array can be obtained via ndarray = true which is handy when the expression has multiple infinite parameter dependencies.

Example

julia> supports(cref)
2-element Array{Tuple{Float64},1}:
 (0.0,)
 (1.0,)

InfiniteOpt.expression_supports — Function

expression_supports(optimizer_model::JuMP.Model, expr,
                    key::Val{ext_key_name}; [kwargs...])

Return the supports associated with the mappings of expr in optimizer_model. This dispatches off of key which permits optimizer model extensions. This should throw an error if expr is not associated with the variable mappings stored in optimizer_model. Keyword arguments can be added as needed. Note that if expr is a GeneralVariableRef this just dispatches to variable_supports.

InfiniteOpt.optimizer_model_constraint — Method

optimizer_model_constraint(cref::InfOptConstraintRef; 
                           [label::Type{<:AbstractSupportLabel} = PublicLabel, 
                           ndarray::Bool = false,
                           kwargs...])

Return the reformulation constraint(s) stored in the optimizer model that correspond to cref. Errors if no such constraint can be found in the optimizer model.

The keyword arugments label and ndarray are what TranscriptionOpt employ and kwargs denote extra ones that user extensions may employ in accordance with their implementation of optimizer_model_constraint. Errors if such an extension has not been written.

By default only the constraints associated with public supports are returned, the full set can be accessed via label = All. Moreover, infinite constraints are returned as a list corresponding to their supports. However, a n-dimensional array can be obtained via ndarray = true which is handy when the constraint has multiple infinite parameter dependencies. The corresponding supports are obtained via supports using the same keyword arguments.

Example

julia> optimizer_model_constraint(c1) # finite constraint
c1 : x(support: 1) - y <= 3.0

InfiniteOpt.optimizer_model_constraint — Function

optimizer_model_constraint(cref::InfOptConstraintRef,
                           key::Val{ext_key_name}; [kwargs...])

Return the reformulation constraint(s) stored in the optimizer model that correspond to cref. This needs to be defined for extensions that implement a custom optimizer model type. Principally, this is accomplished by typed the key argument to Val{ext_key_name}. Keyword arguments can be added as needed.

InfiniteOpt.supports — Method

supports(cref::InfOptConstraintRef; 
         [label::Type{<:AbstractSupportLabel} = PublicLabel,
         ndarray::Bool = false,
         kwargs...])

Return the support associated with cref. Errors if cref is not associated with the constraint mappings stored in optimizer_model.

The keyword arugments label and ndarray are what TranscriptionOpt employ and kwargs denote extra ones that user extensions may employ in accordance with their implementation of constraint_supports. Errors if such an extension has not been written.

By default only the public supports are returned, the full set can be accessed via label = All. Moreover, the supports of infinite constraints are returned as a list. However, a n-dimensional array can be obtained via ndarray = true which is handy when the constraint has multiple infinite parameter dependencies.

Example

julia> supports(cref)
2-element Array{Tuple{Float64},1}:
 (0.0,)
 (1.0,)

InfiniteOpt.constraint_supports — Function

constraint_supports(optimizer_model::JuMP.Model, cref::InfOptConstraintRef,
                    key::Val{ext_key_name}; [kwargs...])

Return the supports associated with the mappings of cref in optimizer_model. This dispatches off of key which permits optimizer model extensions. This should throw an error if cref is not associated with the variable mappings stored in optimizer_model. Keyword arguments can be added as needed.

InfiniteOpt.optimizer_model_ready — Function

optimizer_model_ready(model::InfiniteModel)::Bool

Return Bool if the optimizer model is up to date with model.

Example

julia> optimizer_model_ready(model)
false

InfiniteOpt.set_optimizer_model_ready — Function

set_optimizer_model_ready(model::InfiniteModel, status::Bool)

Set the status of the optimizer model to whether it is up to date or not. Note is more intended as an internal function, but is useful for extensions.

Example

julia> set_optimizer_model_ready(model, true)

julia> optimizer_model_ready(model)
true

JuMP.bridge_constraints — Method

JuMP.bridge_constraints(model::InfiniteModel)::Bool

Extend JuMP.bridge_constraints to return if an infinite model model has an optimizer model where the optimizer is set and unsupported constraints are automatically bridged to equivalent supported constraints when an appropriate transformation is available.

Example

julia> bridge_constraints(model)
false

JuMP.add_bridge — Method

JuMP.add_bridge(model::InfiniteModel,
                BridgeType::Type{<:MOI.Bridges.AbstractBridge})

Extend JuMP.add_bridge to add BridgeType to the list of bridges that can be used by the optimizer model to transform unsupported constraints into an equivalent formulation using only constraints supported by the optimizer.