Infinite Parameters

A guide and manual to the definition and use of infinite parameters in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

Infinite parameters are what parameterize the infinite decision spaces for infinite dimensional mathematical programs. In dynamic optimization this corresponds to time and in stochastic optimization this to uncertain parameters that follow a certain underlying statistical distribution. InfiniteOpt considers principally two kinds of infinite parameters, ones defined over continuous intervals and ones characterized by a distribution (others can be added by defining a user-defined type). These can be used to parameterize infinite variables, point variables, measures, and can be used directly inside constraints.

Basic Usage

First, we need to initialize and add infinite parameters to our InfiniteModel. This can be accomplished using @infinite_parameter. For example, let's define a parameter for time in a time interval from 0 to 10:

julia> using InfiniteOpt, JuMP

julia> model = InfiniteModel();

julia> @infinite_parameter(model, t in [0, 10])
t

Now t is a Julia variable that stores a ParameterRef which points to where the time parameter is stored in model. It can now be used with infinite variables, measures, and constraints as described in the appropriate sections.

When the model is optimized, t will be transcribed (discretized) over its domain following its support points. If none are specified by the user than 50 support points are generated that are equidistant over the interval. However, users may wish to employ their own support scheme. This can be done by using the num_supports or supports keyword arguments. For example, if we desire to have only 10 equidistant supports then we could have instead defined t:

julia> @infinite_parameter(model, t in [0, 10], num_supports = 10)
t

More complex support schemes can be specified via supports such as:

julia> @infinite_parameter(model, t in [0, 10], supports = [0; 2; 7; 10])
t

Where we specified t to use 4 supports: 0, 2, 7, and 10.

We can also add supports after t has been initialized. This can be accomplished with add_supports. For example, consider the initial case where t has no supports and we now wish to add 4 supports:

julia> add_supports(t, [0., 2.5, 7.5, 10.])

julia> supports(t)
4-element Array{Number,1}:
  0.0
  2.5
  7.5
 10.0

Here only 4 supports are specified for the sake of example. Alternatively, we could have initialized the parameter and added supports in just one step using the supports keyword argument:

julia> @infinite_parameter(model, t in [0, 10], supports = [0., 2.5, 7.5, 10.])
t

We could also define a random parameter described by a statistical distribution. This can be accomplished using @infinite_parameter in combination with a distribution from Distributions.jl. For example let's define a vector of random parameters described by a Normal distribution:

julia> using Distributions

julia> @infinite_parameter(model, xi[i = 1:3] in Normal())
3-element Array{ParameterRef,1}:
 xi[1]
 xi[2]
 xi[3]

Note that we could have used i as an index to assign a different distribution to each parameter. Supports can also be specified for each parameter as shown above. The key difference is that by default 50 supports are generated randomly following the distribution. Similarly, the num_supports keyword is used to generate random supports. Now we have infinite parameters t and xi that are ready to be used in defining infinite variables and constraints. We also mention here that the @infinite_parameter macro is designed to closely emulate JuMP.@variable and thus handles arrays and keyword arguments in the same way. This is described in more detail below.

Parameter Definition

Defining/initializing an infinite parameter principally involves the following steps:

Define an AbstractInfiniteSet

Define support points within the set to later discretize the parameter

Specify if parameter is independent (only for multi-dimensional parameter groups)

Construct an InfOptParameter to store this information

Add the InfOptParameter object to an InfiniteModel and assign a name

Create a ParameterRef that points to the parameter object

Manual Definition

Infinite set definition is described above in the Infinite Sets section. The supports should be a vector of finite numbers that are drawn from the domain of the infinite set. These supports will be used to transcribe the InfiniteModel in preparation for it to be optimized. If desired, the supports can be specified after the parameter is defined and an empty vector will be used to construct the infinite parameter. We'll discuss independence a little further below. The build_parameter function is used to construct the InfOptParameter. For example, let's create a time parameter $t \in [0, 10]$ with supports [0, 2, 5, 7, 10]:

julia> set = IntervalSet(0, 10)
[0, 10]

julia> t_param = build_parameter(error, set, supports = [0, 2, 5, 7, 10])
InfOptParameter{IntervalSet}([0, 10], [0, 2, 5, 7, 10], false)

Now we have a InfOptParameter that contains an IntervalSet and supports. Note that the num_params and independent arguments are primarily meant to be used by @infinite_parameter to define array parameters and specify if they are correlated or independent.

Let's now add t_param to our InfiniteModel using add_parameter and assign it the name of t:

julia> t_ref = add_parameter(model, t_param, "t")
t

We can also create an anonymous infinite parameter by dropping the name from the add_parameter function call. For example:

julia> t_ref_noname = add_parameter(model, t_param)
noname

Now suppose we want to create an infinite parameter that is a random variable with a given distribution. We follow the same procedure as above, except we use distributions from Distributions.jl to define DistributionSet. For example, let's consider a random variable $x \in \mathcal{N}(0,1)$ with supports [-0.5, 0.5]:

julia> dist = Normal(0., 1.)
Normal{Float64}(μ=0.0, σ=1.0)

julia> set = DistributionSet(dist)
Normal{Float64}(μ=0.0, σ=1.0)

julia> x_param = build_parameter(error, set, supports = [-0.5, 0.5])
InfOptParameter{DistributionSet{Normal{Float64}}}(Normal{Float64}(μ=0.0, σ=1.0), [-0.5, 0.5], false)

Again, we use add_parameter to add x_param to the InfiniteModel and assign it the name x:

julia> x_ref = add_parameter(model, x_param, "x")
x

Note that add_parameter does not register the name of the parameters into the model that it adds to. As shown in Macro Definition, the macro definition does not allow for multiple parameters sharing the same name and will throw an error if it happens.

Macro Definition

One-Dimensional Parameters

One user-friendly way of defining infinite parameters is by macro @infinite_parameter. The macro executes the same process as the manual definition (steps listed in Parameter Definition), but allows the users to manipulate several features of the defined infinite parameters. Again, let's consider a time parameter $t \in [0, 10]$ with supports [0, 2, 5, 7, 10]. Similar to JuMP.@variable, we can use comparison operators to set lower bounds and upper bounds for the infinite parameter:

julia> @infinite_parameter(model, 0 <= t <= 10, supports = [0, 2, 5, 7, 10])
t

More generally, we use in to define the set that an infinite parameter is subject to. The set could be an interval set, or a distribution set. For example, we can define the same parameter t as above in the following way:

julia> @infinite_parameter(model, t in [0, 10], supports = [0, 2, 5, 7, 10])
t

In a similar way, we can define a random infinite parameter subject to some distribution. For example, a Gaussian infinite parameter with mean 0 and standard deviation 1 can be defined by

julia> dist = Normal(0., 1.)
Normal{Float64}(μ=0.0, σ=1.0)

julia> @infinite_parameter(model, x in dist, supports = [-0.5, 0.5])
x

Additional ways of defining infinite parameters are provided using keyword arguments. For example, we can use lower_bound and upper_bound to define an infinite parameter in an interval set:

julia> @infinite_parameter(model, t, lower_bound = 0, upper_bound = 10, supports = [0, 2, 5, 7, 10])
t

A bit more generally, we can also use set to directly input the AbstractInfiniteSet that the parameter is in. For example:

julia> @infinite_parameter(model, t, set = IntervalSet(0, 10), supports = [0, 2, 5, 7, 10])
t

The parameter definition methods using keyword arguments will be useful later when we introduce how to define anonymous parameters. See the part for anonymous parameter definition for more details.

All the definitions above return a ParameterRef that refer to the defined parameter. Note that we can also ignore the supports keyword argument and the macro will define an empty array of supports for that parameter.

Multi-Dimensional Parameter

Using macro definition, we can also define multi-dimensional infinite parameters in a concise way. For example, consider a position parameter x in a 3-dimensional space constrained in a unit cube (i.e. in the interval [0, 1] for all dimensions). This parameter can be defined in one line as follows:

julia> @infinite_parameter(model, x[1:3] in [0, 1], supports = [0.3, 0.7])
3-element Array{ParameterRef,1}:
 x[1]
 x[2]
 x[3]

For multi-dimensional parameters, the macro calls for build_parameter and add_parameter and creates looped codes that construct separate parameters and references for each dimension. If an array of supports is provided, the macro will assign that array of supports to all dimensions. Otherwise, the indexed syntax can be used to feed in different array of supports to each dimension, similar to JuMP.@variable. For example:

julia> points = [0.2 0.8; 0.3 0.7]
2×2 Array{Float64,2}:
 0.2  0.8
 0.3  0.7

julia> @infinite_parameter(model, a[i = 1:2] in [0, 1], supports = points[i, :])
2-element Array{ParameterRef,1}:
 a[1]
 a[2]

julia> supports(a[1])
2-element Array{Float64,1}:
 0.2
 0.8

julia> supports(a[2])
2-element Array{Float64,1}:
 0.3
 0.7

In a similar way we can define an infinite parameter subject to a multivariate distribution. For example, a 2-dimensional parameter xi subject to a 2-D normal distribution can be created as follows:

julia> dist = MvNormal([0., 0.], [1. 0.; 0. 2.])
FullNormal(
dim: 2
μ: [0.0, 0.0]
Σ: [1.0 0.0; 0.0 2.0]
)

julia> @infinite_parameter(model, xi[1:2] in dist)
2-element Array{ParameterRef,1}:
 xi[1]
 xi[2]

Containers for Multi-Dimensional Parameters

Note that for all the cases of multi-dimensional parameter definition above, the macro always returns an Array of ParameterRef. For most cases this is true. However, we can explicitly dictate the kind of containers we want to hold the defined parameters using the keyword container. For example, we use SparseAxisArray from the JuMP package for the space parameter x:

julia> @infinite_parameter(model, x[1:3] in [0, 1], container = SparseAxisArray)
JuMP.Containers.SparseAxisArray{ParameterRef,1,Tuple{Int64}} with 3 entries:
  [3]  =  x[3]
  [2]  =  x[2]
  [1]  =  x[1]

independent for Multi-Dimensional Parameters

For examples up to now we did not specify the value for the keyword independent, which is set as false by default. The keyword independent applies to multi-dimensional infinite parameters and dictates whether the supports for different dimensions are independent. Setting independent as true would be useful if the users want to generate a grid of supports for a multi-dimensional parameter. For example, consider the position parameter x in a 3D space. Say x is bounded in [0, 1] in all three dimensions, and the user wants to generate grid points with interval 0.5 in all three dimensions. In this case, we can define x in the following way:

julia> pts = collect(range(0, stop = 1, length = 3))
3-element Array{Float64,1}:
 0.0
 0.5
 1.0

julia> @infinite_parameter(model, x[1:3] in [0, 1], supports = pts, independent = true)
3-element Array{ParameterRef,1}:
 x[1]
 x[2]
 x[3]

If independent is set as false, the transcription step will generate JuMP variables for values of any variable parameterized by x at [0.0, 0.0, 0.0], [0.5, 0.5, 0.5] and [1.0, 1.0, 1.0], a total of 3 transcribed variables. Instead, if independent is set as true, the transcription step will obtain a unique permutation of these supports and each transcribe parameterized variable accordingly, leading to a total of 27 transcribed variables in this case.

Anonymous Parameter Definition and base_name

As mentioned above, we can define anonymous parameters using keyword arguments in the macro @infinite_parameter. For instance, we can create an anonymous position parameter in a 3D space, referred to by a list of ParameterRef called x:

julia> x = @infinite_parameter(model, [1:3], lower_bound = 0, upper_bound = 1)
3-element Array{ParameterRef,1}:
 noname
 noname
 noname

julia> typeof(x)
Array{ParameterRef,1}

julia> name(x[1])
""

This syntax creates a 1D parameter if the part [1:3] is neglected.

Note that this macro definition automatically assigns an empty string to the base_name. We can also assign a nontrivial base name to an anonymous parameter using the keyword argument base_name. For example,

julia> @infinite_parameter(model, [1:3], lower_bound = 0, upper_bound = 1, base_name = "x")
3-element Array{ParameterRef,1}:
 x[1]
 x[2]
 x[3]

julia> @infinite_parameter(model, [1:3], lower_bound = -1, upper_bound = 0, base_name = "x")
3-element Array{ParameterRef,1}:
 x[1]
 x[2]
 x[3]

We can see that anonymous parameter definition allows for multiple parameters sharing the same base name. This is not permitted with non-anonymous parameter definition. In fact, in anonymous parameter definition, the macro does not register the name of the parameters in the model, so when the model checks for repeated names it will not detect the x. Refer to Detailed Mechanism of Macro Definition if more details are desired.

Detailed Mechanism of Macro Definition

This section is for people who wish to know more about how the macro @infinite_parameter works in the backend. Users who only want to learn about the setting up the model can skip over this part.

In general, the macro @infinite_parameter follows the same steps as the manual definition. First, it parses the arguments and identifies any recognizable keyword arguments. Specifically, the first argument must be the model, and the second argument, if exists, must be an expression that declares the parameter or simply specify the dimension of the parameter if users choose to define it anonymously. If the information in the keyword arguments is not sufficient to define the set the parameter is in, the users also need to specify the sets in the second argument using expressions like a <= x <= b or x in set.

The keyword arguments give users flexibility in how to define their parameters. As mentioned above, the users can choose to specify the set either in the second argument (nonanonymous parameter definition only), or in the keyword arguments. However, the users cannot do both at the same time. The macro will check this behavior and throw an error if this happens. For example,

julia> @infinite_parameter(model,  y in [0, 1], lower_bound = 0, upper_bound = 1)
ERROR: LoadError

julia> @infinite_parameter(model,  y in [0, 1], set = IntervalSet(0, 1))
ERROR: LoadError

Once the check on arguments and keyword arguments is done, the macro will create the AbstractInfiniteSet based on given information, and create the infinite parameter accordingly. If the users create a multi-dimensional parameter, the macro will create looped code to define individual infinite parameter for each dimension. The looped code will also incorporate different supports for different dimensions.

In the end, if the created parameter is not anonymous, the macro will register the name to the model. In this way, we prevent parameters created by @infinite_parameter non-anonymously to share the same name.

Supports

For an infinite parameter, its supports are a finite set of points that the parameter will take (or possibly take, if the parameter is random). During the transcription stage, the supports specified will become part of the grid points that approximate all functions parameterized by the infinite parameter.

Once an infinite parameter is defined, users can access the supports using supports function:

julia> @infinite_parameter(model, 0 <= t <= 10, supports = [0, 2, 5, 7, 10])
t

julia> supports(t)
5-element Array{Int64,1}:
  0
  2
  5
  7
 10

We also provide functions that access other related information about the supports. For example, has_supports checks whether a parameter has supports, while num_supports gives the number of supports associated with a parameter:

julia> has_supports(t)
true

julia> num_supports(t)
5

Now suppose we want to add more supports to the t, which is already assigned with some supports. We can use add_supports function to achieve this goal:

julia> add_supports(t, [3, 8])

julia> supports(t)
7-element Array{Int64,1}:
  0
  2
  5
  7
 10
  3
  8

At times we might want to change the supports completely. In those cases, the function set_supports resets the supports for a certain parameter with new supports provided:

julia> set_supports(t, [0,3,5,8,10], force = true)

julia> supports(t)
5-element Array{Int64,1}:
  0
  3
  5
  8
 10

Note that the keyword argument [force] must be set as [true] if the parameter has been assigned with supports. Users can also delete all the supports of a parameter with delete_supports.

Automatic Support Generation During Parameter Definition

For the examples in the Parameter Definition, we have seen how to manually add supports to an infinite parameter. For a quick automatic generation of support points, though, users do not have to input the support points. Instead, the number of support points generated is supplied.

For an infinite parameter subject to an IntervalSet, uniformly spaced supports including both ends are generated across the interval. For example, defining a time parameter $t \in [0, 10]$ with 4 supports using build_parameter gives

julia> set = IntervalSet(0, 10)
[0, 10]

julia> t_param = build_parameter(error, set, num_supports = 4, sig_fig = 3)
InfOptParameter{IntervalSet}([0, 10], [0.0, 3.33, 6.67, 10.0], false)

Using macro definition we have

julia> @infinite_parameter(model, 0 <= t <= 10, num_supports = 4, sig_fig = 3)
t

julia> supports(t)
4-element Array{Float64,1}:
  0.0   
  3.33
  6.67
 10.0   

Note that the user can use the keyword argument sig_fig to dictate the significant figures for the supports. The default value of sig_fig is 5.

For an infinite parameter that follows a univariate distribution, supports are sampled from the underlying distribution. For example, we can define an infinite parameter subject to a normal distribution with mean 0 and variance 1:

julia> @infinite_parameter(model, x in dist, num_supports = 4)
x

julia> supports(x)
4-element Array{Float64,1}:
  0.67911
  0.82841
 -0.35301
 -0.13485

For multivariate distributions, though, we require support points are provided in the definition. However, we can use fill_in_supports! to generate supports for parameters following multivariate distributions. See Automatic Support Generation For Defined Parameters for details.

Automatic Support Generation For Defined Parameters

So far, we have seen that in both definition methods it is allowed to initialize a parameter with no supports. This is done by not specifying supports and num_supports. However, infinite parameters would not be allowed at the transcription step since it needs information about how to discretize the infinite parameters. In previous examples, we have shown that users can add supports to a defined parameter using methods add_supports and set_supports.

In this section we introduce automatic support generation for defined parameters with no associated supports. This can be done using the fill_in_supports! functions. fill_in_supports! can take as argument a ParameterRef, in which case it will generate supports for the associated infinite parameter. Alternatively, fill_in_supports! can also take an InfiniteModel as an argument, in which case it will generate supports for all infinite parameters of the InfiniteModel with no supports.

The fill_in_supports! method allows users to specify integer keyword arguments num_supports and sig_fig. num_supports dictates the number of supports to be generated, and sig_fig dictates the significant figures of generated supports desired. The default values are 50 and 5, respectively.

The ways by which supports are automatically generated are as follows. If the parameter is in an IntervalSet, then we generate an array of supports that are uniformly distributed along the interval, including the two ends. For example, consider a 3D position parameter x distributed in the unit cube [0, 1]. We can generate supports for that point in the following way:

julia> @infinite_parameter(model, x[1:3] in [0, 1], independent = true);

julia> fill_in_supports!.(x, num_supports = 3);

julia> supports.(x)
3-element Array{Array{Number,1},1}:
 [0.0, 0.5, 1.0]
 [0.0, 0.5, 1.0]
 [0.0, 0.5, 1.0]

Note that the dot syntax because fill_in_supports! takes single ParameterRef as argument. In each dimension, three equally spaced supports ([0.0, 0.5, 1.0]) are generated. Since the independent keyword is set as true, the transcription stage will create a three-dimensional grid for all variables parameterized by x, with each point separated by 0.5 units in each dimension.

If the parameter is in a DistributionSet, fill_in_supports! samples num_supports supports from the distribution. Recall that support generation is not allowed for parameters under multivariate distribution during parameter definition. However, if the parameter is defined first without supports, fill_in_supports! allows for supports generation. For example, for a 2D random variable xi under a multivariate Gaussian distribution, we can generate supports for it in the following way:

julia> dist = MvNormal([0., 0.], [1. 0.; 0. 2.])
FullNormal(
dim: 2
μ: [0.0, 0.0]
Σ: [1.0 0.0; 0.0 2.0]
)

julia> @infinite_parameter(model, xi[1:2] in dist);

julia> fill_in_supports!(xi[1], num_supports = 3)

julia> supports.(xi)
2-element Array{Array{Number,1},1}:
 [0.67911, -0.35301, 0.58662]
 [1.1716, -0.19071, 0.4205]

Note that for parameters under multivariate distribution, the users do not have to call fill_in_supports! for all ParameterRef associated with that parameter. fill_in_supports!, once called upon one ParameterRef, would search for the other ParameterRef and fill in supports for them if they, together with the original ParameterRef, make up for the same multi-dimensional parameter under a multivariate distribution.

Also, fill_in_supports! only fill in supports for parameters with no associated supports. To modify the supports of parameters already associated with some supports, refer to Supports for how to do that.

Parameter Queries

In addition to the modeling framework, this package provides many functions for users to access information about the model. This section will go over basic functions for accessing parameter information.

Once a (possibly large-scale) InfiniteModel is built, the users might want to check if an infinite parameter is actually used in any way. This could be checked by is_used function as follows:

julia> @infinite_parameter(model, x in [0, 1])
x

julia> is_used(x)
false

This function checks if the parameter is used by any constraint, measure, or variable. In a similar way, functions used_by_constraint, used_by_measure and used_by_variable can be applied to find out any dependency of specific types on the infinite parameter.

In addition, sometimes we need to check if a certain ParameterRef is valid with an InfiniteModel model, meaning that the parameter reference actually refers to some parameter associated with the model. We extend the JuMP.is_valid function from JuMP for that purpose. To see how to use this, for example,

julia> pref1 = ParameterRef(model, 1);

julia> pref2 = ParameterRef(model, 2);

julia> is_valid(model, pref1)
true

julia> is_valid(model, pref2)
false

The second call of is_valid returns false because the model does not have parameter with index 2 yet.

We can also access different information about the set that the infinite parameter is in. This is given by infinite_set, which takes a [ParameterRef] as argument. For example, we have

julia> infinite_set(x)
[0, 1]

infinite_set might be more useful if the infinite parameter is in a DistributionSet, by which users can access information about the underlying distribution. On the other hand, if we already know that the parameter is in an interval set, we can use JuMP.has_lower_bound, JuMP.lower_bound, JuMP.has_upper_bound, JuMP.upper_bound to retrieve information about the interval set in a more specific way:

julia> has_lower_bound(x)
true

julia> lower_bound(x)
0.0

julia> has_upper_bound(x)
true

julia> upper_bound(x)
1.0

For multi-dimensional parameters, the definition step would create the parameter for each dimension an add it into the model separately. However, users can find out if different parameters belong to the same group using group_id function. The group ID is simply an index assigned at the parameter definition. See the following example for how to use it:

julia> @infinite_parameter(model, y[1:2] in [0, 5])
2-element Array{ParameterRef,1}:
 y[1]
 y[2]

julia> group_id(x)
1

julia> group_id(y)
2

julia> group_id([x, y[1]])
ERROR: Array contains parameters from multiple groups.

In the example above, the function returns the group ID if it receives one ParameterRef. If the argument is a vector of ParameterRef, however, it will throw an error if the parameters do not belong to the same group. Otherwise, it will return the group ID for all parameters in the vector. Again, this function could be helpful if the users want to verify if a set of parameters are correlated.

Another important piece of information about multi-dimensional parameters is whether different dimensions are independent. This can be checked by is_independent:

julia> is_independent(y[1])
false

A quick way for users to obtain a ParameterRef for a parameter with a known name would be through parameter_by_name function. This function takes an InfiniteModel and the parameter name in string, and returns a ParameterRef for that parameter. For example,

julia> pref = parameter_by_name(model, "x")
x

If there is no parameter associated with that name, the function would return nothing. Otherwise, if multiple parameters share the same name, the function would throw an error.

Now we introduce two additional functions that we can use to access parameter information for an InfiniteModel. The function num_parameters returns the number of infinite parameters associated with a model, while all_parameters returns the list of all infinite parameter references in the model. For a quick example:

julia> num_parameters(model)
3

julia> all_parameters(model)
3-element Array{ParameterRef,1}:
 x   
 y[1]
 y[2]

Parameter Modification

In this section we introduce a few shortcuts for users to modify defined infinite parameters.

First, once an infinite parameter is defined, we can change its name by calling the [JuMP.set_name] function, which takes the [ParameterRef] that needs a name change and the name string as arguments. For example, to change the parameter x to t we can do:

julia> JuMP.set_name(x, "t")

julia> all_parameters(model)
3-element Array{ParameterRef,1}:
 t   
 y[1]
 y[2]

In a similar way, we can also change the infinite set that the parameter is in using the set_infinite_set function as follows:

julia> t = parameter_by_name(model, "t")
t

julia> set_infinite_set(t, IntervalSet(0, 5))

julia> infinite_set(t)
[0, 5]

For parameters in an IntervalSet, we extend JuMP.set_lower_bound and JuMP.set_upper_bound functions for users to modify the lower bounds and upper bounds. For example,

julia> JuMP.set_lower_bound(t, 1)

julia> JuMP.set_upper_bound(t, 4)

julia> infinite_set(t)
[1, 4]

We do not support setting lower bounds and upper bounds for random parameters in a DistributionSet and will throw an error if users attempt to do so. If users want to set lower bound and upper bound for a random infinite parameter, consider using Distributions.Truncated, which creates a truncated distribution from a univariate distribution.

We can also modify the independent status of a parameter using set_independent and unset_independent functions. The set_independent function sets true for the independent status, while unset_independent function sets false for the independent status.

julia> set_independent(y[1])

julia> is_independent(y[1])
true

julia> is_independent(y[2])
false

julia> unset_independent(y[1])

julia> is_independent(y[1])
false

Note that these two functions takes one single [ParameterRef] as argument, and it will not automatically find the other associated parameters and reset their status altogether. This gives more flexibility on how we want to discretize certain parameters. For example, if for a 3D position parameter x we want to have uniform discretization along x[1], but custom supports over the 2D space x[2] and x[3], we can set x[1] only as independent.

Datatypes

InfiniteOpt.InfOptParameter

InfiniteOpt.ParameterRef

InfiniteOpt.InfOptParameter — Type

InfOptParameter{T <: AbstractInfiniteSet} <: JuMP.AbstractVariable

A DataType for storing core infinite parameter information.

Fields

set::T The infinite set that characterizes the parameter.

supports::Vector{<:Number} The support points used to discretize this parameter.

independent::Bool Is independent of other parameters that share its group ID number.

InfiniteOpt.ParameterRef — Type

ParameterRef <: GeneralVariableRef

A DataType for untranscripted infinite parameters references that parameterize the infinite variables.

Fields

model::InfiniteModel Infinite model.

index::Int Index of variable in model.

Methods/Macros

InfiniteOpt.@infinite_parameter

InfiniteOpt.add_parameter

InfiniteOpt.add_supports

InfiniteOpt.all_parameters

InfiniteOpt.build_parameter

InfiniteOpt.delete_supports

InfiniteOpt.fill_in_supports!

InfiniteOpt.fill_in_supports!

InfiniteOpt.generate_and_add_supports!

InfiniteOpt.group_id

InfiniteOpt.group_id

InfiniteOpt.has_supports

InfiniteOpt.infinite_set

InfiniteOpt.is_independent

InfiniteOpt.is_used

InfiniteOpt.num_parameters

InfiniteOpt.num_supports

InfiniteOpt.parameter_by_name

InfiniteOpt.set_independent

InfiniteOpt.set_infinite_set

InfiniteOpt.set_supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.unset_independent

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_variable

JuMP.delete

JuMP.has_lower_bound

JuMP.has_upper_bound

JuMP.is_valid

JuMP.lower_bound

JuMP.name

JuMP.set_lower_bound

JuMP.set_name

JuMP.set_upper_bound

JuMP.upper_bound

InfiniteOpt.@infinite_parameter — Macro

@infinite_parameter(model, kw_args...)

Add an anonymous infinite parameter to the model model described by the keyword arguments kw_args and returns the parameter reference.

@infinite_parameter(model, expr, kw_args...)

Add a parameter to the model model described by the expression expr, the positional arguments args and the keyword arguments kw_args. (note that in the following the symbol <= can be used instead of ≤, the symbol >=can be used instead of ≥, and the symbo in can be used instead of ∈) The expression expr can be of the form:

paramexpr creating parameters described by paramexpr.

lb ≤ paramexpr ≤ ub creating parameters described by paramexpr characterized by a continuous interval set with lower bound lb and upper bound ub.

paramexpr ∈ [lb, ub] creating parameters described by paramexpr characterized by a continuous interval set with lower bound lb and upper bound ub.

paramexpr ∈ dist creating parameters described by paramexpr characterized by the Distributions.jl distribution object dist.

The expression paramexpr can be of the form:

paramname creating a scalar parameter of name paramname

paramname[...] or [...] creating a container of parameters

The recognized keyword arguments in kw_args are the following:

base_name: Sets the name prefix used to generate parameter names. It corresponds to the parameter name for scalar parameter, otherwise, the parameter names are set to base_name[...] for each index ... of the axes axes.

lower_bound: Sets the value of the parameter lower bound for an interval set.

upper_bound: Sets the value of the parameter upper bound for an interval set.

set: The InfiniteSet characterizing the parameters see IntervalSet and DistributionSet.

distribution: Sets the Distributions.jl distribution object that characterizes the parameters.

supports: Sets the support points for the parameters.

num_supports: Specifies the number of supports to be automatically generated. Note that supports takes precedence. Defaults to 50.

sig_figs: Specifies the number of significant digits that should be used in automatic support generation. Defaults to 5.

independent: Specifies if the each parameter is independent from each other or not. Defaults to false.

container: Specify the container type. Defaults to automatic

Examples

julia> @infinite_parameter(m, 0 <= x <= 1)
x

julia> @infinite_parameter(m, y[i = 1:2] in Normal(), num_supports = 10)
2-element Array{ParameterRef,1}:
 y[1]
 y[2]

julia> z = @infinite_parameter(m, ["a", "b"], distribution = Uniform(), independent = true)
2-dimensional DenseAxisArray{ParameterRef,2,...} with index sets:
    Dimension 1, "a"
    Dimension 2, "b"
And data, a 1×1 Array{ParameterRef,2}:
 noname

InfiniteOpt.build_parameter — Method

build_parameter(_error::Function, set::AbstractInfiniteSet,
                [num_params::Int = 1; num_supports::Int = 0,
                supports::Union{Number, Vector{<:Number}} = Number[],
                independent::Bool = false,
                sig_fig::Int = 5])::InfOptParameter

Returns a InfOptParameter given the appropriate information. This is analagous to JuMP.build_variable. Errors if supports violate the bounds associated set. Also errors if set contains a multivariate distribution with a different dimension than num_params. This is meant to primarily serve as a helper method for @infinite_parameter.

Example

julia> build_parameter(error, IntervalSet(0, 3), supports = Vector(0:3))
InfOptParameter{IntervalSet}([0, 3], [0, 1, 2, 3], false)

InfiniteOpt.add_parameter — Method

add_parameter(model::InfiniteModel, p::InfOptParameter,
              [name::String = ""])::ParameterRef

Returns a ParameterRef associated with the parameter p that is added to model. This adds a parameter to the model in a manner similar to JuMP.add_variable. This can be used to add parameters with the use of @infinite_parameter. build_parameter should be used to construct p.

Example

julia> p = build_parameter(error, IntervalSet(0, 3), supports = Vector(0:3))
InfOptParameter{IntervalSet}([0, 3], [0, 1, 2, 3], false)

julia> param_ref = add_parameter(model, p, "name")
name

InfiniteOpt.used_by_constraint — Method

used_by_constraint(pref::ParameterRef)::Bool

Return true if pref is used by a constraint or false otherwise.

Example

julia> used_by_constraint(t)
true

InfiniteOpt.used_by_measure — Method

used_by_measure(pref::ParameterRef)::Bool

Return true if pref is used by a measure or false otherwise.

Example

julia> used_by_measure(t)
false

InfiniteOpt.used_by_variable — Method

used_by_variable(pref::ParameterRef)::Bool

Return true if pref is used by an infinite variable or false otherwise.

Example

julia> used_by_variable(t)
true

InfiniteOpt.is_used — Method

is_used(pref::ParameterRef)::Bool

Return true if pref is used in the model or false otherwise.

Example

julia> is_used(t)
true

JuMP.delete — Method

JuMP.delete(model::InfiniteModel, pref::ParameterRef)

Extend JuMP.delete to delete infinite parameters and their dependencies. All variables, constraints, and measure functions that depend on pref are updated to exclude it. Errors if the parameter is contained in an AbstractMeasureData datatype that is employed by a measure since the measure becomes invalid otherwise. Thus, measures that contain this dependency must be deleted first. Note that _check_param_in_data(pref, measure_data) needs to be extended to allow deletion of parameters when custom AbstractMeasureData datatypes are used.

Example

julia> print(model)
Min measure(g(t, x)*t + x) + z
Subject to
 z >= 0.0
 g(t, x) + z >= 42.0
 g(0.5, x) == 0
 t in [0, 6]
 x in [0, 1]

julia> delete(model, x)

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 g(t) + z >= 42.0
 g(0.5) == 0
 z >= 0.0
 t in [0, 6]

JuMP.is_valid — Method

JuMP.is_valid(model::InfiniteModel, pref::ParameterRef)::Bool

Extend the JuMP.is_valid function to accomodate infinite parameters. Returns true if the InfiniteModel stored in pref matches model and if the parameter index is used by model. It returns false otherwise.

Example

julia> is_valid(model, t)
true

JuMP.name — Method

JuMP.name(pref::ParameterRef)::String

Extend the JuMP.name function to accomodate infinite parameters. Returns the name string associated with pref.

Example

julia> name(t)
"t"

JuMP.set_name — Method

JuMP.set_name(pref::ParameterRef, name::String)

Extend the JuMP.set_name function to accomodate infinite parameters. Set a new base name to be associated with pref.

Example

julia> set_name(t, "time")

julia> name(t)
"time"

InfiniteOpt.num_parameters — Method

num_parameters(model::InfiniteModel)::Int

Return the number of infinite parameters currently present in model.

Example

julia> num_parameters(model)
1

InfiniteOpt.infinite_set — Method

infinite_set(pref::ParameterRef)::AbstractInfiniteSet

Return the infinite set associated with pref.

Example

julia> infinite_set(t)
[0, 1]

InfiniteOpt.set_infinite_set — Method

set_infinite_set(pref::ParameterRef, set::AbstractInfiniteSet)

Specify the infinite set of pref.

Example

julia> set_infinite_set(t, IntervalSet(0, 2))

julia> infinite_set(t)
[0, 2]

JuMP.has_lower_bound — Method

JuMP.has_lower_bound(pref::ParameterRef)::Bool

Extend the JuMP.has_lower_bound function to accomodate infinite parameters. Return true if the set associated with pref has a defined lower bound or if a lower bound can be found. Extensions with user-defined infinite set types should extend JuMP.has_lower_bound(set::NewType).

Example

julia> has_lower_bound(t)
true

JuMP.lower_bound — Method

JuMP.lower_bound(pref::ParameterRef)::Number

Extend the JuMP.lower_bound function to accomodate infinite parameters. Returns the lower bound associated with the infinite set. Errors if such a bound is not well-defined.

Example

julia> lower_bound(t)
0.0

JuMP.set_lower_bound — Method

JuMP.set_lower_bound(pref::ParameterRef, lower::Number)

Extend the JuMP.set_lower_bound function to accomodate infinite parameters. Updates the infinite set lower bound if such an operation is supported. Set extensions that seek to employ this should extend JuMP.set_lower_bound(set::NewType, lower::Number).

Example

julia> set_lower_bound(t, -1)

julia> lower_bound(t)
-1.0

JuMP.has_upper_bound — Method

JuMP.has_upper_bound(pref::ParameterRef)::Bool

Extend the JuMP.has_upper_bound function to accomodate infinite parameters. Return true if the set associated with pref has a defined upper bound or if a upper bound can be found. Extensions with user-defined sets should extend JuMP.has_upper_bound(set::NewType).

Example

julia> has_upper_bound(t)
true

JuMP.upper_bound — Method

JuMP.upper_bound(pref::ParameterRef)::Number

Extend the JuMP.upper_bound function to accomodate infinite parameters. Returns the upper bound associated with the infinite set. Errors if such a bound is not well-defined. Extensions with user-defined set types should extend JuMP.has_upper_bound(set::NewType) and JuMP.upper_bound(set::NewType) if appropriate.

Example

julia> upper_bound(t)
1.0

JuMP.set_upper_bound — Method

JuMP.set_upper_bound(pref::ParameterRef, lower::Number)

Extend the JuMP.set_upper_bound function to accomodate infinite parameters. Updates the infinite set upper bound if and only if it is an IntervalSet. Errors otherwise. Extensions with user-defined infinite sets should extend JuMP.set_upper_bound(set::NewType, upper::Number) if appropriate.

Example

julia> set_upper_bound(t, 2)

julia> upper_bound(t)
2.0

InfiniteOpt.num_supports — Method

num_supports(pref::ParameterRef)::Int

Return the number of support points associated with pref.

Example

julia> num_supports(t)
2

InfiniteOpt.has_supports — Method

has_supports(pref::ParameterRef)::Bool

Return true if pref has supports or false otherwise.

Example

julia> has_supports(t)
true

InfiniteOpt.supports — Method

supports(pref::ParameterRef)::Vector

Return the support points associated with pref. Errors if there are no supports.

Example

julia> supports(t)
2-element Array{Int64,1}:
 0
 1

InfiniteOpt.supports — Method

supports(prefs::AbstractArray{<:ParameterRef})::Vector

Return the support points associated with an array of prefs formatted as a vector of SparseAxisArrays following the format of the input array. If the parameters are not independent then the supports of each parameter are simply spliced together. Alternatively can call supports. to more efficiently obtain an array of the same input format whose parameter references have been replaced with their supports. Errors if all the parameter references do not have the same group ID number (were intialized together as an array) or if the nonindependent parameters have support vectors of different lengths. If the parameters are independent then all the unique combinations are identified and returned as supports. Warning this operation is computationally expensive if there exist a large number of combinations.

Example

julia> x = @infinite_parameter(model, [i = 1:2], set = IntervalSet(-1, 1),
                               base_name = "x", independent = true)
2-element Array{ParameterRef,1}:
 x[1]
 x[2]

julia> for i = 1:length(x)
           set_supports(x[i], [-1, 1])
       end

julia> supports(x)
4-element Array{JuMP.Containers.SparseAxisArray,1}:
   [2]  =  -1
  [1]  =  -1
   [2]  =  1
  [1]  =  -1
   [2]  =  -1
  [1]  =  1
   [2]  =  1
  [1]  =  1

InfiniteOpt.set_supports — Method

set_supports(pref::ParameterRef, supports::Vector{<:Number}; [force = false])

Specify the support points for pref. Errors if the supports violate the bounds associated with the infinite set. Warns if the points are not unique. If force this will overwrite exisiting supports otherwise it will error if there are existing supports.

Example

julia> set_supports(t, [0, 1])

julia> supports(t)
2-element Array{Int64,1}:
 0
 1

InfiniteOpt.add_supports — Method

add_supports(pref::ParameterRef, supports::Union{Number, Vector{<:Number}})

Add additional support points for pref.

Example

julia> add_supports(t, 0.5)

julia> supports(t)
3-element Array{Float64,1}:
 0.0
 1.0
 0.5

julia> add_supports(t, [0.25, 1])

julia> supports(t)
4-element Array{Float64,1}:
 0.0
 1.0
 0.5
 0.25

InfiniteOpt.delete_supports — Method

delete_supports(pref::ParameterRef)

Delete the support points for pref.

Example

julia> delete_supports(t)

julia> supports(t)
ERROR: Parameter t does not have supports.

InfiniteOpt.fill_in_supports! — Method

fill_in_supports!(model::InfiniteModel; [num_supports::Int = 10,
                  sig_fig::Int = 5])

Automatically generate support points for all infinite parameters in model. User can specify the number of significant figures kept after decimal point for the auto-generated supports wtih sig_fig. This calls fill_in_supports! for each parameter in the model. See fill_in_supports! for more information. Errors if one of the infinite set types is unrecognized.

Example

julia> fill_in_supports!(model, num_supports = 4, sig_fig = 3)

julia> supports(x)
4-element Array{Number,1}:
 0.0
 0.333
 0.667
 1.0

InfiniteOpt.fill_in_supports! — Method

fill_in_supports!(pref::ParameterRef; [num_supports::Int = 10,
                                       sig_fig::Int = 5])

Automatically generate support points for a particular infinite parameter pref. Generating num_supports for the parameter. The supports are generated uniformly if the underlying infinite set is an IntervalSet or they are generating randomly accordingly to the distribution if the set is a DistributionSet. User can specify the number of digits kept after decimal point for the auto-generated supports wtih sig_fig. Extensions that use user defined set types should extend generate_and_add_supports! and/or generate_support_values as needed. Errors if the infinite set type is not recognized.

Example

julia> fill_in_supports!(x, num_supports = 4, sig_fig = 3)

julia> supports(x)
4-element Array{Number,1}:
 0.0
 0.333
 0.667
 1.0

InfiniteOpt.generate_and_add_supports! — Function

generate_and_add_supports!(pref::ParameterRef, set::AbstractInfiniteSet;
                           [num_supports::Int = 10, sig_fig::Int = 5])

Generate supports for pref via generate_support_values and add them to pref. This is intended as an extendable internal method for fill_in_supports!. Note that if pref is part of a DistributionSet that features a multivariate distribution, all the associated parameters with pref will also have supports added to them. Most extensions that empoy user-defined infinite sets can typically enable this by extending generate_support_values. However, in some cases it may be necessary to extend this when more complex operations need to take place then just adding supports to a single infinite parameter (e.g., how we enable multivariate distribution sets). Errors if the infinite set type is not recognized.

InfiniteOpt.group_id — Method

group_id(pref::ParameterRef)::Int

Return the group ID number for pref.

Example

julia> group_id(t)
1

InfiniteOpt.group_id — Method

group_id(prefs::AbstractArray{<:ParameterRef})::Int

Return the group ID number for a group of prefs. Error if contains multiple groups.

Example

julia> group_id([x[1], x[2]])
2

InfiniteOpt.is_independent — Method

is_independent(pref::ParameterRef)::Bool

Returns true for pref if it is independent or false otherwise.

Example

julia> is_independent(t)
false

InfiniteOpt.set_independent — Method

set_independent(pref::ParameterRef)

Specify that pref be independent.

Example

julia> set_independent(t)

julia> is_independent(t)
true

InfiniteOpt.unset_independent — Method

unset_independent(pref::ParameterRef)

Specify that pref be not independent.

Example

julia> unset_independent(t)

julia> is_independent(t)
false

InfiniteOpt.parameter_by_name — Method

parameter_by_name(model::InfiniteModel, name::String)::Union{ParameterRef,
                                                             Nothing}

Return the parameter reference assoociated with a parameter name. Errors if multiple parameters have the same name. Returns nothing if no such name exists.

Example

julia> parameter_by_name(model, "t")
t

InfiniteOpt.all_parameters — Method

all_parameters(model::InfiniteModel)::Vector{ParameterRef}

Return all of the infinite parameter references currently in model.

Example

julia> all_parameters(model)
3-element Array{ParameterRef,1}:
 t
 x[1]
 x[2]