Model Transcription

A guide and manual for transcribing infinite models using InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

All infinite models need to be reformulated in such a way that they can be solved using traditional optimization methods. Typically, this involves discretization of the infinite domain via particular parameter support points. By default, InfiniteOpt employs this methodology via the use of transcription models (which comprise the optimizer_model as discussed in the Infinite Models section). InfiniteOpt is built modularly to readily accept other user defined techniques and this is discussed in further detail on the Extensions page. This page will detail transcription models based in InfiniteOpt.TranscriptionOpt which provide the default transcription (reformulation) capabilities of InfiniteOpt.

Basic Usage

Most users will not need to employ the capabilities of TranscriptionOpt directly since they are employed implicitly with the call of optimize! on an infinite model. This occurs since TranscriptionModels are the default optimizer model type that is employed.

However, some users may wish to use TranscriptionOpt to extract a fully discretized/transcribed version of an infinite model that is conveniently output as a typical JuMP model and can then be treated as such. This is principally accomplished via build_optimizer_model!. To illustrate how this is done, let's first define a basic infinite model with a simple support structure for the sake of example:

julia> using InfiniteOpt, JuMP

julia> inf_model = InfiniteModel();

julia> @infinite_parameter(inf_model, t in [0, 10], supports = [0, 5, 10])
t

julia> @infinite_variable(inf_model, g(t) >= 0)
g(t)

julia> @hold_variable(inf_model, z, Bin)
z

julia> @objective(inf_model, Min, 2z + support_sum(g, t))
2 z + support_sum{t}[g(t)]

julia> @BDconstraint(inf_model, initial(t == 0), g == 1)
initial : g(t) = 1.0, ∀ t = 0

julia> @constraint(inf_model, constr, g^2 - z <= 42)
constr : g(t)² - z ≤ 42.0, ∀ t ∈ [0, 10]

julia> print(inf_model)
Min 2 z + support_sum{t}[g(t)]
Subject to
 g(t) ≥ 0.0, ∀ t ∈ [0, 10]
 z binary
 initial : g(t) = 1.0, ∀ t = 0
 constr : g(t)² - z ≤ 42.0, ∀ t ∈ [0, 10]

Now we can make JuMP model containing the transcribed version of inf_model via build_optimizer_model! and then extract it via optimizer_model:

julia> build_optimizer_model!(inf_model)

julia> trans_model = optimizer_model(inf_model)
A JuMP Model
Minimization problem with:
Variables: 4
Objective function type: GenericAffExpr{Float64,VariableRef}
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
`GenericQuadExpr{Float64,VariableRef}`-in-`MathOptInterface.LessThan{Float64}`: 3 constraints
`VariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 3 constraints
`VariableRef`-in-`MathOptInterface.ZeroOne`: 1 constraint
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

julia> print(trans_model)
Min 2 z + g(support: 1) + g(support: 2) + g(support: 3)
Subject to
 initial(support: 1) : g(support: 1) = 1.0
 constr(support: 1) : g(support: 1)² - z ≤ 42.0
 constr(support: 2) : g(support: 2)² - z ≤ 42.0
 constr(support: 3) : g(support: 3)² - z ≤ 42.0
 g(support: 1) ≥ 0.0
 g(support: 2) ≥ 0.0
 g(support: 3) ≥ 0.0
 z binary

Note

Previous versions of InfiniteOpt, employed a TranscriptionModel(model::InfiniteModel) constructor to build transcription models independently of the optimizer model. This has functionality has been removed in favor of internal optimizer model based builds for efficiency reasons and to properly manage MOI optimizer attributes.

Thus, we have a transcribed JuMP model. To be precise this is actually a TranscriptionModel which is a JuMP.Model with some extra data stored in the ext field that retains the mapping between the transcribed variables/constraints and their infinite counterparts. Notice, that multiple finite variables have been introduced to discretize g(t) at supports 1, 2, and 3 which correspond to 0, 5, and 10 as can be queried by supports:

julia> supports(g)
3-element Array{Tuple,1}:
 (0.0,)
 (5.0,)
 (10.0,)

Also, notice how the constraints are transcribed in accordance with these supports except the initial condition which naturally is only invoked for the first support point. Furthermore, the transcription variable(s) of any variable associated with the infinite model can be determined via transcription_variable:

julia> transcription_variable(g)
3-element Array{VariableRef,1}:
 g(support: 1)
 g(support: 2)
 g(support: 3)

julia> transcription_variable(trans_model, z)
z

Similarly, the transcription constraints associated with infinite model constraints can be queried via transcription_constraint and the associated supports and infinite parameters can be found via supports and parameter_refs:

julia> transcription_constraint(initial)
initial(support: 1) : g(support: 1) = 1.0

julia> transcription_constraint(constr)
3-element Array{ConstraintRef,1}:
 constr(support: 1) : g(support: 1)² - z ≤ 42.0
 constr(support: 2) : g(support: 2)² - z ≤ 42.0
 constr(support: 3) : g(support: 3)² - z ≤ 42.0

julia> supports(constr)
3-element Array{Tuple,1}:
 (0.0,)
 (5.0,)
 (10.0,)

julia> parameter_refs(constr)
(t,)

Note the parameter reference tuple corresponds to the support tuples.

Note

Method that query the transcription surrogates (e.g., transcription_variable) and the respective supports via supports also accept the keyword argument label to specify which that transcription objects are desired in accordance to the support labels that are inherited from and/or are equal to label. By default, this will return any supports that are public (i.e., will hide anything solely associated with internal supports). The full query response can always be obtained via label = All.

Now we have a transcribed JuMP model that can be optimized via traditional JuMP methods whose variables and constraints can be accessed using the methods mentioned above.

Transcription Theory

A given infinite dimensional optimization problem is parameterized according to infinite parameters following our abstraction. In general, most solution strategies transcribe the problem according to certain finite parameter values (supports) and thus represent the problem in terms of these supports (e.g., using discrete time points in dynamic optimization). This methodology can be generalized into the following steps:

define supports for each infinite parameter if not already defined,

add any additional support needed for derivative evaluation,

expand any measures according to their underlying numerical representation

using transcribed infinite variables as appropriate,

replace any remaining infinite variables/derivatives with transcribed variables supported

over each unique combination of the underlying parameter supports,

replace any remaining infinite constraints with transcribed ones supported over

all the unique support combinations stemming from the infinite parameters they depend on,

and add on the transcripted versions of the auxiliary derivative evaluation

equations.

For example, let's consider a space-time optimization problem of the form:

\[\begin{aligned} &&\min_{y(t), g(t, x)} &&& \int_0^{10} y^2(t) dt \\ &&\text{s.t.} &&& y(0) = 1 \\ &&&&& \int_{x \in [-1, 1]^2} \frac{\partial g(t, x)}{\partial t} dx = 42, && \forall t \in [0, 10] \\ &&&&& 3g(t, x) + 2y^2(t) \leq 2, && \forall t \in T, \ x \in [-1, 1]^2. \\ \end{aligned}\]

Thus, we have an optimization problem whose decision space is infinite with respect to time $t$ and position $x$. Now let's transcript it following the above steps. First, we need to specify the infinite parameter supports and for simplicity let's choose the following sparse sets:

$t \in \{0, 10\}$

$x \in \{[-1, -1]^T, [-1, 1]^T, [1, -1]^T, [1, 1]^T\}$.

To handle the derivative $\frac{\partial g(t, x)}{\partial t}$, we'll use backward finite difference so no additional supports will need to be added.

Now we expand the two integrals (measures) via a finite approximation using only the above supports and term coefficients of 1 (note this is not numerically correct but is done for conciseness in example). Doing this, we obtain the form:

\[\begin{aligned} &&\min_{y(t), g(t, x)} &&& y^2(0) + y^2(10) \\ &&\text{s.t.} &&& y(0) = 1 \\ &&&&& g(0, x) = 0 \\ &&&&& \frac{\partial g(t, [-1, -1])}{\partial t} + \frac{\partial g(t, [-1, 1])}{\partial t} + \frac{\partial g(t, [1, -1])}{\partial t} + \frac{\partial g(t, [1, 1])}{\partial t} = 42, && \forall t \in [0, 10] \\ &&&&& 3g(t, x) + 2y^2(t) \leq 2, && \forall t \in T, \ x \in [-1, 1]^2. \\ \end{aligned}\]

Notice that the infinite variable $y(t)$ in the objective measure has been replaced with finite transcribed variables $y(0)$ and $y(10)$. Also, the infinite derivative $\frac{\partial g(t, x)}{\partial t}$ was replaced with partially transcribed variables in the second constraint in accordance with the measure over the positional domain $x$.

Now we need to transcribe the remaining infinite and semi-infinite variables with finite variables and duplicate the remaining infinite constraints accordingly. This means that the second constraint needs to be transcribed over the time domain and the third constraint needs to be transcribed for each unique combination of the time and position supports. Applying this transcription yields:

\[\begin{aligned} &&\min_{y(t), g(t, x)} &&& y^2(0) + y^2(10) \\ &&\text{s.t.} &&& y(0) = 1 \\ &&&&& g(0, [-1, -1]) = 0 \\ &&&&& g(0, [-1, 1]) = 0 \\ &&&&& g(0, [1, -1]) = 0 \\ &&&&& g(0, [1, 1]) = 0 \\ &&&&& \frac{\partial g(0, [-1, -1])}{\partial t} + \frac{\partial g(0, [-1, 1])}{\partial t} + \frac{\partial g(0, [1, -1])}{\partial t} + \frac{\partial g(0, [1, 1])}{\partial t} = 42\\ &&&&& \frac{\partial g(10, [-1, -1])}{\partial t} + \frac{\partial g(10, [-1, 1])}{\partial t} + \frac{\partial g(10, [1, -1])}{\partial t} + \frac{\partial g(10, [1, 1])}{\partial t} = 42\\ &&&&& 3g(0, [-1, -1]) + 2y^2(0) \leq 2 \\ &&&&& 3g(0, [-1, 1]) + 2y^2(0) \leq 2 \\ &&&&& \vdots \\ &&&&& 3g(10, [1, 1]) + 2y^2(10) \leq 2. \end{aligned}\]

Now that the variables and constraints are are transcripted, all that remains is to add relations to define the behavior of the transcripted partial derivatives. We can accomplish this via backward finite difference which will just add one infinite equation in this case this we only have 2 supports in the time domain is then transcripted over the spatial domain to yield:

\[\begin{aligned} &&& g(10, [-1, -1]) = g(0, [-1, -1]) + 10\frac{\partial g(10, [-1, -1])}{\partial t} \\ &&& g(10, [-1, 1]) = g(0, [-1, 1]) + 10\frac{\partial g(10, [-1, 1])}{\partial t} \\ &&& g(10, [1, -1]) = g(0, [1, -1]) + 10\frac{\partial g(10, [1, -1])}{\partial t} \\ &&& g(10, [1, 1]) = g(0, [1, 1]) + 10\frac{\partial g(10, [1, 1])}{\partial t} \end{aligned}\]

Now the problem is fully transcripted (discretized) and can be solved as a standard optimization problem. Note that with realistic measure evaluation schemes more supports might be added to the support sets and these will need to be incorporated when transcribing variables and constraints.

It is easy to imagine how the above procedure can get quite involved to do manually, but this is precisely what InfiniteOpt automates behind the scenes. Let's highlight this by repeating the same example using InfiniteOpt (again using the incorrect simple representation for the integrals for conciseness).

using JuMP, InfiniteOpt

# Initialize model
inf_model = InfiniteModel()

# Define parameters and supports
@infinite_parameter(inf_model, t in [0, 10], supports = [0, 10])
@infinite_parameter(inf_model, x[1:2] in [-1, 1], supports = [-1, 1], independent = true)

# Define variables
@infinite_variable(inf_model, y(t))
@infinite_variable(inf_model, g(t, x))

# Set the objective (using support_sum for the integral given our simple example)
# Note: In real problems integral should be used
@objective(inf_model, Min, support_sum(y^2, t))

# Define the constraints
@BDconstraint(inf_model, t == 0, y == 1)
@BDconstraint(inf_model, t == 0, g == 0)
@constraint(inf_model, support_sum(deriv(g, t), x) == 42) # support_sum for simplicity
@constraint(inf_model, 3g + y^2 <= 2)

# Print the infinite model
print(inf_model)

# output
Min support_sum{t}[y(t)²]
Subject to
 y(t) = 1.0, ∀ t = 0
 g(t, x) = 0.0, ∀ t = 0, x[1] ∈ [-1, 1], x[2] ∈ [-1, 1]
 support_sum{x}[∂/∂t[g(t, x)]] = 42.0, ∀ t ∈ [0, 10]
 y(t)² + 3 g(t, x) ≤ 2.0, ∀ t ∈ [0, 10], x[1] ∈ [-1, 1], x[2] ∈ [-1, 1]

Thus, we obtain the infinite problem in InfiniteOpt. As previously noted, transcription would be handled automatically behind the scenes when the model is optimized. However, we can directly extract the transcribed version by building a TranscriptionModel:

julia> build_optimizer_model!(inf_model)

julia> trans_model = optimizer_model(inf_model);

julia> print(trans_model)
Min y(support: 1)² + y(support: 2)²
Subject to
 y(support: 1) = 1.0
 g(support: 1) = 0.0
 g(support: 3) = 0.0
 g(support: 5) = 0.0
 g(support: 7) = 0.0
 ∂/∂t[g(t, x)](support: 1) + ∂/∂t[g(t, x)](support: 3) + ∂/∂t[g(t, x)](support: 5) + ∂/∂t[g(t, x)](support: 7) = 42.0
 ∂/∂t[g(t, x)](support: 2) + ∂/∂t[g(t, x)](support: 4) + ∂/∂t[g(t, x)](support: 6) + ∂/∂t[g(t, x)](support: 8) = 42.0
 g(support: 1) - g(support: 2) + 10 ∂/∂t[g(t, x)](support: 2) = 0.0
 g(support: 3) - g(support: 4) + 10 ∂/∂t[g(t, x)](support: 4) = 0.0
 g(support: 5) - g(support: 6) + 10 ∂/∂t[g(t, x)](support: 6) = 0.0
 g(support: 7) - g(support: 8) + 10 ∂/∂t[g(t, x)](support: 8) = 0.0
 y(support: 1)² + 3 g(support: 1) ≤ 2.0
 y(support: 2)² + 3 g(support: 2) ≤ 2.0
 y(support: 1)² + 3 g(support: 3) ≤ 2.0
 y(support: 2)² + 3 g(support: 4) ≤ 2.0
 y(support: 1)² + 3 g(support: 5) ≤ 2.0
 y(support: 2)² + 3 g(support: 6) ≤ 2.0
 y(support: 1)² + 3 g(support: 7) ≤ 2.0
 y(support: 2)² + 3 g(support: 8) ≤ 2.0

This precisely matches what we found analytically. Note that the unique support combinations are determined automatically and are represented visually as support: #. The precise support values can be looked up via supports:

julia> supports(y)
2-element Array{Tuple,1}:
 (0.0,)
 (10.0,)

julia> supports(g)
8-element Array{Tuple,1}:
 (0.0, [-1.0, -1.0])
 (10.0, [-1.0, -1.0])
 (0.0, [1.0, -1.0])
 (10.0, [1.0, -1.0])
 (0.0, [-1.0, 1.0])
 (10.0, [-1.0, 1.0])
 (0.0, [1.0, 1.0])
 (10.0, [1.0, 1.0])

julia> supports(g, ndarray = true) # format it as an n-dimensional array (t by x[1] by x[2])
2×2×2 Array{Tuple,3}:
[:, :, 1] =
 (0.0, [-1.0, -1.0])   (0.0, [1.0, -1.0])
 (10.0, [-1.0, -1.0])  (10.0, [1.0, -1.0])

[:, :, 2] =
 (0.0, [-1.0, 1.0])   (0.0, [1.0, 1.0])
 (10.0, [-1.0, 1.0])  (10.0, [1.0, 1.0])

TranscriptionOpt

InfiniteOpt.TranscriptionOpt is a sub-module which principally implements TranscriptionModels and its related access/modification methods. Thus, this section will detail what these are and how they work.

TranscriptionModels

A TranscriptionModel is simply a JuMP.Model whose ext field contains TranscriptionData which acts to map the transcribed model back to the original infinite model (e.g., map the variables and constraints). Such models are constructed via a default version of build_optimizer_model! which wraps build_transcription_model!:

julia> model1 = TranscriptionModel() # make an empty model
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

julia> build_optimizer_model!(inf_model); 

julia> model2 = optimizer_model(inf_model) # generate from an InfiniteModel
A JuMP Model
Minimization problem with:
Variables: 4
Objective function type: GenericAffExpr{Float64,VariableRef}
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
`GenericQuadExpr{Float64,VariableRef}`-in-`MathOptInterface.LessThan{Float64}`: 3 constraints
`VariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 3 constraints
`VariableRef`-in-`MathOptInterface.ZeroOne`: 1 constraint
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

Note that the all the normal JuMP.Model arguments can be used with both constructor when making an empty model and they are simply inherited from those specified in the InfiniteModel. The call to build_optimizer_model! is the backbone behind infinite model transcription and is what encapsulates all of the methods to transcribe measures, variables, derivatives, and constraints. This is also the method that enables the use of optimize!.

Queries

In this section we highlight a number of query methods that pertain to TranscriptionModels and their mappings. First, if the optimizer_model of an InfiniteModel is a TranscriptionModel it can be extracted via transcription_model:

julia> transcription_model(inf_model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

Here we observe that such a model is currently empty and hasn't been populated yet. Furthermore, we check that a Model is an TranscriptionModel via is_transcription_model:

julia> is_transcription_model(optimizer_model(inf_model))
true

julia> is_transcription_model(Model())
false

We can also extract the raw TranscriptionData object from a TranscriptionModel via transcription_data.

julia> transcription_data(trans_model);

Next we can retrieve the JuMP variable(s) for a particular InfiniteOpt variable via transcription_variable. For finite variables, this will be a one to one mapping, and for infinite variables a list of supported variables will be returned in the order of the supports. Following the initial example in the basic usage section, this is done:

julia> build_optimizer_model!(inf_model); trans_model = optimizer_model(inf_model);

julia> transcription_variable(trans_model, g)
3-element Array{VariableRef,1}:
 g(support: 1)
 g(support: 2)
 g(support: 3)

julia> transcription_variable(trans_model, z)
z

Note that if the TranscriptionModel is stored as the current optimizer_model then the first argument (specifying the TranscriptionModel can be omitted). Thus, in this case the first argument can be omitted as it was above, but is shown for completeness.

Similarly, the parameter supports corresponding to the transcription variables (in the case of transcribed infinite variables) can be queried via supports:

julia> supports(g)
3-element Array{Tuple,1}:
 (0.0,)
 (5.0,)
 (10.0,)

Note

Note that like supports the transcription_[obj] methods also employ the label::Type{AbstractSupportLabel} = PublicLabel keyword argument that by default will return variables/expressions/constraints associated with public supports. The full set (e.g., ones corresponding to internal collocation nodes) is obtained via label = All.

These methods also employ the ndarray::Bool keyword argument that will cause the output to be formatted as a n-dimensional array where the dimensions correspond to the infinite parameter dependencies. For example, if we have an infinite variable y(t, ξ) and we invoke a query method with ndarray = true then we'll get a matrix whose dimensions correspond to the supports of t and ξ, respectively. Also, if ndarray = true then label correspond to the intersection of supports labels in contrast to its default of invoking the union of the labels.

Likewise, transcription_constraint and supports(@ref) can be used with constraints to find their transcribed equivalents in the JuMP model and determine their supports.

We can also do this with measures and expressions:

julia> meas = support_sum(g^2, t)
support_sum{t}[g(t)²]

julia> build_optimizer_model!(inf_model)

julia> transcription_variable(meas)
g(support: 1)² + g(support: 2)² + g(support: 3)²

julia> supports(meas)
()

julia> transcription_expression(g^2 + z - 42)
3-element Array{AbstractJuMPScalar,1}:
 g(support: 1)² + z - 42
 g(support: 2)² + z - 42
 g(support: 3)² + z - 42

julia> supports(g^2 + z - 42)
3-element Array{Tuple,1}:
 (0.0,)
 (5.0,)
 (10.0,)

julia> parameter_refs(g^2 + z - 42)
(t,)

Datatypes

InfiniteOpt.TranscriptionOpt.TranscriptionData — Type

TranscriptionData

A DataType for storing the data mapping an InfiniteOpt.InfiniteModel that has been transcribed to a regular JuMP.Model that contains the transcribed variables. This is stored in the ext field of a JuMP.Model to make what is called a TranscriptionModel via the TranscriptionModel constructor.

Fields

infvar_lookup::Dict{InfiniteOpt.GeneralVariableRef, Dict{Vector{Float64}, Int}}: A lookup table of infinite variable transcriptions via support value.

infvar_mappings::Dict{InfiniteOpt.GeneralVariableRef, Vector{JuMP.VariableRef}}: Map infinite variables to their transcription variables.

infvar_supports::Dict{InfiniteOpt.GeneralVariableRef, Vector{Tuple}}: Map infinite variables to their support values.

infvar_support_labels::Dict{InfiniteOpt.GeneralVariableRef, Vector{Set{DataType}}}: Map the infinite variables to their support labels.

finvar_mappings::Dict{InfiniteOpt.GeneralVariableRef, JuMP.VariableRef}: Map finite variables to their transcription variables.

reduced_vars::Vector{InfiniteOpt.ReducedVariable{InfiniteOpt.GeneralVariableRef}}: Store the core reduced variable objects of reduced variables formed on transcription.

last_point_index::Int: The last internal point variable index added.

measure_lookup::Dict{InfiniteOpt.GeneralVariableRef, Dict{Vector{Float64}, Int}}: A lookup table of measure transcriptions via support value.

measure_mappings::Dict{InfiniteOpt.GeneralVariableRef, Vector{JuMP.AbstractJuMPScalar}}: Map measures to transcription expressions.

measure_supports::Dict{InfiniteOpt.GeneralVariableRef, Vector{Tuple}}: Map measures to their supports values (if the transcribed measure is still infinite).

measure_support_labels::Dict{InfiniteOpt.GeneralVariableRef, Vector{Set{DataType}}}: Map measures to their support labels if they have any.

constr_mappings::Dict{InfiniteOpt.InfOptConstraintRef, Vector{JuMP.ConstraintRef}}: Map constraints to their transcriptions.

constr_supports::Dict{InfiniteOpt.InfOptConstraintRef, Vector{Tuple}}: Map constraints to their support values.

constr_support_labels::Dict{InfiniteOpt.InfOptConstraintRef, Vector{Set{DataType}}}: Map constraints to their support labels.

supports::Tuple: Store the collected parameter supports here.

support_labels::Tuple: Store the collected parameter labels here.

has_internal_supports::Bool: Where any internal supports collected?

Methods

InfiniteOpt.add_measure_variable

InfiniteOpt.add_measure_variable

InfiniteOpt.build_optimizer_model!

InfiniteOpt.constraint_supports

InfiniteOpt.delete_reduced_variable

InfiniteOpt.expression_supports

InfiniteOpt.internal_reduced_variable

InfiniteOpt.optimizer_model_constraint

InfiniteOpt.optimizer_model_expression

InfiniteOpt.optimizer_model_variable

InfiniteOpt.variable_supports

InfiniteOpt.TranscriptionOpt.TranscriptionModel

InfiniteOpt.TranscriptionOpt.build_transcription_model!

InfiniteOpt.TranscriptionOpt.has_internal_supports

InfiniteOpt.TranscriptionOpt.index_to_labels

InfiniteOpt.TranscriptionOpt.index_to_support

InfiniteOpt.TranscriptionOpt.is_transcription_model

InfiniteOpt.TranscriptionOpt.lookup_by_support

InfiniteOpt.TranscriptionOpt.make_ndarray

InfiniteOpt.TranscriptionOpt.parameter_supports

InfiniteOpt.TranscriptionOpt.set_parameter_supports

InfiniteOpt.TranscriptionOpt.support_index_iterator

InfiniteOpt.TranscriptionOpt.transcribe_constraints!

InfiniteOpt.TranscriptionOpt.transcribe_derivative_evaluations!

InfiniteOpt.TranscriptionOpt.transcribe_derivative_variables!

InfiniteOpt.TranscriptionOpt.transcribe_hold_variables!

InfiniteOpt.TranscriptionOpt.transcribe_infinite_variables!

InfiniteOpt.TranscriptionOpt.transcribe_measures!

InfiniteOpt.TranscriptionOpt.transcribe_objective!

InfiniteOpt.TranscriptionOpt.transcribe_point_variables!

InfiniteOpt.TranscriptionOpt.transcribe_reduced_variables!

InfiniteOpt.TranscriptionOpt.transcription_constraint

InfiniteOpt.TranscriptionOpt.transcription_data

InfiniteOpt.TranscriptionOpt.transcription_expression

InfiniteOpt.TranscriptionOpt.transcription_expression

InfiniteOpt.TranscriptionOpt.transcription_model

InfiniteOpt.TranscriptionOpt.transcription_variable

InfiniteOpt.TranscriptionOpt.TranscriptionModel — Function

TranscriptionModel([optimizer_constructor;
                   caching_mode::MOIU.CachingOptimizerMode = MOIU.AUTOMATIC,
                   bridge_constraints::Bool = true])::JuMP.Model

Return a JuMP.Model with TranscriptionData included in the ext data field. Accepts the same arguments as a typical JuMP Model. More detailed variable and constraint naming can be enabled via verbose_naming.

Example

julia> TranscriptionModel()
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

InfiniteOpt.TranscriptionOpt.is_transcription_model — Function

is_transcription_model(model::JuMP.Model)::Bool

Return true if model is a TranscriptionModel or false otherwise.

Example

julia> is_transcription_model(model)
true

InfiniteOpt.TranscriptionOpt.transcription_data — Function

transcription_data(model::JuMP.Model)::TranscriptionData

Return the TranscriptionData from a TranscriptionModel. Errors if it is not a TranscriptionModel.

InfiniteOpt.TranscriptionOpt.has_internal_supports — Function

has_internal_supports(model::JuMP.Model)::Bool

Return a Bool whether model has any internal supports that were collected.

InfiniteOpt.TranscriptionOpt.transcription_model — Function

transcription_model(model::InfiniteOpt.InfiniteModel)::JuMP.Model

Return the transcription model stored in model if that is what is stored in model.optimizer_model.

InfiniteOpt.TranscriptionOpt.transcription_variable — Method

transcription_variable(model::JuMP.Model,
    vref::InfiniteOpt.GeneralVariableRef;
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
     ndarray::Bool = false])

Return the transcribed variable reference(s) corresponding to vref. Errors if no transcription variable is found. Also can query via the syntax:

transcription_variable(vref::InfiniteOpt.GeneralVariableRef; 
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
     ndarray::Bool = false])

If the infinite model contains a built transcription model. By default, this method returns only transcribed variables associated with public supports. All the variables can be returned by setting label = All.

If vref is infinite and ndarray = true then an n-dimensional array will be returned in accordance with the infinite parameters that have unique object numbers. In this case, label will be used to search the intersection of variable supports that use the label. This is defers from the default behavior which considers the union.

Example

julia> transcription_variable(trans_model, infvar)
2-element Array{VariableRef,1}:
 infvar(support: 1)
 infvar(support: 2)

julia> transcription_variable(trans_model, hdvar)
hdvar

julia> transcription_variable(infvar)
2-element Array{VariableRef,1}:
 infvar(support: 1)
 infvar(support: 2)

julia> transcription_variable(hdvar)
hdvar

InfiniteOpt.optimizer_model_variable — Method

InfiniteOpt.optimizer_model_variable(vref::InfiniteOpt.GeneralVariableRef,
    ::Val{:TransData};
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

Proper extension of InfiniteOpt.optimizer_model_variable for TranscriptionModels. This simply dispatches to transcription_variable.

InfiniteOpt.variable_supports — Method

InfiniteOpt.variable_supports(model::JuMP.Model,
    vref::InfiniteOpt.DecisionVariableRef,
    key::Val{:TransData} = Val(:TransData);
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

Return the support alias mapping associated with vref in the transcription model. Errors if vref does not have transcripted variables. See transcription_variable for an explanation of ndarray.

InfiniteOpt.TranscriptionOpt.lookup_by_support — Method

lookup_by_support(model::JuMP.Model,
                  vref::InfiniteOpt.GeneralVariableRef,
                  support::Vector)

Return the transcription expression of vref defined at its support. This is intended as a helper method for automated transcription.

InfiniteOpt.internal_reduced_variable — Method

InfiniteOpt.internal_reduced_variable(
    vref::InfiniteOpt.ReducedVariableRef,
    ::Val{:TransData}
    )::InfiniteOpt.ReducedVariable{InfiniteOpt.GeneralVariableRef}

Return the internal reduced variable associated with vref, assuming it was added internally during measure expansion at the transcription step. This extends InfiniteOpt.internal_reduced_variable as described in its docstring. Errors, if no such variable can be found.

InfiniteOpt.TranscriptionOpt.transcription_expression — Method

transcription_expression(model::JuMP.Model,
    expr::JuMP.AbstractJuMPScalar;
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

Return the transcribed expression(s) corresponding to expr. Errors if expr cannot be transcribed. Also can query via the syntax:

transcription_expression(expr::JuMP.AbstractJuMPScalar;
                         [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
                         ndarray::Bool = false])

If the infinite model contains a built transcription model. By default, this method returns only transcribed expressions associated with public supports. All the expressions can be returned by setting label = All.

If expr is infinite and ndarray = true then an n-dimensional array will be returned in accordance with the infinite parameters that have unique object numbers. In this case, label will be used to search the intersection of the supports that use the label. This is defers from the default behavior which considers the union.

Example

julia> transcription_expression(trans_model, my_expr)
x(support: 1) - y

julia> transcription_expression(my_expr)
x(support: 1) - y

InfiniteOpt.optimizer_model_expression — Method

InfiniteOpt.optimizer_model_expression(expr::JuMP.AbstractJuMPScalar,
    ::Val{:TransData};
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

Proper extension of InfiniteOpt.optimizer_model_expression for TranscriptionModels. This simply dispatches to transcription_expression.

InfiniteOpt.expression_supports — Method

InfiniteOpt.expression_supports(model::JuMP.Model,
    expr::JuMP.AbstractJuMPScalar,
    key::Val{:TransData} = Val(:TransData);
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

Return the support alias mappings associated with expr. Errors if expr cannot be transcribed.

InfiniteOpt.TranscriptionOpt.transcription_constraint — Method

transcription_constraint(model::JuMP.Model,
    cref::InfiniteOpt.InfOptConstraintRef;
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

Return the transcribed constraint reference(s) corresponding to cref. Errors if cref has not been transcribed. Also can query via the syntax:

transcription_constraint(cref::InfiniteOpt.InfOptConstraintRef;
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

If the infinite model contains a built transcription model. By default, this method returns only transcribed constraints associated with public supports. All the constraints can be returned by setting label = All.

If cref is infinite and ndarray = true then an n-dimensional array will be returned in accordance with the infinite parameters that have unique object numbers. In this case, label will be used to search the intersection of the supports that use the label. This is defers from the default behavior which considers the union.

Example

julia> transcription_constraint(trans_model, fin_con)
fin_con : x(support: 1) - y <= 3.0

julia> transcription_constraint(fin_con)
fin_con : x(support: 1) - y <= 3.0

InfiniteOpt.optimizer_model_constraint — Method

InfiniteOpt.optimizer_model_constraint(cref::InfiniteOpt.InfOptConstraintRef,
    ::Val{:TransData};
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel
    ndarray::Bool = false])

Proper extension of InfiniteOpt.optimizer_model_constraint for TranscriptionModels. This simply dispatches to transcription_constraint.

InfiniteOpt.constraint_supports — Method

InfiniteOpt.constraint_supports(model::JuMP.Model,
    cref::InfiniteOpt.InfOptConstraintRef,
    key::Val{:TransData} = Val(:TransData);
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

Return the support alias mappings associated with cref. Errors if cref is not transcribed.

InfiniteOpt.TranscriptionOpt.parameter_supports — Method

parameter_supports(model::JuMP.Model)::Tuple

Return the collected parameter support tuple that is stored in TranscriptionData.supports.

InfiniteOpt.TranscriptionOpt.support_index_iterator — Function

support_index_iterator(model::JuMP.Model, [obj_nums::Vector{Int}])::CartesianIndices

Return the CartesianIndices that determine the indices of the unique combinations of TranscriptionData.supports stored in model. If obj_nums is specified, then the indices will only include the tuple elements uses indices are included in the object numbers obj_nums and all others will be assigned the last index which should correspond to an appropriately sized placeholder comprised of NaNs. Note this method assumes that set_parameter_supports has already been called and that the last elements of each support vector contains a placeholder value.

InfiniteOpt.TranscriptionOpt.index_to_support — Function

index_to_support(model::JuMP.Model, index::CartesianIndex)::Vector{Float64}

Given a particular support index generated via support_index_iterator using model, return the corresponding support from TranscriptionData.supports using placeholder NaNs as appropriate for tuple elements that are unneeded.

InfiniteOpt.TranscriptionOpt.index_to_labels — Function

index_to_labels(model::JuMP.Model, index::CartesianIndex)::Set{DataType}

Given a particular support index generated via support_index_iterator using model, return the corresponding support label set from TranscriptionData.support_labels.

InfiniteOpt.TranscriptionOpt.make_ndarray — Function

make_narray(model::JuMP.Model, 
            ref::Union{JuMP.AbstractJuMPScalar, InfiniteOpt.InfOptConstraintRef}, 
            info::Vector, 
            label::Type{<:InfiniteOpt.AbstractSupportLabel})::Array

Take the resultsinfo associated with ref and rearrange them into an n-dimensional array where the axes correspond to the infinite parameter dependencies in accordance with their creation. Note that this works by querying the object numbers. Thus, independent infinite parameters will each get their own dimension (even if they are defined at the same time in an array) and each dependent infinite parameter group will have its own dimension.

InfiniteOpt.TranscriptionOpt.set_parameter_supports — Function

set_parameter_supports(trans_model::JuMP.Model,
                       inf_model::InfiniteOpt.InfiniteModel)::Nothing

Collect the infinite parameter supports stored in their respective dictionaries form inf_model and process them into a tuple of vectors where each vector contains the collected supports of a particular infinite parameter. These support collections are ordered in accordance with the definition order of the parameters (i.e., their object numbers). A support collection assocciated with an independent will be a Vector{Float64} and a support collection associated with a group of dependent parameters will be a Vector{Vector{Float64}}. Note that each collection vector will include an extra final placeholder element comprised of NaNs for convenience in generating support indices via support_index_iterator. This also gathers the associated support labels.

Before this is all done, InfiniteOpt.add_derivative_supports is invoked for any parameter that has at least one derivative dependency.

InfiniteOpt.TranscriptionOpt.transcribe_hold_variables! — Function

transcribe_hold_variables!(trans_model::JuMP.Model,
                           inf_model::InfiniteOpt.InfiniteModel)::Nothing

Create a transcription variable (i.e., a JuMP variable) for each HoldVariable stored in inf_model and add it to trans_model. The variable mapping is also stored in TranscriptionData.finvar_mappings which enables transcription_variable and lookup_by_support.

InfiniteOpt.TranscriptionOpt.transcribe_infinite_variables! — Function

transcribe_infinite_variables!(trans_model::JuMP.Model,
                               inf_model::InfiniteOpt.InfiniteModel)::Nothing

Create transcription variables (i.e., JuMP variables) for each InfiniteVariable stored in inf_model and add them to trans_model. The variable mappings are also stored in TranscriptionData.infvar_mappings in accordance with TranscriptionData.infvar_lookup which enable transcription_variable and lookup_by_support. Note that the supports will not be generated until InfiniteOpt.variable_supports is invoked via InfiniteOpt.supports. Note that TranscriptionData.infvar_support_labels is also populated.

InfiniteOpt.TranscriptionOpt.transcribe_derivative_variables! — Function

transcribe_derivative_variables!(trans_model::JuMP.Model,
                                 inf_model::InfiniteOpt.InfiniteModel)::Nothing

Create transcription variables (i.e., JuMP variables) for each Derivative stored in inf_model and add them to trans_model. The variable mappings are also stored in TranscriptionData.infvar_mappings in accordance with TranscriptionData.infvar_lookup which enable transcription_variable and lookup_by_support. Note that the supports will not be generated until InfiniteOpt.variable_supports is invoked via InfiniteOpt.supports. The futher derivative evaluation constraints are added when transcribe_derivative_evaluations! is invoked. Note that TranscriptionData.infvar_support_labels is also populated.

InfiniteOpt.TranscriptionOpt.transcribe_reduced_variables! — Function

transcribe_reduced_variables!(trans_model::JuMP.Model,
                              inf_model::InfiniteOpt.InfiniteModel)::Nothing

Map each ReducedVariable in inf_model to transcription variables stored in trans_model. The variable mappings are also stored in TranscriptionData.infvar_mappings in accordance with TranscriptionData.infvar_lookup which enable transcription_variable and lookup_by_support. Note that transcribe_infinite_variables! must be called first. Note that the supports will not be generated until InfiniteOpt.variable_supports is invoked via InfiniteOpt.supports. Note that TranscriptionData.infvar_support_labels is also populated.

InfiniteOpt.TranscriptionOpt.transcribe_point_variables! — Function

transcribe_point_variables!(trans_model::JuMP.Model,
                           inf_model::InfiniteOpt.InfiniteModel)::Nothing

Map each PointVariable in inf_model to a transcription variable stored in trans_model. The variable mapping is also stored in TranscriptionData.finvar_mappings which enables transcription_variable and lookup_by_support. Note that transcribe_infinite_variables! must be called first and that the info constraints associated with the transcription variable will be updated in accordance with the point variable.

InfiniteOpt.TranscriptionOpt.transcription_expression — Function

transcription_expression(model::JuMP.Model,
    expr::JuMP.AbstractJuMPScalar;
    [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
    ndarray::Bool = false])

Return the transcribed expression(s) corresponding to expr. Errors if expr cannot be transcribed. Also can query via the syntax:

transcription_expression(expr::JuMP.AbstractJuMPScalar;
                         [label::Type{<:InfiniteOpt.AbstractSupportLabel} = InfiniteOpt.PublicLabel,
                         ndarray::Bool = false])

If the infinite model contains a built transcription model. By default, this method returns only transcribed expressions associated with public supports. All the expressions can be returned by setting label = All.

If expr is infinite and ndarray = true then an n-dimensional array will be returned in accordance with the infinite parameters that have unique object numbers. In this case, label will be used to search the intersection of the supports that use the label. This is defers from the default behavior which considers the union.

Example

julia> transcription_expression(trans_model, my_expr)
x(support: 1) - y

julia> transcription_expression(my_expr)
x(support: 1) - y

transcription_expression(trans_model::JuMP.Model, expr, support::Vector{Float64})

Given the expr from an InfiniteModel, form its transcripted version in accordance with the variable mappings available in trans_model defined at support. This should only be used once all variables and measures have been transcribed (e.g., via transcribe_hold_variables!).

InfiniteOpt.TranscriptionOpt.transcribe_measures! — Function

transcribe_measures!(trans_model::JuMP.Model,
                     inf_model::InfiniteOpt.InfiniteModel)::Nothing

For each Measure in inf_model expand it via InfiniteOpt.expand_measure or analytic_expansion as appropriate and transcribe the expanded expression via transcription_expression. Then store the measure to transcripted expression mappings in TranscriptionData.measure_mappings and TranscriptionData.measure_lookup to enable transcription_variable and lookup_by_support. Note that the supports will not be generated until InfiniteOpt.variable_supports is invoked via InfiniteOpt.supports. Note that TranscriptionData.measure_support_labels is also populated.

InfiniteOpt.TranscriptionOpt.transcribe_objective! — Function

transcribe_objective!(trans_model::JuMP.Model,
                      inf_model::InfiniteOpt.InfiniteModel)::Nothing

Form the transcripted version of the objective stored in inf_model and add it to trans_model. Note that all the variables and measures in inf_model must by transcripted first (e.g., via transcribe_infinite_variables!).

InfiniteOpt.TranscriptionOpt.transcribe_constraints! — Function

transcribe_constraints!(trans_model::JuMP.Model,
                        inf_model::InfiniteOpt.InfiniteModel)::Nothing

For each constraint in inf_model form its transcripted version(s) and add them to trans_model. The mappings are stored in TranscriptionData.constr_mappings and the associated supports are stored in TranscriptionData.constr_supports to enable transcription_constraint and InfiniteOpt.constraint_supports. Note that variable info constraints are simply mapped to the existing info constraints already generated along with the transcription variables. Note that the variables and measures must all first be transcripted (e.g., via transcribe_measures!). Note that TranscriptionData.constr_support_labels is also populated.

InfiniteOpt.TranscriptionOpt.transcribe_derivative_evaluations! — Function

transcribe_derivative_evaluations!(trans_model::JuMP.Model, 
                                   inf_model::InfiniteOpt.InfiniteModel)::Nothing

Generate the auxiliary derivative evaluation equations and transcribe them appropriately for all the derivatives in inf_model. These are in turn added to trans_model. Note that no mapping information is recorded since the InfiniteModel won't have any constraints that correspond to these equations. Also Note that the variables and measures must all first be transcripted (e.g., via transcribe_derivative_variables!).

InfiniteOpt.TranscriptionOpt.build_transcription_model! — Function

build_transcription_model!(trans_model::JuMP.Model,
                           inf_model::InfiniteOpt.InfiniteModel)::Nothing

Given an empty trans_model build it using the information stored in inf_model. This is intended for a TranscriptionModel that serves as a internal optimizer model of inf_model. This detail is important to correctly enable internally generated reduced variables during the transcription process such that inf_model is not modified. Note that this will add supports to inf_model via InfiniteOpt.fill_in_supports! for infinite parameters that contain no supports. Also a warning is thrown when the transcription model contains more than 15,000 support points to alert users when they may naively have a few independent supports whose product quickly yields a very large grid. For example having 3 independent parameters with 100 supports each would result in 1,000,000 supports. This behavior can be overcome using dependent parameters.

InfiniteOpt.add_measure_variable — Method

InfiniteOpt.add_measure_variable(model::JuMP.Model,
                                 var::InfiniteOpt.PointVariable,
                                 key::Val{:TransData}
                                 )::InfiniteOpt.GeneralVariableRef

Make a PointVariableRef and map it to the appropriate transcription variable and return the GeneralVariableRef. This is an extension of add_measure_variable for TranscriptionOpt.

InfiniteOpt.add_measure_variable — Method

InfiniteOpt.add_measure_variable(model::JuMP.Model,
                                 var::InfiniteOpt.ReducedVariable,
                                 key::Val{:TransData}
                                 )::InfiniteOpt.GeneralVariableRef

Make a ReducedVariableRef and add var to the transcription data and return the GeneralVariableRef. This is an extension of add_measure_variable for TranscriptionOpt. Note that internal_reduced_variable is also extended to be able to access the var.

InfiniteOpt.delete_reduced_variable — Method

InfiniteOpt.delete_reduced_variable(model::JuMP.Model,
                                    vref::InfiniteOpt.ReducedVariableRef,
                                    key::Val{:TransData})::Nothing

This is an extension of delete_reduced_variable for use in TranscriptionOpt. Here we do not delete reduced variables once they have been used since there is no performance gain for this paradigm and the memory saving is small. Note this may change in the future.

InfiniteOpt.build_optimizer_model! — Method

InfiniteOpt.build_optimizer_model!(model::InfiniteOpt.InfiniteModel,
                                   key::Val{:TransData})::Nothing

Transcribe model and store it as a TranscriptionModel in the model.optimizer_model field which can be accessed with transcription_model. Ths clears the existing TranscriptionModel via InfiniteOpt.clear_optimizer_model_build! and then builds a new one using build_transcription_model!.