Variables

Variables

A guide and manual for the definition and use of variables in InfiniteOpt.

Overview

Decision variables are at the very core of InfiniteOpt as its name alludes to mathematical programs that entail infinite decision spaces (i.e., contain infinite decision variables). Principally, three variable types are employed: infinite, point, and hold. Infinite variables encompass any decision variable that is parameterized by an infinite parameter (e.g., space-time variables and recourse variables). Point variables are infinite variables at a particular infinite parameter value (point). Finally, hold variables are decisions that are made irrespective of the infinite domain (e.g., first stage variables and design variables). Or in other words they hold a particular over the infinite domain or some sub-domain of it.

Basic Usage

Infinite, point, and global variables are typically defined via their respective macros: @infinite_variable, @point_variable, and @hold_variable.

Infinite Variable Definition

Point Variable Definition

Global Variable Definition

Manipulation

Datatypes

InfiniteOpt.FiniteVariableRef

InfiniteOpt.GeneralVariableRef

InfiniteOpt.HoldVariable

InfiniteOpt.HoldVariableRef

InfiniteOpt.InfOptVariable

InfiniteOpt.InfiniteVariable

InfiniteOpt.InfiniteVariableRef

InfiniteOpt.MeasureFiniteVariableRef

InfiniteOpt.ParameterBounds

InfiniteOpt.PointVariable

InfiniteOpt.PointVariableRef

InfiniteOpt.InfOptVariable — Type.

InfOptVariable <: JuMP.AbstractVariable

An abstract type for infinite, point, and hold variables.

source

InfiniteOpt.InfiniteVariable — Type.

InfiniteVariable{S, T, U, V} <: InfOptVariable

A DataType for storing core infinite variable information. Note each element of the parameter reference tuple must contain either a single ParameterRef or an AbstractArray of ParameterRefs where each ParameterRef has the same group ID number.

Fields

info::JuMP.VariableInfo{S, T, U, V} JuMP variable information.

parameter_refs::Tuple The infinite parameters(s) that parameterize the variable.

source

InfiniteOpt.PointVariable — Type.

PointVariable{S, T, U, V} <: InfOptVariable

A DataType for storing point variable information. Note that the elements parameter_values field must match the format of the parameter reference tuple defined in InfiniteVariable

Fields

info::JuMP.VariableInfo{S, T, U, V} JuMP Variable information.

infinite_variable_ref::InfiniteVariableRef The infinite variable associated with the point variable.

parameter_values::Tuple The infinite parameter values defining the point.

source

InfiniteOpt.ParameterBounds — Type.

ParameterBounds

A DataType for storing intervaled bounds of parameters. This is used to define subdomains of HoldVariables and BoundedScalarConstraints.

Fields

intervals::Dict{ParameterRef, IntervalSet} A dictionary of parameter intervals

that are tighter than those already associated those paraticular parameters.

source

InfiniteOpt.HoldVariable — Type.

HoldVariable{S, T, U, V} <: InfOptVariable

A DataType for storing hold variable information.

Fields

info::JuMP.VariableInfo{S, T, U, V} JuMP variable information.

parameter_bounds::ParameterBounds Valid parameter sub-domains

source

InfiniteOpt.InfOptVariableRef — Constant.

InfOptVariableRef

A union type for infinite, point, and hold variable references.

source

InfiniteOpt.GeneralVariableRef — Type.

GeneralVariableRef <: JuMP.AbstractVariableRef

An abstract type to for variable references used with infinite models.

source

InfiniteOpt.MeasureFiniteVariableRef — Type.

MeasureFiniteVariableRef <: GeneralVariableRef

An abstract type to define finite variable and measure references.

source

InfiniteOpt.FiniteVariableRef — Type.

FiniteVariableRef <: GeneralVariableRef

An abstract type to define new finite variable references.

source

InfiniteOpt.InfiniteVariableRef — Type.

InfiniteVariableRef <: GeneralVariableRef

A DataType for untranscripted infinite dimensional variable references (e.g., second stage variables, time dependent variables).

Fields

model::InfiniteModel Infinite model.

index::Int Index of variable in model.

source

InfiniteOpt.PointVariableRef — Type.

PointVariableRef <: FiniteVariableRef

A DataType for variables defined at a transcipted point (e.g., second stage variable at a particular scenario, dynamic variable at a discretized time point).

Fields

model::InfiniteModel Infinite model.

index::Int Index of variable in model.

source

InfiniteOpt.HoldVariableRef — Type.

HoldVariableRef <: FiniteVariableRef

A DataType for finite fixed variable references (e.g., first stage variables, steady-state variables).

Fields

model::InfiniteModel Infinite model.

index::Int Index of variable in model.

source

Methods/Macros

InfiniteOpt.@add_parameter_bounds

InfiniteOpt.@hold_variable

InfiniteOpt.@infinite_variable

InfiniteOpt.@point_variable

InfiniteOpt.@set_parameter_bounds

InfiniteOpt.add_parameter_bound

InfiniteOpt.add_parameter_ref

InfiniteOpt.delete_parameter_bound

InfiniteOpt.delete_parameter_bounds

InfiniteOpt.has_parameter_bounds

InfiniteOpt.infinite_variable_ref

InfiniteOpt.is_used

InfiniteOpt.is_used

InfiniteOpt.parameter_bounds

InfiniteOpt.parameter_refs

InfiniteOpt.parameter_values

InfiniteOpt.set_parameter_bounds

InfiniteOpt.set_parameter_refs

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_objective

InfiniteOpt.used_by_point_variable

InfiniteOpt.used_by_reduced_variable

JuMP.BinaryRef

JuMP.FixRef

JuMP.IntegerRef

JuMP.LowerBoundRef

JuMP.UpperBoundRef

JuMP.add_variable

JuMP.all_variables

JuMP.build_variable

JuMP.delete

JuMP.delete_lower_bound

JuMP.delete_upper_bound

JuMP.fix

JuMP.fix_value

JuMP.has_lower_bound

JuMP.has_upper_bound

JuMP.index

JuMP.is_binary

JuMP.is_fixed

JuMP.is_integer

JuMP.is_valid

JuMP.lower_bound

JuMP.name

JuMP.num_variables

JuMP.owner_model

JuMP.set_binary

JuMP.set_integer

JuMP.set_lower_bound

JuMP.set_name

JuMP.set_name

JuMP.set_name

JuMP.set_start_value

JuMP.set_upper_bound

JuMP.start_value

JuMP.unfix

JuMP.unset_binary

JuMP.unset_integer

JuMP.upper_bound

JuMP.variable_by_name

JuMP.build_variable — Method.

JuMP.build_variable(_error::Function, info::JuMP.VariableInfo,
                    var_type::Symbol;
                    parameter_refs::Union{ParameterRef,
                                          AbstractArray{<:ParameterRef},
                                          Tuple, Nothing} = nothing,
                    infinite_variable_ref::Union{InfiniteVariableRef,
                                                 Nothing} = nothing,
                    parameter_values::Union{Number, AbstractArray{<:Number},
                                            Tuple, Nothing} = nothing,
                    parameter_bounds::Union{Dict{ParameterRef, IntervalSet},
                                            Nothing} = nothing,
                    extra_kw_args...)

Extend the JuMP.build_variable function to accomodate InfiniteOpt variable types. Returns the appropriate variable Datatype (i.e., InfiniteVariable, PointVariable, and HoldVariable). Primarily this method is to be used internally by the appropriate constructor macros @infinite_variable, @point_variable, and @hold_variable. However, it can be called manually to build InfiniteOpt variables. Errors if an unneeded keyword argument is given or if the keywoard arguments are formatted incorrectly (e.g., parameter_refs contains repeated parameter references when an infinite variable is defined). Also errors if needed kewword arguments are negated.

Examples

julia> @infinite_parameter(m, 0 <= t <= 1)
t

julia> info = VariableInfo(false, 0, false, 0, false, 0, false, 0, false, false);

julia> inf_var = build_variable(error, info, Infinite, parameter_refs = t)
InfiniteVariable{Int64,Int64,Int64,Int64}(VariableInfo{Int64,Int64,Int64,Int64}(false, 0, false, 0, false, 0, false, 0, false, false), (t,))

julia> ivref = add_variable(m, inf_var, "var_name")
var_name(t)

julia> pt_var = build_variable(error, info, Point, infinite_variable_ref = ivref,
                               parameter_values = 0.5)
PointVariable{Int64,Int64,Int64,Float64}(VariableInfo{Int64,Int64,Int64,Float64}(false, 0, false, 0, false, 0, true, 0.0, false, false), var_name(t), (0.5,))

julia> hd_var = build_variable(error, info, Hold)
HoldVariable{Int64,Int64,Int64,Int64}(VariableInfo{Int64,Int64,Int64,Int64}(false, 0, false, 0, false, 0, false, 0, false, false), Subdomain bounds (0): )

source

JuMP.add_variable — Method.

JuMP.add_variable(model::InfiniteModel, var::InfOptVariable, name::String = "")

Extend the JuMP.add_variable function to accomodate InfiniteOpt variable types. Adds a variable to an infinite model model and returns an appropriate variable reference (i.e., InfiniteVariableRef, PointVariableRef, or HoldVariableRef). Primarily intended to be an internal function of the constructor macros @infinite_variable, @point_variable, and @hold_variable. However, it can be used in combination with JuMP.build_variable to add variables to an infinite model object. Errors if invalid parameters reference(s) or an invalid infinite variable reference is included in var.

Examples

julia> @infinite_parameter(m, t in [0, 10]);

julia> info = VariableInfo(false, 0, false, 0, false, 0, false, 0, false, false);

julia> inf_var = build_variable(error, info, Infinite, parameter_refs = t);

julia> ivref = add_variable(m, inf_var, "var_name")
var_name(t)

julia> pt_var = build_variable(error, info, Point, infinite_variable_ref = ivref,
                               parameter_values = 0.5);

julia> pvref = add_variable(m, pt_var, "var_alias")
var_alias

julia> hd_var = build_variable(error, info, Hold);

julia> hvref = add_variable(m, hd_var, "var_name")
var_name

source

InfiniteOpt.@infinite_variable — Macro.

@infinte_variable(model, kw_args...)

Add an anonymous infinite variable to the model model described by the keyword arguments kw_args and returns the variable reference. Note that the parameter_refs keyword is required in this case.

@infinite_variable(model, varexpr, args..., kw_args...)

Add an infinite variable to model described by the expression var_expr, the positional arguments args and the keyword arguments kw_args. The expression varexpr can either be (note that in the following the symbol <= can be used instead of ≤ and the symbol >=can be used instead of ≥) of the form:

varexpr creating variables described by varexpr

varexpr ≤ ub (resp. varexpr ≥ lb) creating variables described by varexpr with upper bounds given by ub (resp. lower bounds given by lb)

varexpr == value creating variables described by varexpr with fixed values given by value

lb ≤ varexpr ≤ ub or ub ≥ varexpr ≥ lb creating variables described by varexpr with lower bounds given by lb and upper bounds given by ub

The expression varexpr can be of the form:

varname creating a scalar real variable of name varname

varname(params...) creating a scalar real variable of name varname with infinite parameters params... see parameter_refs for format.

varname[...] or [...] creating a container of variables.

varname[...](params...) or [...] creating a container of variables with infinite parameters params... in the first case.

The recognized positional arguments in args are the following:

Bin: Sets the variable to be binary, i.e. either 0 or 1.

Int: Sets the variable to be integer, i.e. one of ..., -2, -1, 0, 1, 2, ...

The recognized keyword arguments in kw_args are the following:

parameter_refs: This is mandatory if not specified in varexpr. Can be a single parameter reference, a single parameter array with parameters defined in the same call of @infinite_parameter (i.e., have same group ID), or a tuple where each element is either of the first two options listed.

base_name: Sets the name prefix used to generate variable names. It corresponds to the variable name for scalar variable, otherwise, the variable names are set to base_name[...] for each index ... of the axes axes. Furthermore, the parameter reference tuple is appended on the end of the name i.e., base_name(params...) or base_name[...](params...).

lower_bound: Sets the value of the variable lower bound.

upper_bound: Sets the value of the variable upper bound.

start: Sets the variable starting value used as initial guess in optimization.

binary: Sets whether the variable is binary or not.

integer: Sets whether the variable is integer or not.

container: Specify the container type.

Examples

julia> @infinite_parameter(model, 0 <= t <= 1)
t

julia> @infinite_parameter(model, w[1:2] in Normal())
2-element Array{ParameterRef,1}:
 w[1]
 w[2]

julia> @infinite_variable(model, x(t, w) >= 0)
x(t, w)

julia> x = @infinite_variable(model, parameter_refs = (t, w), base_name = "x",
                              lower_bound = 0)
x(t, w)

julia> @infinite_variable(model, lb[i] <= y[i = 1:2](t) <= ub[i], Int)
2-element Array{InfiniteVariableRef,1}:
 y[1](t)
 y[2](t)

source

InfiniteOpt.@point_variable — Macro.

@point_variable(model, kw_args...)

Add an anonymous point variable to the model model described by the keyword arguments kw_args and returns the variable reference. Note that the infinite_variable_ref and parameter_values keywords are required in this case.

@point_variable(model, infvarexpr, varexpr, args..., kw_args...)

Add a point variable to model described by the expression varexpr, the positional arguments args, and the keyword arguments kw_args and the infinite variable expr infvarexpr. The expression infvarexpr specifies the infinite variable this point variable corresponds to and the values at which the parameters are evaluated and must be of the form: infvar(param_values...) where the parameter values param_values... are listed in the same format as they are in teh definition of infvar. The expression varexpr is used to define variable specific bounds and whose name is used as an alias for the point variable which is simply the infinite variable evaluated at the values indicated. The expression varexpr can either be (note that in the following the symbol <= can be used instead of ≤ and the symbol >=can be used instead of ≥) of the form:

varexpr creating variables described by varexpr

varexpr ≤ ub (resp. varexpr ≥ lb) creating variables described by varexpr with upper bounds given by ub (resp. lower bounds given by lb)

varexpr == value creating variables described by varexpr with fixed values given by value

lb ≤ varexpr ≤ ub or ub ≥ varexpr ≥ lb creating variables described by varexpr with lower bounds given by lb and upper bounds given by ub

Note that be default a point variable inherits all of the same properties as the infinite variable it corresponds to, but that these can be overwritten by specifying properties such as lower bounds, fix values, etc.

The expression varexpr can be of the form:

varname creating a scalar real variable of alias name varname

varname[...] or [...] creating a container of variables.

The recognized positional arguments in args are the following:

Bin: Sets the variable to be binary, i.e. either 0 or 1.

Int: Sets the variable to be integer, i.e. one of ..., -2, -1, 0, 1, 2, ...

The recognized keyword arguments in kw_args are the following:

infinite_variable_ref: Sets the infinite variable reference that the point variable is associated with.

parameter_refs: Sets the values of the infinite parameters of the infinite variable at which this poitn variable is evaluated at. Must be of the same format of that specified for the parameters in the definition of the infinite variable.

base_name: Sets the name prefix used to generate variable names. It corresponds to the variable name for scalar variable, otherwise, the variable names are set to base_name[...] for each index ... of the axes axes. This serves as the alias for infvarexpr (the infinite variable evaluated at particular parameter values).

lower_bound: Sets the value of the variable lower bound.

upper_bound: Sets the value of the variable upper bound.

start: Sets the variable starting value used as initial guess in optimization.

binary: Sets whether the variable is binary or not.

integer: Sets whether the variable is integer or not.

container: Specify the container type.

Examples

julia> @infinite_parameter(model, 0 <= t <= 1)
t

julia> @infinite_parameter(model, w[1:2] in Normal())
2-element Array{ParameterRef,1}:
 w[1]
 w[2]

julia> @infinite_variable(model, x(t, w) >= 0)
x(t, w)

julia> @point_variable(model, x(0, [0, 0]), x0 <= 1)
x0

julia> x0 = @point_variable(model, x(0, [0, 0]), upper_bound = 1, base_name = "x0")
x0

julia> x0 = @point_variable(model, upper_bound = 1, base_name = "x0",
                            infinite_variable_ref = x, parameter_values = (0, [0, 0]))
x0

julia> @point_variable(model, x([0, 1][i], [0, 0]), xf[i = 1:2])
2-element Array{PointVariableRef,1}:
 xf[1]
 xf[2]

julia> @infinite_variable(model, lb[i] <= y[i = 1:2](t) <= ub[i], Int)
2-element Array{InfiniteVariableRef,1}:
y[1](t)
y[2](t)

julia> @point_variable(model, y[i](0), y0[i = 1:2], Bin)
2-element Array{PointVariableRef,1}:
 y0[1]
 y0[2]

 julia> y0 = @point_variable(model, [i = 1:2], binary = true, base_name = "y0",
                             infinite_variable_ref = y[i], parameter_values = 0)
 2-element Array{PointVariableRef,1}:
  y0[1]
  y0[2]

source

InfiniteOpt.@hold_variable — Macro.

@hold_variable(model, kw_args...)

Add an anonymous hold variable to the model model described by the keyword arguments kw_args and returns the variable reference.

@hold_variable(model, varexpr, args..., kw_args...)

Add a hold variable to model described by the expression varexpr, the positional arguments args and the keyword arguments kw_args. The expression varexpr can either be (note that in the following the symbol <= can be used instead of ≤ and the symbol >=can be used instead of ≥) of the form:

varexpr creating variables described by varexpr

varexpr ≤ ub (resp. varexpr ≥ lb) creating variables described by varexpr with upper bounds given by ub (resp. lower bounds given by lb)

varexpr == value creating variables described by varexpr with fixed values given by value

lb ≤ varexpr ≤ ub or ub ≥ varexpr ≥ lb creating variables described by varexpr with lower bounds given by lb and upper bounds given by ub

The expression varexpr can be of the form:

varname creating a scalar real variable of name varname

varname[...] or [...] creating a container of variables.

The recognized positional arguments in args are the following:

Bin: Sets the variable to be binary, i.e. either 0 or 1.

Int: Sets the variable to be integer, i.e. one of ..., -2, -1, 0, 1, 2, ...

Specifiying a hold variable which applies only to sub-domain of the model's infinite parameter(s) domain can be done via the parameter_bounds keyword argument. It is specified as a tuple of parameter bound expressions which can be of the form:

(param in [lb, ub], ...) enforcing param to be in a sub-domain from lb to ub (note ∈ can be used in place of in)

(params in [lb, ub], ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(lb <= param <= ub, ...) enforcing param to be in a sub-domain from lb to ub

(lb <= params <= ub, ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(param == value, ...) enforcing param to be equal to value

(params == value, ...) enforcing that all parameter references in params each to be equal to value

Any combination of the above forms. Must be inside parentheses and comma separated.

The other recognized keyword arguments in kw_args are the following:

base_name: Sets the name prefix used to generate variable names. It corresponds to the variable name for scalar variable, otherwise, the variable names are set to base_name[...] for each index ... of the axes axes.

lower_bound: Sets the value of the variable lower bound.

upper_bound: Sets the value of the variable upper bound.

start: Sets the variable starting value used as initial guess in optimization.

binary: Sets whether the variable is binary or not.

integer: Sets whether the variable is integer or not.

container: Specify the container type.

Examples

julia> @hold_variable(model, x)
x

julia> @hold_variable(model, 0 <= y <= 4, Bin)
y

julia> y = @hold_variable(model, lower_bound = 0, upper_bound = 4,
                            binary = true, base_name = "y")
y

julia> @hold_variable(model, z[2:3] == 0)
1-dimensional DenseAxisArray{HoldVariableRef,1,...} with index sets:
    Dimension 1, 2:3
And data, a 2-element Array{HoldVariableRef,1}:
 z[2]
 z[3]

julia> @hold_variable(model, d, parameter_bounds = (t in [0, 5]))
d

source

JuMP.owner_model — Method.

JuMP.owner_model(vref::GeneralVariableRef)::InfiniteModel

Extend JuMP.owner_model function for InfiniteOpt variables. Returns the infinite model associated with vref.

Example

julia> owner_model(vref)
An InfiniteOpt Model
Feasibility problem with:
Variable: 1
`HoldVariableRef`-in-`MathOptInterface.LessThan{Float64}`: 1 constraint
`HoldVariableRef`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
`HoldVariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 1 constraint
Names registered in the model: vref
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

source

JuMP.index — Method.

JuMP.index(v::GeneralVariableRef)::Int

Extent JuMP.index to return the index of a InfiniteOpt variable.

Example

julia> index(vref)
1

source

InfiniteOpt.used_by_constraint — Method.

used_by_constraint(vref::InfOptVariableRef)::Bool

Return a Bool indicating if vref is used by a constraint.

Example

julia> used_by_constraint(vref)
false

source

InfiniteOpt.used_by_measure — Method.

used_by_measure(vref::InfOptVariableRef)::Bool

Return a Bool indicating if vref is used by a measure.

Example

julia> used_by_measure(vref)
true

source

InfiniteOpt.used_by_objective — Method.

used_by_objective(vref::InfOptVariableRef)::Bool

Return a Bool indicating if vref is used by the objective.

Example

julia> used_by_objective(vref)
true

source

InfiniteOpt.is_used — Method.

is_used(vref::InfOptVariableRef)::Bool

Return a Bool indicating if vref is used in the model.

Example

julia> is_used(vref)
true

source

InfiniteOpt.used_by_point_variable — Method.

used_by_point_variable(vref::InfiniteVariableRef)::Bool

Return a Bool indicating if vref is used by a point variable.

Example

julia> used_by_point_variable(vref)
false

source

InfiniteOpt.used_by_reduced_variable — Method.

used_by_reduced_variable(vref::InfiniteVariableRef)::Bool

Return a Bool indicating if vref is used by a reduced infinite variable.

Example

julia> used_by_reduced_variable(vref)
true

source

InfiniteOpt.is_used — Method.

is_used(vref::InfiniteVariableRef)::Bool

Return a Bool indicating if vref is used in the model.

Example

julia> is_used(vref)
false

source

JuMP.delete — Method.

JuMP.delete(model::InfiniteModel, vref::InfOptVariableRef)

Extend JuMP.delete to delete InfiniteOpt variables and their dependencies. Errors if variable is invalid, meaning it has already been deleted or it belongs to another model.

Example

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 z >= 0.0
 g(t) + z >= 42.0
 g(0.5) == 0
 t in [0, 6]

julia> delete(model, g)

julia> print(model)
Min measure(t) + z
Subject to
 z >= 0.0
 z >= 42.0
 t in [0, 6]

source

JuMP.is_valid — Method.

JuMP.is_valid(model::InfiniteModel, vref::InfOptVariableRef)::Bool

Extend JuMP.is_valid to accomodate InfiniteOpt variables.

Example

julia> is_valid(model, ivref)
true

source

JuMP.num_variables — Method.

JuMP.num_variables(model::InfiniteModel)::Int

Extend JuMP.num_variables to return the number of InfiniteOpt variables assigned to model.

Example

julia> num_variables(model)
3

source

JuMP.has_lower_bound — Method.

JuMP.has_lower_bound(vref::InfOptVariableRef)::Bool

Extend JuMP.has_lower_bound to return a Bool whether an InfiniteOpt variable has a lower bound.

Example

julia> has_lower_bound(vref)
true

source

JuMP.lower_bound — Method.

JuMP.lower_bound(vref::InfOptVariableRef)::Float64

Extend JuMP.lower_bound to return the lower bound of an InfiniteOpt variable. Errors if vref doesn't have a lower bound.

Example

julia> lower_bound(vref)
0.0

source

JuMP.set_lower_bound — Method.

JuMP.set_lower_bound(vref::InfOptVariableRef, lower::Number)

Extend JuMP.set_lower_bound to specify the lower bound of an InfiniteOpt variable vref. Errors if vref is fixed.

Example

julia> set_lower_bound(vref, -1)

julia> lower_bound(vref)
-1.0

source

JuMP.LowerBoundRef — Method.

JuMP.LowerBoundRef(vref::InfOptVariableRef)::GeneralConstraintRef

Extend JuMP.LowerBoundRef to extract a constraint reference for the lower bound of vref.

Example

julia> cref = LowerBoundRef(vref)
var >= 0.0

source

JuMP.delete_lower_bound — Method.

JuMP.delete_lower_bound(vref::InfOptVariableRef)

Extend JuMP.delete_lower_bound to delete lower bound of vref. Errors if it doesn't have a lower bound.

Example

julia> delete_lower_bound(vref)

julia> has_lower_bound(vref)
false

source

JuMP.has_upper_bound — Method.

JuMP.has_upper_bound(vref::InfOptVariableRef)::Bool

Extend JuMP.has_upper_bound to return a Bool whether an InfiniteOpt variable has an upper bound.

Example

julia> has_upper_bound(vref)
true

source

JuMP.upper_bound — Method.

JuMP.upper_bound(vref::InfOptVariableRef)::Float64

Extend JuMP.upper_bound to return the upper bound of an InfiniteOpt variable. Errors if vref doesn't have a upper bound.

Example

julia> upper_bound(vref)
0.0

source

JuMP.set_upper_bound — Method.

JuMP.set_upper_bound(vref::InfOptVariableRef, upper::Number)

Extend JuMP.set_upper_bound to specify the upper bound of an InfiniteOpt variable vref. Errors if vref is fixed.

Example

julia> set_upper_bound(vref, 1)

julia> upper_bound(vref)
1.0

source

JuMP.UpperBoundRef — Method.

JuMP.UpperBoundRef(vref::InfOptVariableRef)::GeneralConstraintRef

Extend JuMP.UpperBoundRef to extract a constraint reference for the upper bound of vref.

Example

julia> cref = UpperBoundRef(vref)
var <= 1.0

source

JuMP.delete_upper_bound — Method.

JuMP.delete_upper_bound(vref::InfOptVariableRef)

Extend JuMP.delete_upper_bound to delete the upper bound of vref. Errors if it doesn't have an upper bound.

Example

julia> delete_upper_bound(vref)

julia> has_upper_bound(vref)
false

source

JuMP.is_fixed — Method.

JuMP.is_fixed(vref::InfOptVariableRef)::Bool

Extend JuMP.is_fixed to return Bool whether an InfiniteOpt variable is fixed.

Example

julia> is_fixed(vref)
true

source

JuMP.fix_value — Method.

JuMP.fix_value(vref::InfOptVariableRef)::Float64

Extend JuMP.fix_value to return the fix value of an InfiniteOpt variable. Errors if variable is not fixed.

Example

julia> fix_value(vref)
0.0

source

JuMP.fix — Method.

JuMP.fix(vref::InfOptVariableRef, value::Number; force::Bool = false)

Extend JuMP.fix to fix the value of an InfiniteOpt variable. Errors if variable has lower/upper bound unless force = true.

Examples

julia> fix(vref, 3)

julia> fix_value(vref)
3.0

julia> fix(vref2, 2, force = true)

julia> fix_value(vref2)
2.0

source

JuMP.FixRef — Method.

JuMP.FixRef(vref::InfOptVariableRef)::GeneralConstraintRef

Extend JuMP.FixRef to return the constraint reference of the fix constraint associated with vref. Errors vref is not fixed.

Examples

julia> cref = FixRef(vref)
var == 1.0

source

JuMP.unfix — Method.

JuMP.unfix(vref::InfOptVariableRef)

Extend JuMP.unfix to unfix vref. Errors if it is not fixed.

Example

julia> unfix(vref)

julia> is_fixed(vref)
false

source

JuMP.start_value — Method.

JuMP.start_value(vref::InfOptVariableRef)::Union{Nothing, Float64}

Extend JuMP.start_value to return starting value of InfiniteOpt variable if it has one. Returns nothing otherwise.

Example

julia> start_value(vref)
0.0

source

JuMP.set_start_value — Method.

JuMP.set_start_value(vref::InfOptVariableRef, value::Number)

Extend JuMP.set_start_value to specify the start value of InfiniteOpt variables.

Example

julia> set_start_value(vref, 1)

julia> start_value(vref)
1.0

source

JuMP.is_binary — Method.

JuMP.is_binary(vref::InfOptVariableRef)::Bool

Extend JuMP.is_binary to return Bool whether an InfiniteOpt variable is binary.

Example

julia> is_binary(vref)
true

source

JuMP.set_binary — Method.

JuMP.set_binary(vref::InfOptVariableRef)

Extend JuMP.set_binary to specify an InfiniteOpt variable as a binary variable. Errors if vref is an integer variable.

Example

julia> set_binary(vref)

julia> is_binary(vref)
true

source

JuMP.BinaryRef — Method.

JuMP.BinaryRef(vref::InfOptVariableRef)::GeneralConstraintRef

Extend JuMP.BinaryRef to return a constraint reference to the constraint constrainting vref to be binary. Errors if one does not exist.

Example

julia> cref = BinaryRef(vref)
var binary

source

JuMP.unset_binary — Method.

JuMP.unset_binary(vref::InfOptVariableRef)

Extend JuMP.unset_binary to unset vref as a binary variable. Errors if it is not binary.

julia> unset_binary(vref)

julia> is_binary(vref)
false

source

JuMP.is_integer — Method.

JuMP.is_integer(vref::InfOptVariableRef)::Bool

Extend JuMP.is_integer to return Bool whether an InfiniteOpt variable is integer.

Example

julia> is_integer(vref)
true

source

JuMP.set_integer — Method.

JuMP.set_integer(vref::InfOptVariableRef)

Extend JuMP.set_integer to specify an InfiniteOpt variable as a integer variable. Errors if vref is an binary variable.

Example

julia> set_integery(vref)

julia> is_integer(vref)
true

source

JuMP.IntegerRef — Method.

JuMP.IntegerRef(vref::InfOptVariableRef)::GeneralConstraintRef

Extend JuMP.IntegerRef to return a constraint reference to the constraint constrainting vref to be integer. Errors if one does not exist.

Example

julia> cref = IntegerRef(vref)
var integer

source

JuMP.unset_integer — Method.

JuMP.unset_integer(vref::InfOptVariableRef)

Extend JuMP.unset_integer to unset vref as an integer variable. Errors if it is not an integer variable.

julia> unset_integer(vref)

julia> is_integer(vref)
false

source

JuMP.name — Method.

JuMP.name(vref::InfOptVariableRef)::String

Extend JuMP.name to return the names of InfiniteOpt variables.

Example

julia> name(vref)
"var_name"

source

JuMP.set_name — Method.

JuMP.set_name(vref::InfiniteVariableRef, root_name::String)

Extend JuMP.set_name to set names of infinite variables. Adds on to root_name the ending (prefs...) where the parameter reference names are listed in the same format as input in the parameter reference tuple.

Example

julia> name(vref)
old_name(t, x)

julia> set_name(vref, "new_name")

julia> name(vref)
new_name(t, x)

source

JuMP.set_name — Method.

JuMP.set_name(vref::PointVariableRef, name::String)

Extend JuMP.set_name to set the names of point variables.

Example

julia> name(vref)
old_name

julia> set_name(vref, "new_name")

julia> name(vref)
new_name

source

JuMP.set_name — Method.

JuMP.set_name(vref::HoldVariableRef, name::String)

Extend JuMP.set_name to set names of hold variables.

Example

julia> set_name(hvref, "var_name")

julia> name(t)
"var_name"

source

InfiniteOpt.parameter_refs — Method.

parameter_refs(vref::InfiniteVariableRef)::Tuple

Return the ParameterRef(s) associated with the infinite variable vref. This is formatted as a Tuple of containing the parameter references as they inputted to define vref.

Example

julia> parameter_refs(vref)
(t,   [2]  =  x[2]
  [1]  =  x[1])

source

InfiniteOpt.set_parameter_refs — Method.

set_parameter_refs(vref::InfiniteVariableRef, prefs::Tuple)

Specify a new parameter reference tuple prefs for the infinite variable vref. Note each element must contain a single parameter reference or an array of parameter references. Errors if a parameter is double specified or if an element contains parameters with different group IDs.

Example

julia> set_parameter_refs(vref, (t, x))

julia> parameter_refs(vref)
(t,   [2]  =  x[2]
  [1]  =  x[1])

source

InfiniteOpt.add_parameter_ref — Method.

add_parameter_ref(vref::InfiniteVariableRef,
                  pref::Union{ParameterRef, AbstractArray{<:ParameterRef}})

Add additional parameter reference or group of parameter references to be associated with the infinite variable vref. Errors if the parameter references are already added to the variable or if the added parameters have different group IDs.

julia> name(vref)
T(t)

julia> add_parameter_ref(vref, x)

julia> name(vref)
T(t, x)

source

InfiniteOpt.@set_parameter_bounds — Macro.

@set_parameter_bounds(ref, bound_expr; [force = false])

Specify new parameter bounds for a constraint reference or hold variable reference ref. These bounds correspond to bounding a constraint in an equivalent way to using @BDconstraint or to limiting the scope of a hold variable in an equivalent way to using the parameter_bounds keyword argument in @hold_variable. Here (bound_expr) can be of the form:

(param in [lb, ub], ...) enforcing param to be in a sub-domain from lb to ub (note ∈ can be used in place of in)

(params in [lb, ub], ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(lb <= param <= ub, ...) enforcing param to be in a sub-domain from lb to ub

(lb <= params <= ub, ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(param == value, ...) enforcing param to be equal to value

(params == value, ...) enforcing that all parameter references in params each to be equal to value

Any combination of the above forms. Must be inside parentheses and comma separated.

Errors if the constraint or variable corresponding to ref already has bounds. However, using force = true can be used ignore the current bounds and overwrite them with new ones. Also, note that bounds on dependent constraints of hold variables will be updated to account for changes in hold variable bounds.

Examples

julia> @infinite_parameter(model, t in [0, 10])
t

julia> @infinite_variable(model, x(t))
x(t)

julia> @hold_variable(model, y)
y

julia> @constraint(model, con, x + y == 0)
con : x(t) + y = 0.0

julia> @set_parameter_bounds(y, t in [0, 5])

julia> con
con : x(t) + y = 0.0, ∀ t ∈ [0, 5]

julia> @set_parameter_bounds(con, t == 0, force = true)

julia> con
con : x(t) + y = 0.0, ∀ t = 0

source

InfiniteOpt.@add_parameter_bounds — Macro.

@add_parameter_bounds(ref, bound_expr)

Add new parameter bounds for a constraint reference or hold variable reference ref. These bounds correspond to bounding a constraint in an equivalent way to using @BDconstraint or to limiting the scope of a hold variable in an equivalent way to using the parameter_bounds keyword argument in @hold_variable. Here (bound_expr) can be of the form:

(param in [lb, ub], ...) enforcing param to be in a sub-domain from lb to ub (note ∈ can be used in place of in)

(params in [lb, ub], ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(lb <= param <= ub, ...) enforcing param to be in a sub-domain from lb to ub

(lb <= params <= ub, ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(param == value, ...) enforcing param to be equal to value

(params == value, ...) enforcing that all parameter references in params each to be equal to value

Any combination of the above forms. Must be inside parentheses and comma separated.

Errors if the new bounds cause irreconcilable differences with existing measures and constraints. For example, this occurs when adding hold variable bounds that are outside the domain of a bounded constraint that uses that hold variable. Also, note that bounds on dependent constraints of hold variables will be updated to account for changes in hold variable bounds.

Examples

julia> @infinite_parameter(model, t in [0, 10])
t

julia> @infinite_parameter(model, q in [-2, 2])
q

julia> @infinite_variable(model, x(t, q))
x(t, q)

julia> @hold_variable(model, y)
y

julia> @constraint(model, con, x + y == 0)
con : x(t, q) + y = 0.0

julia> @add_parameter_bounds(y, t in [0, 5])

julia> con
con : x(t, q) + y = 0.0, ∀ t ∈ [0, 5]

julia> @add_parameter_bounds(con, q == 0)

julia> con
con : x(t, q) + y = 0.0, ∀ t ∈ [0, 5], q = 0

source

InfiniteOpt.has_parameter_bounds — Method.

has_parameter_bounds(vref::HoldVariableRef)::Bool

Return a Bool indicating if vref is limited to a sub-domain as defined by parameter bound.

Example

julia> @infinite_parameter(model, t in [0, 10])
t

julia> @hold_variable(model, vref, parameter_bounds = (t in [0, 2]))
vref

julia> has_parameter_bounds(vref)
true

source

InfiniteOpt.parameter_bounds — Method.

parameter_bounds(vref::HoldVariableRef)::ParameterBounds

Return the ParameterBounds object associated with the hold variable vref. It contains a dictionary where each key is a ParameterRef which points to an IntervalSet that that defines a sub-domain for vref relative to that parameter reference.

Example

julia> @infinite_parameter(model, t in [0, 10])
t

julia> @hold_variable(model, vref, parameter_bounds = (t in [0, 2]))
vref

julia> parameter_bounds(vref)
Subdomain bounds (1): t ∈ [0, 2]

source

InfiniteOpt.set_parameter_bounds — Method.

set_parameter_bounds(vref::HoldVariableRef, bounds::ParameterBounds;
                     [force = false])

Specify a new dictionary of parameter bounds bounds for the hold variable vref. These are stored in a ParameterBounds object which contains a dictionary. Note the dictionary keys must be ParameterRefs and the values must be IntervalSets that indicate a particular sub-domain for which vref is defined. This is meant to be primarily used by @set_parameter_bounds which provides a more intuitive syntax.

Example

julia> @infinite_parameter(model, t in [0, 10])
t

julia> @hold_variable(model, vref)
vref

julia> set_parameter_bounds(vref, ParameterBounds(Dict(t => IntervalSet(0, 2))))

julia> parameter_bounds(vref)
Subdomain bounds (1): t ∈ [0, 2]

source

InfiniteOpt.add_parameter_bound — Method.

add_parameter_bound(vref::HoldVariableRef, pref::ParameterRef,
                    lower::Number, upper::Number)

Add an additional parameter bound to vref such that it is defined over the sub-domain based on pref from lower to upper. This is primarily meant to be used by @add_parameter_bounds.

julia> @infinite_parameter(model, t in [0, 10])
t

julia> @hold_variable(model, vref)
vref

julia> add_parameter_bound(vref, t, 0, 2)

julia> parameter_bounds(vref)
Subdomain bounds (1): t ∈ [0, 2]

source

InfiniteOpt.delete_parameter_bound — Method.

delete_parameter_bound(vref::HoldVariableRef, pref::ParameterRef)

Delete the parameter bound of the hold variable vref associated with the infinite parameter pref if vref has such a bound. Note that any other parameter bounds will be unaffected. Any constraints that employ vref will be updated accordingly.

Example

julia> @infinite_parameter(model, x[1:2] in [0, 10])
2-element Array{ParameterRef,1}:
 x[1]
 x[2]

julia> @hold_variable(model, z, parameter_bounds = (x in [0, 1]))
z

julia> delete_parameter_bound(z, x[2])

julia> parameter_bounds(z)
Subdomain bounds (1): x[1] ∈ [0, 1]

source

InfiniteOpt.delete_parameter_bounds — Method.

delete_parameter_bounds(vref::HoldVariableRef)

Delete all the parameter bounds of the hold variable vref. Any constraints that employ vref will be updated accordingly.

Example

julia> @infinite_parameter(model, x[1:2] in [0, 10])
2-element Array{ParameterRef,1}:
 x[1]
 x[2]

julia> @hold_variable(model, z, parameter_bounds = (x in [0, 1]))
z

julia> delete_parameter_bounds(z)

julia> parameter_bounds(z)
Subdomain bounds (0):

source

InfiniteOpt.infinite_variable_ref — Method.

infinite_variable_ref(vref::PointVariableRef)::InfiniteVariableRef

Return the InfiniteVariableRef associated with the point variable vref.

Example

julia> infinite_variable_ref(vref)
T(t, x)

source

InfiniteOpt.parameter_values — Method.

parameter_values(vref::PointVariableRef)::Tuple

Return the support point associated with the point variable vref.

Example

julia> parameter_values(vref)
(0, )

source

JuMP.variable_by_name — Method.

JuMP.variable_by_name(model::InfiniteModel,
                      name::String)::Union{GeneralVariableRef, Nothing}

Extend JuMP.variable_by_name for InfiniteModel objects. Return the varaible reference assoociated with a variable name. Errors if multiple variables have the same name. Returns nothing if no such name exists.

Examples

julia> variable_by_name(m, "var_name")
var_name

julia> variable_by_name(m, "fake_name")

source

JuMP.all_variables — Method.

JuMP.all_variables(model::InfiniteModel)::Vector{GeneralVariableRef}

Extend JuMP.all_variables to return a list of all the variable references associated with model.

Examples

julia> all_variables(m)
4-element Array{GeneralVariableRef,1}:
 ivar(test, θ)
 ivar2(test, x)
 name
 z

source