Library

Method, DataType, and Macro Index

InfiniteOpt.AbstractInfiniteSet

InfiniteOpt.DistributionSet

InfiniteOpt.FiniteVariableRef

InfiniteOpt.GeneralVariableRef

InfiniteOpt.GlobalVariable

InfiniteOpt.GlobalVariableRef

InfiniteOpt.InfOptParameter

InfiniteOpt.InfOptVariable

InfiniteOpt.InfiniteModel

InfiniteOpt.InfiniteModel

InfiniteOpt.InfiniteVariable

InfiniteOpt.InfiniteVariableRef

InfiniteOpt.IntervalSet

InfiniteOpt.IntervalSet

InfiniteOpt.MeasureFiniteVariableRef

InfiniteOpt.ParameterRef

InfiniteOpt.PointVariable

InfiniteOpt.PointVariableRef

JuMP.Model

JuMP.Model

JuMP.Model

InfiniteOpt.add_parameter

InfiniteOpt.add_parameter_ref

InfiniteOpt.add_supports

InfiniteOpt.all_parameters

InfiniteOpt.build_parameter

InfiniteOpt.delete_supports

InfiniteOpt.group_id

InfiniteOpt.group_id

InfiniteOpt.has_supports

InfiniteOpt.infinite_set

InfiniteOpt.infinite_variable_ref

InfiniteOpt.is_independent

InfiniteOpt.is_used

InfiniteOpt.is_used

InfiniteOpt.is_used

InfiniteOpt.num_parameters

InfiniteOpt.num_supports

InfiniteOpt.parameter_by_name

InfiniteOpt.parameter_refs

InfiniteOpt.parameter_values

InfiniteOpt.set_independent

InfiniteOpt.set_infinite_set

InfiniteOpt.set_parameter_refs

InfiniteOpt.set_supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.unset_independent

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_objective

InfiniteOpt.used_by_point_variable

InfiniteOpt.used_by_reduced_variable

InfiniteOpt.used_by_variable

JuMP.BinaryRef

JuMP.FixRef

JuMP.IntegerRef

JuMP.LowerBoundRef

JuMP.UpperBoundRef

JuMP.add_variable

JuMP.all_variables

JuMP.build_variable

JuMP.delete

JuMP.delete

JuMP.delete_lower_bound

JuMP.delete_upper_bound

JuMP.fix

JuMP.fix_value

JuMP.has_lower_bound

JuMP.has_lower_bound

JuMP.has_upper_bound

JuMP.has_upper_bound

JuMP.index

JuMP.is_binary

JuMP.is_fixed

JuMP.is_integer

JuMP.is_valid

JuMP.is_valid

JuMP.lower_bound

JuMP.lower_bound

JuMP.name

JuMP.name

JuMP.num_variables

JuMP.owner_model

JuMP.set_binary

JuMP.set_integer

JuMP.set_lower_bound

JuMP.set_lower_bound

JuMP.set_name

JuMP.set_name

JuMP.set_name

JuMP.set_name

JuMP.set_start_value

JuMP.set_upper_bound

JuMP.set_upper_bound

JuMP.start_value

JuMP.unfix

JuMP.unset_binary

JuMP.unset_integer

JuMP.upper_bound

JuMP.upper_bound

JuMP.variable_by_name

JuMP.with_optimizer

InfiniteOpt.@global_variable

InfiniteOpt.@infinite_parameter

InfiniteOpt.@infinite_variable

InfiniteOpt.@point_variable

JuMP Documentation

JuMP.Model — Type.

Model

A mathematical model of an optimization problem.

JuMP.Model — Method.

Model(; caching_mode::MOIU.CachingOptimizerMode=MOIU.AUTOMATIC,
        bridge_constraints::Bool=true)

Return a new JuMP model without any optimizer; the model is stored the model in a cache. The mode of the CachingOptimizer storing this cache is caching_mode. The optimizer can be set later in the optimize! call. If bridge_constraints is true, constraints that are not supported by the optimizer are automatically bridged to equivalent supported constraints when an appropriate transformation is defined in the MathOptInterface.Bridges module or is defined in another module and is explicitely added.

JuMP.Model — Method.

Model(optimizer_factory::OptimizerFactory;
      caching_mode::MOIU.CachingOptimizerMode=MOIU.AUTOMATIC,
      bridge_constraints::Bool=true)

Return a new JuMP model using the optimizer factory optimizer_factory to create the optimizer. The optimizer factory can be created by the with_optimizer function.

Examples

The following creates a model using the optimizer Ipopt.Optimizer(print_level=0):

model = Model(with_optimizer(Ipopt.Optimizer, print_level=0))

JuMP.with_optimizer — Function.

with_optimizer(constructor, args...; kwargs...)

Return an OptimizerFactory that creates optimizers using the constructor constructor with positional arguments args and keyword arguments kwargs.

Examples

The following returns an optimizer factory that creates Ipopt.Optimizers using the constructor call Ipopt.Optimizer(print_level=0):

with_optimizer(Ipopt.Optimizer, print_level=0)