Constraints

A guide and manual for defining and manipulating constraints in InfiniteOpt.

Overview

Basic Usage

Data Structure

Modification

Datatypes

InfiniteOpt.BoundedScalarConstraint

InfiniteOpt.FiniteConstraintRef

InfiniteOpt.GeneralConstraintRef

InfiniteOpt.InfiniteConstraintRef

InfiniteOpt.MeasureConstraintRef

InfiniteOpt.BoundedScalarConstraint — Type

BoundedScalarConstraint{F <: JuMP.AbstractJuMPScalar,
                        S <: MOI.AbstractScalarSet} <: JuMP.AbstractConstraint

A DataType that stores infinite constraints defined on a subset of the infinite parameters on which they depend.

Fields

func::F The JuMP object.

set::S The MOI set.

bounds::ParameterBounds Set of valid parameter sub-domains that further bound constraint.

orig_bounds::ParameterBounds Set of the constraint's original parameter sub-domains (not considering hold variables)

InfiniteOpt.GeneralConstraintRef — Type

GeneralConstraintRef

An abstract type for constraint references unique to InfiniteOpt.

InfiniteOpt.InfiniteConstraintRef — Type

InfiniteConstraintRef{S <: JuMP.AbstractShape} <: GeneralConstraintRef

A DataType for constraints that contain infinite variables.

Fields

model::InfiniteModel Infinite model.

index::Int Index of constraint in model.

shape::JuMP.AbstractShape Shape of constraint

InfiniteOpt.FiniteConstraintRef — Type

FiniteConstraintRef{S <: JuMP.AbstractShape} <: GeneralConstraintRef

A DataType for constraints that contain finite variables.

Fields

model::InfiniteModel Infinite model.

index::Int Index of constraint in model.

shape::JuMP.AbstractShape Shape of constraint

InfiniteOpt.MeasureConstraintRef — Type

MeasureConstraintRef{S <: JuMP.AbstractShape} <: GeneralConstraintRef

A DataType for constraints that contain finite variables and measures.

Fields

model::InfiniteModel Infinite model.

index::Int Index of constraint in model.

shape::JuMP.AbstractShape Shape of constraint

Methods/Macros

InfiniteOpt.@BDconstraint

InfiniteOpt.add_parameter_bound

InfiniteOpt.delete_parameter_bound

InfiniteOpt.delete_parameter_bounds

InfiniteOpt.has_parameter_bounds

InfiniteOpt.parameter_bounds

InfiniteOpt.set_parameter_bounds

JuMP.add_constraint

JuMP.add_to_function_constant

JuMP.all_constraints

JuMP.all_constraints

JuMP.all_constraints

JuMP.all_constraints

JuMP.build_constraint

JuMP.constraint_by_name

JuMP.constraint_object

JuMP.delete

JuMP.index

JuMP.is_valid

JuMP.list_of_constraint_types

JuMP.name

JuMP.normalized_coefficient

JuMP.normalized_rhs

JuMP.num_constraints

JuMP.num_constraints

JuMP.num_constraints

JuMP.num_constraints

JuMP.owner_model

JuMP.set_name

JuMP.set_normalized_coefficient

JuMP.set_normalized_rhs

InfiniteOpt.@BDconstraint — Macro

@BDconstraint(model::InfiniteModel, [i = ..., ...](bound_expr), constr_expr;
              [kw_args...])

Add an anonymous bounded constraint to model and return an appropriate container of constraint reference(s).

    @BDconstraint(model::InfiniteModel, name[i = ..., ...](bound_expr),
                  constr_expr; [kw_args...])

Add a named bounded constraint to model and return an appropriate container of constraint reference(s). This defines the constraint as expressed in constr_expr over some sub-domain of parameters as indicated by bound_expr. The format of constr_expr must follow the same syntax as that specified in JuMP.@constraint. For example, if we want to express the constraint 2T(t, x) + 3y = 42, the constraint expression constr_expr would be 2 * T + 3 * y == 42.

By default, JuMP.@constraint would express the above example constraint over the whole domain of t and x. However, we can use @BDconstraint to express this constraint over some sub-domain(s) via bound_expr. Here (bound_expr) can be of the form:

(param in [lb, ub], ...) enforcing param to be in a sub-domain from lb to ub (note ∈ can be used in place of in)

(params in [lb, ub], ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(lb <= param <= ub, ...) enforcing param to be in a sub-domain from lb to ub

(lb <= params <= ub, ...) enforcing that all parameter references in params each be a in sub-domain from lb to ub

(param == value, ...) enforcing param to be equal to value

(params == value, ...) enforcing that all parameter references in params each to be equal to value

Any combination of the above forms. Must be inside parentheses and comma separated.

Like typical constraints, the container keyword argument can be used to specify the JuMP container type used to store the constraint references. Note this macro errors if bound_expr is ommited or if some unrecognized syntax is used.

Examples

julia> @BDconstraint(model, c1(t in [0, 1]), T^2 + z <= 1)
c1 : T(x, t)² + z <= 1.0, for all t in [0, 1]

julia> @BDconstraint(model, c2[i = 1:3](x[i] in [0, 1]), T^2 + z + x[i] <= 1)
3-element Array{GeneralConstraintRef,1}:
 c2[1] : T(x, t)² + z + x[1] <= 1.0, for all x[1] in [0, 1]
 c2[2] : T(x, t)² + z + x[2] <= 1.0, for all x[2] in [0, 1]
 c2[3] : T(x, t)² + z + x[3] <= 1.0, for all x[3] in [0, 1]

julia> @BDconstraint(model, (x == 0, t == 0), T^2 + z <= 1)
T(x, t)² + z <= 1.0, for all x[2] == 0, x[3] == 0, t == 0, x[1] == 0

julia> @BDconstraint(model, [i = 1:3](x[i] == 0), T^2 + z <= 1,
                     container = SparseAxisArray)
JuMP.Containers.SparseAxisArray{GeneralConstraintRef,1,Tuple{Any}} with 3 entries:
  [3]  =  T(x, t)² + z <= 1.0, for all x[3] == 0
  [2]  =  T(x, t)² + z <= 1.0, for all x[2] == 0
  [1]  =  T(x, t)² + z <= 1.0, for all x[1] == 0

JuMP.build_constraint — Method

JuMP.build_constraint(_error::Function, expr::InfiniteExpr,
                      set::MOI.AbstractScalarSet;
                      [parameter_bounds::ParameterBounds = ParameterBounds()])

Extend JuMP.build_constraint to accept the parameter_bounds argument and return a BoundedScalarConstraint if the parameter_bounds keyword argument is specifed or return a JuMP.ScalarConstraint otherwise. This is primarily intended to work as an internal function for constraint macros.

Example

julia> constr = build_constraint(error, g + x, MOI.EqualTo(42.0),
              parameter_bounds = ParameterBounds(Dict(t => IntervalSet(0, 1))));

julia> isa(constr, BoundedScalarConstraint)
true

JuMP.add_constraint — Method

JuMP.add_constraint(model::InfiniteModel, c::JuMP.AbstractConstraint,
                    [name::String = ""])

Extend JuMP.add_constraint to add a constraint c to an infinite model model with name name. Returns an appropriate constraint reference whose type depends on what variables are used to define the constraint. Errors if a vector constraint is used, the constraint only constains parameters, or if any variables do not belong to model. This is primarily used as an internal method for the cosntraint macros.

Example

julia> constr = build_constraint(error, g + x, MOI.EqualTo(42));

julia> cref = add_constraint(model, constr, "name")
name : g(t) + x == 42.0

JuMP.owner_model — Method

JuMP.owner_model(cref::GeneralConstraintRef)::InfiniteModel

Extend JuMP.owner_model to return the infinite model associated with cref.

Example

julia> model = owner_model(cref)
An InfiniteOpt Model
Minimization problem with:
Variables: 3
Objective function type: HoldVariableRef
`GenericAffExpr{Float64,FiniteVariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
Names registered in the model: g, t, h, x
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

JuMP.index — Method

JuMP.index(cref::GeneralConstraintRef)::Int

Extend JuMP.index to return the index of an InfiniteOpt constraint cref.

Example

julia> index(cref)
2

JuMP.constraint_object — Method

JuMP.constraint_object(cref::GeneralConstraintRef)::JuMP.AbstractConstraint

Extend JuMP.constraint_object to return the constraint object associated with cref.

Example

julia> obj = constraint_object(cref)
ScalarConstraint{HoldVariableRef,MathOptInterface.LessThan{Float64}}(x,
MathOptInterface.LessThan{Float64}(1.0))

JuMP.name — Method

JuMP.name(cref::GeneralConstraintRef)::String

Extend JuMP.name to return the name of an InfiniteOpt constraint.

Example

julia> name(cref)
constr_name

JuMP.set_name — Method

JuMP.set_name(cref::GeneralConstraintRef, name::String)

Extend JuMP.set_name to specify the name of a constraint cref.

Example

julia> set_name(cref, "new_name")

julia> name(cref)
new_name

JuMP.is_valid — Method

JuMP.is_valid(model::InfiniteModel, cref::GeneralConstraintRef)::Bool

Extend JuMP.is_valid to return Bool whether an InfiniteOpt constraint reference is valid.

Example

julia> is_valid(model, cref)
true

JuMP.delete — Method

JuMP.delete(model::InfiniteModel, cref::GeneralConstraintRef)

Extend JuMP.delete to delete an InfiniteOpt constraint and all associated information. Errors if cref is invalid.

Example

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 z >= 0.0
 g(t) + z >= 42.0
 t in [0, 6]

julia> delete(model, cref)

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 z >= 0.0
 t in [0, 6]

InfiniteOpt.has_parameter_bounds — Method

has_parameter_bounds(cref::GeneralConstraintRef)::Bool

Return a Bool indicating if cref is limited to a sub-domain as defined by a ParameterBounds object.

Example

julia> has_parameter_bounds(cref)
true

InfiniteOpt.parameter_bounds — Method

parameter_bounds(cref::GeneralConstraintRef)::ParameterBounds

Return the ParameterBounds object associated with the constraint cref. Errors if cref does not have parameter bounds.

Example

julia> parameter_bounds(cref)
Subdomain bounds (1): t ∈ [0, 2]

InfiniteOpt.set_parameter_bounds — Method

set_parameter_bounds(cref::GeneralConstraintRef, bounds:ParameterBounds;
                     [force = false])

Specify a new ParameterBounds object bounds for the constraint cref. This is meant to be primarily used by @set_parameter_bounds which provides a more intuitive syntax.

Example

julia> set_parameter_bounds(cref, ParameterBounds(Dict(t => IntervalSet(0, 2))))

julia> parameter_bounds(cref)
Subdomain bounds (1): t ∈ [0, 2]

InfiniteOpt.add_parameter_bound — Method

add_parameter_bound(cref::GeneralConstraintRef, pref::ParameterRef,
                    lower::Number, upper::Number)

Add an additional parameter bound to cref such that it is defined over the sub-domain based on pref from lower to upper. This is primarily meant to be used by @add_parameter_bounds.

julia> add_parameter_bound(cref, t, 0, 2)

julia> parameter_bounds(cref)
Subdomain bounds (1): t ∈ [0, 2]

InfiniteOpt.delete_parameter_bound — Method

delete_parameter_bound(cref::GeneralConstraintRef, pref::ParameterRef)

Delete the parameter bound of the constraint cref associated with the infinite parameter pref if cref has such a bound. Note that any other parameter bounds will be unaffected. Note any bounds that are needed for hold variables inside in cref will be unaffected.

Example

julia> @BDconstraint(model, c1(x == 0), y <= 42)
c1 : y(x) <= 42, for all x[1] == 0, x[2] == 0

julia> delete_parameter_bounds(c1, x[2])

julia> c1
c1 : y(x) <= 42, for all x[1] == 0

InfiniteOpt.delete_parameter_bounds — Method

delete_parameter_bounds(cref::GeneralConstraintRef)

Delete all the parameter bounds of the constraint cref. Note any bounds that are needed for hold variables inside in cref will be unaffected.

Example

julia> @BDconstraint(model, c1(x == 0), y <= 42)
c1 : y(x) <= 42, for all x[1] == 0, x[2] == 0

julia> delete_parameter_bounds(c1)

julia> c1
c1 : y(x) <= 42

JuMP.set_normalized_rhs — Method

JuMP.set_normalized_rhs(cref::GeneralConstraintRef, value::Real)

Set the right-hand side term of constraint to value. Note that prior to this step, JuMP will aggregate all constant terms onto the right-hand side of the constraint. For example, given a constraint 2x + 1 <= 2, set_normalized_rhs(con, 4) will create the constraint 2x <= 4, not 2x + 1 <= 4.

julia> @constraint(model, con, 2x + 1 <= 2)
con : 2 x <= 1.0

julia> set_normalized_rhs(con, 4)

julia> con
con : 2 x <= 4.0

JuMP.normalized_rhs — Method

JuMP.normalized_rhs(cref::GeneralConstraintRef)::Number

Return the right-hand side term of cref after JuMP has converted the constraint into its normalized form. See also JuMP.set_normalized_rhs.

JuMP.add_to_function_constant — Method

JuMP.add_to_function_constant(cref::GeneralConstraintRef, value::Real)

Add value to the function constant term. Note that for scalar constraints, JuMP will aggregate all constant terms onto the right-hand side of the constraint so instead of modifying the function, the set will be translated by -value. For example, given a constraint 2x <= 3, add_to_function_constant(c, 4) will modify it to 2x <= -1. ```

JuMP.set_normalized_coefficient — Method

JuMP.set_normalized_coefficient(cref::GeneralConstraintRef,
                                variable::GeneralVariableRef, value::Real)

Set the coefficient of variable in the constraint constraint to value. Note that prior to this step, JuMP will aggregate multiple terms containing the same variable. For example, given a constraint 2x + 3x <= 2, set_normalized_coefficient(con, x, 4) will create the constraint 4x <= 2.

julia> con
con : 5 x <= 2.0

julia> set_normalized_coefficient(con, x, 4)

julia> con
con : 4 x <= 2.0

JuMP.normalized_coefficient — Method

JuMP.normalized_coefficient(cref::GeneralConstraintRef,
                            variable::GeneralVariableRef)::Number

Return the coefficient associated with variable in constraint after JuMP has normalized the constraint into its standard form. See also JuMP.set_normalized_coefficient.

JuMP.constraint_by_name — Method

JuMP.constraint_by_name(model::InfiniteModel,
                        name::String)::Union{GeneralConstraintRef, Nothing}

Extend JuMP.constraint_by_name to return the constraint reference associated with name if one exists or returns nothing. Errors if more than one constraint uses the same name.

Example

julia> constraint_by_name(model, "constr_name")
constr_name : x + pt == 3.0

JuMP.list_of_constraint_types — Method

JuMP.list_of_constraint_types(model::InfiniteModel)::Vector{Tuple)

Extend JuMP.list_of_constraint_types to return a list of tuples that contain all the used combinations of function types and set types in the model.

julia> all_constraints(model)
5-element Array{Tuple{DataType,DataType},1}:
 (HoldVariableRef, MathOptInterface.LessThan{Float64})
 (PointVariableRef, MathOptInterface.GreaterThan{Float64})
 (HoldVariableRef, MathOptInterface.GreaterThan{Float64})
 (HoldVariableRef, MathOptInterface.Integer)
 (InfiniteVariableRef, MathOptInterface.GreaterThan{Float64})

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel,
                     function_type::Type{<:JuMP.AbstractJuMPScalar},
                     set_type::Type{<:MOI.AbstractSet})::Int

Extend JuMP.num_constraints to return the number of constraints with a partiuclar function type and set type.

Example

julia> num_constraints(model, HoldVariableRef, MOI.LessThan)
1

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel,
                     function_type::Type{<:JuMP.AbstractJuMPScalar})::Int

Extend JuMP.num_constraints to search by function types for all MOI sets and return the total number of constraints with a particular function type.

julia> num_constraints(model, HoldVariableRef)
3

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel,
                     function_type::Type{<:MOI.AbstractSet})::Int

Extend JuMP.num_constraints to search by MOI set type for all function types and return the total number of constraints that use a particular MOI set type.

julia> num_constraints(model, MOI.LessThan)
2

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel)::Int

Extend JuMP.num_constraints to return the total number of constraints in an infinite model model.

julia> num_constraints(model)
4

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel,
                     function_type::Type{<:JuMP.AbstractJuMPScalar},
                     set_type::Type{<:MOI.AbstractSet}
                     )::Vector{<:GeneralConstraintRef}

Extend JuMP.all_constraints to return a list of all the constraints with a particular function type and set type.

julia> all_constraints(model, HoldVariableRef, MOI.LessThan)
1-element Array{GeneralConstraintRef,1}:
 x <= 1.0

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel,
                     function_type::Type{<:JuMP.AbstractJuMPScalar}
                     )::Vector{<:GeneralConstraintRef}

Extend JuMP.all_constraints to search by function types for all MOI sets and return a list of all constraints use a particular function type.

julia> all_constraints(model, HoldVariableRef)
3-element Array{GeneralConstraintRef,1}:
 x >= 0.0
 x <= 3.0
 x integer

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel,
                     set_type::Type{<:MOI.AbstractSet}
                     )::Vector{<:GeneralConstraintRef}

Extend JuMP.all_constraints to search by MOI set type for all function types and return a list of all constraints that use a particular set type.

julia> all_constraints(model, MOI.GreaterThan)
3-element Array{GeneralConstraintRef,1}:
 x >= 0.0
 g(t) >= 0.0
 g(0.5) >= 0.0

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel)::Vector{<:GeneralConstraintRef}

Extend JuMP.all_constraints to return all a list of all the constraints in an infinite model model.

julia> all_constraints(model)
5-element Array{GeneralConstraintRef,1}:
 x >= 0.0
 x <= 3.0
 x integer
 g(t) >= 0.0
 g(0.5) >= 0.0