Constraints

A guide and manual for defining and manipulating constraints in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

Constraints are a key part of infinite dimensional problems and serve as a fundamental utility of InfiniteOpt. In particular, InfiniteOpt supports finite constraints that entail finite variables and/or measures that fully remove any infinite parameter dependencies (e.g., first stage constraints), infinite constraints that are enforced over the entire domain of its infinite parameter dependencies (e.g., path constraints), and restricted constraints which are enforced over some specified sub-domain of its infinite parameter dependencies (e.g., boundary conditions). This page will highlight how to implement these types of constraints in InfiniteOpt.

Note

Nonlinear constraints as defined by JuMP.@NLconstraint are not currently supported by InfiniteOpt. See Nonlinear Expressions for more information and possible workarounds.

Basic Usage

Principally, the @constraint macro is used to define constraints. First, let's setup an infinite model with variables that we can add constraints to:

julia> model = InfiniteModel();

julia> @infinite_parameter(model, t in [0, 10]);

julia> @infinite_parameter(model, x[1:2] in [-2, 2]);

julia> @variable(model, ya, Infinite(t, x));

julia> @variable(model, yb, Infinite(t));

julia> @variable(model, z[1:2]);

Note

Unlike previous versions, InfiniteOpt now supports all of the constraints offered by JuMP, including vector and semi-definite constraints! Please see JuMP's constraint documentation for a thorough explanation of the supported types and syntax.

Scalar Constraints

Scalar constraints use scalar functions of variables. For example, let's define the constraint $||z||^2 + 2y_a(t, x) \leq 0, \ \forall t \in [0, 10], x \in [-2, 2]^2$ using @constraint:

julia> @constraint(model, c1, sum(z[i]^2 for i = 1:2) + 2ya <= 0)
c1 : z[1]² + z[2]² + 2 ya(t, x) ≤ 0.0, ∀ t ∈ [0, 10], x[1] ∈ [-2, 2], x[2] ∈ [-2, 2]

Thus, we added an infinite constraint (which infinite with respect to t and x) to model and stored the corresponding constraint reference to c1. Note that this is enforced over the full infinite domains of the infinite parameters t and x which are implicitly used by c1. For scalar constraints like this one, the allowed constraint operators are ==, <=, ≤, >=, and ≥.

Note

Linear algebra constraints can also be used when defining constraints when . is added in front of the constraint operators (e.g., .<=). This behavior is further explained in JuMP's constraint documentation.

Similarly, we can define an array of constraints with varied indexes by including an additional argument before the constraint expression. For example, let's define $3z_i - 14 = 0, \ \forall i \in \{1,2\}$:

julia> @constraint(model, c2[i = 1:2], 3z[i] - 14 == 0)
2-element Array{InfOptConstraintRef,1}:
 c2[1] : 3 z[1] = 14.0
 c2[2] : 3 z[2] = 14.0

Thus, we added two constraints to model and stored a vector of the corresponding constraint references to the Julia variable c2. To learn more about building containers of constraints please see JuMP's constraint container documentation.

Multi-Dimensional Constraints

Building upon JuMP we support a variety of multi-dimensional constraint types. For example, we can define the vector constraint:

julia> A = [1 2; 3 4]
2×2 Array{Int64,2}:
 1  2
 3  4

julia> b = [5, 6]
2-element Array{Int64,1}:
 5
 6

julia> @constraint(model, A * z - b in MOI.Nonnegatives(2))
[z[1] + 2 z[2] - 5, 3 z[1] + 4 z[2] - 6] ∈ MathOptInterface.Nonnegatives(2)

See JuMP's constraint documentation for a thorough tutorial on the accepted syntax and constraint types.

Restricted Constraints

Restricted constraints entail an infinite domain (determined by the infinite parameters they explicitly/implicitly depend on) that is restricted to a certain sub-domain. Such constraints are common for enforcing initial/boundary conditions and for enforcing path constraints over a certain sub-domain.

Warning

Previous versions of InfiniteOpt referred to restricted constraints as "bounded constraints" and used @BDconstraint to define them. This has been deprecated in favor of the more intuitive domain restricted nomenclature.

# Old syntax
@BDconstraint(model, name_expr(restricts...), constr_expr)

# New syntax
@constraint(model, name_expr, constr_expr, DomainRestrictions(restricts...))

These types of constraints are defined adding DomainRestrictions. For example, let's add the initial condition $y_b(0) = 0$:

julia> @constraint(model, initial, yb == 0, DomainRestrictions(t => 0))
initial : yb(t) = 0.0, ∀ t = 0

Thus, we have added a constraint to model defined over the sub-domain $t = 0$ in accordance with the initial condition.

More complex sub-domains can be specified by simply adding more restrictions. To illustrate this, let's define the constraint $2y_b^2(t, x) + z_1 \geq 3, \ \forall t = 0, \ x \in [-1, 1]^2$:

julia> @constraint(model, 2ya^2 + z[1] >= 3, DomainRestrictions(t => 0, x => [-1, 1]))
2 ya(t, x)² + z[1] ≥ 3.0, ∀ t = 0, x[1] ∈ [-1, 1], x[2] ∈ [-1, 1]

Now we have added constraints to our model and it is ready to be solved!

Data Structure

Here we detail the data structures used to store constraints in InfiniteOpt. In general, constraints in JuMP are of the form: function in set where function corresponds to a JuMP expression and set corresponds to a MOI set. This leads to the following data structures:

Constraint Type	Function Type	Set Type

Scalar	JuMP.AbstractJuMPScalar	MOI.AbstractScalarSet

Vector	Vector{<:JuMP.AbstractJuMPScalar}	MOI.AbstractVectorSet

Matrix	Matrix{<:JuMP.AbstractJuMPScalar}	MOI.AbstractVectorSet via vectorization

The above combos are then stored in JuMP.ScalarConstraints and `JuMP.VectorConstraints.

Restricted constraints are built upon this data structure where the underlying constraint is created in the same manner. Then the specified DomainRestrictions are added by creating a DomainRestrictedConstraint which stores the JuMP.AbstractConstraint and the restrictions.

These constraint objects are what store constraints in InfiniteModels. And these are pointed to by InfOptConstraintRefs.

Definition

In this section, we describe the ins and outs of defining constraints. Note that this process is analogous to the manner in which variables are defined and added to the model.

Manual Definition

Defining a constraint principally involves the following steps:

Define the constraint information (i.e., function, set, and domain restrictions)

Construct a concrete subtype of JuMP.AbstractConstraint to store the constraint information

Add the AbstractConstraint object to an InfiniteModel and assign a name

Create an InfOptConstraintRef that points to the constraint object stored in the model.

The constraint objects are specified via JuMP.build_constraint which requires that the user provides a function, set, and optionally include domain restrictions. For example, let's build a scalar constraint $3y_a(t, x) - y_b^2(t) \leq 0, \ \forall t \in [0, 10], x \in [-2, 2]^2$ over its full infinite domain (i.e., have no DomainRestrictions):

julia> constr = build_constraint(error, 3ya - yb^2, MOI.LessThan(0.0));

Now the built constraint object can be added to the infinite model via add_constraint. Let's do so with our example and assign it the name of c3 (note that adding a name is optional):

julia> cref = add_constraint(model, constr, "c3")
c3 : -yb(t)² + 3 ya(t, x) ≤ 0.0, ∀ t ∈ [0, 10], x[1] ∈ [-2, 2], x[2] ∈ [-2, 2]

Thus, we have made our constraint and added it model and now have a constraint reference cref that we can use to access it.

The @constraint and @SDconstraint macros automate the above steps.

Macro Definition

As mentioned above in the Basic Usage section, the @constraint macro should be used to define constraints with the syntax: @constraint(model::InfiniteModel, [container/name_expr], constr_expr, [rs::DomainRestrictions]).

The second argument is optional and is used to assign a name and/or define indexing variables to be used in the constraint expr. When a name is provided it is registered and cannot be used again for another constraint or variable name. The indexing expression can be used to produce an array of constraints as shown below (notice this is equivalent to looping over individual @constraint calls):

julia> crefs = @constraint(model, [i = 1:2], 2z[i] - yb == 0)
2-element Array{InfOptConstraintRef,1}:
 2 z[1] - yb(t) = 0.0, ∀ t ∈ [0, 10]
 2 z[2] - yb(t) = 0.0, ∀ t ∈ [0, 10]

julia> crefs = Vector{InfOptConstraintRef}(undef, 2);

julia> for i = 1:2
           crefs[i] = @constraint(model, 2z[i] - yb == 0)
       end

julia> crefs
2-element Array{InfOptConstraintRef,1}:
 2 z[1] - yb(t) = 0.0, ∀ t ∈ [0, 10]
 2 z[2] - yb(t) = 0.0, ∀ t ∈ [0, 10]

Please refer to JuMP's constraint container documentation for a thorough tutorial on creating containers of constraints.

Any constraint type supported by JuMP can be specified in the constr_expr argument. This includes a wealth of constraint types including:

Variable constraints

Scalar constraints

Vector constraints

Conic constraints

Indicator constraints

Semi-definite constraints

For example, we could define the following semi-definite constraint using @SDconstraint:

julia> @SDconstraint(model, [yb 2yb; 3yb 4yb] >= ones(2, 2))
[yb(t) - 1    2 yb(t) - 1;
 3 yb(t) - 1  4 yb(t) - 1] ∈ PSDCone(), ∀ t ∈ [0, 10]

See JuMP's constraint documentation for a thorough tutorial on the accepted syntax and constraint types.

Finally, restrictions on the inherent infinite domain of a constraint can be specified via DomainRestrictions with the rs argument. The accepted syntax is DomainRestrictions(restricts...) where each argument of restricts can be any of the following forms:

pref => value

pref => [lb, ub]

pref => IntervalDomain(lb, ub)

prefs => value

prefs => [lb, ub]

prefs => IntervalDomain(lb, ub).

Note that pref and prefs must correspond to infinite parameters.

For example, we can define the constraint $y_a^2(t, x) + z_i \leq 1$ and restrict the infinite domain of $x_i$ to be $[0, 1]$:

julia> @constraint(model, [i = 1:2], ya^2 + z[i] <= 1, DomainRestrictions(x[i] => [0, 1]))
2-element Array{InfOptConstraintRef,1}:
 ya(t, x)² + z[1] ≤ 1.0, ∀ t ∈ [0, 10], x[1] ∈ [0, 1], x[2] ∈ [-2, 2]
 ya(t, x)² + z[2] ≤ 1.0, ∀ t ∈ [0, 10], x[1] ∈ [-2, 2], x[2] ∈ [0, 1]

Queries

In this section, we describe a variety of methods to extract constraint information.

Basic

A number of constraint properties can be extracted via constraint references. Principally, the validity, name, model, index, and constraint object can be queried via is_valid, name, owner_model, index, and constraint_object, respectively. These methods all constitute extensions of JuMP methods and follow exactly the same behavior. Let's try them out with the following example:

julia> is_valid(model, c1) # check if contained in model
true

julia> name(c1) # get the name
"c1"

julia> m = owner_model(c1); # get the model it is added to

julia> index(c1) # get the constraint's index
InfOptConstraintIndex(1)

julia> constr = constraint_object(c1); # get the raw constraint datatype

Also, constraint_by_name can be used to retrieve a constraint reference if only the name is known and its name is unique. For example, let's extract the reference for "c1":

julia> cref = constraint_by_name(model, "c1")
c1 : z[1]² + z[2]² + 2 ya(t, x) ≤ 0.0, ∀ t ∈ [0, 10], x[1] ∈ [-2, 2], x[2] ∈ [-2, 2]

Domain Restrictions

As explained above, restricted constraints serve as a key capability of InfiniteOpt. Information about domain restrictions can be obtained via has_domain_restrictions and domain_restrictions which indicate if a constraint is restricted and what its DomainRestrictions are, respectively. These are exemplified below:

julia> has_domain_restrictions(c1) # check if constraint is bounded
false

julia> has_domain_restrictions(initial)
true

julia> domain_restrictions(initial)
Subdomain restrictions (1): t = 0

General

Constraints can be defined in a number of ways symbolically that differ from how it is actually stored in the model. This principally occurs since like terms and constants are combined together where possible with the variable terms on the left hand side and the constant on the right hand side. For instance, the constraint $2y_b(t) + 3y_b(t) - 2 \leq 1 + z_1$ would be normalized $5y_b(t) - z_1 \leq 3$. In accordance with this behavior, normalized_rhs and normalized_coefficient can be used to query the normalized right hand side and the coefficient of a particular variable reference, respectively. Let's employ the above example to illustrate this:

julia> @constraint(model, constr, 2yb + 3yb - 2 <= 1 + z[1])
constr : 5 yb(t) - z[1] ≤ 3.0, ∀ t ∈ [0, 10]

julia> normalized_rhs(constr)
3.0

julia> normalized_coefficient(constr, yb)
5.0

There also exist a number of methods for querying an infinite model about what constraints it contains. list_of_constraint_types can be used query what types of constraints have been added to a model. This is provided as a list of tuples where the first element is the expression type and the second element is the set type (recall that constraints are stored in the form func-in-set). Thus, for our current model we obtain:

julia> list_of_constraint_types(model)
4-element Array{Tuple{DataType,DataType},1}:
 (GenericQuadExpr{Float64,GeneralVariableRef}, MathOptInterface.LessThan{Float64})
 (GenericQuadExpr{Float64,GeneralVariableRef}, MathOptInterface.GreaterThan{Float64})
 (GenericAffExpr{Float64,GeneralVariableRef}, MathOptInterface.LessThan{Float64})
 (GenericAffExpr{Float64,GeneralVariableRef}, MathOptInterface.EqualTo{Float64})

This information is useful when used in combination with the num_constraints and all_constraints methods which can take the expression type and/or the set type as inputs. Here num_constraints provides the number of constraints that match a certain type and all_constraints returns a list of constraint references matching the criteria provided. These have been extended beyond JuMP functionality such additional methods have been provided for the cases in which one wants to query solely off of set or off expression type. Let's illustrate this with num_constraints:

julia> num_constraints(model) # total number of constraints
15

julia> num_constraints(model, GenericQuadExpr{Float64,GeneralVariableRef})
5

julia> num_constraints(model, MOI.LessThan{Float64})
5

julia> num_constraints(model, GenericQuadExpr{Float64,GeneralVariableRef},
                       MOI.LessThan{Float64})
4                   

Modification

In this section, we highlight a number of methods that can be used to modify existing constraints.

Deletion

All constraints in InfiniteOpt can be removed in like manner to typical JuMP constraints with the appropriate extension of delete. This will remove the corresponding constraint object from the model. However, please note any registered names will remain registered in the infinite model. This means that a constraint with a registered name cannot be repeatedly added and removed using the same name. To exemplify this, let's delete the constraint c1:

julia> delete(model, c1)

General

There also are a number of ways to modify information and characteristics of constraints. First, set_name can be used to specify a new name for a particular constraint. For instance, let's update the name of initial to "init_cond":

julia> set_name(initial, "init_cond")

julia> initial
init_cond : yb(t) = 0.0, ∀ t = 0

We can also update the normalized right hand side constant value or normalized left hand side variable coefficient value using set_normalized_rhs and set_normalized_coefficient, respectively. Let's again consider the constraint $5y_b(t) - z_1 \leq 3$ as an example. Let's change the constant term to -1 and the y_b(t) coefficient to 2.5:

julia> set_normalized_rhs(constr, -1)

julia> set_normalized_coefficient(constr, yb, 2.5)

julia> constr
constr : 2.5 yb(t) - z[1] ≤ -1.0, ∀ t ∈ [0, 10]

Note

In some cases, it may be more convenient to dynamically modify coefficients and other values via the use of finite parameters. This provides an avenue to update parameters without having to be concerned about the normalized form. For more information, see the Finite Parameters page.

Domain Restrictions

Domain Restrictions can be added to, modified, or removed from any constraint in InfiniteOpt. Principally, this is accomplished via add_domain_restrictions, set_domain_restrictions, and delete_domain_restrictions.

Note

Previous versions of InfiniteOpt used @[set/add]_parameter_bounds which have been deprecated in favor of using DomainRestrictions with the the methods described used in this section.

First, domain restrictions can be added to a constraint via add_domain_restrictions. For example, let's add the bound $t \in [0, 1]$ to constr:

julia> add_domain_restrictions(constr, DomainRestrictions(t => [0, 1]))

julia> constr
constr : 2.5 yb(t) - z[1] ≤ -1.0, ∀ t ∈ [0, 1]

In similar manner, set_domain_restrictions can be employed to specify what restrictions a constraint has (overwriting any existing ones if forced). It follows the same syntax, so let's use it to change the bounds on t to $t = 0$:

julia> set_domain_restrictions(constr, DomainRestrictions(t => 0), force = true)

julia> constr
constr : 2.5 yb(t) - z[1] ≤ -1.0, ∀ t = 0

Finally, constraint restrictions can be deleted via delete_domain_restrictions. Now let's delete the domain restrictions associated with our example:

julia> delete_domain_restrictions(constr)

julia> constr
constr : 2.5 yb(t) - z[1] ≤ -1.0, ∀ t ∈ [0, 10]

Datatypes

InfiniteOpt.ConstraintData

InfiniteOpt.DomainRestrictedConstraint

InfiniteOpt.DomainRestrictions

InfiniteOpt.InfOptConstraintIndex

InfiniteOpt.InfOptConstraintRef

InfiniteOpt.DomainRestrictions — Type

DomainRestrictions{P <: GeneralVariableRef}

A DataType for storing interval domains that constrain particular infinite parameters to a subdomain relative to their full domain. This is used to define subdomains of DomainRestrictedConstraints. Note that the GeneralVariableRef must pertain to infinite parameters.

The constructor syntax is

DomainRestrictions(restrictions...)

where each argument of restrictions is one of the following forms:

pref => value

pref => [lb, ub]

pref => IntervalDomain(lb, ub)

prefs => value

prefs => [lb, ub]

prefs => IntervalDomain(lb, ub).

Note that pref and prefs must correspond to infinite parameters.

Fields

intervals::Dict{GeneralVariableRef, IntervalDomain}: A dictionary of interval bounds on infinite parameters.

InfiniteOpt.DomainRestrictedConstraint — Type

DomainRestrictedConstraint{C <: JuMP.AbstractConstraint, 
                           P <: GeneralVariableRef
                           } <: JuMP.AbstractConstraint

A DataType for creating a constraint with enforced DomainRestrictions. For example this may pertain to a boundary condition.

Fields

constraint::C: The optimization constraint.

restrictions::DomainRestrictions{P}: The restrictions that determine the sub-domain of the constraint.

InfiniteOpt.ConstraintData — Type

ConstraintData{C <: JuMP.AbstractConstraint} <: AbstractDataObject

A mutable DataType for storing constraints and their data.

Fields

constraint::C: The constraint.

object_nums::Vector{Int}: The object numbers of the parameter objects that the constraint depends on.

name::String: The name used for printing.

measure_indices::Vector{MeasureIndex}: Indices of dependent measures.

is_info_constraint::Bool: Is this is constraint based on variable info (e.g., lower bound)

InfiniteOpt.InfOptConstraintIndex — Type

InOptConstraintIndex <: ObjectIndex

A DataType for storing the index of a constraint.

Fields

value::Int64: The index value.

InfiniteOpt.InfOptConstraintRef — Type

InfOptConstraintRef

A DataType for constraints that are in InfiniteModels

Fields

model::InfiniteModel: Infinite model.

index::InfOptConstraintIndex: Index of the constraint in model.

Methods/Macros

InfiniteOpt.add_domain_restrictions

InfiniteOpt.delete_domain_restrictions

InfiniteOpt.domain_restrictions

InfiniteOpt.has_domain_restrictions

InfiniteOpt.parameter_refs

InfiniteOpt.set_domain_restrictions

JuMP.add_constraint

JuMP.add_to_function_constant

JuMP.all_constraints

JuMP.build_constraint

JuMP.constraint_by_name

JuMP.constraint_object

JuMP.delete

JuMP.index

JuMP.is_valid

JuMP.list_of_constraint_types

JuMP.name

JuMP.normalized_coefficient

JuMP.normalized_rhs

JuMP.num_constraints

JuMP.owner_model

JuMP.set_name

JuMP.set_normalized_coefficient

JuMP.set_normalized_rhs

JuMP.build_constraint — Method

JuMP.build_constraint(_error::Function, func, set,
                      restrictions::DomainRestrictions{GeneralVariableRef}
                      )::DomainRestrictedConstraint

Extend JuMP.buid_constraint to handle including restrictions to its inherit infinite parameter domains in addition to the traditional func in set setup. This returns a DomainRestrictedConstraint that can then be added via JuMP.add_constraint. Errors if the restrictions are incompadible with infinite parameter domains.

Example

julia> restrictions = DomainRestrictions(t => 0)
Subdomain restrictions (1): t = 0

julia> con = build_constraint(error, y + 2, MOI.LessThan(0.0), restrictions);

JuMP.add_constraint — Method

JuMP.add_constraint(model::InfiniteModel, c::JuMP.AbstractConstraint,
                    [name::String = ""])::InfOptConstraintRef

Extend JuMP.add_constraint to add a constraint c to an infinite model model with name name. Returns an appropriate constraint reference whose type depends on what variables are used to define the constraint. Errors if any variables do not belong to model. This is primarily used as an internal method for the constraint macros.

Example

julia> @infinite_parameter(model, t in [0, 10]);

julia> @variable(model, g, Infinite(t));

julia> @variable(model, x);

julia> constr = build_constraint(error, g + x, MOI.EqualTo(42));

julia> cref = add_constraint(model, constr, "name")
name : g(t) + x = 42.0, ∀ t ∈ [0, 10]

JuMP.owner_model — Method

JuMP.owner_model(cref::InfOptConstraintRef)::InfiniteModel

Extend JuMP.owner_model to return the infinite model associated with cref.

Example

julia> model = owner_model(cref)
An InfiniteOpt Model
Minimization problem with:
Finite Parameters: 0
Infinite Parameters: 3
Variables: 3
Derivatives: 0
Measures: 0
Objective function type: GeneralVariableRef
`GenericAffExpr{Float64,GeneralVariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
Names registered in the model: g, t, h, x
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

JuMP.index — Method

JuMP.index(cref::InfOptConstraintRef)::InfOptConstraintIndex

Extend JuMP.index to return the index of an InfiniteOpt constraint cref.

Example

julia> index(cref)
InfOptConstraintIndex(2)

JuMP.constraint_object — Method

JuMP.constraint_object(cref::InfOptConstraintRef)::JuMP.AbstractConstraint

Extend JuMP.constraint_object to return the constraint object associated with cref.

Example

julia> @infinite_parameter(model, t in [0, 10]);

julia> @variable(model, x <= 1);

julia> cref = UpperBoundRef(x);

julia> obj = constraint_object(cref)
ScalarConstraint{GeneralVariableRef,MathOptInterface.LessThan{Float64}}(x,
MathOptInterface.LessThan{Float64}(1.0))

JuMP.name — Method

JuMP.name(cref::InfOptConstraintRef)::String

Extend JuMP.name to return the name of an InfiniteOpt constraint.

Example

julia> name(cref)
"constr_name"

JuMP.set_name — Method

JuMP.set_name(cref::InfOptConstraintRef, name::String)::Nothing

Extend JuMP.set_name to specify the name of a constraint cref.

Example

julia> set_name(cref, "new_name")

julia> name(cref)
"new_name"

JuMP.is_valid — Method

JuMP.is_valid(model::InfiniteModel, cref::InfOptConstraintRef)::Bool

Extend JuMP.is_valid to return Bool whether an InfiniteOpt constraint reference is valid.

Example

julia> is_valid(model, cref)
true

JuMP.delete — Method

JuMP.delete(model::InfiniteModel, cref::InfOptConstraintRef)::Nothing

Extend JuMP.delete to delete an InfiniteOpt constraint and all associated information. Errors if cref is invalid.

Example

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 z ≥ 0.0
 g(t) + z ≥ 42.0, ∀ t ∈ [0, 6]

julia> delete(model, cref)

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 z ≥ 0.0

InfiniteOpt.parameter_refs — Method

parameter_refs(cref::InfOptConstraintRef)::Tuple

Return the tuple of infinite parameter references that determine the infinite dependencies of cref.

Example

julia> parameter_refs(cref)
(t,)

InfiniteOpt.has_domain_restrictions — Function

has_domain_restrictions(cref::InfOptConstraintRef)::Bool

Return a Bool indicating if cref is limited to a sub-domain as defined by a DomainRestrictions object.

Example

julia> has_domain_restrictions(cref)
true

InfiniteOpt.domain_restrictions — Function

domain_restrictions(cref::InfOptConstraintRef)::DomainRestrictions{GeneralVariableRef}

Return the DomainRestrictions object associated with the constraint cref.

Example

julia> domain_restrictions(cref)
Subdomain restrictions (1): t ∈ [0, 2]

InfiniteOpt.set_domain_restrictions — Function

set_domain_restrictions(cref::InfOptConstraintRef,
                     restrictions:DomainRestrictions{GeneralVariableRef};
                     [force::Bool = false])::Nothing

Specify a new DomainRestrictions object restrictions for the constraint cref. Errors if cref already has restrictions and force = false. Where possible it is recommended to use add_domain_restrictions instead.

Example

julia> set_domain_restrictions(cref, DomainRestrictions(t => [0, 2]))

julia> domain_restrictions(cref)
Subdomain restrictions (1): t ∈ [0, 2]

InfiniteOpt.add_domain_restrictions — Function

add_domain_restrictions(cref::InfOptConstraintRef,
                     new_restrictions::DomainRestrictions{GeneralVariableRef}
                     )::Nothing

Add additional domain restrictions to cref such that it is defined over the sub-domain based on pref from lower to upper.

julia> add_domain_restrictions(cref, DomainRestrictions(t => [0, 2]))

julia> domain_restrictions(cref)
Subdomain restrictions (1): t ∈ [0, 2]

InfiniteOpt.delete_domain_restrictions — Function

delete_domain_restrictions(cref::InfOptConstraintRef)::Nothing

Delete all the domain restrictions of the constraint cref. Note any restrictions that are needed for finite variables inside in cref will be unaffected.

Example

julia> @constraint(model, c1, y <= 42, DomainRestrictions(x => 0))
c1 : y(x) ≤ 42, ∀ x[1] = 0, x[2] = 0

julia> delete_domain_restrictions(c1)

julia> c1
c1 : y(x) ≤ 42, ∀ x[1] ∈ [-1, 1], x[2] ∈ [-1, 1]

JuMP.set_normalized_rhs — Method

JuMP.set_normalized_rhs(cref::InfOptConstraintRef, value::Real)::Nothing

Set the right-hand side term of constraint to value. Note that prior to this step, JuMP will aggregate all constant terms onto the right-hand side of the constraint. For example, given a constraint 2x + 1 <= 2, set_normalized_rhs(con, 4) will create the constraint 2x <= 4, not 2x + 1 <= 4.

julia> @constraint(model, con, 2x + 1 <= 2)
con : 2 x ≤ 1.0

julia> set_normalized_rhs(con, 4)

julia> con
con : 2 x ≤ 4.0

JuMP.normalized_rhs — Method

JuMP.normalized_rhs(cref::InfOptConstraintRef)::Float64

Return the right-hand side term of cref after JuMP has converted the constraint into its normalized form.

JuMP.add_to_function_constant — Method

JuMP.add_to_function_constant(cref::InfOptConstraintRef, value::Real)::Nothing

Add value to the function constant term. Note that for scalar constraints, JuMP will aggregate all constant terms onto the right-hand side of the constraint so instead of modifying the function, the set will be translated by -value. For example, given a constraint 2x <= 3, add_to_function_constant(c, 4) will modify it to 2x <= -1. ```

JuMP.set_normalized_coefficient — Method

JuMP.set_normalized_coefficient(cref::InfOptConstraintRef,
                                variable::GeneralVariableRef,
                                value::Real)::Nothing

Set the coefficient of variable in the constraint constraint to value. Note that prior to this step, JuMP will aggregate multiple terms containing the same variable. For example, given a constraint 2x + 3x <= 2, set_normalized_coefficient(con, x, 4) will create the constraint 4x <= 2.

julia> con
con : 5 x ≤ 2.0

julia> set_normalized_coefficient(con, x, 4)

julia> con
con : 4 x ≤ 2.0

JuMP.normalized_coefficient — Method

JuMP.normalized_coefficient(cref::InfOptConstraintRef,
                            variable::GeneralVariableRef)::Float64

Return the coefficient associated with variable in constraint after JuMP has normalized the constraint into its standard form.

JuMP.constraint_by_name — Method

JuMP.constraint_by_name(model::InfiniteModel,
                        name::String)::Union{InfOptConstraintRef, Nothing}

Extend JuMP.constraint_by_name to return the constraint reference associated with name if one exists or returns nothing. Errors if more than one constraint uses the same name.

Example

julia> constraint_by_name(model, "constr_name")
constr_name : x + pt = 3.0

JuMP.list_of_constraint_types — Method

JuMP.list_of_constraint_types(model::InfiniteModel)::Vector{Tuple{DataType, DataType}}

Extend JuMP.list_of_constraint_types to return a list of tuples that contain all the used combinations of function types and set types in the model.

Example

julia> all_constraints(model)
3-element Array{Tuple{DataType,DataType},1}:
 (GeneralVariableRef, MathOptInterface.LessThan{Float64})
 (GeneralVariableRef, MathOptInterface.GreaterThan{Float64})
 (GeneralVariableRef, MathOptInterface.Integer)

JuMP.num_constraints — Method

JuMP.num_constraints(model::InfiniteModel, [function_type], [set_type])::Int

Extend JuMP.num_constraints to return the number of constraints with a partiuclar function type and set type.

Example

julia> num_constraints(model, FiniteVariableRef, MOI.LessThan)
1

julia> num_constraints(model, FiniteVariableRef)
3

julia> num_constraints(model, MOI.LessThan)
2

julia> num_constraints(model)
4

JuMP.all_constraints — Method

JuMP.all_constraints(model::InfiniteModel, [function_type], [set_type])::Vector{InfOptConstraintRef}

Extend JuMP.all_constraints to return a list of all the constraints with a particular function type and set type.

Example

julia> all_constraints(model, GeneralVariableRef, MOI.LessThan)
1-element Array{InfOptConstraintRef,1}:
 x ≤ 1.0

julia> all_constraints(model, GeneralVariableRef)
3-element Array{InfOptConstraintRef,1}:
 x ≥ 0.0
 x ≤ 3.0
 x integer

julia> all_constraints(model, MOI.GreaterThan)
3-element Array{InfOptConstraintRef,1}:
 x ≥ 0.0
 g(t) ≥ 0.0, ∀ t ∈ [0, 6]
 g(0.5) ≥ 0.0

julia> all_constraints(model)
5-element Array{InfOptConstraintRef,1}:
 x ≥ 0.0
 x ≤ 3.0
 x integer
 g(t) ≥ 0.0, ∀ t ∈ [0, 6]
 g(0.5) ≥ 0.0