Derivatives

A guide and manual for the definition and use of derivatives in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

Basic Usage

Advanced Information

Numerical Evaluation Methods

Datatypes

InfiniteOpt.HoldVariable

InfiniteOpt.HoldVariableIndex

InfiniteOpt.HoldVariableRef

InfiniteOpt.InfOptVariable

InfiniteOpt.InfiniteVariable

InfiniteOpt.InfiniteVariableIndex

InfiniteOpt.InfiniteVariableRef

InfiniteOpt.ParameterBounds

InfiniteOpt.PointVariable

InfiniteOpt.PointVariableIndex

InfiniteOpt.PointVariableRef

InfiniteOpt.VariableData

InfiniteOpt.DerivativeIndex — Type

DerivativeIndex <: ObjectIndex

A DataType for storing the index of a Derivative.

Fields

value::Int64: The index value.

InfiniteOpt.DerivativeRef — Type

DerivativeRef <: DispatchVariableRef

A DataType for untranscripted derivative references.

Fields

model::InfiniteModel: Infinite model.

index::DerivativeIndex: Index of the derivative in model.

InfiniteOpt.Derivative — Type

Derivative{V <: GeneralVariableRef} <: InfOptVariable

A DataType for storing core infinite derivative information. This follows a derivative of the form: $\frac{\partial x(\alpha, \hdots)}{\partial \alpha}$ where $x(\alpha, \hdots)$ is an infinite variable and $\alpha$ is an infinite parameter. Here, both $x$ and $\alpha$ must be scalars.

It is important to note that info.start should contain a start value function that generates the start value for a given infinite parameter support. This function should map a support to a start value using user-formatting if is_vector_start = false, otherwise it should do the mapping using a single support vector as input. Also, the variable reference type V must pertain to infinite variables and parameters.

Fields

info::JuMP.VariableInfo{Float64, Float64, Float64, Function}: JuMP variable information.

is_vector_start::Bool: Does the start function take support values formatted as vectors?

variable_ref::V: The variable reference of the infinite variable argument.

parameter_ref::V: The variable reference of the infinite parameter the defines the differential operator.

InfiniteOpt.AbstractDerivativeMethod — Type

AbstractDerivativeMethod

An abstract type for storing derivative evaluation data that is pertinent to its reformation/transcription.

InfiniteOpt.GenerativeDerivativeMethod — Type

GenerativeDerivativeMethod <: AbstractDerivativeMethod

An abstract type for derivative evaluation method types that will require support generation when employed (e.g., internal node points associated with orthogonal collocation). Such methods can be used with derivatives that on independent infinite parameters, but cannot be used for ones that depend on dependent parameters.

InfiniteOpt.OrthogonalCollocation — Type

OrthogonalCollocation <: GenerativeDerivativeMethod

A DataType for storing information about orthogonal collocation method for derivative evaluation. Note that the constructor for this method is of the form:

    OrthogonalCollocation(num_nodes::Int, technique::Type{<:OCTechnique} = Labatto)

where num_nodes is total number of nodes for each collocation interval. In practice, this corresponds to num_nodes = num_internal_nodes + 2.

Fields

num_internal_nodes::Int: The number of internal collocation points (nodes) between the each support pair.

technique::Type{<:OCTechnique}: The method used to produce the points.

Missing docstring.

Missing docstring for OCQuadrature. Check Documenter's build log for details.

InfiniteOpt.Lobatto — Type

Lobatto <: OCTechnique

A quadrature method label for orthogonal collocation method that generates internal nodes between public supports using Lobatto quadrature method.

Missing docstring.

Missing docstring for NonGenerativeDerivateMethod. Check Documenter's build log for details.

InfiniteOpt.FiniteDifference — Type

FiniteDifference <: NonGenerativeDerivativeMethod

A DataType for information about finite difference method applied to a derivative evaluation.

Fields

technique::Type{<:FDTechnique}: Mathematical technqiue behind finite difference

Missing docstring.

Missing docstring for FDTechniques. Check Documenter's build log for details.

InfiniteOpt.FDForward — Type

FDForward <: FDTechnique

A technique label for finite difference method that implements a forward difference approximation.

InfiniteOpt.FDCentral — Type

FDCentral <: FDTechnique

A technique label for finite difference method that implements a central difference approximation.

InfiniteOpt.FDBackward — Type

FDBackward <: FDTechnique

A technique label for finite difference method that implements a backward difference approximation.

Methods/Macros

InfiniteOpt.@deriv

InfiniteOpt.@derivative_variable

InfiniteOpt.add_derivative

InfiniteOpt.add_derivative_supports

InfiniteOpt.all_derivatives

InfiniteOpt.build_derivative

InfiniteOpt.deriv

InfiniteOpt.derivative_argument

InfiniteOpt.derivative_method

InfiniteOpt.evaluate

InfiniteOpt.evaluate_all_derivatives!

InfiniteOpt.evaluate_derivative

InfiniteOpt.make_reduced_expr

InfiniteOpt.num_derivatives

InfiniteOpt.operator_parameter

InfiniteOpt.parameter_list

InfiniteOpt.parameter_refs

InfiniteOpt.raw_parameter_refs

InfiniteOpt.reset_start_value_function

InfiniteOpt.set_all_derivative_methods

InfiniteOpt.set_derivative_method

InfiniteOpt.set_derivative_method

InfiniteOpt.set_start_value_function

InfiniteOpt.support_label

InfiniteOpt.@deriv — Macro

@deriv(expr, pref_expr1[, ...]
       )::Union{JuMP.AbstractJuMPScalar, Float64}

The macro variant of deriv that is more efficient for expression building and enables symbolic differential operator parameter defintions via pref_exprs. Like deriv expr can be any InfiniteOpt expression and the appropriate calculus rules will applied to expr to take its derivative with respect to the indicated infinite parameters detailed by the pref_exprs. The resulting derivative expression will contain individual derivatives that were created and added to the InfiniteModel as needed. Here each pref_expr arugment can be of the form:

pref::GeneralVariableRef: An indiviudal infinite parameter reference

(pref::GeneralVariableRef)^(p::Int): An infinite parameter applied p times.

Thus, the syntax @deriv(expr, pref^2) is equivalent to @deriv(expr, pref, pref).

This will error if pref_expr is an unrecongnized syntax, no infinite parameter is given, or if any of the specified parameters are not infinite.

Example

julia> @infinite_parameter(m, t in [0, 1])
t

julia> @infinite_variable(m, x(t))
x(t)

julia> @hold_variable(m, z)
z

julia> deriv_expr = @deriv(x^2 + z, t^2)
2 ∂/∂t[∂/∂t[x(t)]]*x(t) + 2 ∂/∂t[x(t)]²

InfiniteOpt.deriv — Function

deriv(expr::JuMP.AbstractJuMPScalar, pref1::GeneralVariableRef[, ....]
      )::Union{JuMP.AbstractJuMPScalar, Float64}

Apply appropriate calculus methods to define and return the derivative expression of expr with respect to the infinite parameter(s) pref1, pref2, etc. in that respective order. This will implicilty build and add individual [Derivative`](@ref)s as appropriate. Errors if no infinite parameter is given or if the parameters are not infinite.

Example

julia> @infinite_parameter(m, t in [0, 1])
t

julia> @infinite_variable(m, x(t))
x(t)

julia> @hold_variable(m, z)
z

julia> deriv_expr = deriv(x^2 + z, t, t)
2 ∂/∂t[∂/∂t[x(t)]]*x(t) + 2 ∂/∂t[x(t)]²

InfiniteOpt.@derivative_variable — Macro

@derivative_variable(model::InfiniteModel, kw_args...)::GeneralVariableRef

Add an anonymous derivative to the model model described by the keyword arguments kw_args and returns the variable reference. Note that the argument and operator_parameter keywords are required in this case.

@derivative_variable(model::InfiniteModel, deriv_expr, var_expr,
                     kw_args...)::GeneralVariableRef

Add a derivative to model described by the expression var_expr, the keyword arguments kw_args, and the derivative expr deriv_expr. The expression deriv_expr specifies the derivative argument and operator parameter and must be of the from: d(arg_vref)/d(pref) where arg_vref is a variable/derivative/measure with an infinite dependence on the infinite parameter pref. Equivalently, the derivative can be expressed ∂(arg_vref)/∂(pref). The expression var_expr is used to define variable specific bounds and whose name is used as an alias for the derivative reference. The expression var_expr can either be (note that in the following the symbol <= can be used instead of ≤ and the symbol >=can be used instead of ≥) of the form:

var_expr creating variables described by varexpr

var_expr ≤ ub (resp. varexpr ≥ lb) creating variables described by var_expr with upper bounds given by ub (resp. lower bounds given by lb)

var_expr == value creating variables described by var_expr with fixed values given by value

lb ≤ var_expr ≤ ub or ub ≥ var_expr ≥ lb creating variables described by var_expr with lower bounds given by lb and upper bounds given by ub

Note that the preferred way to define derivatives is via @deriv which provides a more succinct way to specify more complex derivatives. However, this variable based syntax is provided as a convenient way to specify starting values/functions when needed.

The expression var_expr can be of the form:

varname creating a scalar real variable of alias name varname

varname[...] or [...] creating a container of variables.

The recognized keyword arguments in kw_args are the following:

argument: Sets the argument of the differential operator.

operator_parameter: Sets the infinite parameter the derivative is with respect to.

base_name: Sets the name prefix used to generate variable names. It corresponds to the variable name for scalar variable, otherwise, the variable names are set to base_name[...] for each index ... of the axes axes.

lower_bound: Sets the value of the variable lower bound.

upper_bound: Sets the value of the variable upper bound.

start: Sets the derivative starting value used as initial guess in optimization. This can be a single value enforced over the entire infinite domain or it can be a function that maps a support value to a scalar guess value. Note that the function arguments must match the format of parameter_refs(argument).

container: Specify the container type.

Examples

julia> @infinite_parameter(model, t in [0, 10])
t

julia> @infinite_variable(model, x(t) >= 0)
x(t)

julia> @infinite_variable(model, y[1:2](t))
2-element Array{GeneralVariableRef,1}:
 y[1](t)
 y[2](t)

julia> @derivative_variable(model, d(x)/d(t), dx <= 1, start = 0)
dx

julia> dx = @derivative_variable(model, d(x)/d(t), upper_bound = 1, base_name = "dx")
dx

julia> dx = @derivative_variable(model, upper_bound = 1, base_name = "dx",
                                  argument = x, operator_parameter = t)
dx

julia> @derivative_variable(model, d(y[i])/d(t), dx2[i = 1:2])
2-element Array{GeneralVariableRef,1}:
 dx2[1]
 dx2[2]

julia> dx = @derivative_variable(model, d(y[i])/d(t), [i = 1:2])
2-element Array{GeneralVariableRef,1}:
 ∂/∂t[y[1](t)]
 ∂/∂t[y[2](t)]
 
julia> dx = @derivative_variable(model, [i = 1:2], argument = y[i], 
                                 operator_parameter = t)
2-element Array{GeneralVariableRef,1}:
 ∂/∂t[y[1](t)]
 ∂/∂t[y[2](t)]

InfiniteOpt.build_derivative — Function

build_derivative(_error::Function, info::JuMP.VariableInfo, 
                 argument_ref::GeneralVariableRef, 
                 parameter_ref::GeneralVariableRef
                 )::Derivative

Constructs and returns a Derivative with a differential operator that depends on parameter_ref and operates on argument_ref. Variable info can also be provided to associate this derivative with bounds and a starting value function like that of infinite variables. Errors when argument_ref is not an infinite/reduced variable or derivative that depends on parameter_ref. Also, errors if such a derivative was already added to the model.

Example ```julia-repl julia> @infiniteparameter(m, t in [0, 1]); @infinitevariable(m, x(t));

julia> info = VariableInfo(false, 0, false, 0, false, 0, false, 0, false, false);

julia> buildderivative(error, info, x, t) Derivative{GeneralVariableRef}(VariableInfo{Float64,Float64,Float64,Function}(false, 0.0, false, 0.0, false, 0.0, false, startfunc, false, false), true, x(t), t) ````

InfiniteOpt.add_derivative — Function

add_derivative(model::InfiniteModel, d::Derivative, 
               [name::String = ""])::GeneralVariableRef

Adds a derivative d to model and returns a GeneralVariableRef that points to it. Errors if the derivative dependencies do not belong to model. Note that d should be built using build_derivative to avoid nuance internal errors.

Example

julia> @infinite_parameter(m, t in [0, 1]); @infinite_variable(m, x(t));

julia> info = VariableInfo(false, 0, false, 0, false, 0, false, 0, false, false);

julia> d = build_derivative(error, info, x, t);

julia> dref = add_derivative(m, d)
∂/∂t[x(t)]

InfiniteOpt.derivative_argument — Method

derivative_argument(dref::DerivativeRef)::GeneralVariableRef

Returns the infinite variable/derivative reference that is the input the differential operator (i.e., the dependent variable of the derivative).

Example

julia> derivative_argument(dref) 
x(t)

InfiniteOpt.operator_parameter — Method

operator_parameter(dref::DerivativeRef)::GeneralVariableRef

Returns the infinite parameter reference that is what the differential operator is operating with respect to (i.e., the independent variable of the derivative).

Example

julia> operator_parameter(dref) 
t

InfiniteOpt.derivative_method — Method

derivative_method(dref::DerivativeRef)::AbstractDerivativeMethod

Returns the evaluation method employed by dref that determines the numerical computation scheme that will be used to evaluate the derivative. Note that this is set on by the infinite parameter with respect to which the derivative is defined.

Example

julia> derivative_method(dref) 
FiniteDifference()

InfiniteOpt.raw_parameter_refs — Method

raw_parameter_refs(dref::DerivativeRef)::VectorTuple{GeneralVariableRef}

Return the raw VectorTuple of the parameter references that dref depends on. This is primarily an internal method where parameter_refs is intended as the preferred user function.

InfiniteOpt.parameter_refs — Method

parameter_refs(dref::DerivativeRef)::Tuple

Return the parameter references associated with the infinite derivative dref. This is formatted as a Tuple of containing the parameter references as they inputted to define dref.

Example

julia> parameter_refs(deriv)
(t,)

InfiniteOpt.parameter_list — Method

parameter_list(dref::DerivativeRef)::Vector{GeneralVariableRef}

Return a vector of the parameter references that dref depends on. This is primarily an internal method where parameter_refs is intended as the preferred user function.

InfiniteOpt.set_start_value_function — Method

set_start_value_function(dref::DerivativeRef,
                         start::Union{Real, Function})::Nothing

Set the start value function of dref. If start::Real then a function is generated to such that the start value will be start for the entire infinite domain. If start::Function then this function should map to a scalar start value given a support value arguments matching the format of the parameter elements in parameter_refs(dref).

Example

julia> set_start_value_function(dref, 1) # all start values will be 1

julia> set_start_value_function(dref, my_func) # each value will be made via my_func

InfiniteOpt.reset_start_value_function — Method

reset_start_value_function(dref::DerivativeRef)::Nothing

Remove the existing start value function and return to the default. Generally, this is triggered by deleting an infinite parameter that dref depends on.

Example

julia> reset_start_value_function(dref)

InfiniteOpt.num_derivatives — Function

num_derivatives(model::InfiniteModel)::Int

Returns the number of derivatives that have been defined in model. Note that nested derivatives will be counted in accordance with their components (e.g., $\frac{d^2 x(t)}{dt^2} =$\frac{d}{dt}\left(\frac{d x(t)}{dt} \right)`` will count as 2 derivatives.)

Example

julia> num_derivatives(model)
12

InfiniteOpt.all_derivatives — Function

all_derivatives(model::InfiniteModel)::Vector{GeneralVariableRef}

Returns a list of all the individual derivatives stored in model.

Example

julia> all_derivatives(model)
4-element Array{GeneralVariableRef,1}:
 TODO PUT OUTPUT HERE ONCE PRINTING IS DONE

InfiniteOpt.set_derivative_method — Method

set_derivative_method(pref::IndependentParameterRef, 
                      method::AbstractDerivativeMethod)::Nothing

Specfies the desired derivative evaluation method method for derivatives that are taken with respect to pref. Any internal supports exclusively associated with the previous method will be deleted.

Example

julia> set_derivative_method(d, OrthogonalCollocation(2))

InfiniteOpt.set_derivative_method — Method

set_derivative_method(pref::DependentParameterRef, 
                      method::NonGenerativeDerivativeMethod)::Nothing

Specfies the desired derivative evaluation method method for derivatives that are taken with respect to pref. Errors if method is generative (i.e., it requires the definition of additional supports)

Example

julia> set_derivative_method(d, FiniteDifference())

InfiniteOpt.set_all_derivative_methods — Function

set_all_derivative_methods(model::InfiniteModel, 
                           method::AbstractDerivativeMethod)::Nothing

Sets the desired evaluation method method for all the derivatives currently added to model. Note that this is done with respect to the infinite parameters. Errors if a generative method is specified and the model contains dependent parameters.

Example

julia> set_all_derivative_methods(model, OrthogonalCollocation(2))

InfiniteOpt.evaluate — Method

evaluate(dref::DerivativeRef)::Nothing

Numerically evaluate dref by computing its auxiliary derivative constraints (e.g., collocation equations) and add them to the model. For normal usage, it is recommended that this method not be called directly and instead have TranscriptionOpt handle these equations, since preemptive evaluation can lead to invalid relations if the support structure is modified. Errors if evaluate_derivative is not defined for the derivative method employed.

Example

julia> m = InfiniteModel(); @infinite_parameter(m, t in [0,2]); @infinite_variable(m, T(t));

julia> dref = @deriv(T,t)
∂/∂t[T(t)]

julia> add_supports(t, [0, 0.5, 1, 1.5, 2])

julia> evaluate(dref)

julia> print(m)
Feasibility
Subject to
 0.5 ∂/∂t[T(t)](0) - T(0.5) + T(0) = 0.0
 ∂/∂t[T(t)](0.5) - T(1) + T(0) = 0.0
 ∂/∂t[T(t)](1) - T(1.5) + T(0.5) = 0.0
 ∂/∂t[T(t)](1.5) - T(2) + T(1) = 0.0
 0.5 ∂/∂t[T(t)](2) - T(2) + T(1.5) = 0.0

InfiniteOpt.evaluate_all_derivatives! — Function

evaluate_all_derivatives!(model::InfiniteModel)::Nothing

Evaluate all the derivatives in model by adding the corresponding auxiliary equations to model. See evaluate for more information.

Example

julia> m = InfiniteModel();

julia> @infinite_parameter(m, t in [0,2], supports = [0, 1, 2]);

julia> @infinite_parameter(m, x in [0,1], supports = [0, 0.5, 1]);

julia> @infinite_variable(m, T(x, t));

julia> dref1 = @deriv(T, t); dref2 = @deriv(T, x^2);

julia> evaluate_all_derivatives!(m)

julia> print(m)
Feasibility
Subject to
 ∂/∂t[T(x, t)](x, 0) - T(x, 1) + T(x, 0) = 0.0, ∀ x ∈ [0, 1]
 2 ∂/∂t[T(x, t)](x, 1) - T(x, 2) + T(x, 0) = 0.0, ∀ x ∈ [0, 1]
 ∂/∂t[T(x, t)](x, 2) - T(x, 2) + T(x, 1) = 0.0, ∀ x ∈ [0, 1]
 0.5 ∂/∂x[T(x, t)](0, t) - T(0.5, t) + T(0, t) = 0.0, ∀ t ∈ [0, 2]
 ∂/∂x[T(x, t)](0.5, t) - T(1, t) + T(0, t) = 0.0, ∀ t ∈ [0, 2]
 0.5 ∂/∂x[T(x, t)](1, t) - T(1, t) + T(0.5, t) = 0.0, ∀ t ∈ [0, 2]
 0.5 ∂/∂x[∂/∂x[T(x, t)]](0, t) - ∂/∂x[T(x, t)](0.5, t) + ∂/∂x[T(x, t)](0, t) = 0.0, ∀ t ∈ [0, 2]
 ∂/∂x[∂/∂x[T(x, t)]](0.5, t) - ∂/∂x[T(x, t)](1, t) + ∂/∂x[T(x, t)](0, t) = 0.0, ∀ t ∈ [0, 2]
 0.5 ∂/∂x[∂/∂x[T(x, t)]](1, t) - ∂/∂x[T(x, t)](1, t) + ∂/∂x[T(x, t)](0.5, t) = 0.0, ∀ t ∈ [0, 2]

InfiniteOpt.evaluate_derivative — Function

evaluate_derivative(dref::GeneralVariableRef, 
                    method::AbstractDerivativeMethod,
                    write_model::JuMP.AbstractModel)::Vector{JuMP.AbstractJuMPScalar}

Build expressions for derivative dref evaluated in accordance with method. The expressions are of the form lhs - rhs, where lhs is a function of derivatives evaluated at some supports for certain infinite parameter, and rhs is a function of the derivative arguments evaluated at some supports for certain infinite parameter. For example, for finite difference methods at point t = 1, lhs is Δt * ∂/∂t[T(1)], and rhs could be T(1+Δt) - T(1) in case of forward difference mode. This is intended as a helper function for evaluate, which will take the the expressions generated by this method and generate constraints that approximate the derivative values by setting the expressions as 0. However, one can extend this function to encode custom methods for approximating derivatives.

InfiniteOpt.support_label — Method

support_label(method::GenerativeDerivativeMethod)

Return the support label associated with method if there is one, errors otherwise. This should be extended for any GenerativeDerivativeMethod.

Missing docstring.

Missing docstring for InfiniteOPt.generate_derivative_supports. Check Documenter's build log for details.

InfiniteOpt.add_derivative_supports — Method

add_derivative_supports(pref::Union{IndependentParameterRef, DependentParameterRef})::Nothing

Add any supports pref that are needed for derivative evaluation. This is intended as a helper method for derivative evaluation and depends generate_derivative_supports which will need to be extended for user-defined derivative methods that generate supports. In such cases, it is necessary to also extend support_label Errors if such is not defined for the current derivative method associated with pref.

InfiniteOpt.make_reduced_expr — Function

make_reduced_expr(vref::GeneralVariableRef, pref::GeneralVariableRef, 
                  support::Float64, write_model::Union{InfiniteModel, JuMP.Model})

Given the argument variable vref and the operator parameter pref from a derivative, build and return the reduced expression in accordance to the support support with respect to pref. New point/reduced variables will be written to write_model. This is solely intended as a helper function for derivative evaluation.