Finite Parameters

A guide and manual to the definition and use of finite parameters in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

Often a mathematical model needs to be optimized several times in accordance with a set of fixed parameter values. In such cases, it is typically preferable to modify these values in place without having to redefine the entire model. This ability is provided in InfiniteOpt via @finite_parameter which permits users to define finite parameters whose values can later be modified as needed. Furthermore, at the optimization step these parameters are replaced with their numeric values. Thus, not adding unnecessary decision variables as is typically done in JuMP models using JuMP.fix on placeholder variables.

Basic Usage

Once an InfiniteModel model has been defined we can add a finite parameter via @finite_parameter. For example, let's define a maximum cost parameter called max_cost with an initial value of 42:

julia> @finite_parameter(model, max_cost, 42)
max_cost

Notice that a Julia variable called max_cost is returned that contains a ParameterRef that points to the finite parameter we have just created. An array of parameters can also be defined following standard JuMP macro syntax:

julia> values = [2, 3.2, 1];

julia> @finite_parameter(model, params[i = 1:3], values[i])
3-element Array{ParameterRef,1}:
 params[1]
 params[2]
 params[3]

The @finite_parameter macro emulates all typical JuMP functionality and can define anonymous parameters, use JuMP containers and more. We refer to its documentation below to learn more. Once a finite parameter is defined the corresponding ParameterRef can be used in expressions, objectives, measures, and constraints just like infinite parameters.

The value of a finite parameter can be checked using JuMP.value and can modified using JuMP.set_value. For example, let's update the value of max_cost to be now be 10.2:

julia> value(max_cost)
42

julia> set_value(max_cost, 10.2)

julia> value(max_cost)
10.2

Advanced Details

The ability to implement finite parameters simply stems from InfiniteOpt's framework for infinite parameters. In reality finite parameters are simply infinite parameters defined with an IntervalSet whose lower and upper bounds equal to the desired value. Thus, @finite_parameter simply processes the names, values, and other arguments and then makes the appropriate @infinite_parameter call. For example, the following calls are equivalent:

julia> @finite_parameter(model, param[1:2], 42)
2-element Array{ParameterRef,1}:
 param[1]
 param[2]

julia> @infinite_parameter(model, param[1:2] in [42, 42], supports = [42])
2-element Array{ParameterRef,1}:
 param[1]
 param[2]

Furthermore, the JuMP.value checks the bound value (also ensuring the bounds match) and JuMP.set_value updates the the IntervalSet and updates the support value.

Methods/Macros

InfiniteOpt.@finite_parameter

InfiniteOpt.is_finite_parameter

JuMP.set_value

JuMP.value

InfiniteOpt.@finite_parameter — Macro

@finite_parameter(model::InfiniteModel, value)

Define and add an anonymous finite parameter to model and return its parameter reference. Its value is equal to value.

    @finite_parameter(model::InfiniteModel, param_expr, value_expr)

Define and add a finite parameter(s) to model and return appropriate parameter reference(s). The parameter(s) has/have value(s) indicated by the value_expr. The expression param_expr can be of the form:

paramname creating a scalar parameter of name paramname

paramname[...] or [...] creating a container of parameters

The expression value_expr simply expresses the value of the parameter(s). This is typically a number but could be an array indexed using an index defined in param_expr.

The recognized keyword arguments in kw_args are the following:

base_name: Sets the name prefix used to generate parameter names. It corresponds to the parameter name for scalar parameter, otherwise, the parameter names are set to base_name[...] for each index ... of the axes axes.

container: Specify the container type.

Examples

julia> par = @finite_parameter(model, 2)
noname

julia> vals = [3, 2];

julia> pars = @finite_parameter(model, [i = 1:2], vals[i], base_name = "par")
2-element Array{ParameterRef,1}:
 par[1]
 par[2]

julia> @finite_parameter(model, par2, 42)
par2

InfiniteOpt.is_finite_parameter — Function

is_finite_parameter(pref::ParameterRef)::Bool

Return a Bool indicating if pref is a finite parameter.

Example

julia> is_finite_parameter(cost)
true

JuMP.value — Method

JuMP.value(pref::ParameterRef)::Number

Return the value of pref so long as it is a finite parameter. Errors if it is an infinite parameter.

Example

julia> value(cost)
42

JuMP.set_value — Method

JuMP.set_value(pref::ParameterRef, value::Number)

Set the value of pref so long as it is a finite parameter. Errors if it is an infinite parameter.

Example

julia> set_value(cost, 27)

julia> value(cost)
27