Infinite Models

A guide and manual for initializing infinite models and understanding how they work. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Overview

Infinite models are expressed via the InfiniteModel datatype which is at the core of InfiniteOpt. These model objects are designed to emulate the behavior of Model objects in JuMP. These data objects store the parameters, variables, measures, objective, constraints, and all other data used in InfiniteOpt. This differs from JuMP models which store such information in a MathOptInterface model backend.

Basic Usage

Infinite models can be initialized with no arguments by default:

julia> using InfiniteOpt

julia> model = InfiniteModel()
An InfiniteOpt Model
Feasibility problem with:
Finite Parameters: 0
Infinite Parameters: 0
Variables: 0
Derivatives: 0
Measures: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

The optimizer that will be used to solve the model can also be specified at model definition:

julia> using InfiniteOpt, Ipopt

julia> model = InfiniteModel(Ipopt.Optimizer)
An InfiniteOpt Model
Feasibility problem with:
Finite Parameters: 0
Infinite Parameters: 0
Variables: 0
Derivatives: 0
Measures: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Ipopt

Note that any optimizer currently supported by JuMP v0.19.0 or newer is supported for use in InfiniteOpt. For completeness, the table of currently supported optimizers is provided below in Supported Optimizers.

We can also specify optimizer attributes via optimizer_with_attributes which allows us to append as many attributes as we like, for example:

julia> using InfiniteOpt, Ipopt

julia> model = InfiniteModel(optimizer_with_attributes(Ipopt.Optimizer,
                                                       "output_level" => 0))
An InfiniteOpt Model
Feasibility problem with:
Finite Parameters: 0
Infinite Parameters: 0
Variables: 0
Derivatives: 0
Measures: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Ipopt

Now you have an initialized InfiniteModel that is ready for your mathematical model to be defined and optimized!

Advanced Definition Information

As you may have noticed in the above examples, InfiniteModels contain an optimizer model backend which simply corresponds to a JuMP Model that will be used to store and optimize the reformulation of the infinite mathematical model stored in InfiniteModel. It also will contain a mapping between its optimization model and that of the InfiniteModel (e.g., a mapping between the variables and constraints). By default, InfiniteModels use a TranscriptionModel optimizer model backend which will store a transcribed (discretized) version of the infinite model. More information on the internal use of TranscriptionModels is provided in Model Transcription.

All the arguments used with the InfiniteModel constructor (e.g., the optimizer) are simply passed on and stored in the optimizer model backend. Thus, any argument supported by JuMP.Model can be passed on to the optimizer model by including it in the InfiniteModel constructor. For example, we can specify the caching_mode keyword argument in the InfiniteModel call to use in the definition of the optimizer model:

julia> using InfiniteOpt, Ipopt

julia> model = InfiniteModel(Ipopt.Optimizer,
                             caching_mode = MOIU.MANUAL)
An InfiniteOpt Model
Feasibility problem with:
Finite Parameters: 0
Infinite Parameters: 0
Variables: 0
Derivatives: 0
Measures: 0
Optimizer model backend information:
Model mode: MANUAL
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Ipopt

Notice that the model mode of the optimizer model is now MANUAL.

Moreover, alternative optimizer model types (i.e., not a TranscriptionModel) can be specified via the OptimizerModel keyword argument when initializing the InfiniteModel. Thus, to redundantly specify a TranscriptionModel we would call:

julia> using InfiniteOpt

julia> model = InfiniteModel(OptimizerModel = TranscriptionModel)
An InfiniteOpt Model
Feasibility problem with:
Finite Parameters: 0
Infinite Parameters: 0
Variables: 0
Derivatives: 0
Measures: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

More information on implementing custom optimizer models is located on the Extensions page.

Supported Optimizers

InfiniteOpt can use any optimizer that is supported by JuMP v0.19.0 or newer (i.e., has a MathOptInterface implementation). Please refer to JuMP's current solver documentation to learn what solvers are supported and how to install them.

Object Dictionaries

Like JuMP.Models, InfiniteModels register the name symbols of macro defined objects. This enables us to access such objects by indexing the InfiniteModel with the appropriate symbol. This is particularly useful for function defined models. For example:

julia> function make_model(num_supports)
        model = InfiniteModel()
        @infinite_parameter(model, t ∈ [0, 10], num_supports = num_supports)
        @infinite_variable(model, y(t) >= 42)
        @objective(model, Min, ∫(y, t))
        return model
       end
make_model (generic function with 1 method)

julia> model1 = make_model(2); model2 = make_model(4);

julia> y1 = model1[:y]
y(t)

Note that when macro defined objects are deleted from an InfiniteModel that the corresponding symbols in the object dictionary are not removed by default. This can be accomplished by use of JuMP.unregister (please click on its link for usage information).

Datatypes

InfiniteOpt.InfiniteModel — Type

InfiniteModel <: JuMP.AbstractModel

A DataType for storing all of the mathematical modeling information needed to model an optmization problem with an infinite-dimensional decision space.

Fields

independent_params::MOIUC.CleverDict{IndependentParameterIndex, ScalarParameterData{IndependentParameter}}: The independent parameters and their mapping information.

dependent_params::MOIUC.CleverDict{DependentParametersIndex, MultiParameterData}: The dependent parameters and their mapping information.

finite_params::MOIUC.CleverDict{FiniteParameterIndex, ScalarParameterData{FiniteParameter}}: The finite parameters and their mapping information.

name_to_param::Union{Dict{String, AbstractInfOptIndex}, Nothing}: Field to help find a parameter given the name.

last_param_num::Int: The last parameter number to be used.

param_object_indices::Vector{Union{IndependentParameterIndex, DependentParametersIndex}}: The collection of parameter object indices in creation order.

param_functions::MOIUC.CleverDict{ParameterFunctionIndex, ParameterFunctionData{ParameterFunction}}: The infinite parameter functions and their mapping information.

infinite_vars::MOIUC.CleverDict{InfiniteVariableIndex, <:VariableData{<:InfiniteVariable}}: The infinite variables and their mapping information.

semi_infinite_vars::MOIUC.CleverDict{SemiInfiniteVariableIndex, <:VariableData{<:SemiInfiniteVariable}}: The semi-infinite variables and their mapping information.

point_vars::MOIUC.CleverDict{PointVariableIndex, <:VariableData{<:PointVariable}}: The point variables and their mapping information.

finite_vars::MOIUC.CleverDict{FiniteVariableIndex, <:VariableData{<:FiniteVariable}}: The finite variables and their mapping information.

name_to_var::Union{Dict{String, AbstractInfOptIndex}, Nothing}: Field to help find a variable given the name.

has_finite_var_bounds::Bool: Does any variable have parameter bounds?

derivatives::MOIUC.CleverDict{DerivativeIndex, <:VariableData{<:Derivative}}: The derivatives and their mapping information.

deriv_lookup::Dict{<:Tuple, DerivativeIndex}: Map derivative variable-parameter pairs to a derivative index to prevent duplicates.

measures::MOIUC.CleverDict{MeasureIndex, MeasureData}: The measures and their mapping information.

integral_defaults::Dict{Symbol}: The default keyword arguments for integral.

constraints::MOIUC.CleverDict{ConstraintIndex, ConstraintData}: The constraints and their mapping information.

name_to_constr::Union{Dict{String, ConstraintIndex}, Nothing}: Field to help find a constraint given the name.

objective_sense::MOI.OptimizationSense: Objective sense.

objective_function::JuMP.AbstractJuMPScalar: Finite scalar function.

objective_has_measures::Bool: Does the objective contain measures?

obj_dict::Dict{Symbol, Any}: Store Julia symbols used with InfiniteModel

optimizer_constructor: MOI optimizer constructor (e.g., Gurobi.Optimizer).

optimizer_model::JuMP.Model: Model used to solve InfiniteModel

ready_to_optimize::Bool: Is the optimizer_model up to date.

ext::Dict{Symbol, Any}: Store arbitrary extension information.

InfiniteOpt.AbstractDataObject — Type

AbstractDataObject

An abstract type for DataTypes that store core variable DataTypes and their model specific information (e.g., dependency mappings). These are what are stored in the InfiniteModel CleverDicts.

InfiniteOpt.AbstractInfOptIndex — Type

AbstractInfOptIndex

An abstract type for all index objects used in InfiniteOpt.

InfiniteOpt.ObjectIndex — Type

ObjectIndex <: AbstractInfOptIndex

An abstract type for indices of objects stored in MOI.Utilities.CleverDicts.

Methods

InfiniteOpt.InfiniteModel — Method

InfiniteModel([optimizer_constructor];
              [OptimizerModel::Function = TranscriptionModel,
              caching_mode::MOIU.CachingOptimizerMode = MOIU.AUTOMATIC,
              bridge_constraints::Bool = true, optimizer_model_kwargs...])

Return a new infinite model where an optimizer is specified if an optimizer_constructor is given. The optimizer can also later be set with the JuMP.set_optimizer call. By default the optimizer_model data field is initialized with a TranscriptionModel, but a different type of model can be assigned via set_optimizer_model as can be required by extensions.

Example

julia> using InfiniteOpt, JuMP, Ipopt;

julia> model = InfiniteModel()
An InfiniteOpt Model
Feasibility problem with:
Finite Parameters: 0
Infinite Parameters: 0
Variables: 0
Measures: 0
Derivatives: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

julia> model = InfiniteModel(Ipopt.Optimizer)
An InfiniteOpt Model
Feasibility problem with:
Finite Parameters: 0
Infinite Parameters: 0
Variables: 0
Measures: 0
Derivatives: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Ipopt

JuMP.object_dictionary — Method

JuMP.object_dictionary(model::InfiniteModel)::Dict{Symbol, Any}

Return the dictionary that maps the symbol name of a macro defined object (e.g., a parameter, variable, or constraint) to the corresponding object. Objects are registered to a specific symbol in the macros. For example, @finite_variable(model, x[1:2, 1:2]) registers the array of variables x to the symbol :x.

InfiniteOpt.has_internal_supports — Function

has_internal_supports(pref::Union{IndependentParameterRef, DependentParameterRef})::Bool

Indicate if pref has internal supports that will be hidden from the user by default.

InfiniteOpt.has_internal_supports(prefs; [kwargs...])

Define InfiniteOpt.has_internal_supports for general variable references. It relies on InfiniteOpt.has_internal_supports being defined for the underlying DispatchVariableRef, otherwise an ArugmentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

Base.empty! — Method

Base.empty!(model::InfiniteModel)::InfiniteModel

Clear out model of everything except the optimizer information and return the cleared model.