Infinite Models

Infinite Models

A guide and manual for initializing infinite models and understanding how they work.

Overview

Infinite models are expressed via the InfiniteModel datatype which is at the core of InfiniteOpt. These model objects are designed to emulate the behavior of Model objects in JuMP. These data objects store the parameters, variables, measures, objective, constraints, and all other data used in InfiniteOpt. This differs from JuMP models which store such information in a MathOptInterface model backend.

Basic Usage

Infinite models can be initialized with no arguments by default:

julia> using InfiniteOpt, JuMP

julia> model = InfiniteModel()
An InfiniteOpt Model
Feasibility problem with:
Variables: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

The optimizer that will be used to solve the model can also be specified at model definition using JuMP's with_optimizer constructor:

julia> using InfiniteOpt, JuMP, Gurobi

julia> model = InfiniteModel(with_optimizer(Gurobi.Optimizer, OutputFlag = 0))
An InfiniteOpt Model
Feasibility problem with:
Variables: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Gurobi

Note that any optimizer currently supported by JuMP v0.19.0 or newer is supported for use in InfiniteOpt. For completeness, the table of currently supported optimizers is provided below in Supported Optimizers.

Now you have an initialized InfiniteModel that is ready for your mathematical model to be defined and optimized!

Advanced Definition Information

As you may have noticed in the above examples, InfiniteModels contain an optimizer model backend which simply corresponds to a JuMP Model that will be used to store and optimize the reformulation of the infinite mathematical model stored in InfiniteModel. It also will contain a mapping between its optimization model and that of the InfiniteModel (e.g., a mapping between the variables and constraints). By default, InfiniteModels use a TranscriptionModel optimizer model backend which will store a transcribed (discretized) version of the infinite model. More information on the internal use of TranscriptionModels is provided in Model Transcription.

All the arguments used with the InfiniteModel constructor (e.g., the optimizer) are simply passed on and stored in the optimizer model backend. Thus, any argument supported by JuMP.Model can be passed on to the optimizer model by including it in the InfiniteModel constructor. For example, we can specify the caching_mode keyword argument in the InfiniteModel call to use in the definition of the optimizer model:

julia> using InfiniteOpt, JuMP, Gurobi, MathOptInterface

julia> const MOIU = MathOptInterface.Utilities

julia> model = InfiniteModel(with_optimizer(Gurobi.Optimizer),
                             caching_mode = MOIU.MANUAL)
An InfiniteOpt Model
Feasibility problem with:
Variables: 0
Optimizer model backend information:
Model mode: MANUAL
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Gurobi

Notice that the model mode of the optimizer model is now MANUAL.

Supported Optimizers

InfiniteOpt can use any optimizer that is supported by JuMP v0.19.0 or newer. In spirit of providing complete documentation, the table of optimizers currently supported by JuMP.jl is provided below. This information comes directly from their documentation pages. Please refer to JuMP documentation for additional optimizer information on installation and use.

Solver	Julia Package	License	Supports

Artelys Knitro	KNITRO.jl	Comm.	LP, MILP, SOCP, MISOCP, NLP, MINLP

Cbc	Cbc.jl	EPL	MILP

CDCS	CDCS.jl	GPL	LP, SOCP, SDP

Clp	Clp.jl	EPL	LP

COSMO	COSMO.jl	Apache	LP, QP, SOCP, SDP

CPLEX	CPLEX.jl	Comm.	LP, MILP, SOCP, MISOCP

CSDP	CSDP.jl	EPL	LP, SDP

ECOS	ECOS.jl	GPL	LP, SOCP

FICO Xpress	Xpress.jl	Comm.	LP, MILP, SOCP, MISOCP

GLPK	GLPK.jl	GPL	LP, MILP

Gurobi	Gurobi.jl	Comm.	LP, MILP, SOCP, MISOCP

Ipopt	Ipopt.jl	EPL	LP, QP, NLP

Juniper	Juniper.jl	MIT	MISOCP, MINLP

MOSEK	MosekTools.jl	Comm.	LP, MILP, SOCP, MISOCP, SDP

OSQP	OSQP.jl	Apache	LP, QP

ProxSDP	ProxSDP.jl	MIT	LP, SOCP, SDP

SCIP	SCIP.jl	ZIB	MILP, MINLP

SCS	SCS.jl	MIT	LP, SOCP, SDP

SDPA	SDPA.jl	GPL	LP, SDP

SeDuMi	SeDuMi.jl	GPL	LP, SOCP, SDP

Where:

LP = Linear programming

QP = Quadratic programming

SOCP = Second-order conic programming (including problems with convex quadratic constraints and/or objective)

MILP = Mixed-integer linear programming

NLP = Nonlinear programming

MINLP = Mixed-integer nonlinear programming

SDP = Semidefinite programming

MISDP = Mixed-integer semidefinite programming

You may also use AmplNLWriter to access solvers that support the nl format. Such solvers include Bonmin and Couenne. See a more complete list here.

Datatypes

InfiniteOpt.InfiniteModel — Type.

InfiniteModel <: JuMP.AbstractModel

A DataType for storing all of the mathematical modeling information needed to model an optmization problem with an infinite dimensional decision space.

Fields

next_meas_index::Int Index - 1 of next measure.

measures::Dict{Int, Measure} Measure indices to measure datatypes.

meas_to_name::Dict{Int, String} Measure indices to names.

meas_to_constrs::Dict{Int, Vector{Int}} Measure indices to dependent constraint indices.

meas_to_meas::Dict{Int, Vector{Int}} Measure indices to dependent measure indices.

meas_in_objective::Dict{Int, Bool} Measure indices to if used in objective.

next_param_index::Int Index - 1 of next infinite parameter.

next_param_id::Int Index - 1 of the next infinite parameter group.

params::Dict{Int, InfOptParameter} Infinite parameter indices to parameter datatype.

param_to_name::Dict{Int, String} Infinite parameter indices to names.

name_to_param::Union{Dict{String, Int}, Nothing} Names to infinite parameters.

param_to_group_id::Dict{Int, Int} Infinite parameter indices to group IDs.

param_to_constrs::Dict{Int, Vector{Int}} Infinite parameter indices to list of dependent constraint indices.

param_to_meas::Dict{Int, Vector{Int}} Infinite parameter indices to list of dependent measure indices.

param_to_vars::Dict{Int, Vector{Int}} Infinite parameter indices to list of dependent variable indices.

next_var_index::Int Index - 1 of next variable index.

vars::Dict{Int, Dict{Int, Union{InfOptVariable, ReducedVariable}} Variable indices to variable datatype.

var_to_name::Dict{Int, String} Variable indices to names.

name_to_var::Union{Dict{String, Int}, Nothing} Variable names to indices.

var_to_lower_bound::Dict{Int, Int} Variable indices to lower bound index.

var_to_upper_bound::Dict{Int, Int} Variable indices to upper bound index.

var_to_fix::Dict{Int, Int} Variable indices to fix index.

var_to_zero_one::Dict{Int, Int} Variable indices to binary index.

var_to_integrality::Dict{Int, Int} Variable indices to integer index.

var_to_constrs::Dict{Int, Vector{Int}} Variable indices to dependent constraint indices.

var_to_meas::Dict{Int, Vector{Int}} Variable indices to dependent measure indices.

var_in_objective::Dict{Int, Bool} Variable indices to if used in objective.

infinite_to_points::Dict{Int, Vector{Int}} Infinite variable indices to dependent point variable indices.

infinite_to_reduced::Dict{Int, Vector{Int}} Infinite variable indices to dependent reduced variable indices.

reduced_to_constrs::Dict{Int, Vector{Int}} Reduced variable indices to dependent constraint indices.

reduced_to_meas::Dict{Int, Vector{Int}} Reduced variable indices to dependent measure indices.

reduced_info::Dict{Int, AbstractReducedInfo} Reduced variable indices to reduced variable information.

next_constr_index::Int Index - 1 of next constraint.

constrs::Dict{Int, JuMP.AbstractConstraint} Constraint indices to constraint datatypes.

constr_to_name::Dict{Int, String} Constraint indices to names.

name_to_constr::Union{Dict{String, Int}, Nothing} Constraint names to indices.

constr_in_var_info::Dict{Int, Bool} Constraint indices to if related to variable information constraints.

objective_sense::MOI.OptimizationSense Objective sense.

objective_function::JuMP.AbstractJuMPScalar Finite scalar function.

obj_dict::Dict{Symbol, Any} Store Julia symbols used with InfiniteModel

optimizer_factory::Union{JuMP.OptimizerFactory, Nothing} Optimizer information.

optimizer_model::JuMP.Model Model used to solve InfiniteModel

ready_to_optimize::Bool Is the optimizer_model up to date.

source

Methods

InfiniteOpt.InfiniteModel — Method.

InfiniteModel([optimizer_factory::JuMP.OptimizerFactory];
              [caching_mode::MOIU.CachingOptimizerMode = MOIU.AUTOMATIC,
              bridge_constraints::Bool = true])

Return a new infinite model where an optimizer is specified if an optimizer_factory is given via JuMP.with_optimizer. The optimizer can also later be set with the JuMP.optimize! call. By default the optimizer_model data field is initialized with a TranscriptionModel, but a different type of model can be assigned via set_optimizer_model as can be required by extensions.

Example

julia> model = InfiniteModel()
An InfiniteOpt Model
Feasibility problem with:
Variables: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

julia> model = InfiniteModel(with_optimizer(Gurobi.Optimizer))
An InfiniteOpt Model
Feasibility problem with:
Variables: 0
Optimizer model backend information:
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Gurobi

source