Optimization

A guide and manual for optimizing (solving) InfiniteOpt models. The Methods section at the end comprise the manual, and the above sections comprise the guide.

Overview

Fundamentally, we seek to optimize a given infinite optimization model that we have defined and this is the very reason why InfiniteOpt was created. Thus, InfiniteOpt offers a general and intuitive platform to do just this. This is made up of transforming the InfiniteModel into a standard optimization problem stored as a JuMP.Model (referred to as the optimizer_model) that is then optimized via a compatible optimizer. By default this is done via a TranscriptionModel as described on the previous page. However, user-defined reformulation strategies can readily be implemented as described in the Optimizer Models section on the extensions page.

Basic Usage

For most users, optimize! is the method required to optimize an InfiniteModel. This is exactly analogous to that of any JuMP.Model and is designed to provide a similar user experience. Let's first define an InfiniteModel with an appropriate optimizer:

julia> using InfiniteOpt, JuMP, Ipopt;

julia> model = InfiniteModel(with_optimizer(Ipopt.Optimizer, print_level = 0));

julia> @infinite_parameter(model, t in [0, 10], num_supports = 10);

julia> @infinite_variable(model, x(t) >= 0);

julia> @hold_variable(model, z >= 0);

julia> @objective(model, Min, 2z);

julia> @constraint(model, z >= x);

julia> @BDconstraint(model, t == 0, x == 42);

julia> print(model)
Min 2 z
Subject to
 x(t) ≥ 0.0
 z ≥ 0.0
 z - x(t) ≥ 0.0
 x(t) = 42.0, ∀ t = 0
 t ∈ [0, 10]

Now we optimize the model using optimize!:

julia> optimize!(model)

julia> termination_status(model)
LOCALLY_SOLVED::TerminationStatusCode = 4

Now our model has been solved and we can query the solution. How to query the solution is explained on the Results page.

If no optimizer has been specified for the InfiniteModel, one can be provided via set_optimizer:

julia> set_optimizer(model, with_optimizer(Ipopt.Optimizer))

A number of methods also exist to adjust the optimizer settings such as suppressing output. This is explained below in the Optimizer Settings section.

Optimizer Models

As discussed previously, InfiniteModels contain an optimizer_model field which stores a transformed finite version of the model in a JuMP.Model that contains a data struct (that stores a mapping between the transformed model and the infinite model) in the Model.ext dictionary with an associated key. By default a JuMP.Model using TranscriptionData stored under the key :TransData is used and is referred to as a TranscriptionModel. The optimizer model is then what is used to optimize the infinite model and it provides the information exacted by solution queries mapped backed back to the infinite model using the mapping data structure.

The process for optimizing an InfiniteModel is summarized in the following steps:

fully define the InfiniteModel

build the optimizer model via build_optimizer_model!

optimize the optimizer_model via optimize!.

Here build_optimizer_model! creates a reformulated finite version of the InfiniteModel, stores it in InfiniteModel.optimizer_model via set_optimizer_model, and indicates that the optimizer model is ready via set_optimizer_model_ready. These steps are all automated when optimize! is invoked on the InfiniteModel.

The optimizer_model can be queried/extracted at any time from an InfiniteModel via optimizer_model. For example, let's extract the optimizer model from the example above in the basic usage section:

julia> trans_model = optimizer_model(model)
A JuMP Model
Minimization problem with:
Variables: 11
Objective function type: GenericAffExpr{Float64,VariableRef}
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.GreaterThan{Float64}`: 10 constraints
`VariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 11 constraints
Model mode: AUTOMATIC
CachingOptimizer state: ATTACHED_OPTIMIZER
Solver name: Ipopt

The purpose of this optimizer_model structure is to readily enable user-defined reformulation extensions (e.g., using polynomial chaos expansion theory). However, this is all handled behind the scenes such that most users can interact with InfiniteModels like any JuMP.Model.

Optimizer Settings

A few optimizer settings can be set in a consistent way agnostic of particular solver keywords. One such setting is that of suppressing and unsuppressing optimizer verbose output. This is accomplished via set_silent and unset_silent. The syntax is exemplified below:

julia> set_silent(model)
true

julia> unset_silent(model)
false

Other optimizer specific settings can be set via set_parameter. For example, let's set the maximum CPU time for Ipopt:

julia> set_parameter(model, "max_cpu_time", 60.)
60.0

Methods

InfiniteOpt.build_optimizer_model!

InfiniteOpt.build_optimizer_model!

InfiniteOpt.optimizer_model

InfiniteOpt.optimizer_model_key

InfiniteOpt.optimizer_model_ready

InfiniteOpt.set_optimizer_model

InfiniteOpt.set_optimizer_model_ready

JuMP.add_bridge

JuMP.bridge_constraints

JuMP.optimize!

JuMP.set_optimizer

JuMP.set_parameter

JuMP.set_silent

JuMP.unset_silent

InfiniteOpt.optimizer_model — Function

optimizer_model(model::InfiniteModel)::JuMP.Model

Return the JuMP model stored in model that is used to solve it.

Example

julia> opt_model = optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

InfiniteOpt.set_optimizer_model — Function

set_optimizer_model(inf_model::InfiniteModel, opt_model::JuMP.Model)

Specify the JuMP model that is used to solve inf_model. This is intended for internal use and extensions. Note that opt_model should contain extension data to allow it to map to inf_model in a manner similar to TranscriptionModel.

Example

julia> set_optimizer_model(model, TranscriptionModel())

julia> optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

InfiniteOpt.optimizer_model_key — Function

optimizer_model_key(model::InfiniteModel)::Any

Return the extension key used in the optimizer model of model. Errors if optimizer_model.ext contains more than one key. This is intended for internal use and extensions. For extensions this is used to dispatch to the appropriate optmizer model functions such as extensions to build_optimizer_model!.

Example

julia> optimizer_model_key(model)
:TransData

InfiniteOpt.build_optimizer_model! — Method

 build_optimizer_model!(model::InfiniteModel)

Build the optimizer model stored in model such that it can be treated as a normal JuMP model. Specifically, translate the variables and constraints stored in model into ones that are stored in the optimizer model and can be solved. This is provided generally to accomodate extensions that use custom optimizer model types in accordance with optimizer_model_key. However, it may be useful in certain applications when the user desires to force a build without calling optimize!. Extensions will need to implement their own version of the function build_optimizer_model!(model::InfiniteModel, key::Val{ext_key_name}).

Example

julia> build_optimizer_model!(model)

julia> optimizer_model_ready(model)
true

InfiniteOpt.build_optimizer_model! — Function

build_optimizer_model!(model::InfiniteModel, key; kwargs...)

Build the optimizer model stored in model such that it can be treated as a normal JuMP model, where the Model.ext field contains a key that points to a datastructure that appropriately maps the data between the two models. The key argument should be be typed to Val{ext_key_name}.

InfiniteOpt.optimizer_model_ready — Function

optimizer_model_ready(model::InfiniteModel)::Bool

Return Bool if the optimizer model is up to date with model.

Example

julia> optimizer_model_ready(model)
false

InfiniteOpt.set_optimizer_model_ready — Function

set_optimizer_model_ready(model::InfiniteModel, status::Bool)

Set the status of the optimizer model to whether it is up to date or not. Note is more intended as an internal function, but is useful for extensions.

Example

julia> set_optimizer_model_ready(model, true)

julia> optimizer_model_ready(model)
true

JuMP.bridge_constraints — Method

JuMP.bridge_constraints(model::InfiniteModel)::Bool

Extend JuMP.bridge_constraints to return if an infinite model model has an optimizer model where the optimizer is set and unsupported constraints are automatically bridged to equivalent supported constraints when an appropriate transformation is available.

Example

julia> bridge_constraints(model)
false

JuMP.add_bridge — Method

JuMP.add_bridge(model::InfiniteModel,
                BridgeType::Type{<:MOI.Bridges.AbstractBridge})

Extend JuMP.add_bridge to add BridgeType to the list of bridges that can be used by the optimizer model to transform unsupported constraints into an equivalent formulation using only constraints supported by the optimizer.

JuMP.set_optimizer — Method

JuMP.set_optimizer(model::InfiniteModel,
                   optimizer_factory::JuMP.OptimizerFactory;
                   bridge_constraints::Bool=true)

Extend JuMP.set_optimizer to set optimizer of infinite models. Specifically, the optimizer of the optimizer model is modified.

Example

julia> set_optimizer(model, with_optimizer(Clp.Optimizer))

julia> optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: SolverName() attribute not implemented by the optimizer.

JuMP.set_silent — Method

JuMP.set_silent(model::InfiniteModel)

Extend JuMP.set_silent for infinite models to take precedence over any other attribute controlling verbosity and requires the solver to produce no output.

Example

julia> set_silent(model)
true

JuMP.unset_silent — Method

JuMP.unset_silent(model::InfiniteModel)

Extend JuMP.unset_silent for infinite models to Neutralize the effect of the set_silent function and let the solver attributes control the verbosity.

Example

julia> unset_silent(model)
false

JuMP.set_parameter — Method

JuMP.set_parameter(model::InfiniteModel, name, value)

Sets solver-specific parameter identified by name to value.

Example

julia> set_parameter(model, "max_cpu_time", 60.) # using Ipopt
60.0

JuMP.optimize! — Method

JuMP.optimize!(model::InfiniteModel,
               optimizer_factory::Union{Nothing, JuMP.OptimizerFactory} = nothing;
               bridge_constraints::Bool=true, kwargs...)

Extend JuMP.optimize! to optimize infinite models using the internal optimizer model. Will call build_optimizer_model! if the optimizer model isn't up to date. The kwargs correspond to keyword arguments passed to build_optimizer_model! if any are defined.

Example

julia> optimize!(model, with_optimizer(Clp.Optimizer))

julia> has_values(model)
true