Optimization

Optimization

A guide and manual for optimizing (solving) InfiniteOpt models.

Overview

Basic Usage

Optimizer Models

Optimizer Settings

Methods

InfiniteOpt.build_optimizer_model!

InfiniteOpt.build_optimizer_model!

InfiniteOpt.optimizer_model

InfiniteOpt.optimizer_model_key

InfiniteOpt.optimizer_model_ready

InfiniteOpt.set_optimizer_model

InfiniteOpt.set_optimizer_model_ready

JuMP.add_bridge

JuMP.bridge_constraints

JuMP.optimize!

JuMP.set_optimizer

JuMP.set_parameter

JuMP.set_silent

JuMP.unset_silent

InfiniteOpt.optimizer_model — Function.

optimizer_model(model::InfiniteModel)::JuMP.Model

Return the JuMP model stored in model that is used to solve it.

Example

julia> opt_model = optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

source

InfiniteOpt.set_optimizer_model — Function.

set_optimizer_model(inf_model::InfiniteModel, opt_model::JuMP.Model)

Specify the JuMP model that is used to solve inf_model. This is intended for internal use and extensions. Note that opt_model should contain extension data to allow it to map to inf_model in a manner similar to TranscriptionModel.

Example

julia> set_optimizer_model(model, TranscriptionModel())

julia> optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

source

InfiniteOpt.optimizer_model_key — Function.

optimizer_model_key(model::InfiniteModel)::Any

Return the extension key used in the optimizer model of model. Errors if optimizer_model.ext contains more than one key. This is intended for internal use and extensions. For extensions this is used to dispatch to the appropriate optmizer model functions such as extensions to build_optimizer_model!.

Example

julia> optimizer_model_key(model)
:TransData

source

InfiniteOpt.build_optimizer_model! — Method.

 build_optimizer_model!(model::InfiniteModel)

Build the optimizer model stored in model such that it can be treated as a normal JuMP model. Specifically, translate the variables and constraints stored in model into ones that are stored in the optimizer model and can be solved. This is build generally to accomodate extensions that use custom optimizer model types in accordance with optimizer_model_key. Extensions will need to implement their own version of the function build_optimizer_model!(model::InfiniteModel, key::Val{ext_key_name}).

Example

julia> build_optimizer_model!(model)

julia> optimizer_model_ready(model)
true

source

InfiniteOpt.build_optimizer_model! — Function.

build_optimizer_model!(model::InfiniteModel, key; kwargs...)

Build the optimizer model stored in model such that it can be treated as a normal JuMP model, where the Model.ext field contains a key that points to a datastructure that appropriately maps the data between the two models. The key argument should be be typed to Val{ext_key_name}.

source

InfiniteOpt.optimizer_model_ready — Function.

optimizer_model_ready(model::InfiniteModel)::Bool

Return Bool if the optimizer model is up to date with model.

Example

julia> optimizer_model_ready(model)
false

source

InfiniteOpt.set_optimizer_model_ready — Function.

set_optimizer_model_ready(model::InfiniteModel, status::Bool)

Set the status of the optimizer model to whether it is up to date or not. Note is more intended as an internal function, but is useful for extensions.

Example

julia> set_optimizer_model_ready(model, true)

julia> optimizer_model_ready(model)
true

source

JuMP.bridge_constraints — Method.

JuMP.bridge_constraints(model::InfiniteModel)::Bool

Extend JuMP.bridge_constraints to return if an infinite model model has an optimizer model where the optimizer is set and unsupported constraints are automatically bridged to equivalent supported constraints when an appropriate transformation is available.

Example

julia> bridge_constraints(model)
false

source

JuMP.add_bridge — Method.

JuMP.add_bridge(model::InfiniteModel,
                BridgeType::Type{<:MOI.Bridges.AbstractBridge})

Extend [JuMP.add_bridge] to add BridgeType to the list of bridges that can be used by the optimizer model to transform unsupported constraints into an equivalent formulation using only constraints supported by the optimizer.

source

JuMP.set_optimizer — Method.

JuMP.set_optimizer(model::InfiniteModel,
                   optimizer_factory::JuMP.OptimizerFactory;
                   bridge_constraints::Bool=true)

Extend JuMP.set_optimizer to set optimizer of infinite models. Specifically, the optimizer of the optimizer model is modified.

Example

julia> set_optimizer(model, with_optimizer(Clp.Optimizer))

julia> optimizer_model(model)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: SolverName() attribute not implemented by the optimizer.

source

JuMP.set_silent — Method.

JuMP.set_silent(model::InfiniteModel)

Extend JuMP.set_silent for infinite models to take precedence over any other attribute controlling verbosity and requires the solver to produce no output.

source

JuMP.unset_silent — Method.

JuMP.unset_silent(model::InfiniteModel)

Extend JuMP.unset_silent for infinite models to Neutralize the effect of the set_silent function and let the solver attributes control the verbosity.

source

JuMP.set_parameter — Method.

JuMP.set_parameter(model::InfiniteModel, name, value)

Sets solver-specific parameter identified by name to value.

source

JuMP.optimize! — Method.

JuMP.optimize!(model::InfiniteModel,
               optimizer_factory::Union{Nothing, JuMP.OptimizerFactory} = nothing;
               bridge_constraints::Bool=true, kwargs...)

Extend JuMP.optimize! to optimize infinite models using the internal optimizer model. Will call build_optimizer_model! if the optimizer model isn't up to date. The kwargs correspond to keyword arguments passed to build_optimizer_model! if any are defined.

Example

julia> optimize!(model, with_optimizer(Clp.Optimizer))

julia> has_values(model)
true

source