Infinite Parameters

Infinite Parameters

A guide and manual to the definition and use of infinite parameters in InfiniteOpt.

Overview

Basic Usage

Infinite Sets

Parameter Definition

Parameter Use

Datatypes

InfiniteOpt.AbstractInfiniteSet

InfiniteOpt.DistributionSet

InfiniteOpt.InfOptParameter

InfiniteOpt.IntervalSet

InfiniteOpt.IntervalSet

InfiniteOpt.ParameterRef

InfiniteOpt.AbstractInfiniteSet — Type.

AbstractInfiniteSet

An abstract type for sets that characterize infinite parameters.

source

InfiniteOpt.IntervalSet — Type.

IntervalSet <: AbstractInfiniteSet

A DataType that stores the lower and upper interval bounds for infinite parameters that are continuous over a certain that interval.

Fields

lower_bound::Float64 Lower bound of the infinite parameter.

upper_bound::Float64 Upper bound of the infinite parameter.

source

InfiniteOpt.DistributionSet — Type.

DistributionSet{T <: Distributions.NonMatrixDistribution} <: AbstractInfiniteSet

A DataType that stores the distribution characterizing infinite parameters that are random.

Fields

distribution::T Distribution of the random parameter.

source

InfiniteOpt.InfOptParameter — Type.

InfOptParameter{T <: AbstractInfiniteSet} <: JuMP.AbstractVariable

A DataType for storing core infinite parameter information.

Fields

set::T The infinite set that characterizes the parameter.

supports::Vector{<:Number} The support points used to discretize this parameter.

independent::Bool Is independent of other parameters that share its group ID number.

source

InfiniteOpt.ParameterRef — Type.

ParameterRef <: GeneralVariableRef

A DataType for untranscripted infinite parameters references that parameterize the infinite variables.

Fields

model::InfiniteModel Infinite model.

index::Int Index of variable in model.

source

Methods/Macros

InfiniteOpt.@infinite_parameter

InfiniteOpt.add_parameter

InfiniteOpt.add_supports

InfiniteOpt.all_parameters

InfiniteOpt.build_parameter

InfiniteOpt.delete_supports

InfiniteOpt.group_id

InfiniteOpt.group_id

InfiniteOpt.has_supports

InfiniteOpt.infinite_set

InfiniteOpt.is_independent

InfiniteOpt.is_used

InfiniteOpt.num_parameters

InfiniteOpt.num_supports

InfiniteOpt.parameter_by_name

InfiniteOpt.set_independent

InfiniteOpt.set_infinite_set

InfiniteOpt.set_supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.unset_independent

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_variable

JuMP.delete

JuMP.has_lower_bound

JuMP.has_upper_bound

JuMP.is_valid

JuMP.lower_bound

JuMP.name

JuMP.set_lower_bound

JuMP.set_name

JuMP.set_upper_bound

JuMP.upper_bound

InfiniteOpt.IntervalSet — Method.

IntervalSet(lower_bound::Number, upper_bound::Number)

A constructor for IntervalSet that converts values of type Number to values of type Float64 as required by IntervalSet.

source

InfiniteOpt.add_parameter — Function.

add_parameter(model::InfiniteModel, p::InfOptParameter,
              name::String="")::ParameterRef

Returns a ParameterRef associated with the parameter p that is added to model. This adds a parameter to the model in a manner similar to JuMP.add_variable. This can be used to add parameters with the use of @infinite_parameter. build_parameter should be used to construct p.

Example

julia> p = build_parameter(error, IntervalSet(0, 3), supports = Vector(0:3))
InfOptParameter{IntervalSet}(IntervalSet(0.0, 3.0), [0, 1, 2, 3], false)

julia> param_ref = add_parameter(model, p, "name")
name

source

InfiniteOpt.build_parameter — Function.

build_parameter(_error::Function, set::AbstractInfiniteSet,
                [num_params::Int = 1;
                supports::Union{Number, Vector{<:Number}} = Number[],
                independent::Bool = false])::InfOptParameter

Returns a InfOptParameter given the appropriate information. This is analagous to JuMP.build_variable. Errors if supports violate the bounds associated set. Also errors if set contains a multivariate distribution with a different dimension than num_params. This is meant to primarily serve as a helper method for @infinite_parameter.

Example

julia> build_parameter(error, IntervalSet(0, 3), supports = Vector(0:3))
InfOptParameter{IntervalSet}(IntervalSet(0.0, 3.0), [0, 1, 2, 3], false)

source

InfiniteOpt.@infinite_parameter — Macro.

@infinite_parameter(model, kw_args...)

Add an anonymous infinite parameter to the model model described by the keyword arguments kw_args and returns the parameter reference.

@infinite_parameter(model, expr, kw_args...)

Add a parameter to the model model described by the expression expr, the positional arguments args and the keyword arguments kw_args. (note that in the following the symbol <= can be used instead of ≤, the symbol >=can be used instead of ≥, and the symbo in can be used instead of ∈) The expression expr can be of the form:

paramexpr creating parameters described by paramexpr.

lb ≤ varexpr ≤ ub creating parameters described by paramexpr characterized by a continuous interval set with lower bound lb and upper bound ub.

paramexpr ∈ dist creating parameters described by paramexpr characterized by the Distributions.jl distribution object dist.

The expression varexpr can be of the form:

paramname creating a scalar parameter of name paramname

paramname[...] or [...] creating a container of parameters

The recognized keyword arguments in kw_args are the following:

base_name: Sets the name prefix used to generate parameter names. It corresponds to the parameter name for scalar parameter, otherwise, the parameter names are set to base_name[...] for each index ... of the axes axes.

lower_bound: Sets the value of the parameter lower bound for an interval set.

upper_bound: Sets the value of the parameter upper bound for an interval set.

set: The InfiniteSet characterizing the parameters see IntervalSet and DistributionSet.

distribution: Sets the Distributions.jl distribution object that characterizes the parameters.

supports: Sets the support points for the parameters.

independent: Specifies if the each parameter is independent from each other or not.

container: Specify the container type.

Examples

julia> @infinite_parameter(m, 0 <= x <= 1)
x

julia> supps = [[0, 1, 2], [-1, 1]];

julia> @infinite_parameter(m, y[i = 1:2] in Normal(), supports = supps[i])
2-element Array{ParameterRef,1}:
 y[1]
 y[2]

julia> z = @infinite_parameter(m, ["a", "b"], distribution = Uniform(), independent = true)
2-dimensional DenseAxisArray{ParameterRef,2,...} with index sets:
    Dimension 1, "a"
    Dimension 2, "b"
And data, a 1×1 Array{ParameterRef,2}:
 noname

source

InfiniteOpt.used_by_constraint — Method.

used_by_constraint(pref::ParameterRef)::Bool

Return true if pref is used by a constraint or false otherwise.

Example

julia> used_by_constraint(t)
true

source

InfiniteOpt.used_by_measure — Method.

used_by_measure(pref::ParameterRef)::Bool

Return true if pref is used by a measure or false otherwise.

Example

julia> used_by_measure(t)
false

source

InfiniteOpt.used_by_variable — Method.

used_by_variable(pref::ParameterRef)::Bool

Return true if pref is used by an infinite variable or false otherwise.

Example

julia> used_by_variable(t)
true

source

InfiniteOpt.is_used — Method.

is_used(pref::ParameterRef)::Bool

Return true if pref is used in the model or false otherwise.

Example

julia> is_used(t)
true

source

JuMP.delete — Method.

JuMP.delete(model::InfiniteModel, pref::ParameterRef)

Extend JuMP.delete to delete infinite parameters and their dependencies. All variables, constraints, and measure functions that depend on pref are updated to exclude it. Errors if the parameter is contained in an AbstractMeasureData datatype that is employed by a measure since the measure becomes invalid otherwise. Thus, measures that contain this dependency must be deleted first. Note that _check_param_in_data(pref, measure_data) needs to be extended to allow deletion of parameters when custom AbstractMeasureData datatypes are used.

Example

julia> print(model)
Min measure(g(t, x)*t + x) + z
Subject to
 z >= 0.0
 g(t, x) + z >= 42.0
 g(0.5, x) == 0
 t in [0, 6]
 x in [0, 1]

julia> delete(model, x)

julia> print(model)
Min measure(g(t)*t) + z
Subject to
 g(t) + z >= 42.0
 g(0.5) == 0
 z >= 0.0
 t in [0, 6]

source

JuMP.is_valid — Method.

JuMP.is_valid(model::InfiniteModel, pref::ParameterRef)::Bool

Extend the JuMP.is_valid function to accomodate infinite parameters. Returns true if the InfiniteModel stored in pref matches model and if the parameter index is used by model. It returns false otherwise.

Example

julia> is_valid(model, t)
true

source

JuMP.name — Method.

JuMP.name(pref::ParameterRef)::String

Extend the JuMP.name function to accomodate infinite parameters. Returns the name string associated with pref.

Example

julia> name(t)
"t"

source

JuMP.set_name — Method.

JuMP.set_name(pref::ParameterRef, name::String)

Extend the JuMP.set_name function to accomodate infinite parameters. Set a new base name to be associated with pref.

Example

julia> set_name(t, "time")

julia> name(t)
"time"

source

InfiniteOpt.num_parameters — Method.

num_parameters(model::InfiniteModel)::Int

Return the number of infinite parameters currently present in model.

Example

julia> num_parameters(model)
1

source

InfiniteOpt.infinite_set — Method.

infinite_set(pref::ParameterRef)::AbstractInfiniteSet

Return the infinite set associated with pref.

Example

julia> infinite_set(t)
IntervalSet(0.0, 3.0)

source

InfiniteOpt.set_infinite_set — Method.

set_infinite_set(pref::ParameterRef, set::AbstractInfiniteSet)

Specify the infinite set of pref.

Example

julia> set_infinite_set(t, IntervalSet(0, 1))

julia> infinite_set(t)
IntervalSet(0.0, 1.0)

source

JuMP.has_lower_bound — Method.

JuMP.has_lower_bound(pref::ParameterRef)::Bool

Extend the JuMP.has_lower_bound function to accomodate infinite parameters. Return true if the set associated with pref has a defined lower bound or if a lower bound can be found.

Example

julia> has_lower_bound(t)
true

source

JuMP.lower_bound — Method.

JuMP.lower_bound(pref::ParameterRef)::Number

Extend the JuMP.lower_bound function to accomodate infinite parameters. Returns the lower bound associated with the infinite set. Errors if such a bound is not well-defined.

Example

julia> lower_bound(t)
0.0

source

JuMP.set_lower_bound — Method.

JuMP.set_lower_bound(pref::ParameterRef, lower::Number)

Extend the JuMP.set_lower_bound function to accomodate infinite parameters. Updates the infinite set lower bound if and only if it is an IntervalSet. Errors otherwise.

Example

julia> set_lower_bound(t, -1)

julia> lower_bound(t)
-1.0

source

JuMP.has_upper_bound — Method.

JuMP.has_upper_bound(pref::ParameterRef)::Bool

Extend the JuMP.has_upper_bound function to accomodate infinite parameters. Return true if the set associated with pref has a defined upper bound or if a upper bound can be found.

Example

julia> has_upper_bound(t)
true

source

JuMP.upper_bound — Method.

JuMP.upper_bound(pref::ParameterRef)::Number

Extend the JuMP.upper_bound function to accomodate infinite parameters. Returns the upper bound associated with the infinite set. Errors if such a bound is not well-defined.

Example

julia> upper_bound(t)
1.0

source

JuMP.set_upper_bound — Method.

JuMP.set_upper_bound(pref::ParameterRef, lower::Number)

Extend the JuMP.set_upper_bound function to accomodate infinite parameters. Updates the infinite set upper bound if and only if it is an IntervalSet. Errors otherwise.

Example

julia> set_upper_bound(t, 2)

julia> upper_bound(t)
2.0

source

InfiniteOpt.num_supports — Method.

num_supports(pref::ParameterRef)::Int

Return the number of support points associated with pref.

Example

julia> num_supports(t)
1

source

InfiniteOpt.has_supports — Method.

has_supports(pref::ParameterRef)::Bool

Return true if pref has supports or false otherwise.

Example

julia> has_supports(t)
true

source

InfiniteOpt.supports — Method.

supports(pref::ParameterRef)::Vector

Return the support points associated with pref. Errors if there are no supports.

Example

julia> supports(t)
1-element Array{Int64,1}:
 1

source

InfiniteOpt.supports — Method.

supports(prefs::AbstractArray{<:ParameterRef})::Vector

Return the support points associated with an array of prefs formatted as a vector of SparseAxisArrays following the format of the input array. If the parameters are not independent then the supports of each parameter are simply spliced together. Alternatively can call supports. to more efficiently obtain an array of the same input format whose parameter references have been replaced with their supports. Errors if all the parameter references do not have the same group ID number (were intialized together as an array) or if the nonindependent parameters have support vectors of different lengths. If the parameters are independent then all the unique combinations are identified and returned as supports. Warning this operation is computationally expensive if there exist a large number of combinations.

Example

julia> x = @infinite_parameter(model, [i = 1:2], set = IntervalSet(-1, 1),
                               base_name = "x", independent = true)
2-element Array{ParameterRef,1}:
 x[1]
 x[2]

julia> for i = 1:length(x)
           set_supports(x[i], [-1, 1])
       end

julia> supports(x)
4-element Array{JuMP.Containers.SparseAxisArray,1}:
   [2]  =  -1
  [1]  =  -1
   [2]  =  1
  [1]  =  -1
   [2]  =  -1
  [1]  =  1
   [2]  =  1
  [1]  =  1

source

InfiniteOpt.set_supports — Method.

set_supports(pref::ParameterRef, supports::Vector{<:Number})

Specify the support points for pref. Errors if the supports violate the bounds associated with the infinite set. Warns if the points are not unique. Note that this will overwrite existing supports.

Example

julia> set_supports(t, [0, 1])

julia> supports(t)
2-element Array{Int64,1}:
 0
 1

source

InfiniteOpt.add_supports — Method.

add_supports(pref::ParameterRef, supports::Union{Number, Vector{<:Number}})

Add additional support points for pref.

Example

julia> add_supports(t, 0.5)

julia> supports(t)
3-element Array{Float64,1}:
 0.0
 0.5
 1.0

julia> add_supports(t, [0.25, 1])

julia> supports(t)
4-element Array{Float64,1}:
 0.0
 0.25
 0.5
 1.0

source

InfiniteOpt.delete_supports — Method.

delete_supports(pref::ParameterRef)

Delete the support points for pref.

Example

julia> delete_supports(t)

julia> supports(t)
ERROR: Parameter test does not have supports.

source

InfiniteOpt.group_id — Method.

group_id(pref::ParameterRef)::Int

Return the group ID number for pref.

Example

julia> group_id(t)
1

source

InfiniteOpt.group_id — Method.

group_id(prefs::AbstractArray{<:ParameterRef})::Int

Return the group ID number for a group of prefs. Error if contains multiple groups.

Example

julia> group_id([x[1], x[2]])
2

source

InfiniteOpt.is_independent — Method.

is_independent(pref::ParameterRef)::Bool

Returns true for pref if it is independent or false otherwise.

Example

julia> is_independent(t)
false

source

InfiniteOpt.set_independent — Method.

set_independent(pref::ParameterRef)

Specify that pref be independent.

Example

julia> set_independent(t)

julia> is_independent(t)
true

source

InfiniteOpt.unset_independent — Method.

unset_independent(pref::ParameterRef)

Specify that pref be not independent.

Example

julia> unset_independent(t)

julia> is_independent(t)
false

source

InfiniteOpt.parameter_by_name — Method.

parameter_by_name(model::InfiniteModel, name::String)::Union{ParameterRef,
                                                             Nothing}

Return the parameter reference assoociated with a parameter name. Errors if multiple parameters have the same name. Returns nothing if no such name exists.

Example

julia> parameter_by_name(model, "t")
t

source

InfiniteOpt.all_parameters — Method.

all_parameters(model::InfiniteModel)::Vector{ParameterRef}

Return all of the infinite parameter references currently in model.

Example

julia> all_parameters(model)
3-element Array{ParameterRef,1}:
 t
 x[1]
 x[2]

source