Results

A guide and manual to querying optimized InfiniteOpt models. The Methods section at the bottom comprises the manual and the above sections form the guide.

Overview

So far we have covered defining, transforming, and optimizing InfiniteModels. Now comes the point to extract information from our optimized model. This is done following extended versions of JuMPs querying functions in combination with the mapping information stored in the optimizer model. Thus, this page will walk through the use of these result query functions.

Basic Usage

Let's revisit the example from the optimization page to get us started:

julia> using InfiniteOpt, JuMP, Ipopt;

julia> model = InfiniteModel(with_optimizer(Ipopt.Optimizer, print_level = 0));

julia> @infinite_parameter(model, t in [0, 10], num_supports = 10);

julia> @infinite_variable(model, x(t) >= 0);

julia> @hold_variable(model, z >= 0);

julia> @objective(model, Min, 2z);

julia> @constraint(model, c1, z >= x);

julia> @BDconstraint(model, c2(t == 0), x == 42);

julia> print(model)
Min 2 z
Subject to
 x(t) ≥ 0.0
 z ≥ 0.0
 z - x(t) ≥ 0.0
 x(t) = 42.0, ∀ t = 0
 t ∈ [0, 10]

julia> optimize!(model)

Now that the model has been optimized, let's find out what happened. To determine why the optimizer stopped, we can use termination_status to report the corresponding MathOptInterface termination code (possible codes are explained here.

julia> termination_status(model)
LOCALLY_SOLVED::TerminationStatusCode = 4

Here we see that our model was locally solved via Ipopt and that is why it stopped. Furthermore, we can query the primal and dual problem optimalities via primal_status and dual_status, respectively.

julia> primal_status(model)
FEASIBLE_POINT::ResultStatusCode = 1

julia> dual_status(model)
FEASIBLE_POINT::ResultStatusCode = 1

The possible statuses are detailed here. These results are useful in knowing if information can be drawn from the primal and/or dual and what it means. We can also verify that we indeed have answers via has_values which indicates if our model has optimized variable values.

julia> has_values(model)
true

And indeed we do have values.

Now we can query the objective value via objective_value which reports the optimal objective value.

julia> objective_value(model)
83.99999998250514

Great now we can inquire about variable values via value. First, let's retrieve the value of z:

julia> value(z)
41.99999999125257

We get a single value since z is a HoldVariable and therefore finite. Now let's retrieve the "value" of x(t) which is infinite with respect to t:

julia> value(x)
10-element Array{Float64,1}:
 42.0
 20.999999995633615
 20.999999995633615
 20.999999995633615
 20.999999995633615
 20.999999995633615
 20.999999995633615
 20.999999995633615
 20.999999995633615
 20.999999995633615

Notice here we obtain an array of values since these correspond to the transcribed finite (discretized) variables used to solve the problem. We obtain the corresponding support (discretized t) values via supports:

julia> supports(x)
10-element Array{Tuple{Float64},1}:
 (0.0,)
 (1.1111,)
 (2.2222,)
 (3.3333,)
 (4.4444,)
 (5.5556,)
 (6.6667,)
 (7.7778,)
 (8.8889,)
 (10.0,)

There is 1-to-1 correspondence between these supports and the values reported above. Note that these are stored in tuples to facilitate multiple infinite parameter dependencies.

Note

The values for an array of variables is obtained via the vectorized call of value following the syntax:

value.(::AbstractArray{<:GeneralVariableRef})

This also holds true for many other methods in InfiniteOpt. For example, adding the dot also vectorizes dual and set_binary.

We can also query the dual of a constraint via dual if a model has duals available as indicated by has_duals:

julia> has_duals(model)
true

julia> dual(c1)
10-element Array{Float64,1}:
 1.9999999988666093
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10

c1 is an infinite constraint and thus we obtain the duals of its transcribed versions. The underlying infinite parameter(s) and support values are queried via parameter_refs and supports:

julia> parameter_refs(c1)
(t,)

julia> supports(c1)
10-element Array{Tuple{Float64},1}:
 (0.0,)
 (1.1111,)
 (2.2222,)
 (3.3333,)
 (4.4444,)
 (5.5556,)
 (6.6667,)
 (7.7778,)
 (8.8889,)
 (10.0,)

These again all have a 1-to-1 correspondence.

Termination Queries

Termination queries are those that question about how the infinite model was solved and what its optimized state entails. Programmatically, such queries on the InfiniteModel are simply routed to its optimizer model.

The commonly used queries include termination_status, primal_status, dual_status, objective_value, and solve_time. The first four are well exemplified in the Basic Usage section above and are helpful in quickly understanding the optimality status of a given model following the many possible statuses reported by MathOptInterface which are documented here. We use solve_time to determine the time in seconds used by the optimizer until it terminated its search.

julia> solve_time(model)
0.004999876022338867

Note that this query might not be supported with all solvers.

The above status queries are designed to report information in a consistent format irrespective of the chosen optimizer. However, raw_status will provide the optimality status verbatim as reported by the optimizer. Thus, following our example with Ipopt we obtain:

julia> raw_status(model)
"Solve_Succeeded"

Also, we obtain the best objective bound via objective_bound which becomes particularly useful solutions that are suboptimal. However, this method is not supported by all optimizers and in this case Ipopt is one such optimizer.

Variable Queries

Information about the optimized variables is gathered consistently in comparison to typical JuMP models. With InfiniteModels this is done by querying the optimizer model and using its stored variable mappings to return the correct information. Thus, here the queries are extended to work with the specifics of the optimizer model to return the appropriate info.

First, we should verify that the optimized model in fact has variable values via has_values. In our example, we have:

julia> has_values(model)
true

So we have values readily available to be extracted.

Now value can be used to query the values as shown above in the Basic Usage section. This works by calling the appropriate map_value defined by the optimizer model. For TranscriptionModels map_value and/or map_value are used. Details on how to extend these methods for user-defined optimizer models is explained on the Extensions page.

Finally, the optimizer index of a variable is queried via optimizer_index which reports back the index of the variable as used in the MathOptInterface backend:

julia> optimizer_index(z)
MathOptInterface.VariableIndex(1)

julia> optimizer_index(x)
10-element Array{MathOptInterface.VariableIndex,1}:
 MathOptInterface.VariableIndex(2)
 MathOptInterface.VariableIndex(3)
 MathOptInterface.VariableIndex(4)
 MathOptInterface.VariableIndex(5)
 MathOptInterface.VariableIndex(6)
 MathOptInterface.VariableIndex(7)
 MathOptInterface.VariableIndex(8)
 MathOptInterface.VariableIndex(9)
 MathOptInterface.VariableIndex(10)
 MathOptInterface.VariableIndex(11)

As noted previously, an array is returned for x(t) in accordance with its transcription variables. In similar manner to value, this is enabled by appropriate versions of map_optimizer_index enabled by TranscriptionOpt.

Constraint Queries

Like variables, a variety of information can be queried about constraints.

First, recall that constraints are stored in the form function-in-set where generally function contains the variables and coefficients and the set contains the relational operator and the constant value. With this understanding, we query the value of a constraint's function via value:

julia> constraint_object(c1).func # show the function expression of c1
z - x(t)

julia> value(c1)
10-element Array{Float64,1}:
 -8.747427671096375e-9
 20.999999995618957
 20.999999995618957
 20.999999995618957
 20.999999995618957
 20.999999995618957
 20.999999995618957
 20.999999995618957
 20.999999995618957
 20.999999995618957

Again, we obtain an array of values since c1 is infinite due to its dependence on x(t). Behind the scenes this is implemented via the appropriate extensions of map_value in TranscriptionOpt by default.

Next the optimizer index(es) of the transcribed constraints in the MathOptInterface backend provided via optimizer_index.

julia> optimizer_index(c1)
10-element Array{MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}},1}:
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(1)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(2)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(3)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(4)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(5)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(6)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(7)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(8)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(9)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(10)

Here 10 indices are given in accordance with the transcription constraints. The mapping between these and the original infinite constraints is managed via the appropriate extensions of map_optimizer_index in TranscriptionOpt.

We can also query dual information from our constraints if it is available. First, we should verify that dual information is available via has_duals:

julia> has_duals(model)
true

Now we can query the duals via dual.

julia> dual(c1)
10-element Array{Float64,1}:
 1.9999999988666093
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10
 1.1930560126841273e-10

Here we obtain the optimal dual values for each transcribed version of c1. This is enabled via the proper extensions of map_dual in TranscriptionOpt by default.

Finally, we query the shadow price of a constraint via shadow_price. This denotes the change in the objective value due to an infinitesimal relaxation of the constraint. For c1 we get:

julia> shadow_price(c1)
10-element Array{Float64,1}:
 -1.9999999988666093
 -1.1930560126841273e-10
 -1.1930560126841273e-10
 -1.1930560126841273e-10
 -1.1930560126841273e-10
 -1.1930560126841273e-10
 -1.1930560126841273e-10
 -1.1930560126841273e-10
 -1.1930560126841273e-10
 -1.1930560126841273e-10

Similarly, the mapping to the transcription constraints is enabled via the appropriate version of map_shadow_price.

Methods

JuMP.dual

JuMP.dual_status

JuMP.has_duals

JuMP.has_values

JuMP.objective_bound

JuMP.objective_value

JuMP.optimizer_index

JuMP.optimizer_index

JuMP.primal_status

JuMP.raw_status

JuMP.shadow_price

JuMP.solve_time

JuMP.termination_status

JuMP.value

JuMP.value

InfiniteOpt.map_dual

InfiniteOpt.map_dual

InfiniteOpt.map_dual

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_shadow_price

InfiniteOpt.map_shadow_price

InfiniteOpt.map_shadow_price

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

JuMP.termination_status — Method

JuMP.termination_status(model::InfiniteModel)

Return the reason why the solver stopped (i.e., the MathOptInterface model attribute TerminationStatus).

Example

julia> termination_status(model)
LOCALLY_SOLVED::TerminationStatusCode = 4

JuMP.raw_status — Method

JuMP.raw_status(model::InfiniteModel)

Return the reason why the solver stopped in its own words (i.e., the MathOptInterface model attribute RawStatusString).

Example

julia> raw_status(model) # Ipopt
"Solve_Succeeded"

JuMP.primal_status — Method

 JuMP.primal_status(model::InfiniteModel)

Return the status of the most recent primal solution of the solver (i.e., the MathOptInterface model attribute PrimalStatus).

Example

julia> primal_status(model)
FEASIBLE_POINT::ResultStatusCode = 1

JuMP.dual_status — Method

JuMP.dual_status(model::InfiniteModel)

Return the status of the most recent dual solution of the solver (i.e., the MathOptInterface model attribute DualStatus).

Example

julia> dual_status(model)
FEASIBLE_POINT::ResultStatusCode = 1

JuMP.solve_time — Method

JuMP.solve_time(model::InfiniteModel)

If available, returns the solve time (in seconds) reported by the solver. Returns "ArgumentError: ModelLike of type Solver.Optimizer does not support accessing the attribute MathOptInterface.SolveTime()" if the attribute is not implemented.

Example

julia> solve_time(model)
0.004999876022338867

JuMP.has_values — Method

JuMP.has_values(model::InfiniteModel)

Return true if the solver has a primal solution available to query, otherwise return false.

Example

julia> has_values(model)
true

JuMP.has_duals — Method

JuMP.has_duals(model::InfiniteModel)

Return true if the solver has a dual solution available to query, otherwise return false.

Example

julia> has_duals(model)
true

JuMP.objective_bound — Method

JuMP.objective_bound(model::InfiniteModel)::Float64

Return the best known bound on the optimal objective value after a call to optimize!(model).

Example

julia> objective_bound(model)
42.0

JuMP.objective_value — Method

JuMP.objective_value(model::InfiniteModel)::Float64

Return the objective value after a call to optimize!(model).

Example

julia> objective_value(model)
42.0

JuMP.value — Method

JuMP.value(vref::GeneralVariableRef)

Get the value of this variable in the result returned by a solver. Use JuMP.has_values to check if a result exists before asking for values.

Example

julia> value(z)
42.0

JuMP.value — Method

JuMP.value(cref::GeneralConstraintRef)

Get the value of this constraint in the result returned by a solver. Use JuMP.has_values to check if a result exists before asking for values. This returns the primal value of the constraint function.

Example

julia> value(c1)
4-element Array{Float64,1}:
 -0.0
 20.9
 20.9
 20.9

InfiniteOpt.map_value — Function

map_value(vref::GeneralVariableRef, key)

Map the value of vref to its counterpart in the optimizer model type is distininguished by its extension key key as type Val{ext_key_name}.

InfiniteOpt.map_value — Method

InfiniteOpt.map_value(fvref::InfiniteOpt.FiniteVariableRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription variable in the transcription model to fvref.

InfiniteOpt.map_value — Method

InfiniteOpt.map_value(ivref::InfiniteOpt.InfiniteVariableRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription variables in the transcription model to ivref.

InfiniteOpt.map_value — Method

InfiniteOpt.map_value(fcref::InfiniteOpt.FiniteConstraintRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription constraint function in the transcription model to fcref.

InfiniteOpt.map_value — Method

InfiniteOpt.map_value(icref::InfiniteOpt.GeneralConstraintRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription constraint functions in the transcription model to icref.

JuMP.optimizer_index — Method

JuMP.optimizer_index(vref::GeneralVariableRef)

Return the index of the variables that corresponds to vref in the optimizer model. It throws JuMP.NoOptimizer if no optimizer is set and throws an ErrorException if the optimizer is set but is not attached.

Example

julia> optimizer_index(x)
4-element Array{MathOptInterface.VariableIndex,1}:
 MathOptInterface.VariableIndex(2)
 MathOptInterface.VariableIndex(3)
 MathOptInterface.VariableIndex(4)
 MathOptInterface.VariableIndex(5)

JuMP.optimizer_index — Method

JuMP.optimizer_index(cref::GeneralConstraintRef)

Return the index of the constraints that corresponds to cref in the optimizer model. It throws JuMP.NoOptimizer if no optimizer is set and throws an ErrorException if the optimizer is set but is not attached.

Example

julia> optimizer_index(c1)
4-element Array{MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}},1}:
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(1)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(2)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(3)
 MathOptInterface.ConstraintIndex{MathOptInterface.ScalarAffineFunction{Float64},MathOptInterface.GreaterThan{Float64}}(4)

InfiniteOpt.map_optimizer_index — Function

map_optimizer_index(ref, key)

Map the optimizer index of ref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}. Here ref need refer to methods for both variable references and constraint references.

InfiniteOpt.map_optimizer_index — Method

InfiniteOpt.map_optimizer_index(fvref::InfiniteOpt.FiniteVariableRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription variable in the transcription model to fvref.

InfiniteOpt.map_optimizer_index — Method

InfiniteOpt.map_optimizer_index(ivref::InfiniteOpt.InfiniteVariableRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription variables in the transcription model to ivref.

InfiniteOpt.map_optimizer_index — Method

InfiniteOpt.map_optimizer_index(fcref::InfiniteOpt.FiniteConstraintRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription constraint in the transcription model to fcref.

InfiniteOpt.map_optimizer_index — Method

InfiniteOpt.map_optimizer_index(cref::InfiniteOpt.GeneralConstraintRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription constraints in the transcription model to cref.

JuMP.dual — Method

JuMP.dual(cref::GeneralConstraintRef)

Get the dual value of this constraint in the result returned by a solver. Use has_dual to check if a result exists before asking for values. See also JuMP.shadow_price.

Example

julia> dual(c1)
4-element Array{Float64,1}:
 -42.0
 -42.0
 32.3
 0.0

InfiniteOpt.map_dual — Function

map_dual(cref::GeneralConstraintRef, key)

Map the dual of cref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}.

InfiniteOpt.map_dual — Method

InfiniteOpt.map_dual(fcref::InfiniteOpt.FiniteConstraintRef,
                     key::Val{:TransData})

Map the dual of the appropriate transcription constraint in the transcription model to fcref.

InfiniteOpt.map_dual — Method

InfiniteOpt.map_dual(icref::InfiniteOpt.InfiniteConstraintRef,
                     key::Val{:TransData})

Map the duals of the appropriate transcription constraints in the transcription model to icref.

JuMP.shadow_price — Method

JuMP.shadow_price(cref::GeneralConstraintRef)

The change in the objective from an infinitesimal relaxation of the constraint. This value is computed from JuMP.dual and can be queried only when has_duals is true and the objective sense is MIN_SENSE or MAX_SENSE (not FEASIBILITY_SENSE). For linear constraints, the shadow prices differ at most in sign from the dual value depending on the objective sense.

Example

julia> shadow_price(c1)
4-element Array{Float64,1}:
 42.0
 42.0
 -32.3
 -0.0

InfiniteOpt.map_shadow_price — Function

map_shadow_price(cref::GeneralConstraintRef, key)

Map the shadow price of cref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}.

InfiniteOpt.map_shadow_price — Method

InfiniteOpt.map_shadow_price(fcref::InfiniteOpt.FiniteConstraintRef,
                             key::Val{:TransData})

Map the shadow price of the appropriate transcription constraint in the transcription model to fcref.

InfiniteOpt.map_shadow_price — Method

InfiniteOpt.map_shadow_price(icref::InfiniteOpt.InfiniteConstraintRef,
                             key::Val{:TransData})

Map the shadow prices of the appropriate transcription constraints in the transcription model to icref.