Results

A guide and manual to querying optimized InfiniteOpt models.

Overview

Basic Usage

Termination Queries

Variable Queries

Constraint Queries

Methods

JuMP.dual

JuMP.dual_status

JuMP.has_values

JuMP.objective_bound

JuMP.objective_value

JuMP.optimizer_index

JuMP.optimizer_index

JuMP.primal_status

JuMP.raw_status

JuMP.shadow_price

JuMP.solve_time

JuMP.termination_status

JuMP.value

JuMP.value

InfiniteOpt.map_dual

InfiniteOpt.map_dual

InfiniteOpt.map_dual

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_shadow_price

InfiniteOpt.map_shadow_price

InfiniteOpt.map_shadow_price

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

JuMP.termination_status — Method

JuMP.termination_status(model::InfiniteModel)

Return the reason why the solver stopped (i.e., the MathOptInterface model attribute TerminationStatus).

JuMP.raw_status — Method

JuMP.raw_status(model::InfiniteModel)

Return the reason why the solver stopped in its own words (i.e., the MathOptInterface model attribute RawStatusString).

JuMP.primal_status — Method

 JuMP.primal_status(model::InfiniteModel)

Return the status of the most recent primal solution of the solver (i.e., the MathOptInterface model attribute PrimalStatus).

JuMP.dual_status — Method

JuMP.dual_status(model::InfiniteModel)

Return the status of the most recent dual solution of the solver (i.e., the MathOptInterface model attribute DualStatus).

JuMP.solve_time — Method

JuMP.solve_time(model::InfiniteModel)

If available, returns the solve time reported by the solver. Returns "ArgumentError: ModelLike of type Solver.Optimizer does not support accessing the attribute MathOptInterface.SolveTime()" if the attribute is not implemented.

JuMP.has_values — Method

JuMP.has_values(model::InfiniteModel)

Return true if the solver has a primal solution available to query, otherwise return false. See also JuMP.value.

JuMP.objective_bound — Method

JuMP.objective_bound(model::InfiniteModel)::Float64

Return the best known bound on the optimal objective value after a call to optimize!(model).

JuMP.objective_value — Method

JuMP.objective_value(model::InfiniteModel)::Float64

Return the objective value after a call to optimize!(model).

JuMP.value — Method

JuMP.value(vref::GeneralVariableRef)

Get the value of this variable in the result returned by a solver. Use JuMP.has_values to check if a result exists before asking for values.

JuMP.value — Method

JuMP.value(cref::GeneralConstraintRef)

Get the value of this constraint in the result returned by a solver. Use JuMP.has_values to check if a result exists before asking for values. This returns the primal value of the constraint function.

InfiniteOpt.map_value — Function

map_value(vref::GeneralVariableRef, key)

Map the value of vref to its counterpart in the optimizer model type is distininguished by its extension key key as type Val{ext_key_name}.

InfiniteOpt.map_value — Method

InfiniteOpt.map_value(fvref::InfiniteOpt.FiniteVariableRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription variable in the transcription model to fvref.

InfiniteOpt.map_value — Method

InfiniteOpt.map_value(ivref::InfiniteOpt.InfiniteVariableRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription variables in the transcription model to ivref.

InfiniteOpt.map_value — Method

InfiniteOpt.map_value(fcref::InfiniteOpt.FiniteConstraintRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription constraint function in the transcription model to fcref.

InfiniteOpt.map_value — Method

InfiniteOpt.map_value(icref::InfiniteOpt.GeneralConstraintRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription constraint functions in the transcription model to icref.

JuMP.optimizer_index — Method

JuMP.optimizer_index(vref::GeneralVariableRef)

Return the index of the variables that corresponds to vref in the optimizer model. It throws JuMP.NoOptimizer if no optimizer is set and throws an ErrorException if the optimizer is set but is not attached.

JuMP.optimizer_index — Method

JuMP.optimizer_index(cref::GeneralConstraintRef)

Return the index of the constraints that corresponds to cref in the optimizer model. It throws JuMP.NoOptimizer if no optimizer is set and throws an ErrorException if the optimizer is set but is not attached.

InfiniteOpt.map_optimizer_index — Function

map_optimizer_index(ref, key)

Map the optimizer index of ref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}. Here ref need refer to methods for both variable references and constraint references.

InfiniteOpt.map_optimizer_index — Method

InfiniteOpt.map_optimizer_index(fvref::InfiniteOpt.FiniteVariableRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription variable in the transcription model to fvref.

InfiniteOpt.map_optimizer_index — Method

InfiniteOpt.map_optimizer_index(ivref::InfiniteOpt.InfiniteVariableRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription variables in the transcription model to ivref.

InfiniteOpt.map_optimizer_index — Method

InfiniteOpt.map_optimizer_index(fcref::InfiniteOpt.FiniteConstraintRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription constraint in the transcription model to fcref.

InfiniteOpt.map_optimizer_index — Method

InfiniteOpt.map_optimizer_index(cref::InfiniteOpt.GeneralConstraintRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription constraints in the transcription model to cref.

JuMP.dual — Method

JuMP.dual(cref::GeneralConstraintRef)

Get the dual value of this constraint in the result returned by a solver. Use has_dual to check if a result exists before asking for values. See also JuMP.shadow_price.

InfiniteOpt.map_dual — Function

map_dual(cref::GeneralConstraintRef, key)

Map the dual of cref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}.

InfiniteOpt.map_dual — Method

InfiniteOpt.map_dual(fcref::InfiniteOpt.FiniteConstraintRef,
                     key::Val{:TransData})

Map the dual of the appropriate transcription constraint in the transcription model to fcref.

InfiniteOpt.map_dual — Method

InfiniteOpt.map_dual(icref::InfiniteOpt.InfiniteConstraintRef,
                     key::Val{:TransData})

Map the duals of the appropriate transcription constraints in the transcription model to icref.

JuMP.shadow_price — Method

JuMP.shadow_price(cref::GeneralConstraintRef)

The change in the objective from an infinitesimal relaxation of the constraint. This value is computed from JuMP.dual and can be queried only when has_duals is true and the objective sense is MIN_SENSE or MAX_SENSE (not FEASIBILITY_SENSE). For linear constraints, the shadow prices differ at most in sign from the dual value depending on the objective sense.

InfiniteOpt.map_shadow_price — Function

map_shadow_price(cref::GeneralConstraintRef, key)

Map the shadow price of cref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}.

InfiniteOpt.map_shadow_price — Method

InfiniteOpt.map_shadow_price(fcref::InfiniteOpt.FiniteConstraintRef,
                             key::Val{:TransData})

Map the shadow price of the appropriate transcription constraint in the transcription model to fcref.

InfiniteOpt.map_shadow_price — Method

InfiniteOpt.map_shadow_price(icref::InfiniteOpt.InfiniteConstraintRef,
                             key::Val{:TransData})

Map the shadow prices of the appropriate transcription constraints in the transcription model to icref.