Results

Results

A guide and manual to querying optimized InfiniteOpt models.

Overview

Basic Usage

Termination Queries

Variable Queries

Constraint Queries

Methods

JuMP.add_bridge

JuMP.bridge_constraints

JuMP.optimize!

JuMP.set_optimizer

JuMP.set_parameter

JuMP.set_silent

JuMP.unset_silent

InfiniteOpt.build_optimizer_model!

InfiniteOpt.build_optimizer_model!

InfiniteOpt.optimizer_model

InfiniteOpt.optimizer_model_key

InfiniteOpt.optimizer_model_ready

InfiniteOpt.set_optimizer_model

InfiniteOpt.set_optimizer_model_ready

JuMP.termination_status — Method.

JuMP.termination_status(model::InfiniteModel)

Return the reason why the solver stopped (i.e., the MathOptInterface model attribute TerminationStatus).

source

JuMP.raw_status — Method.

JuMP.raw_status(model::InfiniteModel)

Return the reason why the solver stopped in its own words (i.e., the MathOptInterface model attribute RawStatusString).

source

JuMP.primal_status — Method.

 JuMP.primal_status(model::InfiniteModel)

Return the status of the most recent primal solution of the solver (i.e., the MathOptInterface model attribute PrimalStatus).

source

JuMP.dual_status — Method.

JuMP.dual_status(model::InfiniteModel)

Return the status of the most recent dual solution of the solver (i.e., the MathOptInterface model attribute DualStatus).

source

JuMP.solve_time — Method.

JuMP.solve_time(model::InfiniteModel)

If available, returns the solve time reported by the solver. Returns "ArgumentError: ModelLike of type Solver.Optimizer does not support accessing the attribute MathOptInterface.SolveTime()" if the attribute is not implemented.

source

JuMP.has_values — Method.

JuMP.has_values(model::InfiniteModel)

Return true if the solver has a primal solution available to query, otherwise return false. See also JuMP.value.

source

JuMP.objective_bound — Method.

JuMP.objective_bound(model::InfiniteModel)::Float64

Return the best known bound on the optimal objective value after a call to optimize!(model).

source

JuMP.objective_value — Method.

JuMP.objective_value(model::InfiniteModel)::Float64

Return the objective value after a call to optimize!(model).

source

JuMP.value — Method.

JuMP.value(vref::GeneralVariableRef)

Get the value of this variable in the result returned by a solver. Use JuMP.has_values to check if a result exists before asking for values.

source

JuMP.value — Method.

JuMP.value(cref::GeneralConstraintRef)

Get the value of this constraint in the result returned by a solver. Use JuMP.has_values to check if a result exists before asking for values. This returns the primal value of the constraint function.

source

InfiniteOpt.map_value — Function.

map_value(vref::GeneralVariableRef, key)

Map the value of vref to its counterpart in the optimizer model type is distininguished by its extension key key as type Val{ext_key_name}.

source

InfiniteOpt.map_value — Method.

InfiniteOpt.map_value(fvref::InfiniteOpt.FiniteVariableRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription variable in the transcription model to fvref.

source

InfiniteOpt.map_value — Method.

InfiniteOpt.map_value(ivref::InfiniteOpt.InfiniteVariableRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription variables in the transcription model to ivref.

source

InfiniteOpt.map_value — Method.

InfiniteOpt.map_value(fcref::InfiniteOpt.FiniteConstraintRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription constraint function in the transcription model to fcref.

source

InfiniteOpt.map_value — Method.

InfiniteOpt.map_value(icref::InfiniteOpt.GeneralConstraintRef,
                      key::Val{:TransData})

Map the value of the appropriate transcription constraint functions in the transcription model to icref.

source

JuMP.optimizer_index — Method.

JuMP.optimizer_index(vref::GeneralVariableRef)

Return the index of the variables that corresponds to vref in the optimizer model. It throws JuMP.NoOptimizer if no optimizer is set and throws an ErrorException if the optimizer is set but is not attached.

source

JuMP.optimizer_index — Method.

JuMP.optimizer_index(cref::GeneralConstraintRef)

Return the index of the constraints that corresponds to cref in the optimizer model. It throws JuMP.NoOptimizer if no optimizer is set and throws an ErrorException if the optimizer is set but is not attached.

source

InfiniteOpt.map_optimizer_index — Function.

map_optimizer_index(ref, key)

Map the optimizer index of ref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}. Here ref need refer to methods for both variable references and constraint references.

source

InfiniteOpt.map_optimizer_index — Method.

InfiniteOpt.map_optimizer_index(fvref::InfiniteOpt.FiniteVariableRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription variable in the transcription model to fvref.

source

InfiniteOpt.map_optimizer_index — Method.

InfiniteOpt.map_optimizer_index(ivref::InfiniteOpt.InfiniteVariableRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription variables in the transcription model to ivref.

source

InfiniteOpt.map_optimizer_index — Method.

InfiniteOpt.map_optimizer_index(fcref::InfiniteOpt.FiniteConstraintRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription constraint in the transcription model to fcref.

source

InfiniteOpt.map_optimizer_index — Method.

InfiniteOpt.map_optimizer_index(cref::InfiniteOpt.GeneralConstraintRef,
                                key::Val{:TransData})

Map the optimizer model index of the appropriate transcription constraints in the transcription model to cref.

source

JuMP.dual — Method.

JuMP.dual(cref::GeneralConstraintRef)

Get the dual value of this constraint in the result returned by a solver. Use has_dual to check if a result exists before asking for values. See also JuMP.shadow_price.

source

InfiniteOpt.map_dual — Function.

map_dual(cref::GeneralConstraintRef, key)

Map the dual of cref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}.

source

InfiniteOpt.map_dual — Method.

InfiniteOpt.map_dual(fcref::InfiniteOpt.FiniteConstraintRef,
                     key::Val{:TransData})

Map the dual of the appropriate transcription constraint in the transcription model to fcref.

source

InfiniteOpt.map_dual — Method.

InfiniteOpt.map_dual(icref::InfiniteOpt.InfiniteConstraintRef,
                     key::Val{:TransData})

Map the duals of the appropriate transcription constraints in the transcription model to icref.

source

JuMP.shadow_price — Method.

JuMP.shadow_price(cref::GeneralConstraintRef)

The change in the objective from an infinitesimal relaxation of the constraint. This value is computed from JuMP.dual and can be queried only when has_duals is true and the objective sense is MIN_SENSE or MAX_SENSE (not FEASIBILITY_SENSE). For linear constraints, the shadow prices differ at most in sign from the dual value depending on the objective sense.

source

InfiniteOpt.map_shadow_price — Function.

map_shadow_price(cref::GeneralConstraintRef, key)

Map the shadow price of cref to its counterpart in the optimizer model type distininguished by its extension key key as type Val{ext_key_name}.

source

InfiniteOpt.map_shadow_price — Method.

InfiniteOpt.map_shadow_price(fcref::InfiniteOpt.FiniteConstraintRef,
                             key::Val{:TransData})

Map the shadow price of the appropriate transcription constraint in the transcription model to fcref.

source

InfiniteOpt.map_shadow_price — Method.

InfiniteOpt.map_shadow_price(icref::InfiniteOpt.InfiniteConstraintRef,
                             key::Val{:TransData})

Map the shadow prices of the appropriate transcription constraints in the transcription model to icref.

source