Infinite Sets

A guide and manual to the definition and use of infinite sets in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Basic Usage

Interval sets are what characterize the behavior of infinite parameters in InfiniteOpt, since they comprise the domains of infinite parameters. However, most users will not need to work directly with infinite sets and can instead focus on the use of infinite parameters as defined via @infinite_parameter and as discussed on the Infinite Parameters page.

However, for convenience below we summary the infinite sets associated with InfiniteOpt:

Set Type	Domain	Type

IntervalSet	$[lb, ub]$	InfiniteScalarSet

UniDistributionSet	$\sim \mathcal{D} \subseteq \mathbb{R}$	InfiniteScalarSet

MultiDistributionSet	$\sim \mathcal{D} \subseteq \mathbb{R}^n$	InfiniteArraySet

CollectionSet	Combination of Univariate Domains	InfiniteArraySet

Infinite Set Classes

The domain of a given infinite parameter(s) is described by an infinite set inherited from AbstractInfiniteSet. InfiniteOpt natively supports two set sub-groups, namely InfiniteScalarSets and InfiniteArraySets. These correspond to a single independent infinite parameter and a dependent multi-dimensional group of infinite parameters, respectively. We describe each group's natively supported sets below.

Univariate Sets

Univariate infinite sets (i.e., InfiniteScalarSets) are one-dimensional domains ($\subseteq \mathbb{R}$) that describe the behavior of one single independent infinite parameter (i.e., infinite parameters made using independent = true). The two natively supported concrete types are IntervalSets and UniDistributionSets.

IntervalSets describe a continuous interval from some lower bound up to some upper bound. Where the range is inclusive of the bounds. Such sets often arise for parameters that pertain to time and/or spatial position. For example, to define a position interval $[-2, 2]$ we would use:

julia> set = IntervalSet(-2, 2)
[-2, 2]

Note that (semi-)infinite bounds are acceptable, as shown in the following example:

julia> infinite_set = IntervalSet(-Inf, Inf)
[-Inf, Inf]

UniDistributionSets pertain to the co-domain of a univariate distribution. In other words, these correspond to the underlying distributions that characterize uncertain scalar parameters. These sets are compatible with any univariate distribution native to Distributions.jl. For example, let's make a UniDistributionSet that depends on a Beta distribution:

julia> using Distributions;

julia> set = UniDistributionSet(Beta(2,2))
Beta{Float64}(α=2.0, β=2.0)

User-defined distributions are also permissible so long as they are created in accordance with Distributions.jl.

Multivariate Sets

Multivariate infinite sets (i.e., [InfiniteArraySet])(@ref)s are multi-dimensional domains that define the behavior of a group of dependent infinite parameters (i.e., an array of infinite parameters where independent = false). This is a unique feature to InfiniteOpt that enables a much richer set of possibilities for modeling infinite domain. Natively two set types are supported: MultiDistributionSets and CollectionSets.

MultiDistributionSets correspond to the co-domain of a multi-variate (or matrix-variate) distribution which characterizes the behavior of multi-dimensional uncertain parameters. Again, these correspond to any appropriate distribution defined in Distributions.jl. For example, we can make a MultiDistributionSet that depends on a 2-D normal distribution as follows:

julia> using Distributions;

julia> dist = MvNormal([0., 0.], [1. 0.; 0. 2.]);

julia> set = MultiDistributionSet(dist)
FullNormal(
dim: 2
μ: [0.0, 0.0]
Σ: [1.0 0.0; 0.0 2.0]
)

Note

The dimensions (shape) of a chosen distribution used in an MultiDistriubtionSet must match those of the corresponding infinite parameter array.

Finally, CollectionSets are a dependent collection of InfiniteScalarSets that correspond to a group of infinite parameters that are treated dependently. This can be useful when the user wishes to have complete control over how the supports are generated for a group independent parameters where the default combinatorial approach in not wanted. For example, let's make a set of IntervalSets:

julia> set = CollectionSet([IntervalSet(-2, 2), IntervalSet(-1, 4)])
CollectionSet with 2 sets:
 [-2, 2]
 [-1, 4]

Now we could use this set in define a two-dimensional infinite parameter of which we can have the freedom to define a non-combinatorial support grid.

Bound Query/Modification Methods for Infinite Sets

Once an infinite set is created, one can query the lower bound and upper bound of the set similar to how one queries the bounds of a JuMP variable. Thus, the functions JuMP.has_lower_bound, JuMP.has_upper_bound, JuMP.lower_bound, JuMP.upper_bound are all applicable to infinite sets mentioned above. For example, for an IntervalSet [-2, 2] we can query the bound information as follows:

julia> set = IntervalSet(-2, 2);

julia> has_lower_bound(set)
true

julia> has_upper_bound(set)
true

julia> lower_bound(set)
-2.0

julia> upper_bound(set)
2.0

In addition, we can also apply JuMP.set_lower_bound and JuMP.set_upper_bound to IntervalSets to generate a new set with updated bounds. Note that this will not modify the original set. For example, we can change the bounds of the set [-2, 2] as follows:

julia> set_lower_bound(set, -1)
[-1, 2]

julia> set_upper_bound(set, 1)
[-2, 1]

Support Generation for Infinite Sets

InfiniteOpt provides a systematic interface to generate support points for infinite sets. This is crucial as support generation decides how each infinite-dimensional parameter, which is subject to certain infinite set, is discretized later in the transcription stage. The interface will allow users to automatically generate support points using our default methods. Later we will also show that users can also input support points manually for an infinite parameter. Please note that these methods are called by the @infinite_parameter macro when the num_supports keyword is used. Thus, users typically will not need to use this interface directly.

In InfiniteOpt supports can be generated via generate_supports function. For example, let's generate 5 equidistant support points for the IntervalSet [-2, 2]:

julia> supps, label = generate_supports(set, num_supports = 5)
([-2.0, -1.0, 0.0, 1.0, 2.0], :uniform_grid)

julia> supps
5-element Array{Float64,1}:
 -2.0
 -1.0
  0.0
  1.0
  2.0

julia> label
:uniform_grid

Note that the number of supports generated is specified via num_supports keyword argument, which will take a default value of 10 if not specified. The function generate_supports returns a vector of the supports generated, and a label that symbolizes the underlying method. In this case the label returned is UniformGrid, which is the default support generation method for IntervalSets. Another support generation method implemented for IntervalSets is MCSample, which is to sample from a uniform distribution over the interval. To use this mehtod, users need to specify a second positional argument, as shown in the following example:

julia> generate_supports(set, MCSample, num_supports = 5)
([1.29459, 1.64143, -1.34174, -1.29068, -0.88448], :mc_sample)

In this case, the returned label is MCSample, instead of UniformGrid.

generate_supports can also be applied to DistributionSets. The default (and currently only) method implemented for DistributionSets is WeightedSample, which generates Monte Carlo samples that are weighted based on the underlying probability density function of the distribution. For example, a set of support points for a 2-D normal distribution can be generated as follows:

julia> dist = MvNormal([0., 0.], [1. 0.;0. 2.]);

julia> set = MultiDistributionSet(dist);

julia> supps, label = generate_supports(set, num_supports = 3)
([0.679107426036 -0.353007400301 0.586617074633; 1.17155358277 -0.190712174623 0.420496392851], :weighted_sample)

julia> supps
2×3 Array{Float64,2}:
 0.679107  -0.353007  0.586617
 1.17155   -0.190712  0.420496

For those who are interested in coding up their own support generation functions, generate_supports is an inteface that calls the proper generate_support_values function based on the type of set and value of method. Therefore, to use custom support generation methods, users can implement extensions for generate_support_values with a different method label from the existing methods. See Extensions for full details.

User Defined Sets

Furthermore, custom infinite sets that inherit AbstractInfiniteSet can also be defined. See Extensions for more information.

Datatypes

InfiniteOpt.AbstractInfiniteSet

InfiniteOpt.CollectionSet

InfiniteOpt.InfiniteArraySet

InfiniteOpt.InfiniteScalarSet

InfiniteOpt.IntervalSet

InfiniteOpt.MultiDistributionSet

InfiniteOpt.UniDistributionSet

InfiniteOpt.AbstractInfiniteSet — Type

AbstractInfiniteSet

An abstract type for sets that characterize infinite parameters.

InfiniteOpt.InfiniteScalarSet — Type

InfiniteScalarSet <: AbstractInfiniteSet

An abstract type for infinite sets that are one-dimensional.

InfiniteOpt.IntervalSet — Type

IntervalSet <: InfiniteScalarSet

A DataType that stores the lower and upper interval bounds for infinite parameters that are continuous over a certain that interval. This is for use with a IndependentParameter.

Fields

lower_bound::Float64 Lower bound of the infinite parameter.

upper_bound::Float64 Upper bound of the infinite parameter.

InfiniteOpt.UniDistributionSet — Type

UniDistributionSet{T <: Distributions.UnivariateDistribution} <: InfiniteScalarSet

A DataType that stores the distribution characterizing an infinite parameter that is random. This is for use with a IndependentParameter.

Fields

distribution::T Distribution of the random parameter.

InfiniteOpt.InfiniteArraySet — Type

InfiniteArraySet <: AbstractInfiniteSet

An abstract type for multi-dimensional infinite sets.

InfiniteOpt.MultiDistributionSet — Type

MultiDistributionSet{T <: NonUnivariateDistribution} <: InfiniteArraySet

A DataType that stores the distribution characterizing a collection of infinite parameters that follows its form. This is for use with DependentParameters.

Fields

distribution::T Distribution of the random parameters.

InfiniteOpt.CollectionSet — Type

CollectionSet{T <: InfiniteScalarSet} <: InfiniteArraySet

A DataType that stores a collection of InfiniteScalarSets characterizing a collection of infinite parameters that follows its form. This is for use with DependentParameters.

Fields

sets::Array{T} The collection of scalar sets.

Methods

InfiniteOpt.collection_sets

InfiniteOpt.generate_support_values

InfiniteOpt.generate_supports

InfiniteOpt.supports_in_set

JuMP.has_lower_bound

JuMP.has_upper_bound

JuMP.lower_bound

JuMP.set_lower_bound

JuMP.set_upper_bound

JuMP.upper_bound

InfiniteOpt.collection_sets — Function

collection_sets(set::AbstractInfiniteSet)

Return the array of sets associated with a CollectionSet. Error if the input set is not a CollectionSet.

JuMP.has_lower_bound — Method

JuMP.has_lower_bound(set::AbstractInfiniteSet)::Bool

Return Bool indicating if set has a lower bound that can be determined. This should be extended for user-defined infinite sets. It defaults to false for unrecognized set types.

Example

julia> set = InfiniteSet(0, 1);

julia> has_lower_bound(set)
true

JuMP.lower_bound — Method

JuMP.lower_bound(set::AbstractInfiniteSet)::Union{Real, Vector{<:Real}}

Return the lower bound of set if one exists. This should be extended for user-defined infinite sets if appropriate. Errors if JuMP.has_lower_bound returns false. Extensions are enabled by JuMP.has_lower_bound(set) and JuMP.lower_bound(set).

Example

julia> set = InfiniteSet(0, 1);

julia> lower_bound(set)
0.0

JuMP.set_lower_bound — Method

JuMP.set_lower_bound(set::AbstractInfiniteSet,
                     lower::Union{Real, Vector{<:Real}})::AbstractInfiniteSet

Set and return the lower bound of set if such an operation makes sense. Errors if the type of set does not support this operation or has not been extended. User-defined set types should extend this if appropriate.

Example

julia> set = InfiniteSet(0, 1);

julia> set_lower_bound(set, 0.5)
[0.5, 1]

JuMP.has_upper_bound — Method

JuMP.has_upper_bound(set::AbstractInfiniteSet)::Bool

Return Bool indicating if set has a upper bound that can be determined. This should be extended for user-defined infinite sets. It defaults to false for unrecognized set types.

Example

julia> set = InfiniteSet(0, 1);

julia> has_upper_bound(set)
true

JuMP.upper_bound — Method

JuMP.upper_bound(set::AbstractInfiniteSet)::Union{Real, Vector{<:Real}}

Return the upper bound of set if one exists. This should be extended for user-defined infinite sets if appropriate. Errors if JuMP.has_upper_bound returns false. Extensions are enabled by JuMP.has_upper_bound(set) and JuMP.upper_bound(set).

Example

julia> set = InfiniteSet(0, 1);

julia> upper_bound(set)
1.0

JuMP.set_upper_bound — Method

JuMP.set_upper_bound(set::AbstractInfiniteSet,
                     upper::Real)::AbstractInfiniteSet

Set and return the upper bound of set if such an aoperation makes sense. Errors if the type of set does not support this operation or has not been extended. User-defined set types should extend this if appropriate.

Example

julia> set = InfiniteSet(0, 1);

julia> set_upper_bound(set, 0.5)
[0, 0.5]

InfiniteOpt.supports_in_set — Function

supports_in_set(supports::Union{Real, Vector{<:Real}, Array{<:Real, 2}},
                set::AbstractInfiniteSet)::Bool

Used to check if supports are in the domain of set. Returns true if supports are in domain of set and returns false otherwise. This is primarily an internal method for performing checks but can be extended for user-defined set types. Extending this is optional, but recommended where possible. Note by fallback, this returns true for unrecognized set types such that an error won't be thrown.

InfiniteOpt.generate_supports — Function

generate_supports(set::AbstractInfiniteSet
                  method::Union{Symbol, Nothing} = Nothing;
                  [num_supports::Int = DefaultNumSupports,
                  sig_digits::Int = DefaultSigDigits]
                  )::Tuple{Array{<:Real}, Symbol}

Generate num_supports support values with sig_digits significant digits in accordance with set and return them along with the correct generation label(s). IntervalSets generate supports uniformly with label UniformGrid and distribution sets generate them randomly accordingly to the underlying distribution. Moreover, method indicates the generation method that should be used. These methods correspond to parameter support labels. Current labels that can be used as generation methods include (but may not be defined for certain set types):

MCSample: Uniformly distributed Monte Carlo samples.

WeightedSample: Monte Carlo samples that are weighted by an underlying PDF.

UniformGrid: Samples that are generated uniformly over the set domain.

Extensions that employ user-defined infinite set types and/or methods should extend generate_support_values to enable this. Errors if the set type and /or methods are unrecognized. This is intended as an internal method to be used by methods such as generate_and_add_supports!.

InfiniteOpt.generate_support_values — Function

generate_support_values(set::AbstractInfiniteSet,
                        [method::Val{:MyMethod} = Val(:MyMethod)];
                        num_supports::Int = DefaultNumSupports,
                        sig_digits::Int = DefaultSigDigits,
                        )::Tuple{Array{<:Real}, Symbol}

A multiple dispatch method for generate_supports. This will return a tuple where the first element are the supports and the second is their label. This can be extended for user-defined infinite sets and/or generation methods. When defining a new set type the default method dispatch should make method an optional argument (making it the default). Otherwise, other method dispatches for a given set must ensure that method is positional argument without a default value (contrary to the definition above).