Infinite Sets

A guide and manual to the definition and use of infinite sets in InfiniteOpt. The Datatypes and Methods sections at the end comprise the manual, and the above sections comprise the guide.

Basic Usage

Interval sets are what characterize the behavior of infinite parameters in InfiniteOpt, since they comprise the domains of infinite parameters. However, most users will not need to work directly with infinite sets and can instead focus on the use of infinite parameters as defined via @infinite_parameter and as discussed on the Infinite Parameters page.

However, for convenience below we summary the infinite sets associated with InfiniteOpt:

Set Type	Domain

IntervalSet	$[lb, ub]$

DistributionSet	$\sim \mathcal{D}$

IntervalSets

The domain of a given infinite parameter is described by an infinite set inherited from AbstractInfiniteSet. InfiniteOpt natively supports two such sets. The first is IntervalSet which describes a continuous interval from some lower bound up to some upper bound. Typically, this range is inclusive of the bounds. Such sets often arise for parameters that pertain to time and/or spatial position. For example, to define a position interval $[-2, 2]$ we would use:

julia> set = IntervalSet(-2, 2)
[-2, 2]

Note that infinite bounds are acceptable.

DistributionSets

The second kind of set is that of DistributionSet which contains distribution from Distributions.jl that characterizes an infinite parameter. Univariate and multivariate distributions can be used, but multivariate distributions are only valid when defining an array of parameters with equivalent dimensions. For example, let's make a DistributionSet that depends on a Beta distribution:

julia> set = DistributionSet(Beta(2,2))
Beta{Float64}(α=2.0, β=2.0)

User-defined distributions are also permissible so long as they are created in accordance with Distributions.jl.

User Defined Sets

Furthermore, custom infinite sets that inherit AbstractInfiniteSet can also be defined. See Extensions for more information.

Datatypes

InfiniteOpt.AbstractInfiniteSet

InfiniteOpt.DistributionSet

InfiniteOpt.IntervalSet

InfiniteOpt.IntervalSet

InfiniteOpt.AbstractInfiniteSet — Type

AbstractInfiniteSet

An abstract type for sets that characterize infinite parameters.

InfiniteOpt.IntervalSet — Type

IntervalSet <: AbstractInfiniteSet

A DataType that stores the lower and upper interval bounds for infinite parameters that are continuous over a certain that interval.

Fields

lower_bound::Float64 Lower bound of the infinite parameter.

upper_bound::Float64 Upper bound of the infinite parameter.

InfiniteOpt.DistributionSet — Type

DistributionSet{T <: Distributions.NonMatrixDistribution} <: AbstractInfiniteSet

A DataType that stores the distribution characterizing infinite parameters that are random.

Fields

distribution::T Distribution of the random parameter.

Methods

InfiniteOpt.generate_support_values

InfiniteOpt.supports_in_set

JuMP.has_lower_bound

JuMP.has_upper_bound

JuMP.lower_bound

JuMP.set_lower_bound

JuMP.set_upper_bound

JuMP.upper_bound

InfiniteOpt.IntervalSet — Method

IntervalSet(lower_bound::Number, upper_bound::Number)

A constructor for IntervalSet that converts values of type Number to values of type Float64 as required by IntervalSet.

JuMP.has_lower_bound — Method

JuMP.has_lower_bound(set::AbstractInfiniteSet)::Bool

Return Bool indicating if set has a lower bound that can be determined. This should be extended for user-defined infinite sets. It defaults to false for unrecognized set types.

Example

julia> set = InfiniteSet(0, 1);

julia> has_lower_bound(set)
true

JuMP.lower_bound — Method

JuMP.lower_bound(set::AbstractInfiniteSet)::Number

Return the lower bound of set if one exists. This should be extended for user-defined infinite sets if appropriate. Errors if JuMP.has_lower_bound returns false. Extensions are enabled by JuMP.has_lower_bound(set) and JuMP.lower_bound(set).

Example

julia> set = InfiniteSet(0, 1);

julia> lower_bound(set)
0.0

JuMP.set_lower_bound — Method

JuMP.set_lower_bound(set::AbstractInfiniteSet,
                     lower::Number)::AbstractInfiniteSet

Set and return the lower bound of set if such an aoperation makes sense. Errors if the type of set does not support this operation or has not been extended. User-defined set types should extend this if appropriate.

Example

julia> set = InfiniteSet(0, 1);

julia> set_lower_bound(set, 0.5)
[0.5, 1]

JuMP.has_upper_bound — Method

JuMP.has_upper_bound(set::AbstractInfiniteSet)::Bool

Return Bool indicating if set has a upper bound that can be determined. This should be extended for user-defined infinite sets. It defaults to false for unrecognized set types.

Example

julia> set = InfiniteSet(0, 1);

julia> has_upper_bound(set)
true

JuMP.upper_bound — Method

JuMP.upper_bound(set::AbstractInfiniteSet)::Number

Return the upper bound of set if one exists. This should be extended for user-defined infinite sets if appropriate. Errors if JuMP.has_upper_bound returns false. Extensions are enabled by JuMP.has_upper_bound(set) and JuMP.upper_bound(set).

Example

julia> set = InfiniteSet(0, 1);

julia> upper_bound(set)
1.0

JuMP.set_upper_bound — Method

JuMP.set_upper_bound(set::AbstractInfiniteSet,
                     upper::Number)::AbstractInfiniteSet

Set and return the upper bound of set if such an aoperation makes sense. Errors if the type of set does not support this operation or has not been extended. User-defined set types should extend this if appropriate.

Example

julia> set = InfiniteSet(0, 1);

julia> set_upper_bound(set, 0.5)
[0, 0.5]

InfiniteOpt.supports_in_set — Function

supports_in_set(supports::Union{Number, Vector{<:Number}},
                set::AbstractInfiniteSet)::Bool

Used to check if supports are in the domain of set. Returns true if supports are in domain of set and returns false otherwise. This is primarily an internal method for performing checks but can be extended for user-defined set types. Extending this is optional, but recommended where possible. Note by fallback, this returns true for unrecognized set types such that an error won't be thrown.

InfiniteOpt.generate_support_values — Function

generate_support_values(set::AbstractInfiniteSet; [num_supports::Int = 10,
                        sig_fig::Int = 5])

Generate num_supports support values with sig_figs significant digits in accordance with set and return them. IntervalSets generate supports uniformly and DistributionSets generate them randomly accordingly to the underlyign distribution. Extensions that employ user-defined infinite set types should extend according to the new set type. Errors if the set is a type that has not been explicitly extended. This is intended as an internal method to be used by generate_and_add_supports! and build_parameter.