Model Transcription

Model Transcription

A guide and manual for transcribing infinite models using InfiniteOpt.

Overview

Basic Usage

TranscriptionOpt

Datatypes

InfiniteOpt.TranscriptionOpt.TranscriptionData — Type.

TranscriptionData

A DataType for storing the data mapping an InfiniteOpt.InfiniteModel that has been transcribed to a regular JuMP.Model that contains the transcribed variables. This is stored in the ext field of a JuMP.Model to make what is called a TranscriptionModel via the TranscriptionModel constructor.

Fields

infinite_to_vars::Dict{InfiniteOpt.InfiniteVariableRef, Vector{JuMP.VariableRef}}: Infinite variables to their transcribed variables.

global_to_var::Dict{InfiniteOpt.GlobalVariableRef, JuMP.VariableRef}: Global variables to model variables.

point_to_var::Dict{InfiniteOpt.PointVariableRef, JuMP.VariableRef}: Point variables to model variables.

infvar_to_supports::Dict{InfiniteOpt.InfiniteVariableRef, Dict}: Infinite variables to transcribed supports indexed by their numeric aliases.

infinite_to_constrs::Dict{InfiniteOpt.InfiniteConstraintRef, Vector{JuMP.ConstraintRef}}: Infinite constraints to their transcribed constraints.

measure_to_constrs::Dict{InfiniteOpt.MeasureConstraintRef, Vector{JuMP.ConstraintRef}}: Measure constraints to model constraints.

finite_to_constr::Dict{InfiniteOpt.FiniteConstraintRef, JuMP.ConstraintRef}: Finite constraints to model constraints.

infconstr_to_supports::Dict{InfiniteOpt.InfiniteConstraintRef, Dict}: Infinite constraints to the transcribed supports indxed by their numeric aliases.

measconstr_to_supports::Dict{InfiniteOpt.MeasureConstraintRef, Dict}: Measure constraints to the transcribed supports indxed by their numeric aliases.

infconstr_to_params::Dict{InfiniteOpt.InfiniteConstraintRef, Tuple}: Infinite constraints to the parameter tuples associated with each transcribed support.

measconstr_to_params::Dict{InfiniteOpt.MeasureConstraintRef, Tuple}: Measure constraints to the parameter tuples associated with each transcribed support.

source

Methods

InfiniteOpt.build_optimizer_model!

InfiniteOpt.parameter_refs

InfiniteOpt.parameter_refs

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.TranscriptionOpt.TranscriptionModel

InfiniteOpt.TranscriptionOpt.TranscriptionModel

InfiniteOpt.TranscriptionOpt.is_transcription_model

InfiniteOpt.TranscriptionOpt.transcription_constraint

InfiniteOpt.TranscriptionOpt.transcription_data

InfiniteOpt.TranscriptionOpt.transcription_model

InfiniteOpt.TranscriptionOpt.transcription_variable

InfiniteOpt.TranscriptionOpt.transcription_model — Function.

transcription_model(model::InfiniteOpt.InfiniteModel)::JuMP.Model

Return the transcription model stored in model if that is what is stored in model.optimizer_model.

source

InfiniteOpt.build_optimizer_model! — Method.

InfiniteOpt.build_optimizer_model!(model::InfiniteOpt.InfiniteModel,
                                   key::Val{:TransData})

Transcribe model and store it as a TranscriptionModel in the model.optimizer_model field which can be accessed with transcription_model.

source

InfiniteOpt.TranscriptionOpt.TranscriptionModel — Method.

TranscriptionModel([optimizer_factory::JuMP.OptimizerFactory];
                   [caching_mode::MOIU.CachingOptimizerMode = MOIU.AUTOMATIC,
                    bridge_constraints::Bool = true])::JuMP.Model

Return a JuMP.Model with TranscriptionData included in the ext data field. Accepts the same arguments as a typical JuMP Model.

Example

julia> TranscriptionModel()
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

source

InfiniteOpt.TranscriptionOpt.TranscriptionModel — Method.

TranscriptionModel(model::InfiniteModel, args...)

Return a TranscriptionModel of model. This transcribes all of the variables, constraints, and objective.

Example

julia> TranscriptionModel(model)
A JuMP Model
Feasibility problem with:
Variables: 130
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 25 constraints
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.GreaterThan{Float64}`: 100 constraint
`GenericAffExpr{Float64,VariableRef}`-in-`MathOptInterface.LessThan{Float64}`: 84 constraints
`VariableRef`-in-`MathOptInterface.EqualTo{Float64}`: 40 constraints
`VariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 40 constraints
`VariableRef`-in-`MathOptInterface.LessThan{Float64}`: 25 constraints
`VariableRef`-in-`MathOptInterface.Integer`: 40 constraints
`VariableRef`-in-`MathOptInterface.ZeroOne`: 40 constraints
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

source

InfiniteOpt.TranscriptionOpt.is_transcription_model — Function.

is_transcription_model(model::JuMP.Model)::Bool

Return true if model is a TranscriptionModel or false otherwise.

Example

julia> is_transcription_model(model)
true

source

InfiniteOpt.TranscriptionOpt.transcription_data — Function.

transcription_data(model::JuMP.Model)::TranscriptionData

Return the TranscriptionData from a TranscriptionModel. Errors if it is not a TranscriptionModel.

source

InfiniteOpt.TranscriptionOpt.transcription_variable — Function.

transcription_variable(model::JuMP.Model,
                       vref::InfiniteOpt.InfOptVariableRef)

Return the transcribed variable reference(s) corresponding to vref. Errors if no transcription variable is found.

Example

julia> transcription_variable(trans_model, infvar)
2-element Array{VariableRef,1}:
 infvar(support: 1)
 infvar(support: 2)

julia> transcription_variable(trans_model, gbvar)
gbvar

source

InfiniteOpt.supports — Method.

InfiniteOpt.supports(model::JuMP.Model,
                     vref::InfiniteOpt.InfiniteVariableRef)::Vector

Return the support alias mapping associated with vref in the transcribed model. Errors if vref does not have transcribed variables.

source

InfiniteOpt.supports — Method.

InfiniteOpt.supports(vref::InfiniteOpt.InfiniteVariableRef)::Vector

Return the support alias mapping associated with vref in the transcription model. Errors if the infinite model does not contain a transcription model or if vref is not transcribed.

Example

julia> supports(vref)
Dict{Int64,Tuple{Float64}} with 2 entries:
  2 => (1.0,)
  1 => (0.0,)

source

InfiniteOpt.TranscriptionOpt.transcription_constraint — Function.

transcription_constraint(model::JuMP.Model,
                         cref::InfiniteOpt.GeneralConstraintRef)

Return the transcribed constraint reference(s) corresponding to cref. Errors if cref has not been transcribed.

Example

julia> transcription_constraint(trans_model, fin_con)
fin_con : x(support: 1) - y <= 3.0

source

InfiniteOpt.supports — Method.

InfiniteOpt.supports(model::JuMP.Model,
                     cref::InfiniteOpt.GeneralConstraintRef)::Vector

Return the support alias mappings associated with cref. Errors if cref is not transcribed.

source

InfiniteOpt.supports — Method.

InfiniteOpt.supports(cref::InfiniteOpt.GeneralConstraintRef)::Vector

Return the support alias mappings associated with cref. Errors if cref is not transcribed or if the infinite model does not have a transcription model.

Example

julia> supports(cref)
Dict{Int64,Tuple{Float64}} with 2 entries:
  2 => (1.0,)
  1 => (0.0,)

source

InfiniteOpt.parameter_refs — Method.

InfiniteOpt.parameter_refs(model::JuMP.Model,
                           cref::InfiniteOpt.GeneralConstraintRef)::Tuple

Return the a parameter reference tuple of all the parameters that parameterize cref and correspond to the supports. Errors if cref has not been transcribed.

source

InfiniteOpt.parameter_refs — Method.

InfiniteOpt.parameter_refs(cref::InfiniteOpt.GeneralConstraintRef)::Tuple

Return the a parameter reference tuple of all the parameters that parameterize cref and correspond to the supports. Errors if cref has not been transcribed or if the infinite model does not have a transcription model associated with it.

Example

julia> parameter_refs(cref)
(t, x)

source