Library

Method, DataType, and Macro Index

InfiniteOpt.AbstractInfiniteSet

InfiniteOpt.AbstractMeasureData

InfiniteOpt.AbstractReducedInfo

InfiniteOpt.BoundedScalarConstraint

InfiniteOpt.DiscreteMeasureData

InfiniteOpt.DiscreteMeasureData

InfiniteOpt.DiscreteMeasureData

InfiniteOpt.DistributionSet

InfiniteOpt.FiniteConstraintRef

InfiniteOpt.FiniteVariableRef

InfiniteOpt.GeneralConstraintRef

InfiniteOpt.GeneralVariableRef

InfiniteOpt.GlobalVariable

InfiniteOpt.GlobalVariableRef

InfiniteOpt.InfOptParameter

InfiniteOpt.InfOptVariable

InfiniteOpt.InfiniteConstraintRef

InfiniteOpt.InfiniteModel

InfiniteOpt.InfiniteModel

InfiniteOpt.InfiniteVariable

InfiniteOpt.InfiniteVariableRef

InfiniteOpt.IntervalSet

InfiniteOpt.IntervalSet

InfiniteOpt.Measure

InfiniteOpt.MeasureConstraintRef

InfiniteOpt.MeasureFiniteVariableRef

InfiniteOpt.MeasureRef

InfiniteOpt.MultiDiscreteMeasureData

InfiniteOpt.ParameterRef

InfiniteOpt.PointVariable

InfiniteOpt.PointVariableRef

InfiniteOpt.ReducedInfiniteInfo

InfiniteOpt.ReducedInfiniteVariableRef

InfiniteOpt.TranscriptionOpt.TranscriptionData

JuMP.Model

JuMP.Model

JuMP.Model

JuMP.NoOptimizer

JuMP.PSDCone

JuMP.RotatedSecondOrderCone

JuMP.ScalarConstraint

JuMP.SecondOrderCone

InfiniteOpt.TranscriptionOpt.TranscriptionModel

InfiniteOpt.TranscriptionOpt.TranscriptionModel

InfiniteOpt.TranscriptionOpt.is_transcription_model

InfiniteOpt.TranscriptionOpt.transcription_constraint

InfiniteOpt.TranscriptionOpt.transcription_data

InfiniteOpt.TranscriptionOpt.transcription_model

InfiniteOpt.TranscriptionOpt.transcription_variable

InfiniteOpt.add_measure

InfiniteOpt.add_parameter

InfiniteOpt.add_parameter_ref

InfiniteOpt.add_supports

InfiniteOpt.all_parameters

InfiniteOpt.build_optimizer_model!

InfiniteOpt.build_optimizer_model!

InfiniteOpt.build_optimizer_model!

InfiniteOpt.build_parameter

InfiniteOpt.delete_supports

InfiniteOpt.eval_supports

InfiniteOpt.expand

InfiniteOpt.expand_all_measures!

InfiniteOpt.group_id

InfiniteOpt.group_id

InfiniteOpt.has_supports

InfiniteOpt.infinite_set

InfiniteOpt.infinite_variable_ref

InfiniteOpt.infinite_variable_ref

InfiniteOpt.is_independent

InfiniteOpt.is_used

InfiniteOpt.is_used

InfiniteOpt.is_used

InfiniteOpt.is_used

InfiniteOpt.map_dual

InfiniteOpt.map_dual

InfiniteOpt.map_dual

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_optimizer_index

InfiniteOpt.map_shadow_price

InfiniteOpt.map_shadow_price

InfiniteOpt.map_shadow_price

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.map_value

InfiniteOpt.measure

InfiniteOpt.measure_data

InfiniteOpt.measure_function

InfiniteOpt.num_parameters

InfiniteOpt.num_supports

InfiniteOpt.optimizer_model

InfiniteOpt.optimizer_model_key

InfiniteOpt.optimizer_model_ready

InfiniteOpt.parameter_by_name

InfiniteOpt.parameter_refs

InfiniteOpt.parameter_refs

InfiniteOpt.parameter_refs

InfiniteOpt.parameter_refs

InfiniteOpt.parameter_values

InfiniteOpt.set_independent

InfiniteOpt.set_infinite_set

InfiniteOpt.set_optimizer_model

InfiniteOpt.set_optimizer_model_ready

InfiniteOpt.set_parameter_refs

InfiniteOpt.set_supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.supports

InfiniteOpt.unset_independent

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_constraint

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_measure

InfiniteOpt.used_by_objective

InfiniteOpt.used_by_objective

InfiniteOpt.used_by_point_variable

InfiniteOpt.used_by_reduced_variable

InfiniteOpt.used_by_variable

JuMP.BinaryRef

JuMP.BinaryRef

JuMP.FixRef

JuMP.FixRef

JuMP.IntegerRef

JuMP.IntegerRef

JuMP.LowerBoundRef

JuMP.LowerBoundRef

JuMP.UpperBoundRef

JuMP.UpperBoundRef

JuMP.add_bridge

JuMP.add_constraint

JuMP.add_to_expression!

JuMP.add_to_function_constant

JuMP.add_variable

JuMP.all_constraints

JuMP.all_constraints

JuMP.all_constraints

JuMP.all_constraints

JuMP.all_variables

JuMP.bridge_constraints

JuMP.build_constraint

JuMP.build_variable

JuMP.constraint_by_name

JuMP.constraint_object

JuMP.delete

JuMP.delete

JuMP.delete

JuMP.delete

JuMP.delete

JuMP.delete_lower_bound

JuMP.delete_upper_bound

JuMP.dual

JuMP.dual_status

JuMP.fix

JuMP.fix_value

JuMP.fix_value

JuMP.has_lower_bound

JuMP.has_lower_bound

JuMP.has_lower_bound

JuMP.has_upper_bound

JuMP.has_upper_bound

JuMP.has_upper_bound

JuMP.has_values

JuMP.index

JuMP.index

JuMP.is_binary

JuMP.is_binary

JuMP.is_fixed

JuMP.is_fixed

JuMP.is_integer

JuMP.is_integer

JuMP.is_valid

JuMP.is_valid

JuMP.is_valid

JuMP.is_valid

JuMP.is_valid

JuMP.list_of_constraint_types

JuMP.lower_bound

JuMP.lower_bound

JuMP.lower_bound

JuMP.name

JuMP.name

JuMP.name

JuMP.name

JuMP.name

JuMP.normalized_coefficient

JuMP.normalized_rhs

JuMP.num_constraints

JuMP.num_constraints

JuMP.num_constraints

JuMP.num_constraints

JuMP.num_variables

JuMP.objective_bound

JuMP.objective_function

JuMP.objective_function_type

JuMP.objective_sense

JuMP.objective_value

JuMP.optimize!

JuMP.optimizer_index

JuMP.optimizer_index

JuMP.owner_model

JuMP.owner_model

JuMP.primal_status

JuMP.raw_status

JuMP.set_binary

JuMP.set_integer

JuMP.set_lower_bound

JuMP.set_lower_bound

JuMP.set_name

JuMP.set_name

JuMP.set_name

JuMP.set_name

JuMP.set_name

JuMP.set_name

JuMP.set_normalized_coefficient

JuMP.set_normalized_rhs

JuMP.set_objective

JuMP.set_objective_coefficient

JuMP.set_objective_function

JuMP.set_objective_function

JuMP.set_objective_sense

JuMP.set_optimizer

JuMP.set_parameter

JuMP.set_silent

JuMP.set_start_value

JuMP.set_upper_bound

JuMP.set_upper_bound

JuMP.shadow_price

JuMP.solve_time

JuMP.start_value

JuMP.start_value

JuMP.termination_status

JuMP.unfix

JuMP.unset_binary

JuMP.unset_integer

JuMP.unset_silent

JuMP.upper_bound

JuMP.upper_bound

JuMP.upper_bound

JuMP.value

JuMP.value

JuMP.variable_by_name

JuMP.with_optimizer

InfiniteOpt.@global_variable

InfiniteOpt.@infinite_parameter

InfiniteOpt.@infinite_variable

InfiniteOpt.@point_variable

JuMP.@constraint

JuMP.@expression

JuMP.@objective

JuMP Documentation

JuMP.Model — Type.

Model

A mathematical model of an optimization problem.

JuMP.Model — Method.

Model(; caching_mode::MOIU.CachingOptimizerMode=MOIU.AUTOMATIC,
        bridge_constraints::Bool=true)

Return a new JuMP model without any optimizer; the model is stored the model in a cache. The mode of the CachingOptimizer storing this cache is caching_mode. The optimizer can be set later in the optimize! call. If bridge_constraints is true, constraints that are not supported by the optimizer are automatically bridged to equivalent supported constraints when an appropriate transformation is defined in the MathOptInterface.Bridges module or is defined in another module and is explicitely added.

JuMP.Model — Method.

Model(optimizer_factory::OptimizerFactory;
      caching_mode::MOIU.CachingOptimizerMode=MOIU.AUTOMATIC,
      bridge_constraints::Bool=true)

Return a new JuMP model using the optimizer factory optimizer_factory to create the optimizer. The optimizer factory can be created by the with_optimizer function.

Examples

The following creates a model using the optimizer Ipopt.Optimizer(print_level=0):

model = Model(with_optimizer(Ipopt.Optimizer, print_level=0))

JuMP.with_optimizer — Function.

with_optimizer(constructor, args...; kwargs...)

Return an OptimizerFactory that creates optimizers using the constructor constructor with positional arguments args and keyword arguments kwargs.

Examples

The following returns an optimizer factory that creates Ipopt.Optimizers using the constructor call Ipopt.Optimizer(print_level=0):

with_optimizer(Ipopt.Optimizer, print_level=0)

JuMP.@expression — Macro.

@expression(args...)

Efficiently builds a linear or quadratic expression but does not add to model immediately. Instead, returns the expression which can then be inserted in other constraints. For example:

@expression(m, shared, sum(i*x[i] for i=1:5))
@constraint(m, shared + y >= 5)
@constraint(m, shared + z <= 10)

The ref accepts index sets in the same way as @variable, and those indices can be used in the construction of the expressions:

@expression(m, expr[i=1:3], i*sum(x[j] for j=1:3))

Anonymous syntax is also supported:

expr = @expression(m, [i=1:3], i*sum(x[j] for j=1:3))

JuMP.add_to_expression! — Function.

add_to_expression!(expression, terms...)

Updates expression in place to expression + (*)(terms...). This is typically much more efficient than expression += (*)(terms...). For example, add_to_expression!(expression, a, b) produces the same result as expression += a*b, and add_to_expression!(expression, a) produces the same result as expression += a.

Only a few methods are defined, mostly for internal use, and only for the cases when (1) they can be implemented efficiently and (2) expression is capable of storing the result. For example, add_to_expression!(::AffExpr, ::VariableRef, ::VariableRef) is not defined because a GenericAffExpr cannot store the product of two variables.

JuMP.@objective — Macro.

@objective(model::Model, sense, func)

Set the objective sense to sense and objective function to func. The objective sense can be either Min, Max, MathOptInterface.MIN_SENSE, MathOptInterface.MAX_SENSE or MathOptInterface.FEASIBILITY_SENSE; see MathOptInterface.ObjectiveSense. In order to set the sense programatically, i.e., when sense is a Julia variable whose value is the sense, one of the three MathOptInterface.ObjectiveSense values should be used. The function func can be a single JuMP variable, an affine expression of JuMP variables or a quadratic expression of JuMP variables.

Examples

To minimize the value of the variable x, do as follows:

julia> model = Model()
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.

julia> @variable(model, x)
x

julia> @objective(model, Min, x)
x

To maximize the value of the affine expression 2x - 1, do as follows:

julia> @objective(model, Max, 2x - 1)
2 x - 1

To set a quadratic objective and set the objective sense programatically, do as follows:

julia> sense = MOI.MIN_SENSE
MIN_SENSE::OptimizationSense = 0

julia> @objective(model, sense, x^2 - 2x + 1)
x² - 2 x + 1

JuMP.ScalarConstraint — Type.

struct ScalarConstraint

The data for a scalar constraint. The func field containts a JuMP object representing the function and the set field contains the MOI set. See also the documentation on JuMP's representation of constraints for more background.

JuMP.@constraint — Macro.

@constraint(m::Model, expr)

Add a constraint described by the expression expr.

@constraint(m::Model, ref[i=..., j=..., ...], expr)

Add a group of constraints described by the expression expr parametrized by i, j, ...

The expression expr can either be

of the form func in set constraining the function func to belong to the set set which is either a MathOptInterface.AbstractSet or one of the JuMP shortcuts SecondOrderCone, RotatedSecondOrderCone and PSDCone, e.g. @constraint(model, [1, x-1, y-2] in SecondOrderCone()) constrains the norm of [x-1, y-2] be less than 1;

of the form a sign b, where sign is one of ==, ≥, >=, ≤ and <= building the single constraint enforcing the comparison to hold for the expression a and b, e.g. @constraint(m, x^2 + y^2 == 1) constrains x and y to lie on the unit circle;

of the form a ≤ b ≤ c or a ≥ b ≥ c (where ≤ and <= (resp. ≥ and >=) can be used interchangeably) constraining the paired the expression b to lie between a and c;

of the forms @constraint(m, a .sign b) or @constraint(m, a .sign b .sign c) which broadcast the constraint creation to each element of the vectors.

Note for extending the constraint macro

Each constraint will be created using add_constraint(m, build_constraint(_error, func, set)) where

_error is an error function showing the constraint call in addition to the error message given as argument,

func is the expression that is constrained

and set is the set in which it is constrained to belong.

For expr of the first type (i.e. @constraint(m, func in set)), func and set are passed unchanged to build_constraint but for the other types, they are determined from the expressions and signs. For instance, @constraint(m, x^2 + y^2 == 1) is transformed into add_constraint(m, build_constraint(_error, x^2 + y^2, MOI.EqualTo(1.0))).

To extend JuMP to accept new constraints of this form, it is necessary to add the corresponding methods to build_constraint. Note that this will likely mean that either func or set will be some custom type, rather than e.g. a Symbol, since we will likely want to dispatch on the type of the function or set appearing in the constraint.

JuMP.SecondOrderCone — Type.

SecondOrderCone

Second order cone object that can be used to constrain the euclidean norm of a vector x to be less than or equal to a nonnegative scalar t. This is a shortcut for the MathOptInterface.SecondOrderCone.

Examples

The following constrains $\|(x-1, x-2)\|_2 \le t$ and $t \ge 0$:

julia> model = Model();

julia> @variable(model, x)
x

julia> @variable(model, t)
t

julia> @constraint(model, [t, x-1, x-2] in SecondOrderCone())
[t, x - 1, x - 2] ∈ MathOptInterface.SecondOrderCone(3)

JuMP.RotatedSecondOrderCone — Type.

RotatedSecondOrderCone

Rotated second order cone object that can be used to constrain the square of the euclidean norm of a vector x to be less than or equal to $2tu$ where t and u are nonnegative scalars. This is a shortcut for the MathOptInterface.RotatedSecondOrderCone.

Examples

The following constrains $\|(x-1, x-2)\|_2 \le 2tx$ and $t, x \ge 0$:

julia> model = Model();

julia> @variable(model, x)
x

julia> @variable(model, t)
t

julia> @constraint(model, [t, x, x-1, x-2] in RotatedSecondOrderCone())
[t, x, x - 1, x - 2] ∈ MathOptInterface.RotatedSecondOrderCone(4)

JuMP.PSDCone — Type.

PSDCone

Positive semidefinite cone object that can be used to constrain a square matrix to be positive semidefinite in the @constraint macro. If the matrix has type Symmetric then the columns vectorization (the vector obtained by concatenating the columns) of its upper triangular part is constrained to belong to the MOI.PositiveSemidefiniteConeTriangle set, otherwise its column vectorization is constrained to belong to the MOI.PositiveSemidefiniteConeSquare set.

Examples

Consider the following example:

julia> model = Model();

julia> @variable(model, x)
x

julia> a = [ x 2x
            2x  x];

julia> b = [1 2
            2 4];

julia> cref = @SDconstraint(model, a ⪰ b)
[x - 1    2 x - 2;
 2 x - 2  x - 4  ] ∈ PSDCone()

julia> jump_function(constraint_object(cref))
4-element Array{GenericAffExpr{Float64,VariableRef},1}:
 x - 1
 2 x - 2
 2 x - 2
 x - 4

julia> moi_set(constraint_object(cref))
MathOptInterface.PositiveSemidefiniteConeSquare(2)

We see in the output of the last command that the matrix the vectorization of the matrix is constrained to belong to the PositiveSemidefiniteConeSquare.

julia> using LinearAlgebra # For Symmetric

julia> cref = @constraint(model, Symmetric(a - b) in PSDCone())
[x - 1    2 x - 2;
 2 x - 2  x - 4  ] ∈ PSDCone()

julia> jump_function(constraint_object(cref))
3-element Array{GenericAffExpr{Float64,VariableRef},1}:
 x - 1
 2 x - 2
 x - 4

julia> moi_set(constraint_object(cref))
MathOptInterface.PositiveSemidefiniteConeTriangle(2)

As we see in the output of the last command, the vectorization of only the upper triangular part of the matrix is constrained to belong to the PositiveSemidefiniteConeSquare.

JuMP.NoOptimizer — Type.

struct NoOptimizer <: Exception end

No optimizer is set. The optimizer can be provided at the Model constructor or at the optimize! call with with_optimizer.