Expressions

A technical manual for variable expressions in InfiniteOpt. See the respective guide for more information.

Parameter Functions

Definition

InfiniteOpt.@parameter_function — Macro

@parameter_function(model::InfiniteModel, func_expr, kwargs...)

Add an anonymous parameter function to the model model described by the keyword arguments kw_args and returns the object reference.

@parameter_function(model::InfiniteModel, var_expr == func_expr, kwargs...)

Add a parameter function to model described by the expression var_expr, the function expression func_expr, and the keyword arguments kwargs. The expression var_expr is used to define the parameter function references of the form varname[...] where the indexing matches the container syntax of other macros.

The expression func_expr determines the concrete Julia function that defines the behavior of parameter function and also specifies the infinite parameters it depends on. The accepted forms are:

func(params...): where func is the function that takes supports of the infinite parameters params as input and outputs a scalar value.

(params...) -> my_func_expr: where params are the infinite parameters and my_func_expr is the source code of the anonymous function.

The recognized keyword arguments in kwargs are the following:

base_name: Sets the name prefix used to generate object names. It corresponds to the object name for scalar parameter function, otherwise, the object names are set to base_name[...] for each index ... of the axes axes.

container: Specify the container type. Defaults to :Auto.

Examples

julia> @parameter_function(model, sin(t))
sin(t)

julia> func_vect = [sin, cos];

julia> @parameter_function(model, [i = 1:2] == func_vect[i](t))
2-element Array{GeneralVariableRef,1}:
 sin(t)
 cos(t)

julia> f(t_val, x_vals) = t_val + sum(x_vals)
f (generic function with 1 method)

julia> @parameter_function(model, pf == f(t, x))
pf(t, x)

julia> g(t_val, a; b = 0) = t_val + a + b
g (generic function with 1 method)

julia> @parameter_function(model, pf2[i = 1:2] == t -> g(t, i, b = 2 * i ))
2-element Array{GeneralVariableRef,1}:
 pf2[1](t)
 pf2[2](t)

InfiniteOpt.parameter_function — Function

parameter_function(
    func::Function, 
    pref_inputs::Union{GeneralVariableRef, Array{GeneralVariableRef}, Tuple}; 
    [name::String = [the name of `func`]]
)::GeneralVariableRef

Make a parameter function and return a GeneralVariableRef that can be embedded in InfiniteOpt expressions. This serves as a convenient wrapper for build_parameter_function and add_parameter_function. For an even more convenient definition method see @parameter_function.

Here func denotes the function that will take a support of infinite parameters as input (formatted like pref_inputs) and will return a scalar value. Specifically, func should be of the form:

func(paramvals...)::Union{Real, Bool}

where the formatting of paramvals is analagous to point variables (and will be based on the tuple of infinite parameter references given in parameter_refs). Moreover, func must be a function that returns a scalar numeric value.

Errors if func will not take a support formatted like pref_inputs in combination with the fargs and fkwargs specified. Also errors if pref_inputs follow an invalid input format.

Example

julia> p_func = parameter_function(sin, t)
sin(t)

julia> p_func3 = parameter_function((t_supp) -> 2 * sin(2 * t_supp), t, name = "mysin")
mysin(t)

julia> p_func4 = parameter_function(t, name = "mysin") do t_supp
                    if t_supp <= 5
                        return sin(t_supp)
                    else 
                        return 2 * sin(2 * t_supp)
                    end
                 end

mysin(t)

InfiniteOpt.build_parameter_function — Function

build_parameter_function(
    _error::Function, 
    func::Function, 
    parameter_refs::Union{GeneralVariableRef, Array{<:GeneralVariableRef}, Tuple}
)::ParameterFunction

Build an ParameterFunction object that employs a parameter function func that takes instances of the infinite parameter(s) as input. This can ultimately by incorporated into expressions to enable nonlinear infinite parameter behavior and/or incorporate data over infinite domains.

Here func should be of the form:

func(paramvals...)::Union{Real, Bool}

where the formatting of paramvals is analagous to point variables (and will be based on the tuple of infinite parameter references given in parameter_refs).

Errors if the infinite parameter tuple is formatted incorrectly. The allowed format follows that of infinite variables. Also errors if the function doesn't accept a support realization of the parameter_refs as input.

Example

julia> f = build_parameter_function(error, sin, t);

InfiniteOpt.ParameterFunction — Type

ParameterFunction{F <: Function, T <: JuMP.AbstractVariableRef}

A DataType for storing known functions of infinite parameters. These equate to arbitrary functions that take support instances of infinite parameters parameter_refs in as input and compute a scalar value as output via func. These can then can incorporated in expressions via ParameterFunctionRefs.

Once constructed, these can be called directly and treated like hte functions they wrap:

(pfunc::ParameterFunction)(support...)

where support::Tuple matches the format of the corresponding tuple of infinite parameters (e.g., (t, x)). For developers, it is also convenient to pass the vectorized form of the support that comes from support::VectorTuple via:

(pfunc::ParameterFunction)(vect_support::Vector{<:Real})

Fields

func::F: The function the takes infinite parameters as input and provide a scalar number as output.

parameter_refs::Collections.VectorTuple{T}: The infinite parameter references that serve as inputs to func. Their formatting is analagous to those of infinite variables.

group_int_idxs::Vector{Int}: The parameter group integer indices associated with parameter_refs.

InfiniteOpt.add_parameter_function — Function

add_parameter_function(
    model::InfiniteModel,
    pfunc::ParameterFunction, 
    [name::String]
)::GeneralVariableRef

Add an ParameterFunction pfunc to the model using name for printing and return a GeneralVariableRef such that it can be embedded in expressions. Errors if the parameter function pfunc points to do not belong to model. Note that pfunc should be created using build_parameter_function.

Example

julia> f = build_parameter_function(error, sin, t);

julia> fref = add_parameter_function(model, f)
sin(t)

InfiniteOpt.ParameterFunctionData — Type

ParameterFunctionData{F <: ParameterFunction} <: AbstractDataObject

A mutable DataType for storing ParameterFunctions and their data.

Fields

func::F: The parameter function.

name::String: The name used for printing.

measure_indices::Vector{MeasureIndex}: Indices of dependent measures.

constraint_indices::Vector{InfOptConstraintIndex}: Indices of dependent constraints.

semi_infinite_var_indices::Vector{SemiInfiniteVariableIndex}: Indices of dependent semi-infinite variables.

point_var_indices::Vector{PointVariableIndex}: Indices of dependent point variables.

InfiniteOpt.ParameterFunctionIndex — Type

ParameterFunctionIndex <: ObjectIndex

A DataType for storing the index of a ParameterFunction.

Fields

value::Int64: The index value.

InfiniteOpt.ParameterFunctionRef — Type

ParameterFunctionRef <: DispatchVariableRef

A DataType for infinite parameter function references.

Fields

model::InfiniteModel: Infinite model.

index::ParameterFunctionIndex: Index of the infinite parameter function.

Queries

JuMP.name — Method

JuMP.name(fref::ParameterFunctionRef)::String

Extend JuMP.name to return the base name of fref.

Example

julia> name(fref)
"func_name"

InfiniteOpt.raw_function — Method

raw_function(fref::ParameterFunctionRef)::Function
raw_function(pfunc::ParameterFunction)::Function

Returns the raw function behind fref that takes a particular support of fref's infinite parameters as input.

JuMP.parameter_value — Method

JuMP.parameter_value(fref::ParameterFunctionRef)::ParameterFunction

Return the current function assigned to fref. See also raw_function.

InfiniteOpt.call_function — Function

call_function(fref::ParameterFunctionRef, support...)::Float64

Safely evaluates the raw_function of fref at a particular support support point that matches the format of the infinite parameter tuple given when the fref was defined. This is essentially equivalent to raw_function(fref)(supps...).

call_function(fref::GeneralVariableRef, support...)::Float64

Call the parameter function of fref at support. An ArgumentError is thrown if fref is not a parameter function.

InfiniteOpt.parameter_refs — Method

parameter_refs(fref::ParameterFunctionRef)::Tuple

Return the parameter references associated with fref. This is formatted as a Tuple of containing the parameter references as they inputted to define fref.

Example

julia> parameter_refs(p_func)
(t,)

InfiniteOpt.parameter_list — Method

parameter_list(fref::ParameterFunctionRef)::Vector{GeneralVariableRef}

Return a vector of the parameter references that fref depends on. This is primarily an internal method where parameter_refs is intended as the preferred user function.

InfiniteOpt.raw_parameter_refs — Method

raw_parameter_refs(fref::ParameterFunctionRef)::VectorTuple

Return the raw VectorTuple of the parameter references that fref depends on. This is primarily an internal method where parameter_refs is intended as the preferred user function.

InfiniteOpt.is_used — Method

is_used(fref::ParameterFunctionRef)::Bool

Return a Bool indicating if fref is used in the model.

Example

julia> is_used(fref)
true

InfiniteOpt.used_by_semi_infinite_variable — Method

used_by_semi_infinite_variable(fref::ParameterFunctionRef)::Bool

Return a Bool indicating if fref is used by a semi-infinite infinite variable.

Example

julia> used_by_semi_infinite_variable(fref)
false

InfiniteOpt.used_by_point_variable — Method

used_by_point_variable(fref::ParameterFunctionRef)::Bool

Return a Bool indicating if fref is used by a point variable.

Example

julia> used_by_point_variable(vref)
true

InfiniteOpt.used_by_measure — Method

used_by_measure(fref::ParameterFunctionRef)::Bool

Return a Bool indicating if fref is used by a measure.

Example

julia> used_by_measure(fref)
true

InfiniteOpt.used_by_constraint — Method

used_by_constraint(fref::ParameterFunctionRef)::Bool

Return a Bool indicating if fref is used by a constraint.

Example

julia> used_by_constraint(fref)
false

InfiniteOpt.parameter_group_int_indices — Method

parameter_group_int_indices(fref::ParameterFunctionRef)::Vector{Int}

Return the list of infinite parameter group integer indices used by fref.

InfiniteOpt.num_parameter_functions — Function

num_parameter_functions(model::InfiniteModel)::Int

Returns the number of parameter functions that have been defined in model.

Example

julia> num_parameter_functions(model)
2

InfiniteOpt.all_parameter_functions — Function

all_parameter_functions(model::InfiniteModel)::Vector{GeneralVariableRef}

Returns a list of all the individual parameter functions stored in model.

Example

julia> all_parameter_functions(model)
3-element Array{GeneralVariableRef,1}:
 sin(t)
 cos(t)
 exp(t)

Modification

JuMP.set_name — Method

JuMP.set_name(fref::ParameterFunctionRef, name::String)::Nothing

Extend JuMP.set_name to set the name of a parameter function.

Example

julia> set_name(fref, "func_name")

julia> name(fref)
"func_name"

JuMP.set_parameter_value — Method

JuMP.set_parameter_value(
    fref::ParameterFunctionRef, 
    func::Function
)::Nothing

Update the function used by fref to func. Errors if func does not accept the same infinite parameter format as the original function of fref. If possible, this will also update the transformation backend to enable efficient resolves.

JuMP.delete — Method

JuMP.delete(model::InfiniteModel, fref::ParameterFunctionRef)::Nothing

Extend JuMP.delete to delete parameter functions and their dependencies. Errors if fref is invalid, meaning it has already been deleted or it belongs to another model.

Nonlinear Expressions

DataTypes

InfiniteOpt.NLPOperator — Type

NLPOperator{F <: Function, G <: Union{Function, Nothing}, 
            H <: Union{Function, Nothing}}

A type for storing new nonlinear operators and their information that is ultimately for automatic differentiation. The constructor is of the form:

    NLPOperator(name::Symbol, dim::Int, f::Function, 
                       [∇f::Function, ∇²f::Function])

Fields

name::Symbol: The name of the operator that is used.

dim::Int: The number of function arguments.

f::F: The function to evaluate the operator.

∇f::G: The gradient function if one is given.

∇²f::H: The hessian function if one is given.

Methods

JuMP.add_nonlinear_operator — Function

JuMP.add_nonlinear_operator(
    model::InfiniteModel,
    dim::Int,
    f::Function,
    [∇f::Function,]
    [∇²f::Function];
    [name::Symbol = Symbol(f)]
)

Extend add_nonlinear_operator for InfiniteModels.

Add a new nonlinear operator with dim input arguments to model and associate it with the name name. Alternatively, @operator can be used for a more convenient syntax.

The function f evaluates the operator. The optional function ∇f evaluates the first derivative, and the optional function ∇²f evaluates the second derivative. ∇²f may be provided only if ∇f is also provided.

julia> @variable(model, y);

julia> g(x) = x^2;

julia> new_op = add_nonlinear_operator(model, 1, g)
NonlinearOperator(g, :g)

julia> @expression(model, new_op(y))
g(y)

InfiniteOpt.all_nonlinear_operators — Function

all_nonlinear_operators(model::InfiniteModel)::Vector{Symbol}

Retrieve all the operators that are currently added to model.

InfiniteOpt.name_to_operator — Function

name_to_operator(model::InfiniteModel, name::Symbol)::Union{Function, Nothing}

Return the nonlinear operator that corresponds to name. Returns nothing if no such operator exists.

Warning

Currently, this does not return functions for default operators.

InfiniteOpt.added_nonlinear_operators — Function

user_defined_operators(model::InfiniteModel)::Vector{NLPOperator}

Return all the operators (and their associated information) that the user has added to model. Each is stored as a NLPOperator.

InfiniteOpt.add_operators_to_jump — Function

add_operators_to_jump(opt_model::JuMP.Model, inf_model::InfiniteModel)::Nothing

Add the additional nonlinear operators in inf_model to a JuMP model opt_model. This is intended as an internal method, but it is provided for developers that extend InfiniteOpt to use new JuMPBackends.

Expression Methods

InfiniteOpt.parameter_refs — Method

parameter_refs(expr)::Tuple

Return the tuple of parameter references that determine the infinite dependencies of expr.

Example

julia> parameter_refs(my_expr)
(t,)

InfiniteOpt.restrict — Method

restrict(expr::JuMP.AbstractJuMPScalar, supps...)::JuMP.AbstractJuMPScalar

Restrict an infinite expression expr to be enforced over infinite parameter supports supps. This is limited to expressions only contain infinite variables with the same kind of infinite parameter dependencies. Note that more conveniently the expression can be treated as a function for the syntax expr(supps...).

Example

julia> ex = @expression(model, 3y - 2)
3 y(t) - 2

julia> restrict(ex, 0)
3 y(0) - 2

julia> ex(0)
3 y(0) - 2

InfiniteOpt.map_expression — Function

map_expression(
    transform::Function, 
    expr::JuMP.AbstractJuMPScalar,
    [write_model::Union{InfiniteModel, AbstractTransformationBackend}]
)::JuMP.AbstractJuMPScalar

Map and return a new expression of expr where each variable is transformed via transform. This can be helpful for writing user extensions. For this to work with new AbstractTransformationBackends, JuMP.variable_ref_type needs to be defined for the backend. If write_model is not provided, the JuMP.owner_model(expr) is used. JuMP.owner_model(expr) is used.

InfiniteOpt.all_expression_variables — Function

all_expression_variables(expr::JuMP.AbstractJuMPScalar)::Vector

Returns a vector of all the variable references contained in expr.

Example

julia> all_expr_variables(y^2 + z - t)
3-element Array{GeneralVariableRef,1}:
 y(t)
 z
 t

InfiniteOpt.parameter_group_int_indices — Method

parameter_group_int_indices(expr::JuMP.AbstractJuMPScalar)::Vector{Int}

Return the list of infinite parameter group integer indices used by expr.

GeneralVariableRef User Methods

InfiniteOpt.GeneralVariableRef — Type

GeneralVariableRef <: JuMP.AbstractVariableRef

A DataType that serves as the principal variable reference in InfiniteOpt for building variable expressions. It contains the needed information to create a variable type specifc reference (e.g., InfiniteVariableRef) via dispatch_variable_ref to obtain the correct subtype of DispatchVariableRef based off of index_type. This allows us to construct expressions using concrete containers unlike previous versions of InfiniteOpt which provides us a significant performance boost.

A convenient constructor is:

    GeneralVariableRef(model::InfiniteModel, index::ObjectIndex)

Fields

model::InfiniteModel: Infinite model.

raw_index::Int64: The raw index to be used in the index_type constructor.

index_type::DataType: The concrete AbstractInfOptIndex type/constructor.

param_index::Int: The index of a parameter in DependentParameters. This is ignored for other variable types.

InfiniteOpt.DispatchVariableRef — Type

DispatchVariableRef <: JuMP.AbstractVariableRef

An abstract type for variable references that are created from GeneralVariableRefs and are used to dispatch to the appropriate methods for that particular variable/parameter/measure type.

InfiniteOpt.FiniteRef — Type

FiniteRef <: DispatchVariableRef

An abstract type for variable references that are finite.

JuMP.owner_model — Method

JuMP.owner_model(vref::GeneralVariableRef)::InfiniteModel

Extend JuMP.owner_model to return the model where vref is stored.

Example

julia> owner_model(vref)
An InfiniteOpt Model
Minimization problem with:
  Finite parameters: 0
  Infinite parameter: 1
  Variables: 3
  Derivatives: 0
  Measures: 0
  Objective function type: GenericAffExpr{Float64, GeneralVariableRef}
  `GenericAffExpr{Float64, GeneralVariableRef}`-in-`MathOptInterface.GreaterThan{Float64}`: 1 constraint
  `GenericAffExpr{Float64, GeneralVariableRef}`-in-`MathOptInterface.EqualTo{Float64}`: 1 constraint
  `GeneralVariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 3 constraints
Names registered in the model: c1, c2, t, y, z
Transformation backend information:
  Backend type: TranscriptionBackend
  `t` transcribed over 10 supports
  Solver: Ipopt
  Transformation built and up-to-date: true

JuMP.owner_model — Method

JuMP.owner_model(vref::DispatchVariableRef)::InfiniteModel

Extend JuMP.owner_model to return the model where vref is stored.

JuMP.index — Method

JuMP.index(vref::GeneralVariableRef)::AbstractInfOptIndex

Extend JuMP.index to return the appropriate index of vref.

Example

julia> index(vref)
FiniteVariableIndex(1)

JuMP.index — Method

JuMP.index(vref::DispatchVariableRef)::AbstractInfOptIndex

Extend JuMP.index to return the appropriate index of vref.

InfiniteOpt.dispatch_variable_ref — Method

dispatch_variable_ref(vef::GeneralVariableRef)::DispatchVariableRef

Return the concrete DispatchVariableRef this associated with vref. This relies on dispatch_variable_ref being extended for the index type, otherwise an MethodError is thrown.

InfiniteOpt.dispatch_variable_ref — Function

dispatch_variable_ref(model::InfiniteModel, index::AbstractInfOptIndex)

Return the variable reference associated the type of index. This needs to be defined for each variable reference type.

JuMP.name — Method

JuMP.name(vref::GeneralVariableRef)::String

Extend JuMP.name to return the name of vref. It relies on JuMP.name being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown.

JuMP.set_name — Method

JuMP.set_name(vref::GeneralVariableRef, name::String)::Nothing

Extend JuMP.set_name to set the name of vref. It relies on JuMP.set_name being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown.

JuMP.is_valid — Method

JuMP.is_valid(model::InfiniteModel, vref::GeneralVariableRef)::Bool

Extend JuMP.is_valid to return Bool if vref is a valid reference.

Example

julia> is_valid(model, vref)
true

JuMP.is_valid — Method

JuMP.is_valid(model::InfiniteModel, vref::DispatchVariableRef)::Bool

Extend JuMP.is_valid to return Bool if vref is a valid reference.

InfiniteOpt.used_by_infinite_variable — Method

used_by_infinite_variable(vref::GeneralVariableRef)::Bool

Define used_by_infinite_variable for general variable references. It relies on used_by_infinite_variable being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.used_by_point_variable — Method

used_by_point_variable(vref::GeneralVariableRef)::Bool

Define used_by_point_variable for general variable references. It relies on used_by_point_variable being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.used_by_semi_infinite_variable — Method

used_by_semi_infinite_variable(vref::GeneralVariableRef)::Bool

Define used_by_semi_infinite_variable for general variable references. It relies on used_by_semi_infinite_variable being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.used_by_parameter_function — Method

used_by_parameter_function(vref::GeneralVariableRef)::Bool

Define used_by_parameter_function for general variable references. It relies on used_by_parameter_function being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.used_by_derivative — Method

used_by_derivative(vref::GeneralVariableRef)::Bool

Define used_by_derivative for general variable references. It relies on used_by_derivative being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.used_by_measure — Method

used_by_measure(vref::GeneralVariableRef)::Bool

Define used_by_measure for general variable references. It relies on used_by_measure being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.used_by_objective — Method

used_by_objective(vref::GeneralVariableRef)::Bool

Define used_by_objective for general variable references. It relies on used_by_objective being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.used_by_constraint — Method

used_by_constraint(vref::GeneralVariableRef)::Bool

Define used_by_constraint for general variable references. It relies on used_by_constraint being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.is_used — Method

is_used(vref::GeneralVariableRef)::Bool

Define is_used for general variable references. It relies on is_used being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.has_derivative_constraints — Method

has_derivative_constraints(vref::GeneralVariableRef)::Bool

Define has_derivative_constraints for general variable references. It relies on has_derivative_constraints being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.parameter_value — Method

JuMP.parameter_value(vref::DispatchVariableRef)::Union{Real, Function}

Extend JuMP.parameter_value to query the value of vref. It relies on JuMP.parameter_value being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown.

JuMP.set_parameter_value — Method

JuMP.set_parameter_value(vref::DispatchVariableRef, value::Union{Real, Function})::Nothing

Extend JuMP.set_parameter_value to set the value of vref. It relies on JuMP.set_parameter_value being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown.

InfiniteOpt.infinite_domain — Method

infinite_domain(prefs; [kwargs...])

Define infinite_domain for general variable references. It relies on infinite_domain being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.infinite_domain — Method

infinite_domain(prefs; [kwargs...])

Define infinite_domain for general variable references. It relies on infinite_domain being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.set_infinite_domain — Method

set_infinite_domain(pref::GeneralVariableRef, domain::InfiniteScalarDomain)::Nothing

Specify the scalar infinite domain of the infinite parameter pref to domain. Note this will reset/delete all the supports contained in the underlying parameter object. Also, errors if pref is used by a measure. An ArgumentError is thrown if pref is not an infinite parameter.

InfiniteOpt.set_infinite_domain — Method

set_infinite_domain(prefs::Array{<:GeneralVariableRef},
                 domain::InfiniteArrayDomain)::Nothing

Specify the multi-dimensional infinite domain of the dependent infinite parameters prefs to domain. Note this will reset/delete all the supports contained in the underlying DependentParameters object. This will error if the not all of the dependent infinite parameters are included or if any of them are used by measures. An ArgumentError is thrown if prefs are not dependent infinite parameters.

InfiniteOpt.num_supports — Method

num_supports(prefs; [kwargs...])

Define num_supports for general variable references. It relies on num_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.num_supports — Method

num_supports(prefs; [kwargs...])

Define num_supports for general variable references. It relies on num_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.has_supports — Method

has_supports(prefs; [kwargs...])

Define has_supports for general variable references. It relies on has_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.has_supports — Method

has_supports(prefs; [kwargs...])

Define has_supports for general variable references. It relies on has_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.supports — Method

supports(
    expr::JuMP.AbstractJuMPScalar; 
    [label::Type{<:AbstractSupportLabel} = PublicLabel,
    kwargs...]
    )

Return the support associated with expr. Errors if expr is not associated with the constraint mappings stored in the transformation backend.

The keyword arguments label is what TranscriptionOpt employs and kwargs denote extra ones that user extensions may employ in accordance with their implementation of expression_supports. Errors if such an extension has not been written.

By default only the public supports are returned, the full set can be accessed via label = All. Where possible, all the supports of an infinite expression are returned as an n-dimensional array where each dimension is determined by the each independent group of infinite parameters it depends on.

Example

julia> supports(cref)
2-element Array{Tuple{Float64},1}:
 (0.0,)
 (1.0,)

supports(prefs; [kwargs...])

Define supports for general variable references. It relies on supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.supports — Method

supports(prefs; [kwargs...])

Define supports for general variable references. It relies on supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.set_supports — Method

set_supports(
    pref::GeneralVariableRef,
    supports::Union{Real, Vector{<:Real}, UnitRange{<:Real}, StepRange{<:Real}, NTuple, Base.Generator};
    [force::Bool = false]
    )::Nothing

Set the support points associated with a single infinite parameter pref. An ArgumentError is thrown if pref is not an independent infinite parameter.

InfiniteOpt.set_supports — Method

set_supports(
    prefs::Union{Vector{GeneralVariableRef}, Array{<:GeneralVariableRef}},
    supports::Union{Array{<:Real, 2}, Vector{<:Array{<:Real}}};
    [force::Bool = false]
    )::Nothing

Set the support points associated with dependent infinite parameters prefs. An ArgumentError is thrown if prefs is are not dependent infinite parameters.

InfiniteOpt.add_supports — Method

add_supports(
    pref::GeneralVariableRef,
    supports::Union{Real, Vector{<:Real}, UnitRange{<:Real}, StepRange{<:Real}, NTuple, Base.Generator}
    )::Nothing

Add the support points supports to a single infinite parameter pref. An ArgumentError is thrown if pref is not an independent infinite parameter.

InfiniteOpt.add_supports — Method

add_supports(
    prefs::Union{Vector{GeneralVariableRef}, Array{<:GeneralVariableRef}},
    supports::Union{Array{<:Real, 2}, Vector{<:Array{<:Real}}}
    )::Nothing

Add the support points supports to the dependent infinite parameters prefs. An ArgumentError is thrown if prefs is are not dependent infinite parameters.

InfiniteOpt.delete_supports — Method

delete_supports(prefs; [kwargs...])

Define delete_supports for general variable references. It relies on delete_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.delete_supports — Method

delete_supports(prefs; [kwargs...])

Define delete_supports for general variable references. It relies on delete_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.fill_in_supports! — Method

fill_in_supports!(prefs; [kwargs...])

Define fill_in_supports! for general variable references. It relies on fill_in_supports! being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.fill_in_supports! — Method

fill_in_supports!(prefs; [kwargs...])

Define fill_in_supports! for general variable references. It relies on fill_in_supports! being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.raw_parameter_refs — Method

raw_parameter_refsc(vref::GeneralVariableRef)

Define raw_parameter_refs for general variable references. It relies on raw_parameter_refs being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.parameter_refs — Method

parameter_refsc(vref::GeneralVariableRef)

Define parameter_refs for general variable references. It relies on parameter_refs being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.parameter_list — Method

parameter_listc(vref::GeneralVariableRef)

Define parameter_list for general variable references. It relies on parameter_list being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.raw_function — Method

raw_function(prefs; [kwargs...])

Define raw_function for general variable references. It relies on raw_function being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.infinite_variable_ref — Method

infinite_variable_refc(vref::GeneralVariableRef)

Define infinite_variable_ref for general variable references. It relies on infinite_variable_ref being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.eval_support — Method

eval_supportc(vref::GeneralVariableRef)

Define eval_support for general variable references. It relies on eval_support being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.raw_parameter_values — Method

raw_parameter_valuesc(vref::GeneralVariableRef)

Define raw_parameter_values for general variable references. It relies on raw_parameter_values being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.parameter_values — Method

parameter_valuesc(vref::GeneralVariableRef)

Define parameter_values for general variable references. It relies on parameter_values being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.significant_digits — Method

significant_digits(prefs; [kwargs...])

Define significant_digits for general variable references. It relies on significant_digits being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.measure_function — Method

measure_function(mref::GeneralVariableRef)

Define measure_function for general variable references. Errors if mref does not correspond to a MeasureRef. See the underlying docstrings for more information.

InfiniteOpt.measure_data — Method

measure_data(mref::GeneralVariableRef)

Define measure_data for general variable references. Errors if mref does not correspond to a MeasureRef. See the underlying docstrings for more information.

InfiniteOpt.is_analytic — Method

is_analytic(mref::GeneralVariableRef)

Define is_analytic for general variable references. Errors if mref does not correspond to a MeasureRef. See the underlying docstrings for more information.

InfiniteOpt.derivative_argument — Method

derivative_argument(dref::GeneralVariableRef)

Define derivative_argument for general variable references. Errors if dref does not correspond to a DerivativeRef. See the underlying docstrings for more information.

InfiniteOpt.operator_parameter — Method

operator_parameter(dref::GeneralVariableRef)

Define operator_parameter for general variable references. Errors if dref does not correspond to a DerivativeRef. See the underlying docstrings for more information.

InfiniteOpt.derivative_order — Method

derivative_order(dref::GeneralVariableRef)

Define derivative_order for general variable references. Errors if dref does not correspond to a DerivativeRef. See the underlying docstrings for more information.

InfiniteOpt.derivative_method — Method

derivative_method(prefs; [kwargs...])

Define derivative_method for general variable references. It relies on derivative_method being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.evaluate — Method

evaluate(dref::GeneralVariableRef)

Define evaluate for general variable references. Errors if dref does not correspond to a DerivativeRef. See the underlying docstrings for more information.

InfiniteOpt.derivative_constraints — Method

derivative_constraints(dref::GeneralVariableRef)

Define derivative_constraints for general variable references. Errors if dref does not correspond to a DerivativeRef. See the underlying docstrings for more information.

InfiniteOpt.delete_derivative_constraints — Method

delete_derivative_constraints(dref::GeneralVariableRef)

Define delete_derivative_constraints for general variable references. Errors if dref does not correspond to a DerivativeRef. See the underlying docstrings for more information.

InfiniteOpt.add_generative_supports — Method

add_generative_supports(prefs; [kwargs...])

Define add_generative_supports for general variable references. It relies on add_generative_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.set_derivative_method — Method

set_derivative_method(pref::GeneralVariableRef,
                      method::AbstractDerivativeMethod
                      )::Nothing

Specify the numerical derivative evaluation technique associated with pref. An ArgumentError is thrown if pref is not an infinite parameter.

InfiniteOpt.has_generative_supports — Method

has_generative_supports(prefs; [kwargs...])

Define has_generative_supports for general variable references. It relies on has_generative_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

InfiniteOpt.has_internal_supports — Method

has_internal_supports(prefs; [kwargs...])

Define has_internal_supports for general variable references. It relies on has_internal_supports being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information. Note that this is a auto generated wrapper and the underlying method may or may not use kwargs.

JuMP.delete — Method

JuMP.delete(model::InfiniteModel, vref::GeneralVariableRef)::Nothing

Extend JuMP.delete to delete vref and its dependencies. It relies on JuMP.delete being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown.

JuMP.delete — Method

JuMP.delete(model::InfiniteModel,
            prefs::Array{<:GeneralVariableRef})::Nothing

Extend JuMP.delete to delete a group of dependent infinite parameters and their dependencies. An ArgumentError is thrown if prefs are not dependent infinite parameters.

JuMP.has_lower_bound — Method

JuMP.has_lower_bound(vref::GeneralVariableRef)

Define JuMP.has_lower_bound for general variable references. It relies on JuMP.has_lower_bound being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.lower_bound — Method

JuMP.lower_bound(vref::GeneralVariableRef)

Define JuMP.lower_bound for general variable references. It relies on JuMP.lower_bound being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.set_lower_bound — Method

JuMP.set_lower_bound(vref::GeneralVariableRef, value::Real)::Nothing

Define JuMP.set_lower_bound for general variable references. It relies on JuMP.set_lower_bound being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.LowerBoundRef — Method

JuMP.LowerBoundRef(vref::GeneralVariableRef)

Define JuMP.LowerBoundRef for general variable references. It relies on JuMP.LowerBoundRef being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.delete_lower_bound — Method

JuMP.delete_lower_bound(vref::GeneralVariableRef)

Define JuMP.delete_lower_bound for general variable references. It relies on JuMP.delete_lower_bound being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.has_upper_bound — Method

JuMP.has_upper_bound(vref::GeneralVariableRef)

Define JuMP.has_upper_bound for general variable references. It relies on JuMP.has_upper_bound being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.upper_bound — Method

JuMP.upper_bound(vref::GeneralVariableRef)

Define JuMP.upper_bound for general variable references. It relies on JuMP.upper_bound being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.set_upper_bound — Method

JuMP.set_upper_bound(vref::GeneralVariableRef, value::Real)::Nothing

Define JuMP.set_upper_bound for general variable references. It relies on JuMP.set_upper_bound being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.UpperBoundRef — Method

JuMP.UpperBoundRef(vref::GeneralVariableRef)

Define JuMP.UpperBoundRef for general variable references. It relies on JuMP.UpperBoundRef being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.delete_upper_bound — Method

JuMP.delete_upper_bound(vref::GeneralVariableRef)

Define JuMP.delete_upper_bound for general variable references. It relies on JuMP.delete_upper_bound being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.is_fixed — Method

JuMP.is_fixed(vref::GeneralVariableRef)

Define JuMP.is_fixed for general variable references. It relies on JuMP.is_fixed being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.fix_value — Method

JuMP.fix_value(vref::GeneralVariableRef)

Define JuMP.fix_value for general variable references. It relies on JuMP.fix_value being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.fix — Method

JuMP.fix(vref::GeneralVariableRef, value::Real; force::Bool = false)::Nothing

Define JuMP.fix for general variable references. It relies on JuMP.fix being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.FixRef — Method

JuMP.FixRef(vref::GeneralVariableRef)

Define JuMP.FixRef for general variable references. It relies on JuMP.FixRef being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.unfix — Method

JuMP.unfix(vref::GeneralVariableRef)

Define JuMP.unfix for general variable references. It relies on JuMP.unfix being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.start_value — Method

JuMP.start_value(vref::GeneralVariableRef)

Define JuMP.start_value for general variable references. It relies on JuMP.start_value being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.set_start_value — Method

JuMP.set_start_value(vref::GeneralVariableRef, value::Real)::Nothing

Define JuMP.set_start_value for general variable references. It relies on JuMP.set_start_value being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.is_binary — Method

JuMP.is_binary(vref::GeneralVariableRef)

Define JuMP.is_binary for general variable references. It relies on JuMP.is_binary being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.set_binary — Method

JuMP.set_binary(vref::GeneralVariableRef)

Define JuMP.set_binary for general variable references. It relies on JuMP.set_binary being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.BinaryRef — Method

JuMP.BinaryRef(vref::GeneralVariableRef)

Define JuMP.BinaryRef for general variable references. It relies on JuMP.BinaryRef being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.unset_binary — Method

JuMP.unset_binary(vref::GeneralVariableRef)

Define JuMP.unset_binary for general variable references. It relies on JuMP.unset_binary being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.is_integer — Method

JuMP.is_integer(vref::GeneralVariableRef)

Define JuMP.is_integer for general variable references. It relies on JuMP.is_integer being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.set_integer — Method

JuMP.set_integer(vref::GeneralVariableRef)

Define JuMP.set_integer for general variable references. It relies on JuMP.set_integer being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.IntegerRef — Method

JuMP.IntegerRef(vref::GeneralVariableRef)

Define JuMP.IntegerRef for general variable references. It relies on JuMP.IntegerRef being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

JuMP.unset_integer — Method

JuMP.unset_integer(vref::GeneralVariableRef)

Define JuMP.unset_integer for general variable references. It relies on JuMP.unset_integer being defined for the underlying DispatchVariableRef, otherwise an ArgumentError is thrown. See the underlying docstrings for more information.

InfiniteOpt.constant_over_collocation — Method

constant_over_collocation(vref::GeneralVariableRef, pref::GeneralVariableRef)::Nothing

Define constant_over_collocation for general variable references. It wraps around the method defined for InfiniteVariableRefs.

InfiniteOpt.core_object — Function

core_object(vref::DispatchVariableRef)::Union{InfOptParameter, InfOptVariable, Measure}

Return the core object that vref points to. This needs to be extended for type of vref. This should use _data_object to access the data object where the variable object is stored.

InfiniteOpt.core_object — Method

core_object(vref::GeneralVariableRef)::Union{InfOptParameter, InfOptVariable, Measure}

Return the core object that vref points to. This is enabled with appropriate definitions of core_object for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt.parameter_group_int_indices — Method

parameter_group_int_indices(vref::GeneralVariableRef)::Vector{Int}

Return the list of infinite parameter group integer indices used by vref.

InfiniteOpt.parameter_group_int_index — Function

parameter_group_int_index(pref::DispatchVariableRef)::Int

Return the parameter group integer index for pref assuming it is an infinite parameter. This needs to be defined for the type of pref. This should use the _data_object to get the number.

InfiniteOpt.parameter_group_int_index — Method

parameter_group_int_index(pref::GeneralVariableRef)::Int

Return the group integer index for pref assuming it is an infinite parameter. It relies on parameter_group_int_index being properly defined for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt.reformulate_high_order_derivatives! — Method

reformulate_high_order_derivatives!(dref::GeneralVariableRef)

Define reformulate_high_order_derivatives! for general variable references. Errors if dref does not correspond to a DerivativeRef. See the underlying docstrings for more information.

Developer Internal Methods

InfiniteOpt._add_data_object — Function

_add_data_object(model::InfiniteModel, object::AbstractDataObject)::ObjectIndex

Add object to the appropriate CleverDict in model and return the its index. This needs to be defined for the type of object. These definitions need to use MOIUC.add_item to add the object to the CleverDict.

InfiniteOpt._data_dictionary — Function

_data_dictionary(vref::DispatchVariableRef)::MOIUC.CleverDict

Return the CleverDict that stores data objects for the type of vref. This needs to be defined for the type of vref.

InfiniteOpt._data_object — Function

_data_object(vref::DispatchVariableRef)::AbstractDataObject

Return the data object associated with vref, in other words the object its index points to in the InfiniteModel. This needs to be defined for the type of vref. This should use _data_dictionary to access the CleverDict that the object is stored in.

InfiniteOpt._delete_data_object — Function

_delete_data_object(vref::DispatchVariableRef)::Nothing

Delete the concrete AbstractDataObject associated with vref.

InfiniteOpt._set_core_object — Function

_set_core_object(vref::DispatchVariableRef, object)::Nothing

Sets the core object that vref points to object. This needs to be extended for types of vref and object. This should use _data_object to access the data object where the variable object is stored.

InfiniteOpt._infinite_variable_dependencies — Function

_infinite_variable_dependencies(vref::DispatchVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This needs to be extended for type of vref. This should use _data_object to access the data object where the name is stored if appropriate.

InfiniteOpt._infinite_variable_dependencies — Method

_infinite_variable_dependencies(vref::GeneralVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This is enabled with appropriate definitions of _infinite_variable_dependencies for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt._semi_infinite_variable_dependencies — Function

_semi_infinite_variable_dependencies(vref::DispatchVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This needs to be extended for type of vref. This should use _data_object to access the data object where the name is stored if appropriate.

InfiniteOpt._semi_infinite_variable_dependencies — Method

_semi_infinite_variable_dependencies(vref::GeneralVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This is enabled with appropriate definitions of _semi_infinite_variable_dependencies for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt._point_variable_dependencies — Function

_point_variable_dependencies(vref::DispatchVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This needs to be extended for type of vref. This should use _data_object to access the data object where the name is stored if appropriate.

InfiniteOpt._point_variable_dependencies — Method

_point_variable_dependencies(vref::GeneralVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This is enabled with appropriate definitions of _point_variable_dependencies for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt._derivative_dependencies — Function

_derivative_dependencies(vref::DispatchVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This needs to be extended for type of vref. This should use _data_object to access the data object where the name is stored if appropriate.

InfiniteOpt._derivative_dependencies — Method

_derivative_dependencies(vref::GeneralVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This is enabled with appropriate definitions of _derivative_dependencies for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt._measure_dependencies — Function

_measure_dependencies(vref::DispatchVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This needs to be extended for type of vref. This should use _data_object to access the data object where the name is stored if appropriate.

InfiniteOpt._measure_dependencies — Method

_measure_dependencies(vref::GeneralVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This is enabled with appropriate definitions of _measure_dependencies for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt._generative_measures — Function

_generative_measures(vref::DispatchVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This needs to be extended for type of vref. This should use _data_object to access the data object where the name is stored if appropriate.

InfiniteOpt._generative_measures — Method

_generative_measures(vref::GeneralVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This is enabled with appropriate definitions of _generative_measures for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt._constraint_dependencies — Function

_constraint_dependencies(vref::DispatchVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This needs to be extended for type of vref. This should use _data_object to access the data object where the name is stored if appropriate.

InfiniteOpt._constraint_dependencies — Method

_constraint_dependencies(vref::GeneralVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This is enabled with appropriate definitions of _constraint_dependencies for the underlying DispatchVariableRef, otherwise an MethodError is thrown.

InfiniteOpt._derivative_constraint_dependencies — Function

_derivative_constraint_dependencies(vref::DispatchVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This needs to be extended for type of vref. This should use _data_object to access the data object where the name is stored if appropriate.

InfiniteOpt._derivative_constraint_dependencies — Method

_derivative_constraint_dependencies(vref::GeneralVariableRef)::Vector{AbstractInfOptIndex}

Return the indices of these entities that depend on vref. This is enabled with appropriate definitions of _derivative_constraint_dependencies for the underlying DispatchVariableRef, otherwise an MethodError is thrown.